% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfilereader.R
\name{acceptMSFileReaderLicense}
\alias{acceptMSFileReaderLicense}
\title{License agreement.}
\usage{
acceptMSFileReaderLicense(sys_path)
}
\arguments{
\item{sys_path}{The system path of mzion.}
}
\description{
License agreement.
}
