% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_aamasses_motifs}
\alias{add_aamasses_motifs}
\title{Adds the motifs attribute to aa_masses}
\usage{
add_aamasses_motifs(aa_masses, mod_motifs, positions_sites)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{mod_motifs}{The motifs to restrict \code{Anywhere} variable
  modification. For example, provided the \code{Anywhere} variable
  modifications containing \code{c("Oxidation (M)", "Deamidated (N)")} and

  \code{mod_motifs = list(`Deamidated (N)` = c("NG", "NM"), `Oxidation (M)` =
  c("NM", "MP"))}

  variable modifications will only be considered at sites that satisfy the
  motifs.}

\item{positions_sites}{Named list of positions (in names) and sites (in
values).}
}
\description{
Adds the motifs attribute to aa_masses
}
