% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_aamasses_neulosses}
\alias{add_aamasses_neulosses}
\title{Adds the attribute of neuloss.}
\usage{
add_aamasses_neulosses(aa_masses, neulosses, mod_type)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{neulosses}{Neutral losses}

\item{mod_type}{The type of modification}
}
\description{
Adds the attribute of neuloss.
}
