% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{add_aas_pos}
\alias{add_aas_pos}
\title{Add site indexes of matched residues from an amino-acid sequence (aas).}
\usage{
add_aas_pos(vmods, pos)
}
\arguments{
\item{vmods}{A set of permutated variable modifications (at the levels of
same MS1 label).}

\item{pos}{Lists of position indexes of matched sites from an amino-acid
sequence.}
}
\description{
Incorporates the aas site indexes (pos, from \link{combi_byvmodsM}) to the
labels of modifications (vmods).
}
