% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{add_comp_elements}
\alias{add_comp_elements}
\title{Adds a series of nodes of \code{umod:element}.}
\usage{
add_comp_elements(node = NULL, composition = "0")
}
\arguments{
\item{node}{A node of either \code{NeutralLoss} or \code{delta}.}

\item{composition}{The chemical composition of a modification.}
}
\description{
Adds a series of nodes of \code{umod:element}.
}
