% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{add_delta}
\alias{add_delta}
\title{Adds a node of \code{delta}.}
\usage{
add_delta(
  node = NULL,
  mod_mono_mass = "0",
  mod_avge_mass = "0",
  mod_composition = "0"
)
}
\arguments{
\item{node}{A node of \code{umod:mod}.}

\item{mod_mono_mass}{The mono-isotopic mass delta of a modification.}

\item{mod_avge_mass}{The average mass delta of a modification.}

\item{mod_composition}{The chemical composition of a modification.}
}
\description{
Adds a node of \code{delta}.
}
