% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_fixed_masses}
\alias{add_fixed_masses}
\title{Adds fixed masses}
\usage{
add_fixed_masses(mods, aa_masses, mod_motifs = NULL)
}
\arguments{
\item{mods}{A list of modifications.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{mod_motifs}{The motifs to restrict \code{Anywhere} variable
modification. For example, provided the \code{Anywhere} variable
modifications containing \code{c("Oxidation (M)", "Deamidated (N)")} and
\code{mod_motifs = list(`Deamidated (N)` = c("NG", "NM"), `Oxidation (M)` =
c("NM", "MP"))},
variable modifications will only be considered at sites that satisfy the
motifs.}
}
\description{
Adds fixed masses
}
