% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{add_fixedlab_masses}
\alias{add_fixedlab_masses}
\title{Adds fixedlab masses to aa_masses}
\usage{
add_fixedlab_masses(fixedlabs, aa_masses)
}
\arguments{
\item{fixedlabs}{Character string(s) of fixed isotopic labels. See examples
of SILAC for details. Can be but not typically used in standard alone
searches of labeled residues.}

\item{aa_masses}{A look-up of amino-acid residue masses.}
}
\description{
Adds fixedlab masses to aa_masses
}
