% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{add_ion_matches}
\alias{add_ion_matches}
\title{Helper of \link{padd_ion_matches}.}
\usage{
add_ion_matches(df, pep_grp, out_path)
}
\arguments{
\item{df}{A data frame of psmC.txt at a given pep_mod_group.}

\item{pep_grp}{A given pep_mod_group.}

\item{out_path}{A file path of outputs.}
}
\description{
By pep_mod_group. No need of pep_mod_group being part of the \code{uniq_id.}
as df is already pep_mod_group specific
}
