% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{add_ion_mathces}
\alias{add_ion_mathces}
\title{Helper of \link{padd_ion_mathces}.}
\usage{
add_ion_mathces(df, pep_grp, out_path)
}
\arguments{
\item{df}{A data frame of psmC.txt at a given pep_mod_group.}

\item{pep_grp}{A given pep_mod_group.}

\item{out_path}{A file path of outputs.}
}
\description{
By pep_mod_group.
}
