% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_ms1_13c}
\alias{add_ms1_13c}
\title{Adds Carbon-13 masses.}
\usage{
add_ms1_13c(peps, n_13c = 1L)
}
\arguments{
\item{peps}{A named vector of peptide sequences. Sequences in names and
masses in values.}

\item{n_13c}{The maximum number of 13C off-sets for consideration in MS1
masses.}
}
\description{
Adds Carbon-13 masses.
}
