% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_ms1_notches}
\alias{add_ms1_notches}
\title{Adds offsets of MS1 masses.}
\usage{
add_ms1_notches(peps, masses = 0, max_mass = 4500L)
}
\arguments{
\item{peps}{A named vector of peptide sequences. Sequences in names and
masses in values.}

\item{masses}{The masses of notches.}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}
}
\description{
Not currently used.
}
