% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{add_neuloss}
\alias{add_neuloss}
\title{Adds a node of \code{NeutralLoss}.}
\usage{
add_neuloss(
  node = NULL,
  neuloss_mono_mass = "0",
  neuloss_avge_mass = "0",
  neuloss_composition = "0"
)
}
\arguments{
\item{node}{A node of \code{umod:specificity}.}

\item{neuloss_mono_mass}{The mono-isotopic mass delta of neutral loss in a
positive value.}

\item{neuloss_avge_mass}{The average mass delta of neutral loss in a positive
value.}

\item{neuloss_composition}{The chemical composition of a neutral loss.}
}
\description{
Including the \code{0} NeutralLoss.
}
