% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{add_nulllist}
\alias{add_nulllist}
\title{Adds a plain NULL entry to a list.}
\usage{
add_nulllist(l, nm)
}
\arguments{
\item{l}{A list.}

\item{nm}{A character vector of name.}
}
\description{
This is different to, for example, \code{x[["a"]] <- list(NULL)}.
}
\examples{
\dontrun{
library(mzion)
mzion::add_nulllist(list(a = 2, b = 3), "m")
} 
}
