% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{add_one_label}
\alias{add_one_label}
\title{Helper of \link{find_perm_sets}.}
\usage{
add_one_label(M, x)
}
\arguments{
\item{M}{A permutation table with unique sets by rows.}

\item{x}{A duplicated label that can be found in M.}
}
\description{
Adds one more labels to a permutation table.
}
