% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{add_one_permlab}
\alias{add_one_permlab}
\title{Helper in making permuation table.}
\usage{
add_one_permlab(M, x)
}
\arguments{
\item{M}{A permutation matrix with non-redundant labels.}

\item{x}{A new label not in M for permutation.}
}
\description{
Helper in making permuation table.
}
\examples{
\donttest{
unilabs <- c("A", "B", "C")
nu <- length(unilabs)

G <- gtools::permutations(nu, nu, unilabs)

M  <- matrix(unilabs[1])
for (ulab in unilabs[-1]) M <- add_one_permlab(M, ulab)

cns <- paste0("p", 1:nu)
M <- data.frame(M)
G <- data.frame(G)
colnames(G) <- colnames(M) <- cns

stopifnot(identical(dplyr::arrange_all(G), dplyr::arrange_all(M)))
}
}
