% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{add_pasef_precursors}
\alias{add_pasef_precursors}
\title{Adds PASEF precursor IDs to \code{iso_info}}
\usage{
add_pasef_precursors(
  iso_info,
  keys = c("MS1Frame", "ScanNumBegin", "IsolationMz")
)
}
\arguments{
\item{iso_info}{A data frame contains MS2 isolation information.}

\item{keys}{The column keys in \code{iso_info} for combinations of unique
identifiers.}
}
\description{
Not used.
}
