% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{add_prot_acc2}
\alias{add_prot_acc2}
\title{Adds prot_acc to a peptide table}
\usage{
add_prot_acc2(df, out_path)
}
\arguments{
\item{df}{The results after scoring.}

\item{out_path}{An output path.}
}
\description{
Cached results are under sub dirs.
}
