% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{add_protacc2}
\alias{add_protacc2}
\title{Adds prot_acc to a peptide table}
\usage{
add_protacc2(
  df = NULL,
  out_path = NULL,
  .path_cache = NULL,
  .path_fasta = NULL
)
}
\arguments{
\item{df}{The results after scoring.}

\item{out_path}{An output path.}

\item{.path_cache}{The file path of cached search parameters. The parameter
  is for the users' awareness of the underlying structure of file folders and
  the use of default is suggested. Occasionally experimenters may remove the
  file folder for disk space or under infrequent events of modified framework
  incurred by the developer.

  Started from version 1.2.1.5, a new data structure of protein-peptide
  lookups are used. Version 1.2.1.4.1 is the final version using the old data
  structure.}

\item{.path_fasta}{The parent file path to the theoretical masses of MS1
precursors. At the NULL default, the path is \code{gsub("(.*)\\.[^\\.]*$",
"\\1", get("fasta", envir = environment())[1])}. The parameter is for the
users' awareness of the structure of file folders and the use of default is
suggested. Occasionally experimenters may remove the file folder for disk
space or under infrequent events of modified framework incurred by the
developer.}
}
\description{
Cached results are under sub dirs.
}
