% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{add_rptrs}
\alias{add_rptrs}
\title{Adds back reporter-ion intensities}
\usage{
add_rptrs(df = NULL, quant = "none", out_path = NULL)
}
\arguments{
\item{df}{Results from \link{calc_protfdr}.}

\item{quant}{A quantitation method. The default is "none". Additional choices
include \code{tmt6, tmt10, tmt11, tmt16 and tmt18}. For other
multiplicities of \code{tmt}, use the compatible higher plexes. For
example, apply \code{tmt16} for \code{tmt12} provided a set of 12-plexes
being constructed from a 16-plex TMTpro (7 * 13C + 2 * 15N). It is also
possible that an experimenter may construct a \code{tmt12} from a 18-plex
TMTpro (8 *13C + 1 * 15N) and thus \code{quant = tmt18} is suitable.}

\item{out_path}{A file path of outputs.}
}
\description{
Adds back reporter-ion intensities
}
