% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{add_specificy}
\alias{add_specificy}
\title{Adds a node of \code{specificity}.}
\usage{
add_specificy(
  node = NULL,
  site = "C",
  position = "Anywhere",
  neuloss_mono_mass = "0",
  neuloss_avge_mass = "0",
  neuloss_composition = "0"
)
}
\arguments{
\item{node}{A node of \code{umod:mod}.}

\item{site}{The site of a modification.}

\item{position}{The position of a modification.}

\item{neuloss_mono_mass}{The mono-isotopic mass delta of neutral loss in a
positive value.}

\item{neuloss_avge_mass}{The average mass delta of neutral loss in a positive
value.}

\item{neuloss_composition}{The chemical composition of a neutral loss.}
}
\description{
Adds a node of \code{specificity}.
}
