% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_term_mass}
\alias{add_term_mass}
\title{Helper in calculating peptide masses.}
\usage{
add_term_mass(peps, aa_masses, min_mass = 700L, max_mass = 4500L)
}
\arguments{
\item{peps}{A list of peptide sequences.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value. The primary guard
against low molecular-weight precursors is \code{min_len}.}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation. The default is 4500.}
}
\description{
(2) "amods- tmod+ vnl- fnl-".
}
\details{
No needs of checking \code{is_empty(ntmod) && is_empty(ctmod)}. 
\code{aa_masses} is one of \code{aa_masses_ms1} (coerced).
}
