% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_term_mass2}
\alias{add_term_mass2}
\title{Helper in calculating peptide masses.}
\usage{
add_term_mass2(aa_masses, peps)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
(2) "amods- tmod+ vnl- fnl-".
}
