% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{add_var_masses}
\alias{add_var_masses}
\title{Helper to add modification masses to amino-acid residues.}
\usage{
add_var_masses(
  mods,
  aa_masses,
  varlabs = NULL,
  mod_motifs = NULL,
  anywhere_coerce_sites = NULL,
  nt_coerce_site = NULL,
  ct_coerce_site = NULL
)
}
\arguments{
\item{mods}{A list of modifications.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{varlabs}{Character string(s) of variable isotopic labels. See examples
of SILAC for details. Can be but not typically used in standard alone
searches of labeled residues.}

\item{mod_motifs}{The motifs to restrict \code{Anywhere} variable
  modification. For example, provided the \code{Anywhere} variable
  modifications containing \code{c("Oxidation (M)", "Deamidated (N)")} and

  \code{mod_motifs = list(`Deamidated (N)` = c("NG", "NM"), `Oxidation (M)` =
  c("NM", "MP"))}

  variable modifications will only be considered at sites that satisfy the
  motifs.}

\item{anywhere_coerce_sites}{The sites of coerced Anywhere modifications; for
example, \code{K}.}

\item{anywhere_f_to_v}{Anywhere modifications coerced from fixed to variable;
for example, \code{TMT6plex (K)}.}
}
\value{
Lists of of amino-acid residues with modified mono-isotopic masses
  being incorporated. Returns NULL if \code{is.null(varmods_comb)}.
}
\description{
It adds the masses of fixed, variable, and neutral-loss modifications to
amino-acid residues.
}
