% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{addChim}
\alias{addChim}
\title{Preparation for scoring of chimeric spectra.}
\usage{
addChim(df)
}
\arguments{
\item{df}{A data frame for scoring.}
}
\description{
Preparation for scoring of chimeric spectra.
}
