% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{adj_anywhere_masses}
\alias{adj_anywhere_masses}
\title{Adjusts the masses for the coerced sites}
\usage{
adj_anywhere_masses(peps, aa_masses)
}
\arguments{
\item{peps}{The list of forward peptides with masses.}

\item{aa_masses}{An amino-acid masses look-up table.}
}
\description{
Adjusts the masses for the coerced sites
}
