% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{adj_masses1}
\alias{adj_masses1}
\title{Adjust X values}
\usage{
adj_masses1(mgfs, mdiff, is_listmass = TRUE, is_tmt = FALSE)
}
\arguments{
\item{mdiff}{The mass difference (the median of differences).}

\item{is_listmass}{Logical; are the X values in list for scalar.}

\item{is_tmt}{Logical; is a TMT experiment or not.}

\item{A}{table of peak lists.}
}
\description{
Adjust X values
}
