% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{aggr_prot_es}
\alias{aggr_prot_es}
\title{Function factories for descriptive statistics.}
\usage{
aggr_prot_es(f)
}
\arguments{
\item{f}{One of \code{max, mean, median, min}. Data frame \code{df} is a PSM
table.}
}
\description{
Function factories for descriptive statistics.
}
