% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{are_ovlap_ranges}
\alias{are_ovlap_ranges}
\title{Checks overlaps in start and end positions}
\usage{
are_ovlap_ranges(sta_b, end_b, stas_a, ends_a, margin = 0L)
}
\arguments{
\item{sta_b}{Scalar; the start index of b.}

\item{end_b}{Scalar; the end index of b. Note that \code{end_b >= sta_b}.}

\item{stas_a}{Vector; the start indexes of a.}

\item{ends_a}{Vector; the end indexes of a.}

\item{margin}{The margin for considering being intersecting.}
}
\description{
Checks overlaps in start and end positions
}
\examples{
oks <- mzion:::are_ovlap_ranges(sta_b = 450, end_b = 774, stas_a = c(412, 775, 850), ends_a = c(455, 843, 900), margin = 5)
# stopifnot(identical(oks, c(TRUE, TRUE, FALSE)))
oks <- mzion:::are_ovlap_ranges(sta_b = 470, end_b = 774, stas_a = c(412, 775, 850), ends_a = c(455, 843, 900), margin = 5)
# stopifnot(identical(oks, c(FALSE, TRUE, FALSE)))
oks <- mzion:::are_ovlap_ranges(sta_b = 470, end_b = 600, stas_a = c(412, 775, 850), ends_a = c(455, 843, 900), margin = 5)
# stopifnot(identical(oks, c(FALSE, FALSE, FALSE)))
oks <- mzion:::are_ovlap_ranges(sta_b = 440, end_b = 600, stas_a = c(412, 775, 850), ends_a = c(455, 843, 900), margin = 5)
# stopifnot(identical(oks, c(TRUE, FALSE, FALSE)))
oks <- mzion:::are_ovlap_ranges(sta_b = 440, end_b = 850, stas_a = c(412, 775, 850), ends_a = c(455, 843, 900), margin = 5)
# stopifnot(identical(oks, c(TRUE, TRUE, TRUE)))
}
