% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms1_labels.R
\name{bacth_vmods_combi}
\alias{bacth_vmods_combi}
\title{Helper in cycling through \code{2:maxn_vmods_per_pep} numbers of positions.}
\usage{
bacth_vmods_combi(
  resmods = NULL,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L
)
}
\arguments{
\item{resmods}{A vector of amino-acid residues with Unimod in names. For
example, c(`Oxidation (M)` = "M", `Carbamyl` = "M").The residues are the
same and the names (modifications) are different.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}
}
\description{
For the same residue (M) at different names of modifications.
}
\examples{
\donttest{
resmods <- c(`Oxidation (M)` = "M", `Carbamidomethyl (M)` = "M")
ans <- bacth_vmods_combi(resmods)
}
}
