% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_ms2ions.R
\name{batch_ms2ions}
\alias{batch_ms2ions}
\title{Batch processing of MS2 ions}
\usage{
batch_ms2ions(
  fasta = c("~/proteoM/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta",
    "~/proteoM/dbs/fasta/crap/crap.fasta"),
  fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)"),
  varmods = c("Acetyl (Protein N-term)", "Oxidation (M)", "Deamidated (N)",
    "Gln->pyro-Glu (N-term = Q)"),
  include_insource_nl = FALSE,
  exclude_phospho_nl = TRUE,
  type_ms2ions = "by",
  maxn_vmods_setscombi = 64L,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  .path_cache = NULL,
  .path_fasta = NULL,
  digits = 5L
)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The experimenters need to supply the files.}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{include_insource_nl}{Logical Logical; if TRUE, includes MS1 precursor
masses with the losses of neutral species prior to MS2 fragmentation. The
default is FALSE. The setting at TRUE remains experimenting by allowing
additional masses in the universe of MS1 precursors. The offsets in NLs (by
adding back precursor masses) have not yet taken into account in MS2 ion
searches. A more systemically approach such as \code{open} MS1 masses might
be developed in the future.}

\item{exclude_phospho_nl}{If TRUE, excludes neutral losses in the MS2 ion
searches against variable modifications of phospho sites.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial
variable modifications and neutral losses.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence. The default is
64. May consider a smaller value, i.e. 32, when searching against
phosphopeptide data.}

\item{.path_cache}{The file path of cached search parameters. At the NULL
default, the path is \code{"~/proteoM/.MSearches/Cache/Calls/"}. The
parameter is for users' awareness of the structure of file folders and the
default is suggested. Occasionally experimenters may remove the file folder
for disk space or (hopefully not) in events of (improper) structural change
incurred by the developer.}

\item{.path_fasta}{The parent file path to the theoretical masses of MS1
precursors. At the NULL default, the path is \code{gsub("(.*)\\.[^\\.]*$",
"\\1", get("fasta", envir = environment())[1])}. The parameter is for
users' awareness of the structure of file folders and the default is
suggested. Occasionally experimenters may remove the file folder for disk
space or (hopefully) in rare events of structural change incurred by the
developer.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Batch processing of MS2 ions
}
