% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{bin_ms1masses}
\alias{bin_ms1masses}
\title{Helper in binning precursor masses.}
\usage{
bin_ms1masses(
  res = NULL,
  min_mass = 500L,
  max_mass = 6000L,
  ppm_ms1 = 20L,
  .path_cache = NULL,
  .path_fasta = NULL,
  is_ms1_three_frame = TRUE
)
}
\arguments{
\item{res}{The results from \link{calc_pepmasses2}.}

\item{min_mass}{A minimum mass of precursors.}

\item{max_mass}{A maximum mass of precursors.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{.path_cache}{The file path of cached search parameters.}

\item{.path_fasta}{The file path to the theoretical masses of MS1 precursors.}

\item{is_ms1_three_frame}{Logical; is the searches by the three frames of
preceeding, current and following.}
}
\description{
For target peptides.
}
