% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{bin_ms1masses_td}
\alias{bin_ms1masses_td}
\title{Helper of \link{bin_ms1masses} by the separate sets of \code{target}s and
\code{decoy}s.}
\usage{
bin_ms1masses_td(
  bins = NULL,
  type = c("target", "decoy"),
  res = NULL,
  min_mass = 500L,
  max_mass = 10000L,
  ppm_ms1 = 20L,
  out_path = NULL
)
}
\arguments{
\item{bins}{The names of bins in the format of
"binned_theopeps_(rev_)*\\d+\\.rds$".}

\item{type}{The type of data: "target" or "decoy".}

\item{res}{Lists of data containing theoretical peptides and masses from
\link{readRDS}.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm_ms1}{Numeric; the error tolerance of MS1 mass in ppm.}

\item{out_path}{The output path.}
}
\description{
Helper of \link{bin_ms1masses} by the separate sets of \code{target}s and
\code{decoy}s.
}
