% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{bin_theoseqs}
\alias{bin_theoseqs}
\title{Separates theoretical peptides into mass groups.}
\usage{
bin_theoseqs(
  peps = NULL,
  out_nm = NULL,
  min_mass = 700L,
  max_mass = 4500L,
  ppm_ms1 = 20L
)
}
\arguments{
\item{peps}{A list of theoretical peptides with masses.}

\item{out_nm}{A output name with prepending file path.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm_ms1}{Numeric; the error tolerance of MS1 mass in ppm.}
}
\description{
Separates theoretical peptides into mass groups.
}
