% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{bind_dfs}
\alias{bind_dfs}
\title{Binds data frames}
\usage{
bind_dfs(dfs)
}
\arguments{
\item{dfs}{A list of data frames.}
}
\description{
Assume identical column names across data frames. Also different to
\link[dplyr]{bind_rows} by row names.
}
