% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{binTheoPeps}
\alias{binTheoPeps}
\title{Separates theoretical peptides into mass groups.}
\usage{
binTheoPeps(
  idxes = NULL,
  res = NULL,
  min_mass = 500L,
  max_mass = 10000L,
  ppm_ms1 = 20L,
  out_path = NULL
)
}
\arguments{
\item{idxes}{A set of indexes, e.g. "1" for \code{pepmasses_1.rds} and
"rev_1" for \code{pepmasses_rev_1.rds}.}

\item{res}{Lists of data containing theoretical peptides and masses from
\link{readRDS}.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm_ms1}{Numeric; the error tolerance of MS1 mass in ppm.}

\item{out_path}{The output path.}
}
\value{
Lists of theoretical peptides binned by MS1 masses. The lists
  correspond to the lists of \code{res}.
}
\description{
Separates theoretical peptides into mass groups.
}
\examples{
\donttest{
res <- readRDS("~/proteoM/dbs/fasta/uniprot/pepmass/uniprot_hs_2020_05_2miss.rds")
theopeps <- proteoM:::binTheoPeps(res)
}
}
