% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{binTheoPeps_bysets}
\alias{binTheoPeps_bysets}
\title{Helper of \link{bin_ms1masses_td}.}
\usage{
binTheoPeps_bysets(
  idxes = NULL,
  min_mass = 500L,
  max_mass = 10000L,
  ppm_ms1 = 20L,
  out_path = NULL
)
}
\arguments{
\item{idxes}{A set of indexes, e.g. "1" for \code{pepmasses_1.rds} and
"rev_1" for \code{pepmasses_rev_1.rds}.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm_ms1}{Numeric; the error tolerance of MS1 mass in ppm.}

\item{out_path}{The output path.}
}
\description{
Bin peptide masses by sets of data.
}
