% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{binTheoPeps_i}
\alias{binTheoPeps_i}
\title{Helper of \link{binTheoPeps2}.}
\usage{
binTheoPeps_i(
  idx = 1L,
  min_mass = 500L,
  max_mass = 10000L,
  ppm_ms1 = 20L,
  out_path = NULL
)
}
\arguments{
\item{idx}{An index, e.g. "1" for \code{pepmasses_1.rds} and "rev_1" for
\code{pepmasses_rev_1.rds}.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm_ms1}{Numeric; the error tolerance of MS1 mass in ppm.}

\item{out_path}{The output path.}
}
\description{
Helper of \link{binTheoPeps2}.
}
