% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{binTheoSeqs2}
\alias{binTheoSeqs2}
\title{Separates theoretical peptides into mass groups.}
\usage{
binTheoSeqs2(
  idx = 1L,
  res = NULL,
  min_mass = 700L,
  max_mass = 4500L,
  ppm_ms1 = 20L,
  out_path = NULL
)
}
\arguments{
\item{idx}{An index, e.g. "1" for \code{pepmasses_1.rds} and "rev_1" for
\code{pepmasses_rev_1.rds}.}

\item{res}{Lists of data containing theoretical peptides and masses from
\link{readRDS}.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm_ms1}{Numeric; the error tolerance of MS1 mass in ppm.}

\item{out_path}{The output path.}
}
\description{
Separates theoretical peptides into mass groups.
}
