% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{binTheoSeqs_i}
\alias{binTheoSeqs_i}
\title{Helper of \link{binTheoSeqs2}.}
\usage{
binTheoSeqs_i(
  idx = 1L,
  min_mass = 700L,
  max_mass = 4500L,
  ppm_ms1 = 10L,
  in_path = NULL,
  out_path = NULL
)
}
\arguments{
\item{idx}{An index, e.g. "1" for \code{pepmasses_1.rds} and "rev_1" for
\code{pepmasses_rev_1.rds}.}

\item{min_mass}{Numeric; the minimum MS1 mass.}

\item{max_mass}{Numeric; the maximum MS1 mass.}

\item{ppm_ms1}{Numeric; (half of) the error tolerance of MS1 mass in ppm.}

\item{in_path}{An input path of \code{pepmasses_}.}

\item{out_path}{The output path.}
}
\description{
Helper of \link{binTheoSeqs2}.
}
