% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_ladder.R
\name{byions}
\alias{byions}
\alias{czions}
\alias{axions}
\alias{bions_base}
\alias{yions_base}
\alias{cions_base}
\alias{zions_base}
\alias{c2ions}
\alias{z2ions}
\alias{aions_base}
\alias{xions_base}
\alias{a2ions}
\alias{astarions}
\alias{astar2ions}
\alias{a0ions}
\alias{a02ions}
\alias{x2ions}
\title{Masses of singly-charged b- and y-ions.}
\usage{
byions(ntmass, ctmass, aam)

czions(ntmass, ctmass, aam)

axions(ntmass, ctmass, aam)

bions_base(aam, ntmass)

yions_base(aam, ctmass)

cions_base(aam, ntmass)

zions_base(aam, ctmass)

c2ions(aam, ntmass, n = 2L)

z2ions(aam, ctmass, n = 2L)

aions_base(aam, ntmass)

xions_base(aam, ctmass)

a2ions(aam, ntmass, n = 2L)

astarions(aam, ntmass)

astar2ions(aam, ntmass, n = 2L)

a0ions(aam, ntmass)

a02ions(aam, ntmass, n = 2L)

x2ions(aam, ctmass, n = 2L)
}
\arguments{
\item{ntmass}{The mass of a fixed or variable N-term modification.}

\item{ctmass}{The mass of a fixed or variable C-term modification.}

\item{aam}{A sequence of amino-acid residues with \emph{masses}. Residues
  are in names and masses in values.

  The masses reflects fixed/variable modifications, and/or fixed/variable
  neutral losses.}

\item{n}{The charge state.}
}
\description{
b-ions first, then y-ions

For (1) "amods- tmod- vnl- fnl-", (2) "amods- tmod+ vnl- fnl-".

# (1) OH (C-term), + H (neutralizes the N-term on a fragment) + H+ 
# (2) Other C-term (other than OH) + H + H+: X + 1.007825 + 1.00727647

\code{NH3 = 17.026549}

\code{CO = 27.9949146}

\code{+CO -H2 = 27.9949146 - 2 * 1.007825}

\code{-CO -NH3 = -(27.9949146 + 17.026549)}

\code{-CO -H2O = -(27.9949146 + 18.010565)}
}
\examples{
\donttest{
library(mzion)
library(stringr)

## (1) "amods- tmod- vnl- fnl-"
# (Fixed N-term mods; also for no N-term mod)

fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("Oxidation (M)", "Deamidated (N)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

aa_masses <- aa_masses_all[[1]]

ntmod <- attr(aa_masses, "ntmod", exact = TRUE)
ctmod <- attr(aa_masses, "ctmod", exact = TRUE)

if (!length(ntmod)) {
  ntmass <- aa_masses["N-term"] - 0.000549 # - electron
} else {
  ntmass <- aa_masses[names(ntmod)] - 0.000549
}

if (!length(ctmod)) {
  ctmass <- aa_masses["C-term"] + 2.01510147 # + (H) + (H+)
} else {
  ctmass <- aa_masses[names(ctmod)] + 2.01510147
}

aa_seq <- "MAKEMASSPECFUN"
aas <- stringr::str_split(aa_seq, "", simplify = TRUE)
aam <- aa_masses[aas]

b <- mzion:::bions_base(aam, ntmass)
y <- mzion:::yions_base(aam, ctmass)


## (2) "amods- tmod+ vnl- fnl-"
# (2a, N-term)
fixedmods <- c("TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("TMT6plex (N-term)", "Acetyl (Protein N-term)", "Oxidation (M)",
             "Deamidated (N)", "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

aa_masses <- aa_masses_all[[3]]

# (Fixed or variable C-term mods +/- makes no difference on b-ions;
# and vice versa for y-ions)
ntmod <- attr(aa_masses, "ntmod", exact = TRUE)
ctmod <- attr(aa_masses, "ctmod", exact = TRUE)

if (!length(ntmod)) {
  ntmass <- aa_masses["N-term"] - 0.000549
} else {
  ntmass <- aa_masses[names(ntmod)] - 0.000549
}

if (!length(ctmod)) {
  ctmass <- aa_masses["C-term"] + 2.01510147
} else {
  ctmass <- aa_masses[names(ctmod)] + 2.01510147
}

aa_seq <- "MAKEMASSPECFUN"
aas <- stringr::str_split(aa_seq, "", simplify = TRUE)
aam <- aa_masses[aas]

b <- mzion:::bions_base(aam, ntmass)
y <- mzion:::yions_base(aam, ctmass)


# (2b, C-term)
fixedmods <- c("TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("TMT6plex (N-term)", "Amidated (Protein C-term)", "Oxidation (M)",
             "Deamidated (N)", "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

# `TMT6plex (N-term)`; `Amidated (Protein C-term)`
aa_masses <- aa_masses_all[[7]]

ntmod <- attr(aa_masses, "ntmod", exact = TRUE)
ctmod <- attr(aa_masses, "ctmod", exact = TRUE)

if (!length(ntmod)) {
  ntmass <- aa_masses["N-term"] - 0.000549
} else {
  ntmass <- aa_masses[names(ntmod)] - 0.000549
}

if (!length(ctmod)) {
  ctmass <- aa_masses["C-term"] + 2.01510147
} else {
  ctmass <- aa_masses[names(ctmod)] + 2.01510147
}

b <- mzion:::bions_base(aam, ntmass)
y <- mzion:::yions_base(aam, ctmass)
}
}
