% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{bracket_pasef_ms2info}
\alias{bracket_pasef_ms2info}
\title{Separates PASEF MS2Info by a range of frames}
\usage{
bracket_pasef_ms2info(iso_info, mdata, offset_fr = 3L)
}
\arguments{
\item{iso_info}{A data frame containing ms2info.}

\item{mdata}{Line data of MS frames (containing both MS1 and MS2).}

\item{offset_fr}{The offset from an empty line to the line of a FRAME number.}
}
\description{
Separates PASEF MS2Info by a range of frames
}
