% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calc_auc}
\alias{calc_auc}
\title{Calculate area under a curve}
\usage{
calc_auc(ys, ts, interpol = FALSE, yco = 2)
}
\arguments{
\item{ys}{Y values.}

\item{ts}{Time values.}

\item{interpol}{Not yet used. Logical; interpolation of small values or not.}

\item{yco}{Not yet used. The cut-off values in Y for interpolation.}
}
\description{
Calculate area under a curve
}
