% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calc_fwxm}
\alias{calc_fwxm}
\title{Calculate FWXM}
\usage{
calc_fwxm(
  ys,
  ts,
  ymax,
  ymin,
  len,
  imax,
  max_fwhm = 25,
  h = 0.5,
  min_noks = 5L,
  gap_size = 4L,
  n_margin = 4L,
  n_bimod = 4L
)
}
\arguments{
\item{ys}{A vector of intensity values.}

\item{ts}{A vector of retention times corresponding to \code{ys}.}

\item{ymax}{The maximum of \code{ys}.}

\item{ymin}{The minimum of \code{ys}.}

\item{len}{The length of \code{ys}.}

\item{imax}{The index of \code{ymax} along \code{ys}.}

\item{max_fwhm}{The maximum FWHM for considering as a peak.}

\item{h}{The height (from the apex); the height from baseline: \code{1 - h}.
The value must be \eqn{\le{.5}}.}

\item{min_noks}{The minimum number of \code{ys} values \eqn{\ge} the value at
\code{h} for consideration as a peak.}

\item{gap_size}{The maximum allowance in discontinuous indexes of the
\code{ys} sequence.}

\item{n_bimod}{The gap size for biomodality in \code{ys}.}

\item{margin}{The margin allowance for the maximum.}
}
\description{
For a height \eqn{\ge{.5}}.
}
