% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{calc_monopep}
\alias{calc_monopep}
\title{Calculates the mono-isotopic mass of a peptide sequence.}
\usage{
calc_monopep(
  aa_seq,
  aa_masses,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  min_mass = 200L,
  max_mass = 4500L,
  digits = 4L
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Only used for calc_monopeptide at a user's interface. Typically coupled to
\link{subpeps_by_vmods} for automatic dispatching of peptide sequences by
sets of fixed and variable modifications. For manual calculations, uses
\link{calc_monopeptide}.
}
