% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{calc_monopeptide}
\alias{calc_monopeptide}
\title{Calculates the mono-isotopic mass of a peptide sequence.}
\usage{
calc_monopeptide(
  aa_seq,
  fixedmods,
  varmods,
  maxn_vmods_setscombi = 64L,
  maxn_vmods_per_pep = Inf,
  maxn_sites_per_vmod = Inf,
  min_mass = 200L,
  max_mass = 4500L,
  digits = 4L
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial variable
modifications and neutral losses.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value. The primary guard
against low molecular-weight precursors is \code{min_len}.}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation. The default is 4500.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Only for direct uses from an R console (with trade-offs in speed).
}
\examples{
\donttest{
## No variable modifications
# (1)
x <- calc_monopeptide(aa_seq = "MAKEMASSPECFUN",
                      fixedmods = NULL,
                      varmods = NULL)

stopifnot((unlist(x$mass) - 1594.5369) < 1e-4)

# (2-a)
x <- calc_monopeptide(aa_seq = "MAKEMASSPECFUN",
                      fixedmods = "Oxidation (M)",
                      varmods = NULL)

m <- unlist(x$mass)

stopifnot((m[1] - 1626.5267) < 1e-4)

# (2-b) combinatorial NL for fixed modifications
x <- calc_monopeptide(aa_seq = "MAKEMASSPECFUN",
                      fixedmods = "Oxidation (M)",
                      varmods = NULL)

m <- unlist(x$mass)

stopifnot(length(m) == 2L)
stopifnot((m[1] - m[2] - 127.9965) < 1e-4)

## With variable modifications
# (3-a)
x <- calc_monopeptide(aa_seq = "MAKEMASSPECFUN",
                      fixedmods = NULL,
                      varmods = "Oxidation (M)")

m <- unlist(x$mass)

stopifnot((m[1] - 1594.5369) < 1e-4)
stopifnot((m[2] - m[1]) == (m[3] - m[2]))

# x$vmods_ps

# (3-b)
x <- calc_monopeptide(aa_seq = "MAKEMASSPECFUN",
                      fixedmods = NULL,
                      varmods = "Oxidation (M)")

x$mass

# (4-a)
x <- calc_monopeptide(aa_seq = "MAKEMASSPECFUN",
                      fixedmods = c("TMT6plex (N-term)",
                                    "TMT6plex (K)",
                                    "Carbamidomethyl (C)"),
                      varmods = c("Acetyl (N-term)",
                                  "Gln->pyro-Glu (N-term = Q)",
                                   "Oxidation (M)"))

x$mass

# The N-term M realizes with acetylation
x$vmods_ps[[1]]

# The N-term M realizes with TMT
x$vmods_ps[[2]]


# (4-b)
x <- calc_monopeptide(aa_seq = "MAKEMASSPECFUN",
                      fixedmods = c("TMT6plex (N-term)",
                                    "TMT6plex (K)",
                                    "Carbamidomethyl (C)"),
                      varmods = c("Acetyl (N-term)",
                                  "Gln->pyro-Glu (N-term = Q)",
                                  "Oxidation (M)"))

x$mass
}
}
