% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{calc_ms2ionseries}
\alias{calc_ms2ionseries}
\title{Calculates the mono-isotopic mass of a MS2 ions.}
\usage{
calc_ms2ionseries(
  aa_seq,
  fixedmods,
  varmods,
  type_ms2ions = "by",
  ms1_mass = NULL,
  maxn_vmods_setscombi = 512L,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  digits = 5L
)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{ms1_mass}{The mass of a theoretical MS1 (for subsetting).}

\item{maxn_vmods_setscombi}{Integer; the maximum number of combinatorial variable
modifications and neutral losses.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
For direct uses from an R console (with trade-offs in speed).
}
\examples{
\donttest{
## No variable modifications
# (1)
x <- calc_ms2ionseries("MAKEMASSPECFUN", 
                       fixedmods = NULL, 
                       varmods = NULL)

x$mass

# (2) no combinatorial NL for fixed modifications
x <- calc_ms2ionseries("MAKEMASSPECFUN", 
                       fixedmods = "Oxidation (M)", 
                       varmods = NULL)
                       
x$mass

## With variable modifications
# (3) combinatorial sites and NL available
aa_seq <- "MAKEMASSPECFUN"
fixedmods <- NULL
varmods <- "Oxidation (M)"
ms <- calc_monopeptide(aa_seq, fixedmods, varmods)

x <- calc_ms2ionseries(aa_seq, fixedmods = fixedmods, varmods = varmods, 
                       ms1_mass = ms$mass[[2]][1])

x$mass
# x$vmods_ps

# (4)
aa_seq <- "MAKEMASSPECFUN"
fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("Acetyl (N-term)", "Gln->pyro-Glu (N-term = Q)", "Oxidation (M)")
ms <- calc_monopeptide(aa_seq, fixedmods, varmods)

x <- calc_ms2ionseries(aa_seq, fixedmods = fixedmods, varmods = varmods, 
                       ms1_mass = ms$mass[[4]][[1]])
                      
x$mass

# (5) Neutral losses for occurrences of both fixed 
aa_seq <- "MAKEMASSPECFUN"
fixedmods <- c("TMT6plex (N-term)", "Oxidation (M)", "Deamidated (N)")
varmods <- c("dHex (S)")
ms <- calc_monopeptide(aa_seq, fixedmods, varmods)

# and variable modifications ignored
x <- calc_ms2ionseries(aa_seq, fixedmods = fixedmods, varmods = varmods, 
                       ms1_mass = 2002.7316)
                      
stopifnot(is.null(x$mass[[2]]))

# Change from fixed to variable for full combinatorials
fixedmods <- c("TMT6plex (N-term)", "Deamidated (N)")
varmods <- c("Acetyl (Protein N-term)", "Oxidation (M)", "dHex (S)")

x <- calc_ms2ionseries(aa_seq, fixedmods = fixedmods, varmods = varmods, 
                       ms1_mass = 2002.7316)
                      
x$mass[[8]]
x$vmods_ps[[8]]

# (6) A lot of S
aa_seq <- "MAKEMASSSSSSPECFUNSS"
fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("Acetyl (N-term)", "Oxidation (M)", "Deamidated (N)", 
             "Phospho (S)", "Phospho (T)", "Phospho (Y)", 
             "Gln->pyro-Glu (N-term = Q)")
ms <- calc_monopeptide(aa_seq, fixedmods, varmods)

x <- calc_ms2ionseries(aa_seq, fixedmods = fixedmods, varmods = varmods, 
                       ms1_mass = 2649.0553)

# (7) A lot of S and Y
aa_seq <- "MAKEMASSSSSSPECFUNSSYYYYYYY"
fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("Acetyl (N-term)", "Oxidation (M)", "Deamidated (N)", 
             "Phospho (S)", "Phospho (T)", "Phospho (Y)", 
             "Gln->pyro-Glu (N-term = Q)")
ms <- calc_monopeptide(aa_seq, fixedmods, varmods)

x <- calc_ms2ionseries(aa_seq, fixedmods = fixedmods, varmods = varmods, 
                       ms1_mass = ms$mass[[21]][[3]])
}

}
