% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepfdr}
\alias{calc_pepfdr}
\title{Calculates the cut-off score at a peptide FDR.}
\usage{
calc_pepfdr(
  target_fdr = 0.01,
  fdr_type = "psm",
  min_len = 7L,
  max_len = 50L,
  out_path
)
}
\arguments{
\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{fdr_type}{Character string; the type of FDR for controlling.}

\item{min_len}{A positive integer; the minimum length of peptides. Shorter
peptides will be excluded.}

\item{max_len}{A positive integer; the maximum length of peptides. Longer
peptides will be excluded.}

\item{out_path}{A file path of outputs.}

\item{out}{The output from \link{calc_pepscores}.}

\item{nms}{The name(s) of \code{out} that correspond(s) to decoy results.}
}
\description{
Needs \code{min_len} and \code{max_len} since the target-decoy pair may not
cover all \code{pep_len} values.
}
\examples{
\donttest{
prob_cos <- calc_pepfdr(target_fdr = .01, 
                        fdr_type = "protein", 
                        min_len = 7L, 
                        max_len = 50L, 
                        out_path = "~/proteoM/bi_1")
}
}
