% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_peploc}
\alias{calc_peploc}
\title{Calculates the delta scores of \code{pep_seq}.}
\usage{
calc_peploc(
  x = NULL,
  out_path = NULL,
  mod_indexes = NULL,
  locmods = c("Phospho (S)", "Phospho (T)", "Phospho (Y)"),
  topn_mods_per_seq = 3L,
  topn_seqs_per_query = 3L
)
}
\arguments{
\item{x}{The results from \link{calc_pepscores}.}

\item{out_path}{An output path.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{locmods}{Among \code{varmods} for the consideration of localization
  probabilities; for instance, \code{locmods = NULL} for nothing,
  \code{locmods = c("Phospho (S)", "Phospho (T)", "Phospho (Y)")} for
  phosphopeptides, \code{locmods = "Acetyl (K)"} for lysine acetylation.
  \code{fixedmods} that were coerced to \code{varmods} will be added
  automatically to \code{locmods}.

  For convenience, the default is set to look for applicable peptide
  phosphorylation (and may encounter warning messages if the data type is
  different to the default).}

\item{topn_mods_per_seq}{Positive integer; a threshold to discard variable
  modifications under the same peptide match with scores beyond the top-n.

  The same \code{peptide match} refers to matches with identities in \code{MS
  scan number}, \code{MS raw file name} and \code{peptide sequence}. Target
  and decoys matches are treated separately.

  For a variable modification with multiple neutral losses (NL), the
  best-scored NL will be used in the ranking.}

\item{topn_seqs_per_query}{Positive integer; a threshold to discard peptide
  matches under the same MS query with scores beyond the top-n.

  The same \code{MS query} refers to the identity in \code{MS scan number}
  and \code{MS raw file name}. Target and decoys matches are treated
  separately.}
}
\description{
A score delta between the best and the second best.
}
\details{
There should not be any duplicated rows at the combination of
c("pep_isdecoy", "scan_num", "raw_file", "pep_seq", "pep_ivmod")
}
