% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_probi_byvmods}
\alias{calc_probi_byvmods}
\title{Helper for score calculations}
\usage{
calc_probi_byvmods(
  df,
  nms,
  expt_moverzs,
  expt_ints,
  N,
  type_ms2ions = "by",
  topn_ms2ions = 100L,
  ppm_ms2 = 25L,
  burn_ins = c(1:2),
  digits = 5L
)
}
\arguments{
\item{df}{Two lists of \code{theo} and matched \code{expt} m-over-z.}

\item{nms}{The names (character strings indicating the names and position of
variable modifications).}

\item{expt_moverzs}{Nested list of match and unmatched experimental m-over-z.}

\item{expt_ints}{Nested list of match and unmatched experimental intensity.}

\item{N}{Numeric; the number of MS2 features in an MGF query.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches. The default is to use the top-100 ions in an MS2 event.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 25.}

\item{burn_ins}{The range of burn-ins where inputs will be excluded from
probablity assessments.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
By the positions of variable modifications.
}
\section{Model}{
 
N - the total number of features (white and black balls) \cr
k - the number of sampled features \cr
m - the numbers of theoretical features (white balls) \cr
n - the number of noise (black balls) \cr

* Subtracts \code{m} and the counts of secondary b0, y0 matches etc. from noise
((N < m) -> (n < 0L); OK if n < 0L) \cr
* \code{ith} and \code{ith2} in ascending order, \code{iex} and \code{iex2} in adaptive order \cr
* One-to-one correspondence: \code{ith} <-> \code{iex}; \code{ith2} <-> \code{iex2} \cr
}

\section{Check matches}{

\code{abs(df2$theo[df2$ith] - expt_moverzs[df2$iex]) <= ppm_ms2} \cr
\code{identical(df2$int[df2$ith], expt_ints[df2$iex])}
}

\section{y}{

y$expt - experimental m/z's \cr
y$int  - experimental intensities (primary) \cr
y$int2 - experimental intensiteis (secondary) \cr
y$theo - matched theoretical m/z's \cr
y$idx  - values: the indexes in theoretical sequence (df$ith)
         positions: the position in expt_moverzs
}

\examples{
\donttest{
##
pep <- "YGPQYGHPPPPPPPPDYGPHADSPVLMVYGLDQSK"
nms <- unlist(stringr::str_split(pep, ""))

theos <- c(393.2335,450.2550,547.3078,675.3663,838.4297,
           895.4511,1032.5100,1129.5628,1226.6156,1323.6683,
           1420.7211,1517.7739,1614.8266,1711.8794,1808.9322,
           1923.9591,2087.0224,2144.0439,2241.0967,2378.1556,
           2449.1927,2564.2196,2651.2517,2748.3044,2847.3728,
           2960.4569,3091.4974,3190.5658,3353.6291,3410.6506,
           3523.7347,3638.7616,3766.8202,3853.8522,4211.1101,
           376.2757,463.3078,591.3663,706.3933,819.4773,
           876.4988,1039.5621,1138.6306,1269.6710,1382.7551,
           1481.8235,1578.8763,1665.9083,1780.9353,1851.9724,
           1989.0313,2086.0840,2143.1055,2306.1688,2421.1958,
           2518.2485,2615.3013,2712.3541,2809.4068,2906.4596,
           3003.5124,3100.5651,3197.6179,3334.6768,3391.6983,
           3554.7616,3682.8202,3779.8729,3836.8944,3999.9577)

expts <- c(NA,NA,NA,675.36646,838.42981,895.45056,1032.51025,
           NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,
           NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,376.27603,463.30823,
           591.36676,706.39380,819.47552,876.50018,1039.56287,
           NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,
           NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA)

names(expts) <- names(theos) <- c(nms, rev(nms))

ith <- c(4,5,6,7,36,37,38,39,40,41,42)
iex <- c(42,57,72,93,20,26,36,44,56,65,96)
ith <- as.integer(ith)
iex <- as.integer(iex)
m <- length(ith)

df <- list(theo = theos, expt = expts, ith = ith, iex = iex, m = m)

expt_moverzs <- c(126.05530,126.12798,127.12505,127.13139,128.12843,
                  128.13475,129.13177,129.13806,130.13516,130.14140,
                  131.13841,136.07588,143.11812,195.11308,221.09239,
                  230.17036,235.14416,244.09303,271.12891,376.27603,
                  377.28003,384.21320,400.54648,441.89435,442.22861,
                  463.30823,464.31158,502.76166,503.26221,516.75928,
                  517.26062,531.28296,565.28522,565.78705,574.33997,
                  591.36676,592.36957,613.81238,614.31451,648.34070,
                  669.80933,675.36646,696.86902,706.39380,707.39600,
                  713.32672,713.82715,714.33093,761.85156,762.33771,
                  762.85907,810.37610,810.88019,811.38434,811.88660,
                  819.47552,838.42981,846.49664,846.99695,855.74426,
                  859.91162,867.92706,868.43024,868.93225,876.50018,
                  877.50354,878.75262,884.42273,884.75665,885.09052,
                  885.42700,895.45056,896.45435,908.43860,908.94092,
                  909.44440,916.95612,949.79688,950.13049,950.46545,
                  950.79791,962.48212,964.98084,965.48273,987.49255,
                  987.82599,988.16028,988.50989,1004.51562,1005.52985,
                  1013.56659,1031.50378,1032.51025,1033.51355,1034.51636,
                  1039.56287,1040.56641,1062.48096,1063.48511,1161.55078)

expt_ints <- c(16921,29468,36904,28121,37829,23537,39307,36194,25192,33532,
               26551,91477,15182,24720,17471,50430,14282,14084,39681,99581,
               13900,52774,16289,17160,14127,115919,22728,29658,16358,77851,
               28386,18132,49422,32030,39155,78946,18520,25765,13728,19688,
               13347,12440,14183,95484,26643,20027,14759,13589,14318,18829,
               12227,19829,16826,31208,22442,18285,14859,31434,13299,16824,
               12843,21389,21024,12491,69633,25811,14946,28967,55172,29773,
               17955,40642,19888,35676,35222,14974,12466,46216,86923,48637,
               30824,12961,26805,31218,36352,49433,40495,31233,40643,18265,
               12316,25125,202241,90877,20903,40353,15008,31908,22554,13634)

calc_probi_byvmods(df, nms = "0000000", expt_moverzs, expt_ints, N = 404)

## 
pep <- "LFEEDEREK"
nms <- unlist(stringr::str_split(pep, ""))

theos <- c(343.2543,490.3227,619.3653,748.4079,863.4348,992.4774,
           1148.5785,1277.6211,1634.8790,376.2757,505.3183,661.4194,
           790.4620,905.4890,1034.5316,1163.5742,1310.6426,1423.7266)

expts <- c(343.25455,490.32318,619.36487,NA,863.43542,NA,NA,
           NA,NA,376.27606,505.31924,NA,NA,NA,NA,NA,NA,NA)

names(expts) <- names(theos) <- c(nms, rev(nms))

ith <- c(1,2,3,5,10,11)
iex <- c(39,57,73,92,41,59)
ith <- as.integer(ith)
iex <- as.integer(iex)
m <- length(ith)

df <- list(theo = theos, expt = expts, ith = ith, iex = iex, m = m)

expt_moverzs <- c(115.08694,120.08116,126.12807,127.12512,127.13135,
                  128.12845,128.13469,129.13177,129.13800,130.13510,
                  130.14134,131.13840,132.14159,136.07581,158.09254,
                  173.15001,175.11917,175.15672,176.15985,186.15297,
                  188.15988,229.16661,230.17041,231.17406,248.18094,
                  249.18439,251.11345,269.12421,273.21259,287.19257,
                  291.17169,301.20755,301.24423,305.16913,315.25937,
                  322.70038,331.21384,331.71515,343.25455,361.71152,
                  376.27606,377.24023,377.27985,386.72955,387.22852,
                  395.23312,395.73502,396.23633,406.26596,407.27014,
                  433.20410,453.24899,470.30963,470.81094,472.27225,
                  487.30685,490.32318,499.81451,505.31924,517.77081,
                  519.27625,519.77734,551.32843,551.63458,551.82617,
                  551.86658,551.96796,552.32220,552.36621,591.37250,
                  592.34991,609.33093,619.36487,633.33594,634.33850,
                  635.34485,636.34497,644.39349,648.31799,655.82574,
                  706.39349,740.87518,741.37933,747.88177,748.39014,
                  748.88788,749.38904,749.89026,750.39099,791.45404,
                  819.47766,863.43542,864.43774,943.52075,944.51770,
                  945.51129,946.51874,947.52203,948.52100,1342.65393)

expt_ints <- c(8508,28501,194673,190890,160012,223742,147623,260388,198943,164580,
               149024,165036,7306,21035,8098,11560,35686,8953,15544,10771,
               17617,7660,227328,16612,88320,10075,8968,9123,20253,7351,
               49432,11202,8013,65251,21790,11174,48669,10941,40270,24887,
               81536,9488,9709,28670,8442,7151,185877,54537,9754,7176,
               10297,12091,42045,9021,14507,11520,27411,7991,45770,7927,
               18996,9520,111456,46109,49064,59457,41824,36185,9619,9065,
               8708,21105,26995,13041,8674,11169,14089,7697,9047,10710,
               12763,7125,7321,7483,31612,18050,26134,31474,11331,8789,
               8672,65913,25918,17819,34367,15767,15221,14225,7419,7284)

calc_probi_byvmods(df, nms = "0000000", expt_moverzs, expt_ints, N = 434)

}
}
