% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_protfdr}
\alias{calc_protfdr}
\title{Calculates the cut-offs of protein scores.}
\usage{
calc_protfdr(
  df = NULL,
  target_fdr = 0.01,
  max_protscores_co = Inf,
  max_protnpep_co = 10L,
  method_prot_es_co = "median",
  out_path = NULL
)
}
\arguments{
\item{df}{An output from upstream steps.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{max_protscores_co}{A positive numeric; the upper limit in the cut-offs
of protein scores for discriminating significant and insignificant
identities.  For higher quality and data-driven thresholds, choose the
default \code{max_protscores_co = Inf}.}

\item{max_protnpep_co}{A positive integer; the maximum number of peptides
under a protein (\code{prot_n_pep}) to warrant the protein significance.
For instance, proteins with \code{prot_n_pep > max_protnpep_co} will have a
protein significance score cutoff of zero and thus are significant. Choose
\code{max_protnpep_co = Inf} to learn automatically the cut-off from data.
Note that the the value of \code{prot_n_pep} includes the counts of shared
peptides.}

\item{method_prot_es_co}{A low-priority setting. A character string; the
method to calculate the cut-offs of protein enrichment scores. The value is
in one of \code{"median", "mean", "max", "min"} with the default of
\code{"median"}. For instance at the default, the median of
\code{peptide_score - pep_score_cutoff} under a protein will be used to
represent the threshold of a protein enrichment score. For more conserved
thresholds, the statistics of \code{"max"} may be considered.}

\item{out_path}{An output path.}
}
\description{
Calculates the cut-offs of protein scores.
}
