% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{calc_rev_ms2}
\alias{calc_rev_ms2}
\title{Calculates the reversed MS2 from the forward}
\usage{
calc_rev_ms2(af, aas)
}
\arguments{
\item{af}{An sequence of answer of the forward.}

\item{aas}{The sequence of amino acid residues.}

\item{l}{The number amino acid residues in a peptide.}
}
\description{
Calculates the reversed MS2 from the forward
}
