% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{calc_semipepmasses}
\alias{calc_semipepmasses}
\title{Finds and calculates the masses of semi-enzymatic sequences.}
\usage{
calc_semipepmasses(
  val,
  pep,
  min_len = 7L,
  max_len = 40L,
  aa_masses,
  ct_offset = 0L
)
}
\arguments{
\item{val}{A mass of a peptide.}

\item{pep}{A character string of peptide.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded.}

\item{ct_offset}{Zero or one to account for the "-" in the C-term of pep.}
}
\description{
Semi-enzymatic sequences are built on full-enzymatic where the N-term
residues are removed sequentially.
}
\details{
N-term "-" has no effect on the semi-enzymatic generation since
semi-enzymatic sequences are always NONE N-term (original enzymatic sequences
keep separately and concatenated later). Thus, the N-term tag of "-" was
removed from \code{pep} in \link{semipeps_byprots}, before calling this
function.
}
\examples{
\donttest{
fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)",
              "Carbamidomethyl (C)")

varmods = c("Acetyl (Protein N-term)", "Oxidation (M)",
            "Deamidated (N)","Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods, varmods = varmods)
aa_masses <- aa_masses_all[[1]]

val <- 4237.89756
pep <- "ALELNQSAEYYYEENEMNYTHDYSQYEVICIK"
ans <- calc_semipepmasses(val, pep, 7, aa_masses)

val <- 2423.1017
pep <- "QNVEEIPFDSEGPTEPTSSFTI-"
ans <- calc_semipepmasses(val, pep, 7, aa_masses, 1L)
}
}
