% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{calc_threeframe_ppm}
\alias{calc_threeframe_ppm}
\title{Calculates the three-frame ppm.}
\usage{
calc_threeframe_ppm(ppm = 20L, is_three_frame = TRUE, fct_ppm = 0.5)
}
\arguments{
\item{ppm}{A positive integer; the mass tolerance of MS1 species. The default
is 20.}

\item{is_three_frame}{Logical; is a three-frame search or not. The value is
always TRUE.}

\item{fct_ppm}{A factor to new ppm error. The value is always 0.5 for a
three-frame search.}
}
\description{
Calculates the three-frame ppm.
}
