% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{calc_tmtint}
\alias{calc_tmtint}
\title{Reporter-ion quantitation.}
\usage{
calc_tmtint(
  data = NULL,
  quant = c("none", "tmt6", "tmt10", "tmt11", "tmt16"),
  ppm_reporters = 10L,
  index_mgf_ms2 = FALSE
)
}
\arguments{
\item{data}{An upstream result from \link{matchMS}.}

\item{quant}{A quantitation method. The default is "none". Additional choices
include \code{tmt6} etc. For other multiplicities of \code{tmt}, use the
compatible higher plexes, for example, \code{tmt16} for \code{tmt12} etc.
and \code{tmt10} for \code{tmt8} etc.}

\item{ppm_reporters}{The mass tolerance of MS2 reporter ions.}

\item{index_mgf_ms2}{A low-priority feature. Logical; if TRUE, converts
  upfrontly MS2 m-over-z values from numeric to integers as opposed to
  \emph{in-situ} conversion during ion matches. The default is FALSE. The
  \code{index_mgf_ms2 = TRUE} might be useful for very large MS files by
  reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}
}
\description{
Not yet used: \code{`134C` = 134.154565}, \code{`135N` = 135.15160}
}
