% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{calc_unimod_compmass}
\alias{calc_unimod_compmass}
\title{Calculates the masses of a chemical formula.}
\usage{
calc_unimod_compmass(composition = "H(4) C O S", digits = 6L)
}
\arguments{
\item{composition}{A chemical composition.}

\item{digits}{A non-negative integer; the number of decimal places to be
used.}
}
\description{
Calculates the masses of a chemical formula.
}
