% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calcFWHM}
\alias{calcFWHM}
\title{Calculate FWHM}
\usage{
calcFWHM(ys, ts, yco = 100)
}
\arguments{
\item{ys}{A vector of Y values.}

\item{ts}{A vector of X values (retention times).}

\item{yco}{The cut-off in Y values.}
}
\description{
Calculate FWHM
}
