% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{calcms1mass_noterm}
\alias{calcms1mass_noterm}
\title{Helper function for parallel calculations peptide masses by proteins.}
\usage{
calcms1mass_noterm(aa_seqs, aa_masses)
}
\arguments{
\item{aa_seqs}{Character string; a vector of peptide sequences with
one-letter representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
For each split of multiple proteins; no terminal masses.
}
