% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{calcms1mass_noterm_bypep}
\alias{calcms1mass_noterm_bypep}
\title{Helper of \link{calcms1mass_noterm_byprot}.}
\usage{
calcms1mass_noterm_bypep(aa_seq, aa_masses)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequence with one-letter
representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
For single protein.
}
