% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{calcms1mass_noterm_byprot}
\alias{calcms1mass_noterm_byprot}
\title{Helper of \link{calcms1mass_noterm}.}
\usage{
calcms1mass_noterm_byprot(prot_peps, aa_masses)
}
\arguments{
\item{prot_peps}{Lists of peptides under a proteins.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
For single protein.
}
