% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{calib_ms1masses}
\alias{calib_ms1masses}
\title{Precursor masses calibration.}
\usage{
calib_ms1masses(
  df = NULL,
  mgf_path = NULL,
  out_path = NULL,
  ppm_ms1 = 20L,
  min_mass = 200L,
  max_mass = 4500L,
  knots = 50L,
  digits = 4L
)
}
\arguments{
\item{df}{A data frame after MS2 matches.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).

  Individuality in MGF files are slightly preferred to take advantage of
  parallel reading of the files.}

\item{out_path}{A file path of outputs.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value. The primary guard
against low molecular-weight precursors is \code{min_len}.}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{knots}{The number of knots for spline fits.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
\code{ppm_ms1} only for the calculation of frame indexes of precursors.
}
