% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{centroid_pasefms2}
\alias{centroid_pasefms2}
\title{Sum PASEF MS2 peak area}
\usage{
centroid_pasefms2(
  xs,
  ys,
  reso = 30000,
  maxn = 500L,
  ymin = 10L,
  grad = 0.25,
  fct_reso = 1.25,
  ms_lev = 2L,
  keep_ohw = TRUE,
  tol = 0.1
)
}
\arguments{
\item{xs}{A vector of ascending m-over-z values.}

\item{ys}{A vector of intensity values.}

\item{reso}{The resolution of a peak.}

\item{maxn}{The maximum number of peaks.}

\item{ymin}{The minimum Y values for considering in peak centroiding.}

\item{grad}{A threshold of gradient between two adjacent peaks (more abundant
over less abundant).}

\item{fct_reso}{A scaling factor for instrument resolution.}

\item{ms_lev}{The level of MS.}

\item{keep_ohw}{Logical; Keep one-hit-wonders or not.}

\item{tol}{The tolerance of Y for defining a peak profile.}
}
\description{
Sum PASEF MS2 peak area
}
