% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{check_aaseq}
\alias{check_aaseq}
\title{Checks the validity of a peptide sequence and dispatched it by fixedmods and
varmods.}
\usage{
check_aaseq(aa_seq, aa_masses_all, fixedmods, varmods)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses_all}{All the amino acid lookup tables.}

\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}
}
\description{
A sequence may be invalide at a given set of fixedmods and varmods. For
example, "MAKEMASSPECFUN" cannot have a mod of "Gln->pyro-Glu (N-term = Q)".
}
