% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{check_anywhere_fmods_coercion}
\alias{check_anywhere_fmods_coercion}
\title{Checks fixed modifications after coercion to variable modifications.}
\usage{
check_anywhere_fmods_coercion(site, aa_masses_all, vmods_ps)
}
\arguments{
\item{site}{An amino acid site; e.g. site = "M".}

\item{aa_masses_all}{All the amino acid lookup tables.}

\item{vmods_ps}{All of the "vmods_ps" attributes from "aa_masses_all".}
}
\description{
The coerced site needs to be present in final \code{aa_masses_all}.
}
