% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{check_chduo}
\alias{check_chduo}
\title{Checks the duplicacy detween charge states 2 and 4.}
\usage{
check_chduo(duo = c(1L, 3L), chs, ps, msxints, cutoff = 1.5)
}
\arguments{
\item{duo}{The potentially duplicated charge states. The values are off by
one less in that the input charge states are derived from searches
beginning charge states 2 instead of 1.}

\item{chs}{The charge states for checking duplicacy.}

\item{ps}{The corresponding positions of the observed \code{chs} in a
m-over-z sequence.}

\item{msxints}{A sequence of intensity values.}

\item{cutoff}{A discriminating cut-off threshold.}
}
\description{
Checks the duplicacy detween charge states 2 and 4.
}
