% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{check_cs_next}
\alias{check_cs_next}
\title{Checks the goodness of a next charge state}
\usage{
check_cs_next(ch, mass, xsub, ysub, sta, msxints, n_ms1s, ppm = 8, f = `+`)
}
\arguments{
\item{ch}{The current charge state.}

\item{mass}{The current mass.}

\item{xsub}{A subset of moverzs.}

\item{ysub}{A subset of msxints.}

\item{sta}{The start position.}

\item{msxints}{MS2 intensities.}

\item{n_ms1s}{The underlying counts that have contributed to \code{moverzs}
and \code{maxints}.}

\item{ppm}{Allowance in mass error when deisotoping.}

\item{f}{A function of plus or minus.}
}
\description{
Next charge state: ch - 1
}
