% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{check_dup_term_any}
\alias{check_dup_term_any}
\title{Check duplicated variable terminal and anywhere modifications with the same
site.}
\usage{
check_dup_term_any(mods)
}
\arguments{
\item{mods}{A set of variable modifications.}
}
\description{
Combinations with variable sites occur on both terminal and anywhere will be
flagged as \code{FALSE} (for removals).
}
\examples{
mods <- c("Oxidation (M)", "Deamidated (N)", "Gln->pyro-Glu (N-term = Q)", "Deamidated (Q)")
mzion:::check_dup_term_any(mods)

mods <- c("TMT6plex (K)", "Acetyl (K)", "Carbamyl (K)")
}
