% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{check_dupfvmods}
\alias{check_dupfvmods}
\title{Checks duplicated modifications between fixedmods and varmods}
\usage{
check_dupfvmods(fixedmods, varmods)
}
\arguments{
\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}
}
\description{
Checks duplicated modifications between fixedmods and varmods
}
