% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{check_existed_psms}
\alias{check_existed_psms}
\title{Checks the existing PSM files.}
\usage{
check_existed_psms(x)
}
\arguments{
\item{x}{A PSM file type.}
}
\description{
Required: psmQ, psmC; Optional: psmT2, psmT3
}
