% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{check_fdr_group}
\alias{check_fdr_group}
\title{Checks the values of \code{fdr_group}}
\usage{
check_fdr_group(fdr_group = c("base", "all", "top3"), oks = c("base", "all"))
}
\arguments{
\item{fdr_group}{A character string; the modification group(s) for uses in
peptide FDR controls. The value is in one of \code{c("all", "base")}. The
\code{base} corresponds to the modification group with the largest number
of matches.}

\item{oks}{A vector of allowed modification groups.}
}
\description{
Not yet used. Takes values of integers or character strings.
}
