% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{check_fmods_pos_site}
\alias{check_fmods_pos_site}
\title{Checks the positions_sites in fixedmods}
\usage{
check_fmods_pos_site(positions_sites)
}
\arguments{
\item{positions_sites}{Lists of positions and sites.}
}
\description{
Checks the positions_sites in fixedmods
}
