% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{check_ggname}
\alias{check_ggname}
\title{Checks file names for ggsave}
\usage{
check_ggname(filename)
}
\arguments{
\item{filename}{Character string; An output file name.}
}
\description{
The same as proteoQ::gg_imgname.
}
