% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{check_half_cs}
\alias{check_half_cs}
\title{Check the goodness of a half charge state.}
\usage{
check_half_cs(ch, mass, xsub, ysub, sta, msxints, ppm = 8, f = `+`)
}
\arguments{
\item{ch}{The current charge state.}

\item{mass}{The current mass.}

\item{xsub}{A subset of moverzs.}

\item{ysub}{A subset of msxints.}

\item{sta}{The start position.}

\item{msxints}{MS2 intensities.}

\item{ppm}{Allowance in mass error when deisotoping.}

\item{f}{A function of plus or minus.}
}
\description{
Check the goodness of a half charge state.
}
