% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{check_locmods}
\alias{check_locmods}
\title{Checks \code{locmods}}
\usage{
check_locmods(fixedmods, varmods, locmods)
}
\arguments{
\item{fixedmods}{Character string(s) of fixed modifications.}

\item{varmods}{Character string(s) of variable modifications. Multiple
  modifications to the same residue are allowed, for example, both
  \code{Carbamyl (M)} and \code{Oxidation (M)}.

  For both \code{fixedmods} and \code{varmods}, the modification title,
  \code{TMT6plex}, applies to all of TMT-6, TMT-10, TMT-11. It is also
  possible to use aliased: (1) \code{TMT10plex} for TMT-10 and
  \code{TMT11plex} for TMT-11 and (2) \code{TMT16plex} for TMTpro. See also
  \link{parse_unimod} for grammars of modification \code{title},
  \code{position} and \code{site}.}

\item{locmods}{Among \code{varmods} for the consideration of localization
  probabilities; for instance, \code{locmods = NULL} for nothing,
  \code{locmods = c("Phospho (S)", "Phospho (T)", "Phospho (Y)")} for
  phosphopeptides, \code{locmods = "Acetyl (K)"} for lysine acetylation.
  \code{fixedmods} that were coerced to \code{varmods} will be added
  automatically to \code{locmods}.

  For convenience, the default is set to look for applicable peptide
  phosphorylation (and may encounter warning messages if the data type is
  different to the default).}
}
\description{
Coerced \code{fixedmods} not considered.
}
