% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl0_fnl0.R
\name{check_ms1_mass_vmods2}
\alias{check_ms1_mass_vmods2}
\title{Checks the MS1 mass for proceeding with MS2 matches.}
\usage{
check_ms1_mass_vmods2(
  vmods_combi,
  aam,
  aa_masses,
  ntmod,
  ctmod,
  ms1_mass,
  tol = 0.001
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{aam}{A sequence of amino-acid residues with \emph{masses}. Residues
are in names and masses in values (note that argument \code{aas}
corresponds to residues without masses).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmod}{The attribute \code{ntmod} from a \code{aa_masses} (for MS1
calculations).}

\item{ctmod}{The attribute \code{ctmod} from a \code{aa_masses} (for MS1
calculations).}

\item{ms1_mass}{The mass of a theoretical MS1 (for subsetting).}

\item{tol}{The tolerance in mass.}
}
\value{
A logical vector.
}
\description{
Maybe missed if a "match" is beyond `maxn_vmods_sitescombi_per_pep`.
}
\details{
'bare + fixed terminals (aa_masses["N-term"] + aa_masses["C-term"]) +
variable terminals + anywhere'.
}
