% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{check_ms1calib}
\alias{check_ms1calib}
\title{Checks the status of a prior execution of precursor mass calibration.}
\usage{
check_ms1calib(out_path = NULL, calib_ms1mass = FALSE)
}
\arguments{
\item{out_path}{A file path of outputs.}

\item{calib_ms1mass}{Logical; if TRUE, calibrates precursor masses.}
}
\value{
TRUE if without precursor mass calibration.
}
\description{
Checks the status of a prior execution of precursor mass calibration.
}
