% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMS2.R
\name{check_ms2frames}
\alias{check_ms2frames}
\title{Checks for pre-existed MS2 bins.}
\usage{
check_ms2frames(
  .path_fasta = "~/proteoM/dbs/fasta/uniprot/uniprot_hs_2020_05",
  ms1_time = NULL,
  .path_cache = "~/proteoM/.MSearches/Cache/Calls",
  path_bin2 = file.path(.path_fasta, "ms2masses", ms1_time, "bin_ms2masses")
)
}
\arguments{
\item{.path_fasta}{The parent file path to the theoretical masses of MS1
precursors. At the NULL default, the path is \code{gsub("(.*)\\.[^\\.]*$",
"\\1", get("fasta", envir = environment())[1])}. The parameter is for the
users' awareness of the structure of file folders and the use of default is
suggested. Occasionally experimenters may remove the file folder for disk
space or under infrequent events of modified framework incurred by the
developer.}

\item{ms1_time}{A cached MS1 time (directory).}

\item{.path_cache}{The file path of cached search parameters. The parameter
is for the users' awareness of the underlying structure of file folders and
the use of default is suggested. Occasionally experimenters may remove the
file folder for disk space or under infrequent events of modified framework
incurred by the developer.}

\item{path_bin2}{A file path to binned MS2 results.}
}
\value{
Time stamps of MS2 that are yet to be binned.
}
\description{
Checks for pre-existed MS2 bins.
}
