% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{check_notches}
\alias{check_notches}
\title{Checks the compatibility between ms1_notches and ms1_neulosses.}
\usage{
check_notches(ms1_notches, ms1_neulosses)
}
\arguments{
\item{ms1_notches}{A numeric vector; notches (off-sets) in precursor masses,
e.g., \code{c(-97.976896)} to account fo the loss of a phosphoric acid in
precursor masses.}

\item{ms1_neulosses}{Character string(s) specifying the neutral losses of
  precursors. The nomenclature is the same as those in argument
  \code{varmods}, e.g., \code{c("Phospho (S)", "Phospho (T)", "Phospho
  (Y)")}.

  The argument is a simplified (narrower) usage of argument
  \code{ms1_notches} with additional specificity in modifications and sites.}
}
\description{
Checks the compatibility between ms1_notches and ms1_neulosses.
}
