% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{check_peak_borders}
\alias{check_peak_borders}
\title{Check the lower intensities in a peak borders}
\usage{
check_peak_borders(ys)
}
\arguments{
\item{ys}{A vector of intensity values.}
}
\description{
Check the lower intensities in a peak borders
}
