% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{check_peak_convex}
\alias{check_peak_convex}
\title{Check the convexness of a peak}
\usage{
check_peak_convex(ys, i_sta, i_midsta, i_midend, i_end, min_n = 15L)
}
\arguments{
\item{ys}{A vector of intensity values.}

\item{i_sta}{The left-start index of a peak along \code{ys}.}

\item{i_midsta}{The middle-start index of a peak along \code{ys}.}

\item{i_midend}{The middle-end index of a peak along \code{ys}.}

\item{i_end}{The right-end index of a peak along \code{ys}.}

\item{min_n}{The minimum points across the peak profile of \code{ys} for
consideration.}
}
\description{
Check the convexness of a peak
}
