% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{check_peak_convex_sub}
\alias{check_peak_convex_sub}
\title{Check the convexness of a peak}
\usage{
check_peak_convex_sub(ys, i_midsta, i_midend, min_n = 15L)
}
\arguments{
\item{ys}{A vector of intensity values.}

\item{i_midsta}{The middle-start index of a peak along \code{ys}.}

\item{i_midend}{The middle-end index of a peak along \code{ys}.}

\item{min_n}{The minimum points across the peak profile of \code{ys} for
consideration.}
}
\description{
Use a sub-sequence of intensity values with a pre-calculated left and right
margins (not much faster than \link{check_peak_convex}).
}
