% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{check_resunimod}
\alias{check_resunimod}
\title{Checks the results from find_unimod}
\usage{
check_resunimod(res)
}
\arguments{
\item{res}{Results from \link{find_unimod}}
}
\description{
Shows warning if the same \code{site} with different fixedmods.
}
