% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{check_tmt_pars}
\alias{check_tmt_pars}
\title{Checks the compatibility of TMT names and plexes.}
\usage{
check_tmt_pars(fixedmods, varmods, quant)
}
\arguments{
\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications. Multiple
modifications to the same residue are allowed, for example, both a less
common \code{Carbamyl (M)} and a common \code{Oxidation (M)}.}

\item{quant}{A quantitation method. The default is "none". Additional choices
include \code{tmt6, tmt10, tmt11, tmt16 and tmt18}. For other
multiplicities of \code{tmt}, use the compatible higher plexes. For
example, apply \code{tmt16} for \code{tmt12} provided a set of 12-plexes
being constructed from a 16-plex TMTpro (7 * 13C + 2 * 15N). It is also
possible that an experimenter may construct a \code{tmt12} from a 18-plex
TMTpro (8 *13C + 1 * 15N) and thus \code{quant = tmt18} is suitable.}
}
\description{
Checks the compatibility of TMT names and plexes.
}
