% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{chunksplit_spmat}
\alias{chunksplit_spmat}
\title{Splits sparse matrix by chunks.}
\usage{
chunksplit_spmat(mat, peps = NULL, n_chunks = 4L)
}
\arguments{
\item{mat}{A sparse matrix.}

\item{peps}{The names of peptide sequences.}

\item{n_chunks}{The number of chunks.}
}
\description{
The same peptide sequence spans multiple consecutive rows will stay in the
same chunk.
}
