% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{clean_flat_mgfs}
\alias{clean_flat_mgfs}
\title{Cleans up peak lists.}
\usage{
clean_flat_mgfs(
  mgfs,
  min_mass = 200L,
  max_mass = 4500L,
  ppm_ms1_bin = 10L,
  ms1_offset = 0L,
  is_ms1notched = FALSE,
  n_13c = 0L,
  mass_co_13c = 1500,
  proton = 1.00727647
)
}
\arguments{
\item{mgfs}{Peak lists with \emph{flat} precursor masses.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{ppm_ms1_bin}{The tolerance in precursor mass error after mass binning.}

\item{ms1_offset}{A MS1 mass off-set.}

\item{is_ms1notched}{Logical; is there an off-set (notch) in MS1 mass.}

\item{n_13c}{Number(s) of 13C off-sets in precursor masses, for example, over
the range of \code{-1:2}. The default is 0.}

\item{mass_co_13c}{The mass cut-off per number of 13C for considering 13C
off-sets.}

\item{proton}{The mass of proton.}
}
\description{
Cleans up peak lists.
}
