% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{coerce_fvmods}
\alias{coerce_fvmods}
\title{Coerces conflicting fixedmods to varmods}
\usage{
coerce_fvmods(fixedmods, varmods)
}
\arguments{
\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}
}
\description{
Coerces conflicting fixedmods to varmods
}
