% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{coll_cent_pasefms1}
\alias{coll_cent_pasefms1}
\title{Collapse PASEF MS1}
\usage{
coll_cent_pasefms1(data, pasef_slice_size = 1L)
}
\arguments{
\item{data}{List data in an MS1 frame.}

\item{pasef_slice_size}{The gap of scan indexes for data binning to a slice.}
}
\description{
By every interval of scan indexes according to \code{gap}.
}
