% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{coll_cent_pasefms1_v1}
\alias{coll_cent_pasefms1_v1}
\title{Collapse PASEF MS1 and keep track of slices with MS2 occurrence.}
\usage{
coll_cent_pasefms1_v1(data, ms1_iso)
}
\arguments{
\item{data}{List data in an MS1 frame.}

\item{ms1_iso}{MS1 \code{iso_info} corresponding to \code{data}.}
}
\description{
Not currently used but may be an interesting approach.
}
