% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{collapse_mms1ints_orig}
\alias{collapse_mms1ints_orig}
\title{Collapse MS1 intensities.}
\usage{
collapse_mms1ints_orig(xs, ys, lwr, step = 1e-05)
}
\arguments{
\item{xs}{Vectors of m-over-z values.}

\item{ys}{Vectors of intensity values.}

\item{lwr}{The lower mass limit.}

\item{step}{The step size for mass bins.}
}
\description{
Not currently used. The original summing approach.
}
