% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{collapse_pasef_fullms1scans}
\alias{collapse_pasef_fullms1scans}
\title{Collapses slices in an MS1 frame.}
\usage{
collapse_pasef_fullms1scans(data, iso_info, min_ms2n = 0L)
}
\arguments{
\item{data}{One frame of MS1 data before combining slices by IM groups.}

\item{iso_info}{A subset of MS1 isolation information for the current
\code{frame}.}

\item{min_ms2n}{The minimum number of MS2 in a slices for considerations.}
}
\value{
A vector of lists. Each list corresponding to a slice of MS1 between
  ScanNumBegin and ScanNumEnd in a frame.
}
\description{
Collapses slices in an MS1 frame.
}
