% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{collapse_pasef_ms1scans}
\alias{collapse_pasef_ms1scans}
\title{Collapses PASEF MS scans}
\usage{
collapse_pasef_ms1scans(
  xss,
  yss,
  iso_info,
  slice_starts,
  slice_ends,
  mobils,
  ret_times
)
}
\arguments{
\item{xss}{Lists of X values (aggregated by every ten indexes of slices) in a
frame.}

\item{yss}{Lists of Y values in a frame.}

\item{iso_info}{Metadata in a frame.}

\item{slice_starts}{The start indexes of scans.}

\item{slice_ends}{The end indexes of scans.}

\item{mobils}{Mobilities.}

\item{ret_times}{Retention times.}
}
\description{
Collapses PASEF MS scans
}
