% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{collapse_pasef_ms2scans}
\alias{collapse_pasef_ms2scans}
\title{Collapses slices in an MS2 frame.}
\usage{
collapse_pasef_ms2scans(data, iso_info, min_ms2n = 0L)
}
\arguments{
\item{data}{One frame of MS2 data before combining slices by IM groups.}

\item{iso_info}{A subset of MS2 isolation information at the current MS2
\code{frame}.}

\item{min_ms2n}{The minimum number of MS2 in a slices for considerations.}
}
\value{
A vector of lists. Each list corresponding to a slice of MS2 between
  ScanNumBegin and ScanNumEnd in a frame, with additional fields of
  IsolationMz, IsolationWidth, Precursor etc.
}
\description{
Collapses slices in an MS2 frame.
}
