% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{collapse_pasef_xys}
\alias{collapse_pasef_xys}
\title{Collapse PASEF X and Y values within an MS1 or MS2 frame.}
\usage{
collapse_pasef_xys(xs, ys)
}
\arguments{
\item{xs}{Slices of m-over-z values.}

\item{ys}{Slices of intensity values.}
}
\description{
Collapse PASEF X and Y values within an MS1 or MS2 frame.
}
