% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{collapse_rawtims_xys}
\alias{collapse_rawtims_xys}
\title{Collapses timsTOF X and Y values.}
\usage{
collapse_rawtims_xys(
  xs,
  ys,
  maxn_peaks = 1000L,
  yco = 20L,
  miny = 30L,
  tol = 2e-05
)
}
\arguments{
\item{xs}{A vector of m-over-z values.}

\item{ys}{A vector of intensity values.}

\item{maxn_peaks}{The maximum number of peaks for consideration.}

\item{yco}{The noise levels of intensity values.}

\item{miny}{The minimum value of intensity for consideration as a peak.}

\item{tol}{The tolerance for masses binning.}
}
\description{
Collapses timsTOF X and Y values.
}
