% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{collapse_sortpeps}
\alias{collapse_sortpeps}
\title{Helper of \link{map_pepprot}.}
\usage{
collapse_sortpeps(mat, ncol = NULL, peps = NULL)
}
\arguments{
\item{mat}{A dgCMatrix object. Column names are SORTED protein accessions.
Rownames are SORTED peptide sequences.}

\item{ncol}{The number of columns in \code{mat}.}

\item{peps}{Peptide sequences as the row names of \code{mat}.}
}
\description{
Collapses the counts the number of peptide sequences under proteins.
}
\details{
The row names in the input matrix need to be \emph{sorted}. The output is a
vector and will be later wrapped into a sparse matrix.
}
