% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{collapse_vecs}
\alias{collapse_vecs}
\title{Pastes vectors to character strings.}
\usage{
collapse_vecs(vecs, nm = "theo", sep = ";")
}
\arguments{
\item{vecs}{A list of vectors.}

\item{nm}{The name of sub list in \code{vecs}.}

\item{sep}{A separator.}
}
\description{
Pastes vectors to character strings.
}
