% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{comb_ms1_offsets}
\alias{comb_ms1_offsets}
\title{Combines off-sets in precursor masses (notches and neutral losses)}
\usage{
comb_ms1_offsets(ms1_offsets = 0, ms1_neulosses = NULL)
}
\arguments{
\item{ms1_offsets}{Precursor mass off-sets (notches, not neutral losses).}

\item{ms1_neulosses}{Character string(s) specifying the neutral losses of
  precursors. The nomenclature is the same as those in argument
  \code{varmods}, e.g., \code{c("Phospho (S)", "Phospho (T)", "Phospho
  (Y)")}.

  The argument is a simplified (narrower) usage of argument
  \code{ms1_notches} with additional specificity in modifications and sites.}
}
\value{
A vector of mass off-sets.
}
\description{
The return contains no information of Unimod titles and positions since it is
only used for pairing with experimental MGF data.
}
