% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{comb_mstraces}
\alias{comb_mstraces}
\title{Combines adjacent MS2 traces.}
\usage{
comb_mstraces(xs, ys, zs, ws, n_dia_ms2bins = 1L)
}
\arguments{
\item{xs}{Vectors of MS2 moverzs.}

\item{ys}{Vectors of MS2 intensities.}

\item{zs}{Vectors of MS2 charge states.}

\item{ws}{Weights.}

\item{n_dia_ms2bins}{Allowance in adjacent DIA-MS2 bins for retention time
alignment with a DIA-MS1.}
}
\description{
Not used.
}
