% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{combi_byvmodsM}
\alias{combi_byvmodsM}
\title{Permutations of a set of MS1 labels.}
\usage{
combi_byvmodsM(M, aas)
}
\arguments{
\item{M}{A matrix of labels of modification names with permutation. Each
  permutation in a row. The corresponding matrix of site permutations in the
  attribute of \code{resids}.

  Note that M is a matrix other than lists of vectors, which allows the
  application of one copy of attributes to all rows.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}
}
\description{
Inputs in a matrix where each row corresponds to a permutation entry. With
one-to-one correspondance between \emph{names} and \emph{site}, the input is
a single-row matrix.
}
\details{
The permutations are first against site matrix M2 (or M3 at a list form for
speed). Note that, e.g., \emph{names} 'Carbamidomethyl (M)' and 'Carbamyl
(M)' are both at \emph{site} 'M'. The redundancy at the levels of
\emph{sites} are tracked and later applied back to form non-redundant
permutations by \emph{names}.
}
\examples{
\donttest{
# One MS1 set of labels:
#    "Carbamidomethyl (M)", "Carbamyl (M)", "Deamidated (N)"
#
# Expands to six name permutation (an input matrix)
#   (each row unique by names but not at sites)
#
#      [,1]                  [,2]                  [,3]
# [1,] "Carbamidomethyl (M)" "Carbamyl (M)"        "Deamidated (N)"
# [2,] "Carbamidomethyl (M)" "Deamidated (N)"      "Carbamyl (M)"
# [3,] "Carbamyl (M)"        "Carbamidomethyl (M)" "Deamidated (N)"
# [4,] "Carbamyl (M)"        "Deamidated (N)"      "Carbamidomethyl (M)"
# [5,] "Deamidated (N)"      "Carbamidomethyl (M)" "Carbamyl (M)"
# [6,] "Deamidated (N)"      "Carbamyl (M)"        "Carbamidomethyl (M)"

# The same 'site' table for both rows [1,] and [3,]
#   (Redundancy at sites tracked)
# 5   9   6
# "M" "M" "N"
# 5  13   6
"M" "M" "N"
9  13   6
# "M" "M" "N"
# 5   9  14
# "M" "M" "N"
# 5  13  14
# "M" "M" "N"
# 9  13  14
# "M" "M" "N"

# Differentiated after applied back to 'names'
#        5                   9               6
# "Carbamidomethyl (M)" "Carbamyl (M)" "Deamidated (N)"
#        5                   9               6
# "Carbamyl (M)" "Carbamidomethyl (M)" "Deamidated (N)"
# ...
}
}
