% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{combi_byvmodsR}
\alias{combi_byvmodsR}
\title{Helper of \link{combi_byvmodsM} (by each rows of labels).}
\usage{
combi_byvmodsR(vec, ps, aas)
}
\arguments{
\item{vec}{A vector of labels.}

\item{ps}{Named vector; counts for each site. Sites in names and counts in
values.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}
}
\description{
Finds the positions of residues (sites) from a given amino acid sequence
(aas).
}
\examples{
\donttest{
vec <- c("M", "M", "N")
ps <- c(M = 2, N = 1)
aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

ans <- combi_byvmodsR(vec, ps, aas)

stopifnot(length(ans) == 6L)
}
}
