% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{combi_mat}
\alias{combi_mat}
\title{Populates the count matrix for combinations.}
\usage{
combi_mat(nb = 5L, ns = 3L)
}
\arguments{
\item{nb}{The number of balls.}

\item{ns}{The number of samplings (the number of columns).}
}
\description{
Populates the count matrix for combinations.
}
