% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl0_fnl0.R
\name{combi_mvmods2}
\alias{combi_mvmods2}
\title{Finds the combinations of variable modifications (multiple sites).}
\usage{
combi_mvmods2(
  amods,
  aas,
  aa_masses,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  digits = 4L
)
}
\arguments{
\item{amods}{Anywhere modifications.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{maxn_vmods_per_pep}{The maximum number of \code{Anywhere}
(non-terminal) variable modifications per peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
\code{Anywhere} (non-terminal) variable modifications per site in a per
peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence. The default is
64. May consider a smaller value, i.e. 32, when searching against
phosphopeptide data.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\value{
Lists by residues in \code{amods}.
}
\description{
For all the \code{Anywhere} modifications specified in \code{amods}.
}
\examples{
\donttest{
## M
fixedmods = c("TMT6plex (K)", "dHex (S)")
varmods = c("Carbamidomethyl (M)", "Carbamyl (M)", "Acetyl (Protein N-term)")

aa_masses_all <- calc_aamasses(fixedmods, varmods,
                               add_varmasses = FALSE,
                               add_nlmasses = FALSE)

aa_masses <- aa_masses_all[[8]]

amods <- list(`Carbamidomethyl (M)` = c(Anywhere = "M"), 
              `Carbamyl (M)` = c(Anywhere = "M"))

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

ans <- combi_mvmods2(amods, aas, aa_masses)

## M and N
fixedmods = c("TMT6plex (K)", "dHex (S)")
varmods = c("Carbamidomethyl (M)", "Carbamyl (M)", 
            "Deamidated (N)", "Acetyl (Protein N-term)")

aa_masses_all <- calc_aamasses(fixedmods, varmods,
                               add_varmasses = FALSE,
                               add_nlmasses = FALSE)

aa_masses <- aa_masses_all[[16]]

amods <- list(`Carbamidomethyl (M)` = c(Anywhere = "M"), 
              `Carbamyl (M)` = c(Anywhere = "M"), 
              `Deamidated (N)` = c(Anywhere = "N"))

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

ans <- combi_mvmods2(amods, aas, aa_masses)

}
}
