% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{combi_namesiteM}
\alias{combi_namesiteM}
\title{Helper of \link{combi_vmodsMat} (by each rows of labels).}
\usage{
combi_namesiteM(M, aas, nrows)
}
\arguments{
\item{M}{A matrix of labels of modification names with permutation. Each
  permutation in a row. The corresponding matrix of site permutations in the
  attribute of \code{resids}.

  Note that M is a matrix other than lists of vectors, which allows the
  application of one copy of attributes to all rows.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{nrows}{The number of rows of M.}
}
\description{
Multiple Names to the same Site. Finds the positions of
residues (sites) from a given amino acid sequence (aas).
}
