% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl0_fnl0.R
\name{combi_np}
\alias{combi_np}
\title{Helper of \link{combi_vmods2}.}
\usage{
combi_np(n, p)
}
\arguments{
\item{n}{The number of modifications.}

\item{p}{The number of positions.}
}
\description{
Helper of \link{combi_vmods2}.
}
