% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{combine_fracs}
\alias{combine_fracs}
\title{Combines results from fractions}
\usage{
combine_fracs(files, tempdir, sc_path)
}
\arguments{
\item{files}{A list of file names}

\item{tempdir}{A temporary directory containing the files}

\item{sc_path}{An output path}
}
\description{
Combines results from fractions
}
