% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{combine_ion_matches}
\alias{combine_ion_matches}
\title{Combines the results of ion matches.}
\usage{
combine_ion_matches(out_path, out_paths, type = "ion_matches_")
}
\arguments{
\item{out_path}{A parent output path.}

\item{out_paths}{Sub output pathes.}

\item{type}{The type of data for combining.}
}
\description{
Combines the results of ion matches.
}
