% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{combine_prisec_matches}
\alias{combine_prisec_matches}
\title{Combines primary and secondary matches.}
\usage{
combine_prisec_matches(th_ex_mgf, type_ms2ions = "by")
}
\arguments{
\item{th_ex_mgf}{The results of theoretical, experimental and MGF.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}
}
\description{
Combines primary and secondary matches.
}
