% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{comine_PSMsubs}
\alias{comine_PSMsubs}
\title{Combines PSMs from sub folders.}
\usage{
comine_PSMsubs(sub_paths, groups, out_path)
}
\arguments{
\item{sub_paths}{A list of sub paths.}

\item{groups}{A character vector of sample groups. The group names will also
be applied to the \code{pep_group} in PSM tables.}

\item{out_path}{An output path.}
}
\description{
Combines PSMs from sub folders.
}
