% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{comp_iso_info}
\alias{comp_iso_info}
\title{Compile PASEF MS1 slices into non-overlapping sections}
\usage{
comp_iso_info(stas, ends)
}
\arguments{
\item{stas}{A vector of start indexes.}

\item{ends}{A vector of end indexes.}
}
\description{
Not yet used; try this later.
}
