% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{concatFracs}
\alias{concatFracs}
\title{Concatenates localization fractions.}
\usage{
concatFracs(x, y, d = 0.001)
}
\arguments{
\item{x}{A vector of counts of MS2 matches between two adjacent STY sites
(for a preceding match: pep_ivmod).}

\item{y}{A vector of counts of MS2 matches between two adjacent STY sites
(for a next match: pep_ivmod).}

\item{d}{A small positive number to handle value \eqn{0}.}
}
\description{
The probability of the second localization with \eqn{y} in \eqn{x = 0; y = 6}
is not any more probable than that in \eqn{x = 0; y = 1}. This is different
to the binomial model in A-score, which will lead to the finding that the
probability of localization 2 at \eqn{x = 0; y = 6} is much more probable.
}
\examples{
\donttest{
library(proteoM)

proteoM:::concatFracs(c(3, 3, 5), c(2, 1, 3))
proteoM:::concatFracs(c(0, 6), c(1, 3))
proteoM:::concatFracs(c(0, 6), c(0, 3))
proteoM:::concatFracs(c(0, 2), c(0, 2))
proteoM:::concatFracs(c(1, 2), c(1, 2))

proteoM:::concatFracs(c(1, 2), c(1, 4))
proteoM:::concatFracs(c(0, 2), c(0, 4))
proteoM:::concatFracs(c(1, 2), c(0, 4)) # near 1 0 0

proteoM:::concatFracs(0, 2)
proteoM:::concatFracs(2, 0)
proteoM:::concatFracs(0, 0) # should not occur
}
}
