% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{count_elements}
\alias{count_elements}
\title{A simplified table utility.}
\usage{
count_elements(vec)
}
\arguments{
\item{vec}{A named vector.}
}
\description{
A faster alternative to \code{table}.
}
\examples{
\dontrun{
library(mzion)
library(microbenchmark)

vec <- c("Carbamidomethyl (M)", "Carbamyl (M)", "Carbamidomethyl (M)")
microbenchmark(mzion:::count_elements(vec), table(vec))
}
}
