% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolator.R
\name{creat_folds}
\alias{creat_folds}
\title{Creates folds for cross validation.}
\usage{
creat_folds(y, k = 10L, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{The labels.}

\item{k}{The number of folds.}

\item{list}{Logical; should the result be a list or not.}

\item{returnTrain}{Logical; return training sets or not (test sets).}
}
\description{
From package caret.
}
