% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{cut_proteinGroups}
\alias{cut_proteinGroups}
\title{Cuts proteins into groups.}
\usage{
cut_proteinGroups(M = NULL, out_path = NULL)
}
\arguments{
\item{out_path}{A file pth to outputs.}

\item{mat}{A logical matrix; peptides in rows and proteins in columns.}
}
\description{
By the number of shared peptides.
}
