% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolator.R
\name{cv_svm}
\alias{cv_svm}
\title{Calculates cross-validation errors.}
\usage{
cv_svm(train, test, costs = c(0.1, 1, 1, 5, 50), ...)
}
\arguments{
\item{train}{Training set.}

\item{test}{Test set.}

\item{costs}{A vector of costs.}

\item{...}{Additional arguments for svm.}
}
\description{
Optimizes the regularization cost for svm.
}
