% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{deisotope}
\alias{deisotope}
\title{De-isotopes precursor masses.}
\usage{
deisotope(
  moverzs,
  ms1ints,
  ppm = 8L,
  offset_upr = 30L,
  offset_lwr = 30L,
  maxn_prec = 400L,
  min_int = 3e+05,
  m13c = 1.003355,
  order_mz = TRUE
)
}
\arguments{
\item{moverzs}{Mass-to-charge ratios.}

\item{ms1ints}{Precursor intensities.}

\item{ppm}{Allowance in mass error.}

\item{offset_upr}{A cardinal number of upper mass off-sets.}

\item{offset_lwr}{A cardinal number of lower mass off-sets.}

\item{maxn_prec}{The maximum number of precursors.}

\item{min_int}{Minimum intensity for considering as a peak.}

\item{m13c}{Mass of 13C.}

\item{order_mz}{Logical; if TRUE, orders peaks from low to high moverzs.}
}
\description{
De-isotopes precursor masses.
}
