% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{estimate_rtgap}
\alias{estimate_rtgap}
\title{Estimates the number of MS1 scans at a 3-min range of retention times}
\usage{
estimate_rtgap(rts, d = 180)
}
\arguments{
\item{rts}{A vector of retention times.}

\item{d}{The allowance of retention time distance in seconds.}
}
\value{
The number of MS1 scans in a 3-min time frame.
}
\description{
Estimates the number of MS1 scans at a 3-min range of retention times
}
