% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{exeReadPASEF}
\alias{exeReadPASEF}
\title{Helper in executing Cpp-compiled timsdataSampleCpp.exe}
\usage{
exeReadPASEF(raw_file, mgf_path)
}
\arguments{
\item{raw_file}{A RAW file name.}

\item{mgf_path}{A file path RAW MS files.}
}
\description{
Requires mono for Linux or Mac OS.
}
