% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfilereader.R
\name{exeReadRAW}
\alias{exeReadRAW}
\title{Helper in executing CSharp-compiled ReadRAW.exe}
\usage{
exeReadRAW(raw_file, mgf_path)
}
\arguments{
\item{raw_file}{A RAW file name.}

\item{mgf_path}{A file path RAW MS files.}
}
\description{
Requires mono for Linux or Mac OS.
}
