% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{expand_gr}
\alias{expand_gr}
\title{Modified from expand.grid}
\usage{
expand_gr(..., nmax = 3L)
}
\arguments{
\item{...}{Lists of data.}

\item{nmax}{The maximum number of combinations allowed.}
}
\description{
Inputs are in matrices
}
