% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{expand_grid}
\alias{expand_grid}
\title{Modified from expand.grid}
\usage{
expand_grid(..., nmax = 3L)
}
\arguments{
\item{...}{Lists of data.}

\item{nmax}{The maximum number of combinations allowed.}
}
\description{
Net yet used.
}
