% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{expand_grid_rows}
\alias{expand_grid_rows}
\title{Based on expand.grid.}
\usage{
expand_grid_rows(..., use.names = TRUE)
}
\arguments{
\item{use.names}{Logical; uses names or not.}
}
\description{
Outputs are row vectors corresponding to the data.frame from the original
expand.grid.
}
\details{
Some overhead in making row vectors but can avoid the expensive row
subsetting from a data.frame.
}
\examples{
x <- list(`Oxidation (M)` = c(0.000000, 63.998285),
          `Carbamidomethyl (M)` = c(0.000000, 105.024835),
          `Oxidation (M)` = c(0.000000, 63.998285))

expand_grid_rows(x)

x <- list(`Bar (M)` = c(0, 3),
          `Foo (M)` = c(0, 5, 7),
          `Bar (M)` = c(0, 3))

expand_grid_rows(x)

x <- list(`Bar (M)` = c(0, 3))
expand_grid_rows(x)
}
