% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{extract_mgf_rptrs}
\alias{extract_mgf_rptrs}
\title{Extracts reporter-ion data from MGF.}
\usage{
extract_mgf_rptrs(
  ms2_moverzs,
  ms2_ints,
  quant = "none",
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE
)
}
\arguments{
\item{ms2_moverzs}{Lists of MS2 m-over-z values.}

\item{ms2_ints}{Lists of MS2 intensity values.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The argument affects only
TMT data. The range of TMT reporter ions is given by
\code{tmt_reporter_lower} and \code{tmt_reporter_upper}.}
}
\description{
Also purges MS2 m-over-z and intensity when applicable.
}
