% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{extract_pasef_frame}
\alias{extract_pasef_frame}
\title{Extracts MS1 or MS2 data from a PASEF frame}
\usage{
extract_pasef_frame(data, coll = FALSE)
}
\arguments{
\item{data}{Line data in a PASEF frame (with multiple slices).}

\item{coll}{Logical; collapses MS data or not.Should be FALSE. Only TRUE at
debugging.}
}
\description{
Extracts MS1 or MS2 data from a PASEF frame
}
