% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{extract_pasef_ms1}
\alias{extract_pasef_ms1}
\title{Extracts PASEF MS1 data.}
\usage{
extract_pasef_ms1(mdata, keys)
}
\arguments{
\item{mdata}{A list of \emph{MS1} PASEF frames. Each list entry corresponding
to one MS1 frame. Note that each frame contains multiple mobility slices.}

\item{keys}{The key names of output lists.}
}
\description{
Extracts PASEF MS1 data.
}
