% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{extract_pasefms1}
\alias{extract_pasefms1}
\title{Extracts PASEF MS1 data.}
\usage{
extract_pasefms1(
  mdata,
  ms_frs,
  iso_info,
  ms_col = "MS1Frame",
  keys,
  title = "",
  step = 1.6e-05
)
}
\arguments{
\item{mdata}{A list of \emph{MS1} PASEF frames. Each list entry corresponding
to one MS1 frame. Note that each frame contains multiple mobility slices.}

\item{iso_info}{The information of MS1 slices.}

\item{ms_col}{The criteria column in \code{iso_info} for subsetting.}

\item{keys}{The key names of output lists.}

\item{title}{The title (.d file  name).}

\item{step}{A step size for mass binning.}

\item{lens}{The number of lines in each MS2 frame.}
}
\description{
Extracts PASEF MS1 data.
}
