% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{extract_vmodsets}
\alias{extract_vmodsets}
\title{Extracts sets of combinatorial vmods labels.}
\usage{
extract_vmodsets(.ms1_vmodsets, .base_ent, p = 4L, labs = c("Carbamyl (M)"))
}
\arguments{
\item{.ms1_vmodsets}{The complete set of combinatorial variable modifications
for MS1.}

\item{.base_ent}{The base entries of .ms1_vmodsets.}

\item{p}{The number of open positions for filling.}

\item{labs}{The labels of balls.}
}
\description{
For faster calculations of precursor masses.
}
\seealso{
make_ms1_vmodsets
}
