% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{extrDDA}
\alias{extrDDA}
\title{Proc mzML data for mDDA workflows.}
\usage{
extrDDA(
  spec = NULL,
  raw_file = NULL,
  temp_dir = NULL,
  idx_sc = 3L,
  idx_osc = 3L,
  idx_mslev = 2L,
  idx_title = 10L,
  idx_scanList_2 = 11L,
  idx_rt_2 = 2L,
  idx_scan_start_2 = 1L,
  idx_precursor_2 = 12L,
  idx_isolationWindow = 1L,
  idx_ctrmz = 1L,
  idx_lwrmz = 2L,
  idx_uprmz = 3L,
  idx_selectedIonList = 2L,
  idx_moverz = 1L,
  idx_charge = 2L,
  idx_ms1int = 3L,
  idx_bin_2 = 13L,
  idx_scanList_1 = 11L,
  idx_rt_1 = 2L,
  idx_scan_start_1 = 1L,
  idx_bin_1 = 12L,
  idx_scan_lwr_1 = 8L,
  idx_scan_upr_1 = 9L,
  idx_scan_lwr_2 = 8L,
  idx_scan_upr_2 = 9L
)
}
\arguments{
\item{spec}{Spectrum list.}

\item{raw_file}{The RAW file name of an mzML.}

\item{temp_dir}{A temporary directory.}

\item{idx_sc}{Index of scan numbers.}

\item{idx_osc}{Index of original scan numbers.}

\item{idx_mslev}{Index of MS levels.}

\item{idx_title}{Index of scan titles.}

\item{idx_scanList_2}{Index of MS2 scanList.}

\item{idx_rt_2}{Index of MS2 retention times.}

\item{idx_scan_start_2}{Index of MS2 scan start.}

\item{idx_precursor_2}{Index of MS2 precursor.}

\item{idx_isolationWindow}{Index of an isolationWindow.}

\item{idx_ctrmz}{Index of center m-over-z of an isolationWindow.}

\item{idx_lwrmz}{Index of lower m-over-z an isolationWindow.}

\item{idx_uprmz}{Index of upper m-over-z an isolationWindow.}

\item{idx_selectedIonList}{Index of selectedIonList.}

\item{idx_moverz}{Index of m-over-z sequences.}

\item{idx_charge}{Index of charge states.}

\item{idx_ms1int}{Index of MS1 intensities.}

\item{idx_bin_2}{Index of MS2 binary data.}

\item{idx_scanList_1}{Index of MS1 scanList.}

\item{idx_rt_1}{Index of precursor retention times.}

\item{idx_scan_start_1}{Index of MS1 scan starts.}

\item{idx_bin_1}{Index of MS1 binary data.}

\item{idx_scan_lwr_1}{Index of lowest MS1 mass.}

\item{idx_scan_upr_1}{Index of highest MS1 mass.}

\item{idx_scan_lwr_2}{Index of lowest MS2 mass.}

\item{idx_scan_upr_2}{Index of highest MS2 mass.}
}
\description{
Proc mzML data for mDDA workflows.
}
