% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{fill_lc_gaps}
\alias{fill_lc_gaps}
\title{Fills LC gaps.}
\usage{
fill_lc_gaps(ys, n_dia_scans = 4L)
}
\arguments{
\item{ys}{A vector of intensity values.}

\item{n_dia_scans}{The number of adjacent MS scans for constructing a peak
profile and thus for determining the apex scan number of an moverz value
along LC.}
}
\description{
Fills LC gaps.
}
\examples{
library(mzion)

ys <- c(10,0,0,0,11,0,15,0,0,12,0,10,0,0,10)
fill_lc_gaps(ys, 3)
fill_lc_gaps(ys, 2)

ys <- c(10,0,0,0,11,15,15,0,0,12,0,10,0,0,10)
fill_lc_gaps(ys, 3)
fill_lc_gaps(ys, 2)
}
