% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{fill_probs}
\alias{fill_probs}
\title{Helper of \link{fill_probco_nas}}
\usage{
fill_probs(nas, prob_cos, target_fdr = 0.01)
}
\arguments{
\item{nas}{A vector of NA or inferior probability cut-offs to be replaced.}

\item{prob_cos}{Named vector; non-NA probability cutoffs.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}
}
\value{
The replacements. Do not return the concatenated prob_cos and nas, to
  prevent the estimated values from affecting the consequent replacement of
  prob_no_uses.
}
\description{
Helper of \link{fill_probco_nas}
}
