% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_aamasses_vmodscombi}
\alias{find_aamasses_vmodscombi}
\title{Finds the combination of varmods}
\usage{
find_aamasses_vmodscombi(
  varmods = NULL,
  f_to_v = NULL,
  anywhere_coerce_sites = NULL
)
}
\arguments{
\item{varmods}{A character vector of variable modifications.}

\item{f_to_v}{Coerced fixed to variable modifications}

\item{anywhere_coerce_sites}{\code{Anywhere} sites coerced from fixed to
variable modifications.}
}
\description{
Finds the combination of varmods
}
