% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{find_apex_scan}
\alias{find_apex_scan}
\title{Find the apex under a mass column}
\usage{
find_apex_scan(
  k,
  xs_k,
  ys_k,
  apexs_k,
  rts_k,
  rngs_k,
  xm,
  scan,
  ss,
  step = 6e-06,
  min_y = 0
)
}
\arguments{
\item{k}{The k-th column in \code{matx} or \code{maty}.}

\item{xs_k}{The X values under \code{matx[, k]}.}

\item{ys_k}{The Y values under \code{maty[, k]}.}

\item{apexs_k}{All apexes scan numbers (\code{scan_apexs[[k]]}) under column
k.}

\item{rts_k}{All apexes retention times (\code{scan_apexs[[k]]}) under column
k.}

\item{rngs_k}{All apexes scan ranges (\code{scan_apexs[[k]]}) under column k.}

\item{xm}{The experimental X value for tracing.}

\item{scan}{The scan number of an MS2 event corresponding to the \code{xm}.}

\item{ss}{All of the scan numbers.}

\item{step}{A step size.}

\item{min_y}{The cut-off of intensity values.}

\item{rngs}{The scan ranges of apexes under column k.}
}
\description{
Find the apex under a mass column
}
