% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_base_aamasses_index}
\alias{find_base_aamasses_index}
\title{Finds the \code{aa_masses} corresponding to the base modification.}
\usage{
find_base_aamasses_index(out_path)
}
\arguments{
\item{out_path}{An output path.}
}
\description{
Finds the \code{aa_masses} corresponding to the base modification.
}
