% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_baseline_x}
\alias{find_baseline_x}
\title{Find the baseline values of Y}
\usage{
find_baseline_x(vals, vmax, perc = 0.02, max_perc = 0.05, y_rpl = 2)
}
\arguments{
\item{vals}{A vector of Y values.}

\item{vmax}{The maximum of vals.}

\item{perc}{The percentage to the base-peak intensity for cut-offs.}

\item{max_perc}{The maximum percentage of baseline levels.}

\item{y_rpl}{A replacement value of intensities.}
}
\description{
Find the baseline values of Y
}
