% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{find_bestfit}
\alias{find_bestfit}
\title{Finds the best fit between ns and bs}
\usage{
find_bestfit(fit_ns, fit_bs)
}
\arguments{
\item{fit_ns}{The fittings from ns.}

\item{fit_bs}{The fitting from bs.}
}
\description{
Finds the best fit between ns and bs
}
