% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{find_bestnotch}
\alias{find_bestnotch}
\title{Finds the best off-set in precursor masses.}
\usage{
find_bestnotch(x)
}
\arguments{
\item{x}{A data.table object}
}
\description{
Finds the best off-set in precursor masses.
}
