% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{find_charage_state}
\alias{find_charage_state}
\title{Finds the charge state of a mass.}
\usage{
find_charage_state(imax, mass, moverzs, max_charge = 4L, ppm = 5L)
}
\arguments{
\item{mass}{The current m-over-z.}

\item{moverzs}{A vector of m-over-z values.}

\item{max_charge}{The maximum charge state for consideration.}

\item{ppm}{Mass error tolerance.}
}
\description{
Finds the charge state of a mass.
}
\examples{
\donttest{
library(mzion)
moverzs <- c(881 + 1:10*.1, 882.0674, 882.0981, 882.4034, 882.60, 882.7372)
mzion:::find_charage_state(1, moverzs[[1]], moverzs)
mzion:::find_charage_state(5, moverzs[[5]], moverzs)
mzion:::find_charage_state(7, moverzs[[7]], moverzs[-10])
}
}
