% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{find_charge_state}
\alias{find_charge_state}
\title{Finds charge state.}
\usage{
find_charge_state(
  mass,
  imax,
  mint,
  moverzs,
  msxints,
  n_ms1s,
  max_charge = 4L,
  n_fwd = 15L,
  n_bwd = 20L,
  ms_lev = 1L,
  is_dda = TRUE,
  tol = 8e-06
)
}
\arguments{
\item{mass}{The mass of a peak for questing its charge state.}

\item{imax}{The index of the peak, which is often the most intense.}

\item{mint}{Not yet used. The intensity of the peak at \code{imax}.}

\item{moverzs}{MS1 or MS2 Mass-to-charge ratios. The inputs are typically at
weighted-mean statistics.}

\item{msxints}{MS1 or MS2 peak intensities. The inputs are typically at mean
statistics.}

\item{n_ms1s}{The underlying counts that have contributed to \code{moverzs}
and \code{maxints}.}

\item{max_charge}{The maximum charge state.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries when determining
the charge state.}

\item{n_bwd}{Backward looking up to \code{n_bwd} mass entries when
determining the charge state.}

\item{ms_lev}{MS level.}

\item{is_dda}{Logical; is DDA or not.}

\item{tol}{The tolerance in mass errors.}
}
\description{
Finds charge state.
}
\examples{
\donttest{
moverzs <- c(907.968018,908.136475,908.872711,909.073690,909.121,
             909.273670,909.280212,909.45,909.454981,909.459717,
             909.473072,909.675781,909.721191,909.78,909.79,
             909.791633,909.840088,909.969744,910.077278,910.091888,
             910.124076,910.130798,910.180115,910.218689,910.341797,
             910.456115,910.463650,910.474152,910.791669,911.126692,
             911.133219,911.471605)
msxints <- c(2971101,3013171,5774301,24508055,40267873,
             1E6,7340374,1E6,154459778,8312258,
             18375220,22909676,2863776,1.1E6,1.2E6,
             191655971,7273611,32089542,9125108,5008687,
             167021373,12912115,3242378,16278636,4731828,
             98601412,22828501,10539991,26255569,8499971,
             8305187,37899662)
n_ms1s <- rep_len(1L, length(msxints))
mass <- 909.7900
imax <- 15L
mint <- 1200000

# xvals <- moverzs[24]
moverzs[24] <- 910.2919

# duo
moverzs[18] <- 910.0410
moverzs[24] <- 910.2917

# chs1 = 2, chs2 = c(2, 3)
moverzs[21] <- 910.1281
moverzs[22] <- 910.1290
moverzs[12] <- 909.5392

# chs1 = 2, chs2 = 3
moverzs[21:22] <- c(910.1281, 910.1290)
moverzs[8:10] <- c(909.4650, 909.4660, 909.4670)
moverzs[12] <- 909.5392

moverzs[21:22] <- c(910.1281, 910.1290)
moverzs[12] <- 909.5392

# no l1,  no l2
moverzs[21:22] <- c(910.1600, 910.1620)
moverzs[8:10] <- c(909.4650, 909.4660, 909.4670)
}
}
