% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{find_chunkbreaks}
\alias{find_chunkbreaks}
\title{Finds the break points}
\usage{
find_chunkbreaks(vals, n_chunks)
}
\arguments{
\item{vals}{An order vector.}

\item{n_chunks}{The number of chunks.}
}
\description{
Finds the break points
}
\examples{
\donttest{
library(mzion)

vals <- c(rep("B", 6), rep("A", 6), rep("E", 5), rep("D", 5), rep("C", 7))
brs <- mzion:::find_chunkbreaks(vals, 3L)
split(vals, brs)
}
}
