% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{find_cmodtree}
\alias{find_cmodtree}
\title{Find and subset peptides.}
\usage{
find_cmodtree(prps, min_n_res, vmods, posns, len, motifs = NULL)
}
\arguments{
\item{prps}{Lists of peptide sequences with a one-letter representation of
amino acid residues. Each list is named by protein accession.}

\item{min_n_res}{The minimum numbers of residues for a given set of sites.}

\item{vmods}{A named list of variable modifications. See also
\link{find_protntsite} for examples of \code{vmods}.}

\item{posns}{The position (e.g., \code{Protein N-term}, \code{Anywhere},
etc.) of \code{vmods}. The argument can be obtained from \code{vmods} but
passed as a parameter for vectorization.}

\item{len}{The count of \code{vmods} (passed as a parameter for
vectorization).}

\item{motifs}{A list of motifs of amino-acide residues. for example,
"MN|NG". See also \link{matchMS}.}
}
\description{
(1-ct) Protein C-term + site -> (2-ct) Any protein C-term -> (3-ct) Any
C-term + site -> (4-ct) Any C-term.
}
