% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{find_decoy}
\alias{find_decoy}
\title{Find the index or name of decoy results.}
\usage{
find_decoy(out_path, pattern = "^ion_matches_")
}
\arguments{
\item{out_path}{A file path of outputs.}

\item{pattern}{The pattern of files.}
}
\description{
Find the index or name of decoy results.
}
