% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_except_sites}
\alias{find_except_sites}
\title{Finds the coerced fixed sites that can be coerced back to fixed sites.}
\usage{
find_except_sites(fixed_sites, positions_sites)
}
\arguments{
\item{fixed_sites}{Sites coerced from fixed to variable modifications.}

\item{positions_sites}{Lists of positions and sites.}
}
\description{
Finds the coerced fixed sites that can be coerced back to fixed sites.
}
\examples{
\donttest{
library(proteoM)

mods <- c("TMT6plex (K)", "Carbamyl (M)", "Oxidation (M)", "Gln->pyro-Glu (N-term = Q)")
res <- proteoM:::extract_umods(mods)
positions_sites <- lapply(res, `[[`, "position_site")
fixed_sites <- c("TMT6plex (K)" = "K", "Carbamyl (M)" = "M")
}
}
