% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_f_to_v}
\alias{find_f_to_v}
\title{Helper of finding the coercion sites.}
\usage{
find_f_to_v(fixedmods, fmods_ps, vmods_ps)
}
\arguments{
\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{fmods_ps}{Positions and sites of fixed modifications.}

\item{vmods_ps}{Positions and sites of variable modifications.}
}
\value{
A named vector, e.g., \code{c("TMT6plex (K) = K", "TMT6plex
  (N-term)" = "N-term")}
}
\description{
Helper of finding the coercion sites.
}
