% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{find_fdr_fits}
\alias{find_fdr_fits}
\title{Helper to find the best regression.}
\usage{
find_fdr_fits(knot, df)
}
\arguments{
\item{knot}{The number of knots.}

\item{df}{A data frame.}
}
\description{
Helper to find the best regression.
}
