% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_free_mem}
\alias{find_free_mem}
\title{Finds the amount of free system memory.}
\usage{
find_free_mem(sys_ram = NULL)
}
\arguments{
\item{sys_ram}{The putative amount of system RAM, e.g., 32L.}
}
\description{
Outputs of free RAM in the unit of MB.
}
