% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_from_hash}
\alias{find_from_hash}
\title{Finds a value through a hash table.}
\usage{
find_from_hash(ht, val, n_bucks = length(ht), offset = 0L)
}
\arguments{
\item{ht}{A hash table.}

\item{val}{A value.}

\item{n_bucks}{The number of buckets in \code{ht}.}

\item{offset}{An offset. Not yet used.}
}
\description{
Not yet used.
}
