% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_gate_edges2}
\alias{find_gate_edges2}
\title{Helper to find the positions of signal edges}
\usage{
find_gate_edges2(vals)
}
\arguments{
\item{vals}{A vector of integers or logical values.}
}
\value{
The indexes of ups and downs along \code{vals}. Indexes of discrete
  values are also included.
}
\description{
Results includes discrete values (one-hit-wonders).
}
\examples{
find_gate_edges2(c(1:3, 5:7, 10, 14:16))
find_gate_edges2(c(3, 5)); find_gate_edges2(c(3, 5:6, 8))
find_gate_edges2(2:5)
find_gate_edges2(c(2:3, 5:6, 8:9))
find_gate_edges2(c(3:5, 7, 12:14, 18))
vals <- c(3, 5)
vals <- c(3:5, 7:10, 12:14); vals <- c(vals, 18)
vals <- c(3:5, 7, 9)
vals <- c(3, 5:7, 10:11, 13)
vals <- c(3, 5, 7, 8, 10, 12:13)

}
