% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_gatepos}
\alias{find_gatepos}
\title{Finds the positions of logical gates.}
\usage{
find_gatepos(vec)
}
\arguments{
\item{vec}{A logical vector.}
}
\description{
Finds the positions of logical gates.
}
\examples{
library(mzion)

# starts at low and ends at low
vec <- c(rep(0L, 2), rep(1L, 4L), rep(0L, 3L), rep(1L, 3L), rep(0L, 1L))
# find_gatepos(vec)

# low-high
vec <- c(rep(0L, 2), rep(1L, 4L), rep(0L, 3L), rep(1L, 3L), rep(0L, 1L), rep(1L, 2L))

# high-low
vec <- c(rep(1L, 4L), rep(0L, 3L), rep(1L, 3L), rep(0L, 1L), rep(1L, 2L), rep(0L, 2))

# high-high
vec <- c(rep(1L, 4L), rep(0L, 3L), rep(1L, 3L), rep(0L, 1L), rep(1L, 3L))

# all-zeros
# find_gatepos(diff(c(6L, 12L, 18L, 164L), 1L) == 1L)

# all-ones
# find_gatepos(rep(1L, 5L))

# up-width == 1L
# find_gatepos(diff(c(143L, 159L, 310L, 311L, 316L), 1L) == 1L)

# single TRUE
# find_gatepos(diff(c(310L, 311L), 1L) == 1L)
}
