% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_gates}
\alias{find_gates}
\title{Finds the positions of logical gates.}
\usage{
find_gates(vals)
}
\arguments{
\item{vals}{A logical vector.}
}
\description{
Finds the positions of logical gates.
}
\examples{
library(mzion)

# starts at low and ends at low
# find_gates(c(100, 102, 104:108, 110, 112, 114:117, 119))

# low-high
# find_gates(c(100, 102, 104:108, 110, 112, 114:117))

# high-low
# find_gates(c(95:100, 102, 104:108, 110:111, 114:117, 119))

# high-high
# find_gates(c(95:100, 102, 104:108, 110, 112:113, 116:119))

# all-zeros (discrete)
# find_gates(c(6L, 12L, 18L, 164L))

# all-ones
# find_gates(1:5)

# up-width == 1L
# find_gates(c(143L, 159L, 310L, 311L, 316L))

# single TRUE
# find_gates(c(310L, 311L))

#' # single value
# find_gates(c(310L))

# find_gates(c(281, 326, 335, 336, 337, 437, 447, 448, 449, 450, 553, 554, 557))
}
