% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{find_group_breaks}
\alias{find_group_breaks}
\title{Chunksplits by groups.}
\usage{
find_group_breaks(vec, fold = 5L, by_rngs = TRUE)
}
\arguments{
\item{vec}{A sorted vector.}

\item{fold}{The number of folds.}

\item{by_rngs}{Logical; if TRUE return ranges of indexes or else return the
indexes of breaks.}
}
\value{
Ranges or indexes of breaks
}
\description{
Chunksplits by groups.
}
\examples{
\donttest{
library(mzion)

vec <- sort(rep(LETTERS[1:5], 1:5))
mzion:::find_group_breaks(vec, 3)
}
}
