% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_lc_edges_bp}
\alias{find_lc_edges_bp}
\title{Find peak edges by intensity threshold relative to the base peak}
\usage{
find_lc_edges_bp(ys, ps, range, n_dia_scans = 6L, y_rpl = 2)
}
\arguments{
\item{ys}{A vector of intensity values.}

\item{ps}{A vector of Y positions.}

\item{range}{A vector Y ranges.}

\item{n_dia_scans}{The number of adjacent MS scans for constructing a peak
profile and thus for determining the apex scan number of an moverz value
along LC.}

\item{y_rpl}{A replacement value of intensities.}
}
\description{
Not yet used.
}
