% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{find_lcpeaks}
\alias{find_lcpeaks}
\title{Detects chromatographic peaks.}
\usage{
find_lcpeaks(
  y,
  lag = 5L,
  min_lag = 3L,
  max_lag = 200L,
  min_len = 10L,
  min_var = 1e-05,
  static = TRUE,
  threshold = 3,
  influence = 0
)
}
\arguments{
\item{y}{Values of y.}

\item{lag}{The lag length.}

\item{min_lag}{The minimum length of the lag.}

\item{max_lag}{The maximum length of the lag.}

\item{min_len}{The minimum length of y for peak-picking algorithm.}

\item{min_var}{The minimum variance.}

\item{static}{Is the time series static (or adaptive).}

\item{threshold}{The threshold.}

\item{influence}{The influence.}
}
\description{
May be over-killing.
}
\details{
https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data
https://stackoverflow.com/questions/58942623/isolating-peaks-in-from-time-series-data-in-r
}
