% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_mdda_mms1s}
\alias{find_mdda_mms1s}
\title{Finds mDDA precursors}
\usage{
find_mdda_mms1s(
  msx_moverzs = NULL,
  msx_ints = NULL,
  iso_ctr = NULL,
  iso_lwr = NULL,
  iso_upr = NULL,
  ppm = 10L,
  maxn_precurs = 5L,
  max_ms1_charge = 4L,
  n_fwd = 15L,
  n_bwd = 20L,
  offset_upr = 20L,
  offset_lwr = 30L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  use_defpeaks = FALSE,
  width_left = 2.01,
  width_right = 1.25,
  margin = 0.5,
  min_mass = 200L,
  step = ppm/1e+06,
  is_pasef = FALSE
)
}
\arguments{
\item{msx_moverzs}{Vectors of bracketing (e.g. +/-6 scans) full-spectrum MS1
m-over-z values. Each \code{msx_moverzs} needs to be in an ascending order.}

\item{msx_ints}{Vectors of bracketing full-spectrum MS1 intensity values
corresponding to \code{msx_moverzs}.}

\item{iso_ctr}{Vectors of isolation centers (e.g. top-12 scans) of MS2.}

\item{iso_lwr}{Vectors of isolation lowers of MS2.}

\item{iso_upr}{Vectors of isolation uppers of MS2. Both \code{iso_lwr} and
\code{iso_upr} are used, in case of asymmetrical isolation to
\code{iso_ctr}.}

\item{ppm}{Mass error tolerance.}

\item{maxn_precurs}{Maximum number of precursors for consideration.}

\item{max_ms1_charge}{Maximum charge state of precursors for consideration.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries when determining
the charge state.}

\item{n_bwd}{Backward looking up to \code{n_bwd} mass entries when
determining the charge state.}

\item{offset_upr}{A cardinal number of upper mass off-sets in relative to a
mass center when finding mono-isotopic species.}

\item{offset_lwr}{A cardinal number of lower mass off-sets in relative to a
mass center when finding mono-isotopic species.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{A multiplication factor for the fuzzy discrimination of a
secondary precursor in parallel to a primary precursor in an isotopic
envelop. The smaller the value, the more stringent it is in calling an
adjacent peak being a mono-isotopic precursor}

\item{use_defpeaks}{Use default peak info or not.}

\item{width_left}{The left-width of an extended MS1 isolation window to
contain a full isotope envelops.}

\item{width_right}{The right-width of an extended MS1 isolation window to
contain a full isotope envelops.}

\item{margin}{The margin of an m-over-z extended to an isolation window (to
account for the imperfection of the voltage gates of isolation).}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{step}{A step size.}

\item{is_pasef}{Logical; is TIMS TOF data or not.}
}
\value{
A list. x: monoisotopic moverzs (weighted mean statistics); y:
  intensities (mean); z: charge states.
}
\description{
Averages of multiple MS1 scans.
}
