% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_mdda_ms1s}
\alias{find_mdda_ms1s}
\title{Finds chimeric precursors.}
\usage{
find_mdda_ms1s(
  df1,
  df2,
  ppm = 8L,
  maxn_precurs = 5L,
  max_ms1_charge = 4L,
  n_fwd = 20L,
  grad_isotope = 2.5,
  fct_iso2 = 3,
  width = 2.01,
  step = ppm/1e+06
)
}
\arguments{
\item{df1}{A one-row data frame of precursor data.}

\item{df2}{The data frame of MS2 corresponding to df1.}

\item{ppm}{Mass error tolerance.}

\item{maxn_precurs}{The maximum number of chimeric precursors.}

\item{max_ms1_charge}{The maximum charge state of precursors.}
}
\description{
Not used; by single MS1 scan.
}
