% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{find_mgf_query}
\alias{find_mgf_query}
\title{Finds matched MGF query.}
\usage{
find_mgf_query(mgf_path, raw_id, scan)
}
\arguments{
\item{mgf_path}{An MGF path.}

\item{raw_id}{The index of raw_file.}

\item{scan}{A scan number or identifier.}
}
\description{
Finds matched MGF query.
}
