% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_min_ncores}
\alias{find_min_ncores}
\title{Find the minimal number of cores.}
\usage{
find_min_ncores(x = 25, n = 6)
}
\arguments{
\item{x}{The number of files.}

\item{n}{The number of cores where \code{x > n > 1}.}
}
\description{
Find the minimal number of cores.
}
