% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_mod_indexes}
\alias{find_mod_indexes}
\title{Find the indexes of modifications.}
\usage{
find_mod_indexes(file)
}
\arguments{
\item{file}{A full-path name of file where modifications are recorded.}
}
\description{
Find the indexes of modifications.
}
