% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_modps}
\alias{find_modps}
\title{Finds the positions and sites of modifications.}
\usage{
find_modps(mods)
}
\arguments{
\item{mods}{A vector of modifications, e.g., \code{c("TMT6plex (K)", "Acetyl
(K)")}}
}
\description{
Finds the positions and sites of modifications.
}
