% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_motif_pat}
\alias{find_motif_pat}
\title{Finds the motifs of modifications.}
\usage{
find_motif_pat(aa_masses)
}
\arguments{
\item{aa_masses}{A list of amino acid lookups.}
}
\description{
Finds the motifs of modifications.
}
