% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{find_ms1_times}
\alias{find_ms1_times}
\title{Finds the time stamp(s) of MS1 precursors.}
\usage{
find_ms1_times(out_path)
}
\arguments{
\item{out_path}{An output path.}
}
\description{
Finds the time stamp(s) of MS1 precursors.
}
