% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_ms1byms2}
\alias{find_ms1byms2}
\title{Finds precursors using MS2 data.}
\usage{
find_ms1byms2(moverzs, msxints, charges, center, iso_lwr, iso_upr)
}
\arguments{
\item{moverzs}{Mass-to-charge ratios.}

\item{msxints}{MS1 or MS2 peak intensities.}

\item{charges}{MS2 charges.}

\item{center}{The mass center of an isolation window (only for MS1).}

\item{iso_lwr}{The lower width of an isolation window.}

\item{iso_upr}{The upper width of an isolation window.}
}
\description{
For absolutely no precursor identification based on MS1 data.
}
