% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{find_ms1stat}
\alias{find_ms1stat}
\title{De-isotopes precursor masses.}
\usage{
find_ms1stat(
  moverzs,
  msxints,
  n_ms1s = 1L,
  center = 0,
  exclude_reporter_region = FALSE,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  is_dda = TRUE,
  ppm = 8L,
  ms_lev = 1L,
  maxn_feats = 300L,
  max_charge = 4L,
  n_fwd = 20L,
  offset_upr = 30L,
  offset_lwr = 30L,
  step = ppm/1e+06,
  backward_mass_co = 800/ms_lev,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  use_defpeaks = FALSE,
  is_pasef = FALSE
)
}
\arguments{
\item{moverzs}{MS1 or MS2 Mass-to-charge ratios. The inputs are typically at
weighted-mean statistics.}

\item{msxints}{MS1 or MS2 peak intensities. The inputs are typically at mean
statistics.}

\item{n_ms1s}{The underlying counts that have contributed to \code{moverzs}
and \code{maxints}.}

\item{center}{The mass center of an isolation window (only for MS1).}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{is_dda}{Logical; is DDA or not.}

\item{ppm}{Allowance in mass error when deisotoping.}

\item{ms_lev}{MS level.}

\item{maxn_feats}{The maximum number of MS features.}

\item{max_charge}{The maximum charge state.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries. The default is
20 for MS1 and 10 for MS2.}

\item{offset_upr}{A cardinal number of upper mass off-sets.}

\item{offset_lwr}{A cardinal number of lower mass off-sets.}

\item{step}{Step size for mass binning.}

\item{backward_mass_co}{A mass cut-off to initiate backward looking of an
isotope envelop.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{The multiplication factor for the second isotopic peak.}

\item{use_defpeaks}{Depreciated. Logical; if TRUE, uses MS1 m-over-z's,
intensities and charge states pre-calculated by other peak-picking
algorithms.}

\item{is_pasef}{Logical; is TIMS TOF data or not.}
}
\description{
De-isotopes precursor masses.
}
\examples{
\donttest{
moverzs <- c(881 + 1:10*.1, 882.0674, 882.0981, 882.4034, 882.60, 882.7372)
msxints <- c(1000 * 1:10, 1652869, 788368, 2043015, 2314111, 4314111)
n_ms1s <- rep_len(1L, length(msxints))

if (FALSE) {
  out <- mzion:::find_ms1stat(moverzs, msxints, n_ms1s, ppm = 5L, ms_lev = 1L,
                              maxn_feats = 5L, max_charge = 4L, offset_upr = 8L,
                              offset_lwr = 8L)
}

moverzs <- c(907.968018,908.136475,908.872711,909.073690,909.121,
             909.273670,909.280212,909.45,909.454981,909.459717,
             909.473072,909.675781,909.721191,909.78,909.79,
             909.791633,909.840088,909.969744,910.077278,910.091888,
             910.124076,910.130798,910.180115,910.218689,910.341797,
             910.456115,910.463650,910.474152,910.791669,911.126692,
             911.133219,911.471605)
msxints <- c(2971101,3013171,5774301,24508055,40267873,
             1E6,7340374,1E6,154459778,8312258,
             18375220,22909676,2863776,1.1E6,1.2E6,
             191655971,7273611,32089542,9125108,5008687,
             167021373,12912115,3242378,16278636,4731828,
             98601412,22828501,10539991,26255569,8499971,
             8305187,37899662)
n_ms1s <- rep_len(1L, length(msxints))

if (FALSE) {
  out <- mzion:::find_ms1stat(moverzs, msxints, n_ms1s, center = 909.788696,
                      exclude_reporter_region = FALSE,
                      ppm = 5L, ms_lev = 1L, maxn_feats = 1L, max_charge = 4L,
                      step = 5/1e6)
  out <- mzion:::find_ms1stat(moverzs, msxints, n_ms1s, # center = 909.788696,
                      exclude_reporter_region = FALSE,
                      ppm = 5L, ms_lev = 1L, maxn_feats = 1L, max_charge = 4L,
                      step = 5/1e6)
}
}
}
