% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{find_ms1stat_rf}
\alias{find_ms1stat_rf}
\title{De-isotopes precursor masses.}
\usage{
find_ms1stat_rf(
  moverzs,
  msxints,
  n_ms1s = 1L,
  center = 0,
  exclude_reporter_region = FALSE,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  ppm = 5L,
  ms_lev = 1L,
  maxn_feats = 300L,
  max_charge = 4L,
  n_fwd = 20L,
  offset_upr = 30L,
  offset_lwr = 30L,
  order_mz = TRUE,
  step = ppm/1e+06,
  backward_mass_co = 800/ms_lev,
  grad_isotope = 2.5,
  fct_iso2 = 3,
  fct_bg = 15,
  fct_bg2 = 5,
  bound = FALSE
)
}
\arguments{
\item{moverzs}{Mass-to-charge ratios.}

\item{msxints}{MS1 or MS2 peak intensities.}

\item{n_ms1s}{The underlying counts that have contributed to \code{moverzs}
and \code{maxints}.}

\item{center}{The mass center of an isolation window (only for MS1).}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{ppm}{Allowance in mass error when deisotoping.}

\item{ms_lev}{MS level.}

\item{maxn_feats}{The maximum number of MS features.}

\item{max_charge}{The maximum charge state.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries. The default is
20 for MS1 and 10 for MS2.}

\item{offset_upr}{A cardinal number of upper mass off-sets.}

\item{offset_lwr}{A cardinal number of lower mass off-sets.}

\item{order_mz}{Logical; if TRUE, orders peaks from low to high m-over-z's.}

\item{step}{Step size for mass binning.}

\item{backward_mass_co}{A mass cut-off to initiate backward looking of an
isotope envelop.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{The multiplication factor for the second isotopic peak.}

\item{fct_bg}{A ratio factor to discriminate background noises from isotopic
features.}

\item{fct_bg2}{A second ratio factor to discriminate background noises from
isotopic features.}

\item{bound}{Not yet used. Logical; if TRUE, removes precursors outside of
the boundary of isolation window.}
}
\description{
The right half first.
}
