% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_ms2ends}
\alias{find_ms2ends}
\title{Finds group indexes with end points at \code{vals = 2L}.}
\usage{
find_ms2ends(vals, n_chunks = 3L)
}
\arguments{
\item{vals}{An integer vector of (MS levels of) 1 or 2.}

\item{n_chunks}{The number of chunks.}
}
\description{
For chunk-splitting with each chuck ending at value 2.
}
\examples{
library(mzion)

vals1 <- c(rep(1, 2), rep(2, 5), 1, rep(2, 4), rep(1, 3), rep(2, 4), 1)
grps1 <- find_ms2ends(vals1, 3)

vals2 <- c(rep(1, 2), rep(2, 5), 1, rep(2, 4), rep(1, 3), rep(2, 4))
grps2 <- find_ms2ends(vals2, 3)
}
