% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{find_ms2resids}
\alias{find_ms2resids}
\title{Helper of \link{make_ms2vmods}.}
\usage{
find_ms2resids(M, vec)
}
\arguments{
\item{M}{A modification-name matrix from \link{make_ms2vmods}.}

\item{vec}{A vector of labels.}
}
\description{
Makes the corresponding residue matrix from \link{make_ms2vmods}.
}
