% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_mslev_brackets}
\alias{find_mslev_brackets}
\title{Finds the bracketing positions of MS1 and MS2 levels along the scan indexes.}
\usage{
find_mslev_brackets(vals, out_type = c("sim_vec", "sim_list", "full_list"))
}
\arguments{
\item{vals}{A vector of MS levels. Assume that \code{vals} contains only
levels 1L or 2L and always start with 1L.}

\item{out_type}{The type of outputs.}
}
\description{
Finds the bracketing positions of MS1 and MS2 levels along the scan indexes.
}
\examples{
# trailing MS1s
vals <- c(rep(1L, 1), rep(2L, 4), rep(1L, 2), rep(2L, 3), rep(1L, 3))
# find_mslev_brackets(vals)

# multiple MS1s at the beginning
vals <- c(rep(1L, 2), rep(2L, 4), rep(1L, 2), rep(2L, 3), rep(1L, 3), rep(2L, 4))
# find_mslev_brackets(vals)

# One MS1 at the beginning
vals <- c(rep(1L, 1), rep(2L, 4), rep(1L, 2), rep(2L, 3), rep(1L, 3), rep(2L, 4))
}
