% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{find_optlens}
\alias{find_optlens}
\title{Find the suitable pep_len values for the fitting of probability cut-offs.}
\usage{
find_optlens(all_lens, counts, min_count = 128L)
}
\arguments{
\item{all_lens}{A vector of all possible pep_len values found from data.}

\item{counts}{A vector of the counts of sequences at given pep_len values.}

\item{min_count}{The minimum counts for consideration in fitting.}
}
\description{
Recursively decrease the value of \code{min_count} by half until some indexes
are found.
}
