% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{find_optns}
\alias{find_optns}
\title{Finds the optimal natural spline.}
\usage{
find_optns(diff_ms1, ret_time, knots = 50L)
}
\arguments{
\item{diff_ms1}{The differences between experimental and theoretical
precursor masses.}

\item{ret_time}{Retention timm.}

\item{knots}{The number of knots.}
}
\description{
Finds the optimal natural spline.
}
