% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{find_perm_sets}
\alias{find_perm_sets}
\title{A (faster alternative) to \link[gtools]{permutations} with duplicated labels.}
\usage{
find_perm_sets(labs = c("A", "A", "A", "B", "B", "C"))
}
\arguments{
\item{labs}{A vector of labels.}
}
\description{
A (faster alternative) to \link[gtools]{permutations} with duplicated labels.
}
\examples{
\donttest{
labs <- c("A", "A", "A", "B", "B", "C")
x <- find_perm_sets(labs)

# Compares to gtools::permutations
len <- length(labs)
y <- gtools::permutations(len, len, labs, set = FALSE)
y <- unique(y)

stopifnot(dim(x) == dim(y))

# Compares to the theoretical number of entries
len_labs <- length(labs)
ns <- count_elements(labs)
theo_cts <- factorial(len_labs)/prod(factorial(ns))
stopifnot(nrow(x) == theo_cts)

## Others
x <- find_perm_sets(rep("A", 3))
}
}
