% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{find_pos_site}
\alias{find_pos_site}
\title{Function factories for finding the amino-acid residue at a position.}
\usage{
find_pos_site(pos)
}
\arguments{
\item{pos}{The UniMod position of a variable modification, which must be one
of \code{c("Protein N-term", "Protein C-term", "Any N-term", "Any C-term",
"Anywhere")}. No need to call stopifnot for the way the function is called
(by the developer).}
}
\value{
A function for finding the residue at the position specified by the
  argument \code{pos}. For each function, it takes a list of variable
  modifications specified by argument \code{vmods} and the corresponding
  "positions" (for vectorization) as inputs.
}
\description{
If the the value in a \code{vmods} is one of the LETTERS -> a site is
specified in the corresponding variable modification. Otherwise, a site is in
one of \code{c("C-term", "N-term")}. For example, \code{`Oxidation (M)` =
"M"} for the former and \code{`Protein N-term` = "N-term"} for the later. In
other words, \code{pos} can be a simple upper case letter or a string of
"\code{[NC]-term}".
}
