% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{find_ppm_outer_bycombi}
\alias{find_ppm_outer_bycombi}
\title{Finds the the outer products for a vector of MS2 ions at a given ion series.}
\usage{
find_ppm_outer_bycombi(theos, expts, ppm_ms2 = 25L)
}
\arguments{
\item{theos}{Numeric vector; one series of theoretical MS2s.}

\item{expts}{Numeric vector; one series experimental MS2s.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species.}
}
\description{
A theoretical peptide at a given position permutation.
}
\examples{
\donttest{
expts <- c(101.0714, 102.0554, 110.0717, 115.0505, 126.1279, 127.0504,
           127.1249, 127.1312, 128.1283, 128.1346, 129.0660, 129.1316,
           129.1379, 130.1350, 130.1412, 131.1383, 133.0431, 133.0609,
           134.0448, 136.0757, 145.0608, 158.0924, 173.1496, 175.1190,
           176.1594, 191.0663, 193.0971, 198.0873, 201.0869, 201.1233,
           202.0821, 230.1702, 248.0876, 248.1806, 257.1527, 298.6711,
           312.6687, 335.1190, 361.1768, 361.6685, 367.2292, 369.6903,
           370.1821, 370.6830, 376.2756, 377.2790, 384.6952, 412.7264,
           420.2143, 423.2597, 475.2341, 475.3442, 476.3469, 484.2382,
           484.7389, 487.7817, 488.2842, 496.2717, 523.3182, 537.3159,
           537.8178, 572.8343, 573.3373, 623.3586, 624.3298, 636.8528,
           637.3558, 637.8575, 638.4089, 645.3453, 645.8500, 646.3501,
           687.8795, 688.3799, 692.8696, 695.4283, 701.3759, 701.8768,
           702.3782, 702.8797, 737.3956, 737.8962, 738.3671, 739.3572,
           740.3575, 745.9085, 746.4121, 752.4502, 801.4272, 809.4722,
           810.4739, 852.4418, 967.4690, 968.4714, 969.5046, 1084.5293,
           1132.5676, 1133.5686, 1134.5735, 1135.5786)

theos <- c(116.0342, 203.0662, 290.0983, 418.1569, 475.1783, 572.2311,
           732.2617, 861.3043, 958.3571, 1071.4412, 1168.4939, 1225.5154,
           1322.5681, 1435.6522, 1536.6999, 1664.7585, 1761.8112, 1917.9123,
           175.1190, 272.1717, 400.2303, 501.2780, 614.3620, 711.4148,
           768.4363, 865.4890, 978.5731, 1075.6259, 1204.6684, 1364.6991,
           1461.7519, 1518.7733, 1646.8319, 1733.8639, 1820.8960, 1935.9229)

names(theos) <- c("D", "S", "S", "Q", "G", "P", "C", "E", "P",
                  "L", "P", "G", "P", "L", "T", "Q", "P", "R",
                  "R", "P", "Q", "T", "L", "P", "G", "P", "L",
                  "P", "E", "C", "P", "G", "Q", "S", "S", "D")

find_ppm_outer_bycombi(theos, expts)

# No secondary matches
theos <- c(56.5233, 235.1522, 284.6864, 326.2050, 361.7235, 
           418.2656, 474.8076, 653.4366, 95.0128, 452.2707, 
           551.3391, 634.3762, 705.4133, 818.4974, 931.5815, 
           1288.8394, 48.0100, 226.6390, 276.1732, 317.6917, 
           353.2103, 409.7523, 466.2944, 644.9233, 94.0287, 
           451.2866, 550.3550, 633.3921, 704.4292, 817.5133, 
           930.5974, 1287.8553, 47.5180, 226.1470, 275.6812, 
           317.1997, 352.7183, 409.2603, 465.8024, 644.4313,
           188.6415, 245.1835, 301.7256, 337.2441, 378.7627, 
           428.2969, 606.9258, 670.9551, 359.2492, 472.3333, 
           585.4174, 656.4545, 739.4916, 838.5600, 1195.8179, 
           1323.8765, 180.1282, 236.6703, 293.2123, 328.7309, 
           370.2494, 419.7836, 598.4126, 662.4419, 358.2651, 
           471.3492, 584.4333, 655.4704, 738.5075, 837.5759, 
           1194.8338, 1322.8924, 179.6362, 236.1783, 292.7203, 
           328.2389, 369.7574, 419.2916, 597.9206, 661.9499)

a <- c("Q", "K", "V", "M", "A", "I", "I", "K")
b <- rev(a)
names(theos) <- c(rep(a, 5), rep(b, 5))

expts <- c(101.01134,110.07162,112.03947,116.01675,118.83968,
           120.08091,126.12779,127.12479,127.13109,128.12816,
           128.13440,129.13150,129.13777,130.13483,130.14053,
           131.13824,150.26431,155.08147,156.07658,159.07658,
           173.14934,175.05345,175.15630,176.13777,176.15942,
           176.72833,187.07147,188.15984,203.04829,203.35188,
           204.05164,219.14928,227.06596,229.16634,230.16991,
           231.17354,232.95274,248.17976,268.16522,315.25906,
           318.07504,319.07828,333.21524,361.11679,361.20996,
           372.28027,376.27521,377.27872,389.11212,390.11490,
           404.25201,432.24750,489.26880,489.35962,504.09521,
           510.33450,515.28448,527.35297,533.29462,542.31720,
           560.34607,577.31445,588.33716,594.34283,599.39276,
           602.44226,604.36749,614.35211,632.36328,659.46362,
           694.44006,694.93970,695.39063,707.20782,734.39813,
           752.41211,753.41321,756.51660,757.52002,830.91998,
           851.48267,852.48303,855.58301,899.52069,956.62952,
           1027.66833,1230.73962,1232.75513,1233.75024)

find_ppm_outer_bycombi(theos, expts)
}
}
