% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{find_psm_rows2}
\alias{find_psm_rows2}
\title{Extracts the first row of matched PSMs from psmC.}
\usage{
find_psm_rows2(file_t0, scan, raw_file, rank = 1L, is_decoy = FALSE)
}
\arguments{
\item{file_t0}{The filename of psmC results.}

\item{scan}{A scan number or identifier.}

\item{raw_file}{Character string; a RAW file name.}

\item{rank}{Positive integer; the rank of a match. The default is one for the
best match.}

\item{is_decoy}{Logical; is the match from decoy. The default is FALSE.}
}
\description{
Extracts the first row of matched PSMs from psmC.
}
