% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{find_psm_rowsQ}
\alias{find_psm_rowsQ}
\title{Extracts the first row of matched PSMs from tiers 1-3.}
\usage{
find_psm_rowsQ(
  file_t1,
  file_t2,
  file_t3,
  scan,
  raw_file,
  rank = 1L,
  is_decoy = FALSE
)
}
\arguments{
\item{file_t1}{The filename of tier-1 PSMs.}

\item{file_t2}{The filename of tier-1 PSMs.}

\item{file_t3}{The filename of tier-1 PSMs.}

\item{scan}{A scan number or identifier.}

\item{raw_file}{Character string; a RAW file name.}

\item{rank}{Positive integer; the rank of a match. The default is one for the
best match.}

\item{is_decoy}{Logical; is the match from decoy. The default is FALSE.}
}
\description{
Extracts the first row of matched PSMs from tiers 1-3.
}
