% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{find_secion_types}
\alias{find_secion_types}
\title{Finds the types of secondary ions.}
\usage{
find_secion_types(type_ms2ions = "by")
}
\arguments{
\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}
}
\description{
The order of secondary ions was defined in \link{add_seions}.
}
