% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{find_targets}
\alias{find_targets}
\title{Find the index or name of target results.}
\usage{
find_targets(out_path, pattern = "^ion_matches_")
}
\arguments{
\item{out_path}{An output path.}

\item{pattern}{The pattern of files.}
}
\description{
Find the index or name of target results.
}
