% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{find_theoexpt_pair}
\alias{find_theoexpt_pair}
\title{Finds the pairs of theoretical and experimental values.}
\usage{
find_theoexpt_pair(psm, out_path, scan, raw_id, is_decoy = FALSE)
}
\arguments{
\item{psm}{Matched PSM.}

\item{out_path}{An output path.}

\item{scan}{A scan number or identifier.}

\item{raw_id}{The index of raw_file.}

\item{is_decoy}{Logical; is the match from decoy. The default is FALSE.}
}
\description{
Finds the pairs of theoretical and experimental values.
}
