% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms1_labels.R
\name{find_unique_sets}
\alias{find_unique_sets}
\title{Finds the sets of unique labels for \code{n} balls in \code{p-positions}.}
\usage{
find_unique_sets(p = 5L, labs = c("A", "B", "C"))
}
\arguments{
\item{p}{The number of positions.}

\item{labs}{The names to be filled into the \code{p}-number of positions.}
}
\description{
At least one occupancy for each balls in \code{ns}.
}
\examples{
\donttest{
library(mzion)
mzion:::find_unique_sets(5, c("Oxidation (M)", 
                           "Carbamidomethyl (M)", 
                           "Carbamyl (M)"))
}
}
