% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{find_vmodposU}
\alias{find_vmodposU}
\title{Helper of \link{combi_namesiteU}.}
\usage{
find_vmodposU(vec, ps, aas)
}
\arguments{
\item{vec}{A vector of labels.}

\item{ps}{Named vector; counts for each site. Sites in names and counts in
values.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}
}
\description{
One-to-one correspondence between Names and Sites.
}
\details{
Custom functions: vec_to_list.
}
