% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{find_vmodscombi}
\alias{find_vmodscombi}
\title{Finds the combinatorial MS2 variable modifications.}
\usage{
find_vmodscombi(
  aas = NULL,
  ms2vmods = NULL,
  maxn_vmods_sitescombi_per_pep = 64L
)
}
\arguments{
\item{aas}{\code{aa_seq} split into a sequence of LETTERS.}

\item{ms2vmods}{The i-th result from: 
lapply(ms1vmods_all, function (x) lapply(x, make_ms2vmods)).}

\item{maxn_vmods_sitescombi_per_pep}{A non-negative integer; the maximum
number of combinatorial variable modifications per peptide sequence. The
default is 64.}
}
\description{
Finds the combinatorial MS2 variable modifications.
}
\examples{
\donttest{
## One-to-one correspondence between Names and Sites
#  (no need to permute MS1 labels)

fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)",
               "Carbamidomethyl (C)")

varmods   <- c("Acetyl (Protein N-term)", "Oxidation (M)",
               "Deamidated (N)",
               "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods,
                               varmods = varmods,
                               out_path = NULL)

maxn_vmods_per_pep  <- 5L
maxn_sites_per_vmod <- 3L

ms1vmods_all <- lapply(aa_masses_all, make_ms1vmod_i,
                       maxn_vmods_per_pep = maxn_vmods_per_pep,
                       maxn_sites_per_vmod = maxn_sites_per_vmod)

ms2vmods_all <- lapply(ms1vmods_all, function (x) lapply(x, make_ms2vmods))

i <- 11L
aa_masses <- aa_masses_all[[i]]
amods <- attr(aa_masses, "amods")

ms1vmods <- ms1vmods_all[[i]]
ms2vmods <- ms2vmods_all[[i]]

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

# Subset from ms1vmods by aas
oks <- match_mvmods(aas = aas, ms1vmods = ms1vmods, amods = amods)$inds
ms2vmods <- ms2vmods[oks]

vmods_combi <- find_vmodscombi(aas, ms2vmods)


## 'Carbamidomethyl (M)',  'Carbamyl (M)' and N
#  (need permutation of MS1 labels)

fixedmods <- c("TMT6plex (K)", "dHex (S)")
varmods   <- c("Carbamidomethyl (M)", "Carbamyl (M)",
               "Deamidated (N)", "Acetyl (Protein N-term)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

maxn_vmods_per_pep  <- 5L
maxn_sites_per_vmod <- 3L

ms1vmods_all <- lapply(aa_masses_all, make_ms1vmod_i,
                       maxn_vmods_per_pep = maxn_vmods_per_pep,
                       maxn_sites_per_vmod = maxn_sites_per_vmod)

ms2vmods_all <- lapply(ms1vmods_all, function (x) lapply(x, make_ms2vmods))

i <- 16L
aa_masses <- aa_masses_all[[i]]
amods <- attr(aa_masses, "amods")
ntmod <- attr(aa_masses, "ntmod")
ctmod <- attr(aa_masses, "ctmod")

ms1vmods <- ms1vmods_all[[i]]
ms2vmods <- ms2vmods_all[[i]]

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))


# Subset from ms1vmods by aas
oks <- match_mvmods(aas = aas, ms1vmods = ms1vmods, amods = amods)$inds
ms2vmods <- ms2vmods[oks]

vmods_combi <- find_vmodscombi(aas, ms2vmods)

n_pos <- lapply(vmods_combi, function (x) names(x[x == "Deamidated (N)"]))
stopifnot(all(sapply(n_pos, function (x) all(x \%in\% c("6", "14")))))
}
}
