% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{findLocFracsDF}
\alias{findLocFracsDF}
\title{Finds the localization fractions of STY.}
\usage{
findLocFracsDF(df, phosmods = NULL)
}
\arguments{
\item{df}{A data frame containing columns \code{pep_ivmod2} and
\code{pep_ms2_ideltas.}.}

\item{phosmods}{A vector to the modification indexes of phosphorylations.}
}
\description{
Counting statistics.
}
