% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{flat_pepseqs}
\alias{flat_pepseqs}
\title{Flattens pep_seqs with the removals of prot_accs.}
\usage{
flat_pepseqs(x)
}
\arguments{
\item{x}{Lists of pep_seqs by prot_accs.}
}
\description{
Flattens pep_seqs with the removals of prot_accs.
}
