% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{flatten_list}
\alias{flatten_list}
\title{Alternative to \link[purrr]{flatten}}
\usage{
flatten_list(data, use_names = TRUE)
}
\arguments{
\item{data}{A list of data.}

\item{use_names}{Logical; to use names or not.}
}
\description{
Alternative to \link[purrr]{flatten}
}
