% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{flattenMSxyz}
\alias{flattenMSxyz}
\title{Flattens MS data.}
\usage{
flattenMSxyz(matx, maty, matz, join_ms = FALSE)
}
\arguments{
\item{matx}{The data matrix of moverzs.}

\item{maty}{The data matrix of intensities.}

\item{matz}{The data matrix of charge states.}

\item{join_ms}{Logical; if TRUE, combine adjacent entries.}
}
\description{
Joins data across adjacent scans af \code{combine = TRUE}.
}
