% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{fold_vec}
\alias{fold_vec}
\title{Folds a vector evenly}
\usage{
fold_vec(vec, fold = 5L)
}
\arguments{
\item{vec}{A vector}

\item{fold}{The number of folds}
}
\description{
Only for the special case that \code{length(vec)} is exactly n-times of the
\code{fold}.
}
\seealso{
\link{fold_vec2} for a more generalized solution.
}
