% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{fold_vec2}
\alias{fold_vec2}
\title{Folds a vector evenly}
\usage{
fold_vec2(vec, fold = 5L, remove_allna_subvec = TRUE)
}
\arguments{
\item{vec}{A vector}

\item{fold}{The number of folds}

\item{remove_allna_subvec}{Logical; if TRUE, removes all NA sub vectors. This
can occur with the rounding of lengths. E.g., at \code{length(vec) == 15}
and \code{fold = 6}, the 6-th subvec will be all NA.}
}
\description{
A more general form of \link{fold_vec} handling fractional
\code{length(vec)/fold}.
}
\examples{
vec <- sort(rep(LETTERS[1:5], 1:5))
mzion:::fold_vec2(vec, 6L)
}
\seealso{
\link{find_group_breaks}
}
