% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_base.R
\name{frames_adv2}
\alias{frames_adv2}
\title{Frames advancement (with cached \code{ms2masses}.}
\usage{
frames_adv2(
  mgf_frames = NULL,
  theopeps = NULL,
  theopeps2 = NULL,
  aa_masses = NULL,
  ms1vmods = NULL,
  ms2vmods = NULL,
  ntmod = NULL,
  ctmod = NULL,
  ntmass = NULL,
  ctmass = NULL,
  amods = NULL,
  vmods_nl = NULL,
  fmods_nl = NULL,
  mod_indexes = NULL,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  minn_ms2 = 7L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  digits = 4L
)
}
\arguments{
\item{mgf_frames}{A group of mgf frames. Each frame contains one to multiple
MGFs whose MS1 masses are in the same interval.}

\item{theopeps}{Binned theoretical peptides corresponding to an i-th
\code{aa_masses}.}

\item{theopeps2}{Binned theoretical MS2 ions corresponding to an i-th
\code{aa_masses}.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ms1vmods}{The set of all possible MS1 vmod labels at a given aa_masses.}

\item{ms2vmods}{Matrices of labels of variable modifications. Each
permutation in a row for each matrix.}

\item{ntmass}{The mass of N-terminal.}

\item{ctmass}{The mass of C-terminal.}

\item{amods}{The attribute of \code{amods} from an \code{aa_masses}.}

\item{vmods_nl}{The attribute of \code{vmods_nl} from an \code{aa_masses}.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{maxn_vmods_sitescombi_per_pep}{A non-negative integer; the maximum
number of combinatorial variable modifications per peptide sequence. The
default is 64.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\value{
Matches to each MGF as a list elements. The length of the output is
  equal to the number of MGFs in the given frame.
}
\description{
Not yet used. readRDS takes about 30 secs and clusterApply(cl, x, readRDS)
takes about 16 seconds for one module and couple minutes for all twelve
modules. For small data sets, the gain in using pre-calcucated ms2masses did
not outweigh the loss in the loading of the large files. Needs a faster
parallel readRDS.
}
