% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_base.R
\name{fuzzy_match_one2}
\alias{fuzzy_match_one2}
\title{Fuzzy matches with a +/-1 window.}
\usage{
fuzzy_match_one2(x, y)
}
\arguments{
\item{x}{A vector to be matched.}

\item{y}{A vector to be matched against.}
}
\description{
Fuzzy matches with a +/-1 window.
}
\examples{
ans1 <- fuzzy_match_one2(c(74953, 74955), rep(74954, 2))
ans2 <- fuzzy_match_one2(c(74953, 74955), 74954)

stopifnot(identical(ans1, ans2))
stopifnot(ans1 == c(FALSE, TRUE))

ans3 <- fuzzy_match_one2(c(74953, 74955, 80000), c(74955, 80000))
}
