% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{gen_isoenvlope}
\alias{gen_isoenvlope}
\title{Generate an isotope envelop}
\usage{
gen_isoenvlope(mass = 560, charge = 2L)
}
\arguments{
\item{mass}{Precursor mass.}

\item{charge}{Charge state.}
}
\description{
Generate an isotope envelop
}
