% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{get_globalvar}
\alias{get_globalvar}
\title{Finds a global variable.}
\usage{
get_globalvar(val = NULL, cache = NULL)
}
\arguments{
\item{val}{Variable name}

\item{cache}{A cache file name.}
}
\description{
Not yet used. The utility at first looks up the global environment then the
indicated file.
}
\examples{
\dontrun{
library(proteoM)

out_path <- "~/proteoM/example/"
cache <- file.path(out_path, "Calls/.cache_info.rds")
proteoM:::get_globalvar(".time_stamp", cache)
}
}
\seealso{
\link{find_ms1_times}
}
