% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{get_ms1charges}
\alias{get_ms1charges}
\title{Gets the MS1 charges.}
\usage{
get_ms1charges(charges)
}
\arguments{
\item{charges}{A vector of \code{2+, 3+} etc.}
}
\description{
Gets the MS1 charges.
}
