% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{get_ms1xs_space}
\alias{get_ms1xs_space}
\title{Obtain the space of monoisotopic MS1}
\usage{
get_ms1xs_space(df)
}
\arguments{
\item{df}{A data frame of MS1 and MS2.}
}
\description{
Followed by de-isotoping and temporarily puts MS1-XYZ of MS2 scans to the
preceding MS1 scans.
}
\details{
Note that the Y-values were averaged from franking MS1 scans, not the real
precursor intensities for for ascribing X-values.
}
