% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{get_topn_vals}
\alias{get_topn_vals}
\title{Gets top-n values}
\usage{
get_topn_vals(vals, n)
}
\arguments{
\item{vals}{a vector of values}

\item{n}{The top-n values to be retained}
}
\description{
Gets top-n values
}
