% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{getMS1xyz}
\alias{getMS1xyz}
\title{Deisotoping DDA-MS1.}
\usage{
getMS1xyz(
  msx_moverzs = NULL,
  msx_ints = NULL,
  ms_level = NULL,
  iso_ctr = NULL,
  iso_lwr = NULL,
  ms1_moverzs = NULL,
  ms1_charges = NULL,
  ms1_ints = NULL,
  maxn_mdda_precurs = 1L,
  n_mdda_flanks = 6L,
  topn_ms2ions = 150L,
  quant = "none",
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  max_ms1_charge = 4L,
  ppm_ms1_deisotope = 8L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  use_defpeaks = FALSE
)
}
\arguments{
\item{msx_moverzs}{Vectors of bracketing (e.g. +/-6 scans) full-spectrum
moverzs.}

\item{msx_ints}{Vectors of bracketing full-spectrum intensities.}

\item{ms_level}{Vectors of MS levels}

\item{iso_ctr}{A vector of isolation centers.}

\item{iso_lwr}{A vecor of isolation lowers.}

\item{maxn_mdda_precurs}{Maximum number of precursors for consideration in a
multi-precursor DDA scan. Note that at \code{maxn_mdda_precurs = 1}, it is
equivalent to DDA with precursor re-deisotoping. At \code{maxn_mdda_precurs
= 0}, it is equivalent to DDA using the original monoisotopic masses and
charge states provided by a peak-picking algorithm (e.g., MSConvert).}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{max_ms1_charge}{Maximum charge state of precursors for consideration.}

\item{ppm_ms1_deisotope}{Mass error tolerance in MS1 deisotoping.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{A multiplication factor for the fuzzy discrimination of a
secondary precursor in parallel to a primary precursor in an isotopic
envelop. The smaller the value, the more stringent it is in calling an
adjacent peak being a mono-isotopic precursor}

\item{use_defpeaks}{Use default peak info or not.}
}
\description{
Deisotoping DDA-MS1.
}
