% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{groupProts}
\alias{groupProts}
\title{Groups proteins by shared peptides.}
\usage{
groupProts(df, out_path = NULL, fct = 4L, out_name = "prot_pep_setcover.rds")
}
\arguments{
\item{df}{A two-column data frame contains \code{prot_acc} and
\code{pep_seq}.}

\item{out_path}{The output path.}

\item{fct}{A factor for data splitting into chunks.}

\item{out_name}{The output filename.}
}
\description{
Adds columns \code{prot_hit_num} and \code{prot_family_member} etc. to
\code{psmQ.txt}.
}
\details{
In general, non-significant and decoy peptides should have been removed from
the input \code{df}, as well as decoy proteins.

In addition, there is no duplicated entries.
}
\examples{
\donttest{
library(mzion)

df <- data.frame(prot_acc = character(2000), pep_seq = character(2000))
set.seed(100)
df$prot_acc <- sample(LETTERS[1:20], 2000, replace = TRUE)
df$pep_seq <- sample(letters[1:26], 20, replace = TRUE)
df <- df[!duplicated(df), ]

out <- mzion:::groupProts(df, "~")

# One peptide, multiple proteins
df <- data.frame(prot_acc = LETTERS[1:3], pep_seq = rep("X", 3))
out <- mzion:::groupProts(df, "~")
stopifnot(nrow(out) == 3L)

# One peptide, one proteins
df <- data.frame(prot_acc = "A", pep_seq = "X")
out <- mzion:::groupProts(df, "~")
stopifnot(nrow(out) == 1L)

# One proteins
df <- data.frame(prot_acc = rep("A", 3), pep_seq = LETTERS[24:26])
out <- mzion:::groupProts(df, "~")
stopifnot(nrow(out) == 3L)
}
}
