% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{hadd_neuloss}
\alias{hadd_neuloss}
\title{Helper of \link{add_neuloss}.}
\usage{
hadd_neuloss(
  node = NULL,
  neuloss_mono_mass = "0",
  neuloss_avge_mass = "0",
  neuloss_composition = "0"
)
}
\arguments{
\item{node}{A node of \code{umod:specificity}.}

\item{neuloss_mono_mass}{The mono-isotopic mass delta of neutral loss in a
positive value.}

\item{neuloss_avge_mass}{The average mass delta of neutral loss in a positive
value.}

\item{neuloss_composition}{The chemical composition of a neutral loss.}
}
\description{
Helper of \link{add_neuloss}.
}
