% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{hadd_prot_acc}
\alias{hadd_prot_acc}
\title{Helper of \link{add_prot_acc}.}
\usage{
hadd_prot_acc(df, fwd_prps, rev_prps)
}
\arguments{
\item{df}{A data frame.}

\item{fwd_prps}{The look-ups of forward protein and peptides.}

\item{rev_prps}{The look-ups of reversed protein and peptides.}
}
\description{
Helper of \link{add_prot_acc}.
}
