% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{hadd_protacc}
\alias{hadd_protacc}
\title{Helper of \link{add_protacc}.}
\usage{
hadd_protacc(df, prps_fwd, prps_rev)
}
\arguments{
\item{df}{A data frame.}

\item{prps_fwd}{The look-ups of forward protein and peptides.}

\item{prps_rev}{The look-ups of reversed protein and peptides.}
}
\description{
Helper of \link{add_protacc}.
}
