% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{hash_frame_nums}
\alias{hash_frame_nums}
\title{Hash table of binned MS1.}
\usage{
hash_frame_nums(data, r = 1.5)
}
\arguments{
\item{data}{Input data.}

\item{r}{Hash ratio.}
}
\description{
Hash table of binned MS1.
}
