% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_ms2ions.R
\name{hbatch_ms2ions}
\alias{hbatch_ms2ions}
\title{Helper of \link{batch_ms2ions} by time directory.}
\usage{
hbatch_ms2ions(
  time = NULL,
  types = NULL,
  mod_indexes = NULL,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 32L,
  .path_fasta = NULL,
  digits = 5L
)
}
\arguments{
\item{time}{A cached time (directory).}

\item{types}{The types of modification, e.g., "amods- tmod+ vnl- fnl-" etc.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence.}

\item{.path_fasta}{The parent file path to the theoretical masses of MS1
precursors. At the NULL default, the path is \code{gsub("(.*)\\.[^\\.]*$",
"\\1", get("fasta", envir = environment())[1])}. The parameter is for
users' awareness of the structure of file folders and the default is
suggested.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper of \link{batch_ms2ions} by time directory.
}
