% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{hextract_pasef}
\alias{hextract_pasef}
\title{Extracts data of PASEF frames.}
\usage{
hextract_pasef(
  file = "spectra_1.txt",
  path = NULL,
  keys = c("msx_moverzs", "msx_ints", "msx_ns", "ms1_fr", "ms1_moverzs", "ms1_ints",
    "ms1_charges", "scan_title", "ms_level", "ret_time", "scan_num", "orig_scan",
    "iso_ctr", "iso_lwr", "iso_upr", "mobility"),
  step = 1.6e-05
)
}
\arguments{
\item{file}{A file name of \code{spectra_[...].txt}.}

\item{path}{The full path to a .d folder.}

\item{keys}{The key names of output lists.}

\item{step}{A step size for mass binning.}
}
\description{
(1) Pairs MS2 data and iso_info by frame IDs; (2) Collapses data by Precursor
groups.
}
