% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{hextract_pasef_frame}
\alias{hextract_pasef_frame}
\title{Helper of \link{extract_pasef_frame}}
\usage{
hextract_pasef_frame(mdata)
}
\arguments{
\item{mdata}{A list of PASEF frame.}
}
\description{
Helper of \link{extract_pasef_frame}
}
