% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{hfind_unimod}
\alias{hfind_unimod}
\title{Helper of \link{find_unimod}.}
\usage{
hfind_unimod(xml_files = c("master.xml", "custom.xml"), unimod)
}
\arguments{
\item{xml_files}{A list of xml file names.}

\item{unimod}{The name of a \href{https://www.unimod.org/}{Unimod} modification.}
}
\description{
Helper of \link{find_unimod}.
}
