% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{hfwd_prps}
\alias{hfwd_prps}
\title{Helper of finding forward protein-peptide map.}
\usage{
hfwd_prps(.time_stamp, .path_ms1masses, uniq_peps)
}
\arguments{
\item{.time_stamp}{An MS1 time stamp.}

\item{.path_ms1masses}{The file path to the theoretical masses of MS1
precursors.}

\item{uniq_peps}{Unique forward peptide sequence in the lookup table.}
}
\description{
Helper of finding forward protein-peptide map.
}
