% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{hloadMZML}
\alias{hloadMZML}
\title{Helper of \link{loadMZML}.}
\usage{
hloadMZML(filelist = NULL, mgf_path = NULL, temp_dir = NULL)
}
\arguments{
\item{filelist}{A list of MGF or mzML files.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}

\item{temp_dir}{A temporary file folder.}
}
\description{
Helper of \link{loadMZML}.
}
