% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{hmake_noenzpeps}
\alias{hmake_noenzpeps}
\title{Helper of \link{make_noenzpeps} (one start positions).}
\usage{
hmake_noenzpeps(
  start = 1L,
  prot = NULL,
  min_len = 7L,
  max_len = 40L,
  len = NULL,
  aa_masses = NULL,
  ftmass = 18.010565
)
}
\arguments{
\item{start}{The staring position of an amino-acid in a \code{prot}.}

\item{prot}{A FASTA entry of protein.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{len}{The number of amino-acid residues in a \code{prot}.}

\item{aa_masses}{A lookup table of the masses of amino-acid residues.}

\item{ftmass}{The sum of masses of \code{fixed} N-term and C-term
modifications.}
}
\description{
For \code{no enzyme} searches.
}
\details{
C-term tagged with "-" after the mass calculations (but not yet N-term).
}
\examples{
\donttest{
fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", 
              "Carbamidomethyl (C)")

varmods = c("Acetyl (Protein N-term)", "Oxidation (M)", 
            "Deamidated (N)","Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods, varmods = varmods)
aa_masses <- aa_masses_all[[1]]

aas <- LETTERS[LETTERS \%in\% names(aa_masses)]
prot <- paste0(aas, collapse = "")
len <- nchar(prot)
masses <- hmake_noenzpeps(1, prot, 7, 40, len, aa_masses)
}
}
