% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{hms1_a0_vnl0_fnl1}
\alias{hms1_a0_vnl0_fnl1}
\title{Helper of \link{ms1_a0_vnl0_fnl1}.}
\usage{
hms1_a0_vnl0_fnl1(masses, fnl_combi, aa_masses, digits = 4L)
}
\arguments{
\item{masses}{A named list of peptide masses.}

\item{fnl_combi}{A data.frame of combinations of neutral losses for fixed
modifications. Each row corresponds to a set of neutral loss. The first row
corresponds to the combination without NLs (all zeros).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Helper of \link{ms1_a0_vnl0_fnl1}.
}
