% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{hms1_a1_vnl0_fnl0}
\alias{hms1_a1_vnl0_fnl0}
\title{Helper of \link{ms1_a1_vnl0_fnl0}.}
\usage{
hms1_a1_vnl0_fnl0(
  masses,
  amods,
  aa_masses,
  vmods_nl = NULL,
  fmods_nl = NULL,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  ms1vmods = NULL,
  min_mass = 200L,
  max_mass = 4500L,
  digits = 4L
)
}
\arguments{
\item{masses}{A named list of peptide masses.}

\item{amods}{\code{Anywhere} variable modifications.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{vmods_nl}{The attribute of \code{vmods_nl} from an \code{aa_masses}.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{ms1vmods}{The set of all possible MS1 vmod labels at a given aa_masses.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper of \link{ms1_a1_vnl0_fnl0}.
}
