% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{hms2match}
\alias{hms2match}
\title{Help of \link{ms2match_all}}
\usage{
hms2match(
  aa_masses_all,
  funs_ms2,
  ms1vmods_all,
  ms2vmods_all,
  ms1_neulosses = NULL,
  maxn_neulosses_fnl = 1L,
  maxn_neulosses_vnl = 1L,
  deisotope_ms2 = TRUE,
  mod_indexes,
  mgf_path,
  out_path,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_fnl_per_seq = 3L,
  maxn_vnl_per_seq = 3L,
  maxn_vmods_sitescombi_per_pep = 64L,
  minn_ms2 = 6L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  by_modules = FALSE,
  df0 = NULL
)
}
\arguments{
\item{aa_masses_all}{All look-ups of \code{aa_masses}}

\item{funs_ms2}{The functions in correspondence to \code{aa_masses_all} for
calculating MS2 ion series.}

\item{ms1vmods_all}{All possible labels of MS1 variable modifications in
correspondence to \code{aa_masses_all}}

\item{ms2vmods_all}{All possible labels of MS2 variable modifications in
correspondence to \code{aa_masses_all}}

\item{ms1_neulosses}{Character string(s) specifying the neutral losses of
  precursors. The nomenclature is the same as those in argument
  \code{varmods}, e.g., \code{c("Phospho (S)", "Phospho (T)", "Phospho
  (Y)")}.

  The argument is a simplified (narrower) usage of argument
  \code{ms1_notches} with additional specificity in modifications and sites.}

\item{maxn_neulosses_fnl}{A positive integer used in conjunction with
arguments \code{ms1_notches} and \code{ms1_neulosses}. The maximum number
of fixed MS2 neutral losses to be considered when searching against peak
lists with MS1 mass off-sets. The default is two (one MS2 neutral loss in
addition to 0).}

\item{maxn_neulosses_vnl}{A positive integer used in conjunction with
arguments \code{ms1_notches} and \code{ms1_neulosses}. The maximum number
of variable MS2 neutral losses to be considered when searching against peak
lists with MS1 mass off-sets. The default is two (one MS2 neutral loss in
addition to 0).}

\item{deisotope_ms2}{Logical; if TRUE, de-isotope MS2 features.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}

\item{out_path}{A file path of outputs.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{maxn_fnl_per_seq}{A non-negative integer; the maximum number of
permutative neutral losses per peptide sequence for fixed modifications. To
bypass the combinatorial of neutral losses, set \code{maxn_fnl_per_seq = 1}
or \code{0}.}

\item{maxn_vnl_per_seq}{A non-negative integer; the maximum number of
permutative neutral losses per peptide sequence for variable modifications.
To bypass the combinatorial of neutral losses, set \code{maxn_vnl_per_seq =
1} or \code{0}.}

\item{maxn_vmods_sitescombi_per_pep}{A non-negative integer; the maximum
number of combinatorial variable modifications per peptide sequence (per
module). The ways include the permutations in neutral losses and
modifications (e.g., \code{Acetyl (K) and TMT (K)}).}

\item{minn_ms2}{A positive integer; the minimum number of matched MS2 ions
for consideration as a hit. Counts of secondary ions, e.g. b0, b* etc., are
not part of the threshold.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{by_modules}{Not used. Logical. At the TRUE default, searches MS data by
individual modules of combinatorial fixed and variable modifications. If
FALSE, search all modules together. The later would probably need more than
32G RAM if the number of modules is over 96.}

\item{df0}{An output template with zero rows}
}
\description{
By MGF chunks
}
