% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{hpair_mgths}
\alias{hpair_mgths}
\title{Helper of \link{pair_mgftheos}.}
\usage{
hpair_mgths(
  ms1_offset = 0,
  notch = NULL,
  mgfs,
  n_modules,
  by_modules = TRUE,
  mgf_path,
  min_mass = 200L,
  max_mass = 4500L,
  ppm_ms1_bin = 10L,
  .path_bin
)
}
\arguments{
\item{ms1_offset}{The ms1 offset.}

\item{notch}{The index assigned to an ms1_offset.}

\item{mgfs}{MGF data.}

\item{n_modules}{The number of modules (\code{length(aa_masses_all)}) or one}

\item{by_modules}{Logical; if TRUE, results are saved with one mgf to one
theo module. At FALSE, results are saved with one mgf paired to all theo
modules}

\item{mgf_path}{The path to MGF files}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{ppm_ms1_bin}{The tolerance in precursor mass error after mass binning.}

\item{.path_bin}{The path to binned theoretical masses}
}
\description{
Helper of \link{pair_mgftheos}.
}
