% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{hproc_pasefs}
\alias{hproc_pasefs}
\title{Processes single .d file}
\usage{
hproc_pasefs(raw_file, mgf_path, temp_dir, n_para = 1L, debug = FALSE)
}
\arguments{
\item{raw_file}{The name of a .d folder.}

\item{mgf_path}{A file path to .d folders.}

\item{temp_dir}{A file path for temporary files.}

\item{n_para}{The number of maximum allowed parallel processes.}

\item{debug}{Debugging mode or not.}
}
\description{
By each .d file.
}
