% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{hsemipeps_byprots}
\alias{hsemipeps_byprots}
\title{Helper of \link{semipeps_byprots}.}
\usage{
hsemipeps_byprots(prots, min_len = 7L, aa_masses = NULL)
}
\arguments{
\item{prots}{Lists of proteins with tryptic peptides. For each entry under a
protein, the value is mass and the name is peptide sequence.}

\item{min_len}{The minimum length of peptide sequences for consideration.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
Helper of \link{semipeps_byprots}.
}
