% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{hsubDIAMS1}
\alias{hsubDIAMS1}
\title{Helper of \link{subDIAMS1} for parallel processing.}
\usage{
hsubDIAMS1(filename = NULL, temp_dir = NULL, n_para = 1L)
}
\arguments{
\item{filename}{A filename.}

\item{temp_dir}{A temporary file folder.}

\item{n_para}{The allowance of parallel processing.}
}
\description{
Helper of \link{subDIAMS1} for parallel processing.
}
