% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{htable_unimods}
\alias{htable_unimods}
\title{Helper of \link{table_unimods}.}
\usage{
htable_unimods(file)
}
\arguments{
\item{file}{A file path to a Unimod ".xml".}
}
\description{
Helper of \link{table_unimods}.
}
