% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{htraceDIA}
\alias{htraceDIA}
\title{Helper of tracing DIA.}
\usage{
htraceDIA(
  filename,
  raw_id,
  temp_dir,
  mgf_path,
  min_ret_time = 0L,
  max_ret_time = Inf,
  min_mass = 200L,
  max_mass = 4500L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  ppm_ms1 = 8L,
  ppm_ms2 = 8L,
  n_dia_ms2bins = 1L,
  n_dia_scans = 4L,
  delayed_diams2_tracing = FALSE,
  topn_ms2ions = 150L,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  quant = "none",
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  n_para = 1L
)
}
\arguments{
\item{filename}{A filename.}

\item{raw_id}{A raw file id.}

\item{temp_dir}{A temporary directory.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}

\item{min_ret_time}{A non-negative numeric; the minimum retention time in
seconds for considerations. The default is 0.}

\item{max_ret_time}{Numeric; the maximum retention time in seconds for
considerations. The default is \code{Inf}. At a negative value, e.g. -800,
data in the last 800s of LC retention time will be excluded.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{n_dia_ms2bins}{Allowance in adjacent DIA-MS2 bins for retention time
alignment with a DIA-MS1.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}

\item{delayed_diams2_tracing}{Logical; if TRUE, delays the tracing of MS1-MS2
after ion matches.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{mgf_cutmzs}{Cut points of MS1 m-over-z values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{n_para}{The allowance of parallel processing.}

\item{gap}{The size of gap.}
}
\description{
Helper of tracing DIA.
}
