% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{htraceXY}
\alias{htraceXY}
\title{Helper of \link{traceXY}.}
\usage{
htraceXY(
  xs,
  ys,
  df,
  gap = 128L,
  type = c("first", "middle", "last"),
  n_dia_scans = 4L,
  from = 200L,
  step = 1e+05
)
}
\arguments{
\item{xs}{Vectors of MS1 moverzs.}

\item{ys}{Vectors of MS1 intensities.}

\item{df}{A data frame of both MS1 and MS2 staggering the range of \code{xs}.}

\item{gap}{A gap size.}

\item{type}{The type of data subtype.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}

\item{from}{The starting point for mass binning.}

\item{step}{The step size for mass binning.}
}
\description{
Helper of \link{traceXY}.
}
