% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{index_mz}
\alias{index_mz}
\title{Converts ms2_moverz to integers.}
\usage{
index_mz(x, from = 115L, d = 1e-05)
}
\arguments{
\item{x}{Numeric; MS2 mass.}

\item{from}{Numeric; the starting MS1 mass.}

\item{d}{Numeric; \eqn{ppm * 10E-6}.}
}
\description{
Converts ms2_moverz to integers.
}
\seealso{
find_ms1_interval
}
