% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{ins_permlab}
\alias{ins_permlab}
\title{Helper of \code{add_one_label} and \code{add_one_permlab}.}
\usage{
ins_permlab(M, x, nrow, ncol, nrow2, ncol2, rm_dup = FALSE)
}
\arguments{
\item{M}{A permutation table with unique sets by rows.}

\item{x}{A duplicated label that can be found in M.}

\item{nrow}{The number of rows of M.}

\item{ncol}{The number of columns of M.}

\item{nrow2}{The number of rows of the output.}

\item{ncol2}{The number of columns of the output.}

\item{rm_dup}{Logical; remove duplicated rows or not.}
}
\description{
No duplicated labels for base permutation and \code{rm_dup = FALSE}.
}
