% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{insVal}
\alias{insVal}
\title{Inserts a value.}
\usage{
insVal(x, sv)
}
\arguments{
\item{x}{A value to be inserted.}

\item{sv}{A sorted vector of indexes(thus without ties).}
}
\description{
Inserts a value.
}
\examples{
library(mzion)
sv <- c(4, 10)

sv <- mzion:::insVal(2, sv)
sv <- mzion:::insVal(6, sv)
sv <- mzion:::insVal(20, sv)
}
