% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{integerize_ms2ints}
\alias{integerize_ms2ints}
\title{Integerizes the MS2 intensities.}
\usage{
integerize_ms2ints(ms2_ints, max_intdbl = 2147483647)
}
\arguments{
\item{ms2_ints}{Lists of MS2 intensities}

\item{max_intdbl}{Maximum 32-bit integer as double:
\code{as.double(.Machine$integer.max)}.}
}
\description{
Also guards against intensity integers above machine maximum (32-bit).
}
