% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{is_nulllist}
\alias{is_nulllist}
\title{Is a NULL list?}
\usage{
is_nulllist(x)
}
\arguments{
\item{x}{A list.}
}
\description{
Is a NULL list?
}
\examples{
\dontrun{
library(mzion)
mzion:::is_nullist(list(a = NULL))
}
}
