% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{is_partial_peak}
\alias{is_partial_peak}
\title{Check partial peaks}
\usage{
is_partial_peak(ys, width = 200L, min_p = 10L)
}
\arguments{
\item{ys}{A vector of intensity values}

\item{width}{The width of a peak (number of points).}

\item{min_p}{The minimum number of points on one side of a peak for excluding
the call for a partial peak.}
}
\description{
Check partial peaks
}
