% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{keep_pepfdr_best}
\alias{keep_pepfdr_best}
\title{Keeps the best entries of targets and decoys.}
\usage{
keep_pepfdr_best(td, cols = c("pep_scan_num", "raw_file"))
}
\arguments{
\item{td}{A data frame of targets and decoys.}

\item{cols}{Columns for grouping.}
}
\description{
Keeps the best entries of targets and decoys.
}
