% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{list_leftmatch}
\alias{list_leftmatch}
\title{Matches two lists without making a data frame..}
\usage{
list_leftmatch(a, b)
}
\arguments{
\item{a}{The left vector.}

\item{b}{The right vector.}
}
\description{
Not currently used.
}
\examples{
\donttest{
library(mzion)

a <- c(3, 4, 1, 2, 5)
b <- 2

mzion:::list_leftmatch(a, b)
}
}
