% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{load_cache_info}
\alias{load_cache_info}
\title{Loads variable to the global environment.}
\usage{
load_cache_info(cache = NULL, overwrite = TRUE)
}
\arguments{
\item{cache}{A cache file name.}

\item{overwrite}{If TRUE, overwrite the pre-existed global values with the
cache values.}
}
\description{
Not yet used.
}
\seealso{
\link{find_ms1_times}
}
