% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{load_mgfs}
\alias{load_mgfs}
\title{Helper in loading MGFs.}
\usage{
load_mgfs(
  out_path,
  mgf_path,
  min_mass = 500L,
  max_mass = 6000L,
  min_ms2mass = 110L,
  topn_ms2ions = 100L,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  index_ms2 = FALSE,
  is_ms1_three_frame = TRUE,
  is_ms2_three_frame = TRUE
)
}
\arguments{
\item{out_path}{A file path of outputs.}

\item{mgf_path}{The file path to a list of MGF files. The experimenters need
to supply the files. Note that the supported MGFs are from MSConvert or
Proteome Discoverer.}

\item{min_mass}{A minimum mass of precursors for considerations.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in MS2
ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{index_ms2}{Logical; if TRUE, converts MS2 m/z values to indices.}

\item{is_ms1_three_frame}{Logical; is the searches by the three frames of
preceeding, current and following.}

\item{is_ms2_three_frame}{Logical; is the searches by the three frames of
preceeding, current and following.}
}
\description{
Currently, \code{index_ms2} is always FALSE. Argument \code{ppm_ms2} is only
effective at \code{index_ms2 = TRUE}.
}
