% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{loadMZML}
\alias{loadMZML}
\title{Loads mzML from MSConvert.}
\usage{
loadMZML(xml_file = NULL, temp_dir = NULL)
}
\arguments{
\item{xml_file}{A file name of mzML.}

\item{temp_dir}{A temporary file folder.}
}
\description{
For single mzML file.
}
