% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{make_dia_mgfs}
\alias{make_dia_mgfs}
\title{Makes DIA MGFs}
\usage{
make_dia_mgfs(mgfs, mgf_path, min_mass = 200L, ppm_ms1_bin = 10L)
}
\arguments{
\item{mgfs}{MGF data.}

\item{mgf_path}{The path to MGF files}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{ppm_ms1_bin}{The tolerance in precursor mass error after mass binning.}
}
\description{
Replicates MS2 data by the multiplicity of precursor masses.
}
