% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{make_ms1_vmodsets}
\alias{make_ms1_vmodsets}
\title{Makes the sets of labels of variable modifications.}
\usage{
make_ms1_vmodsets(
  aa_masses_all = NULL,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L
)
}
\arguments{
\item{aa_masses_all}{All the amino acid lookup tables.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}
}
\description{
No position permutation (for MS1 masses).
}
\examples{
\donttest{
fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)")

varmods <- c("Acetyl (Protein N-term)", "Oxidation (M)", "Carbamidomethyl (M)", 
            "Deamidated (N)", "Carbamyl (M)", 
            "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods,
                               varmods = varmods,
                               maxn_vmods_setscombi = 64,
                               add_varmasses = FALSE,
                               add_nlmasses = FALSE, 
                               exclude_phospho_nl = TRUE, 
                               out_path = NULL)

.ms1_vmodsets <- make_ms1_vmodsets(aa_masses_all)
.base_ent <- lapply(.ms1_vmodsets, `[[`, 1)

extract_vmodsets(.ms1_vmodsets, .base_ent, p = 4, 
                 labs = c("Carbamidomethyl (M)", "Carbamyl (M)"))
}
}
