% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMS2.R
\name{make_ms2frames}
\alias{make_ms2frames}
\title{Bin MS2 ion series by precursor masses.}
\usage{
make_ms2frames(
  ms1_time = NULL,
  ms1s,
  ms2s,
  .path_cache = "~/proteoM/.MSearches/Cache/Calls",
  path_bin2,
  i
)
}
\arguments{
\item{ms1_time}{A ms1_time stamp.}

\item{ms1s}{Lists of named values. Peptide sequences in names and precursor
masses in values.}

\item{ms2s}{Lists of MS2 ion series.}

\item{.path_cache}{The file path of cached search parameters. The parameter
  is for the users' awareness of the underlying structure of file folders and
  the use of default is suggested. Occasionally experimenters may remove the
  file folder for disk space or under infrequent events of modified framework
  incurred by the developer.

  Started from version 1.2.1.5, a new data structure of protein-peptide
  lookups are used. Version 1.2.1.4.1 is the final version using the old data
  structure.}

\item{path_bin2}{A file path to binned MS2 results.}

\item{i}{An index of aa_masses}
}
\description{
The same MS2 results but different frame numbers according to the paramter
set of min_mass, max_mass and ppm_ms1.
}
