% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_ms2ions.R
\name{make_ms2frames_bypars}
\alias{make_ms2frames_bypars}
\title{Bins MS2 ion series at a set of parameters.}
\usage{
make_ms2frames_bypars(
  pars,
  ms1s,
  ms2s,
  path_bin2,
  i,
  is_ms1_three_frame = TRUE
)
}
\arguments{
\item{pars}{A set of cached parameters from \code{.Msearches}.}

\item{ms1s}{Lists of named values. Peptide sequences in names and precursor
masses in values.}

\item{ms2s}{Lists of MS2 ion series.}

\item{path_bin2}{A file path to binned MS2 results.}

\item{i}{An index of aa_masses}

\item{is_ms1_three_frame}{Logical; is the searches by the three frames of
preceeding, current and following.}
}
\description{
The set of parameters include min_mass, max_mass and ppm_ms1.
}
