% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mztab.R
\name{make_mztab}
\alias{make_mztab}
\title{Makes an mzTab file.}
\usage{
make_mztab(
  mzion_path = stop("Provide the path.", call. = FALSE),
  proteoq_path = mzion_path
)
}
\arguments{
\item{mzion_path}{The parent parthwhere \code{mzion} search was performed.}

\item{proteoq_path}{A parent path where the \code{proteoQ} post-processing
was performed.}
}
\description{
With \code{mzion} searches and \code{proteoQ} post-processing.
}
