% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{make_noenzpeps}
\alias{make_noenzpeps}
\title{Makes peptides and masses for a protein.}
\usage{
make_noenzpeps(
  prot = NULL,
  min_len = 7L,
  max_len = 40L,
  aa_masses = NULL,
  ftmass = 18.010565
)
}
\arguments{
\item{prot}{A FASTA entry of protein.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{aa_masses}{A lookup table of the masses of amino-acid residues.}

\item{ftmass}{The sum of masses of \code{fixed} N-term and C-term
modifications.}
}
\description{
For \code{no enzyme} searches.
}
\examples{
\donttest{
library(mzion)

prot <- paste0(
  "MSSKQHCVKLNDGHLIPALGFGTYKPKEVPKSKSLEAACLA", 
  "LDVGYRHVDTAYAYQVEEEIGQAIQSKIKAGVVKREDLFIT", 
  "TKLWCTCFRPELVKPALEKSLKKLQLDYVDLYIMHYPVPMK", 
  "SGDNDFPVNEQGKSLLDTVDFCDTWERLEECKDAGLVKSIG", 
  "VSNFNHRQLERILNKPGLKYKPVCNQVECHLYLNQRKLLDY",
  "CESKDIVLVAYGALGTQRYKEWVDQNSPVLLNDPVLCDVAK", 
  "KNKRSPALIALRYLIQRGIVPLAQSFKENEMRENLQVFGFQ", 
  "LSPEDMKTLDGLNKNFRYLPAEFLVDHPEYPFVEEY")

fixedmods = c("TMT6plex (N-term)", "TMT6plex (K)", 
              "Carbamidomethyl (C)")

varmods = c("Acetyl (Protein N-term)", "Oxidation (M)", 
            "Deamidated (N)","Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods, varmods = varmods)
aa_masses <- aa_masses_all[[1]]

ans <- mzion:::make_noenzpeps(prot, 7, 40, aa_masses)

# short FASTA (both N-term and C-term on a sequence)
prot <- substring(prot, 1, 10)
ans <- mzion:::make_noenzpeps(prot, 7, 40, aa_masses)
}
}
