% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{make_speclib}
\alias{make_speclib}
\title{Makes spectrum library}
\usage{
make_speclib(
  out_path = NULL,
  in_name = "psmQ.txt",
  score_co = 15,
  mdev_co = 5,
  sd_co = 10,
  type_ms2ions = "by"
)
}
\arguments{
\item{out_path}{A file path where outputs of \link{matchMS} can be
identified.}

\item{in_name}{An input file name of PSMs, such as psmQ.txt, psmC.txt,
psmT2.txt, psmT3.txt etc.}

\item{score_co}{The cut-off in PSM scores for library construction.}

\item{mdev_co}{The cut-off in PSM \code{pep_ms2_deltas_mean}. The default is
5 (ppm).}

\item{sd_co}{The cut-off in PSM \code{pep_ms2_deltas_sd}. The default is 10.}
}
\description{
Makes spectrum library
}
