% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms1_labels.R
\name{make_unique_sets}
\alias{make_unique_sets}
\title{Makes the unique sets of modifications by individual AA residues.}
\usage{
make_unique_sets(
  p = 5L,
  n = 2L,
  labs = c("X", "Y"),
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L
)
}
\arguments{
\item{p}{The number of open positions for filling.}

\item{n}{The number of labels.}

\item{labs}{The labels of balls.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}
}
\description{
Makes the unique sets of modifications by individual AA residues.
}
\examples{
\donttest{
library(mzion)

# multiple outs
p <- 5
labs <- c("Oxidation (M)", "Carbamidomethyl (M)", "Carbamyl (M)")
n <- length(labs)
stopifnot(n == length(labs))

ans5 <- mzion:::make_unique_sets(5, n, labs)
ans7 <- mzion:::make_unique_sets(7, n, labs)

stopifnot(identical(ans5, ans7))

# single out
p <- 3
labs <- c("Oxidation (M)", "Carbamidomethyl (M)", "Carbamyl (M)")
n <- length(labs)
stopifnot(n == length(labs))

ans <- mzion:::make_unique_sets(p, n, labs)
}
}
