% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{map_raw_n_scan}
\alias{map_raw_n_scan}
\title{Maps raw_file and scan_title from indexes to real values.}
\usage{
map_raw_n_scan(df, mgf_path)
}
\arguments{
\item{df}{A data frame.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}
}
\description{
Maps raw_file and scan_title from indexes to real values.
}
