% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{mapcoll_xyz}
\alias{mapcoll_xyz}
\title{Sub modules in data collapsion.}
\usage{
mapcoll_xyz(
  vals,
  ps,
  lenx,
  lenu,
  temp_dir,
  icenter = 1L,
  ms_lev = 2L,
  type = "xs",
  direct_out = FALSE
)
}
\arguments{
\item{vals}{Data in one of moverzs, intensities or charge states.}

\item{ps}{Positions of \code{vals} in universe. Should be the same for among
moverzs, intensities and charge states.}

\item{lenx}{The length of \code{vals}.}

\item{lenu}{The number of entries in the universe.}

\item{temp_dir}{A temporary directory.}

\item{icenter}{The index of an isolation center.}

\item{ms_lev}{The level of MS.}

\item{type}{The type of data in one of xs, ys and zs.}

\item{direct_out}{Logical; if TRUE, return outputs directly.}
}
\description{
For RAM efficiency.
}
