% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{mapMS2ions}
\alias{mapMS2ions}
\title{Visualization of matched MS2 ions.}
\usage{
mapMS2ions(
  out_path = "~/proteoM/outs",
  scan = 1234,
  raw_file = "foo.raw",
  rank = 1L,
  is_decoy = FALSE,
  type_ms2ions = "by",
  filename = "bar.png"
)
}
\arguments{
\item{out_path}{A file path where outputs of \link{matchMS} can be
identified.}

\item{scan}{A scan number that can be found from the outputs. Positive
  integer, Thermo's Orbitrap data; character string, Bruker's timsTOF data.

  The searches of scan number will proceed automatically through
  \code{psmQ.txt},  \code{psmT2.txt}, \code{psmT3.txt} and/or
  \code{psmC.txt}.}

\item{raw_file}{Character string; a RAW file name.}

\item{rank}{Positive integer; the rank of a match. The default is one for the
best match.}

\item{is_decoy}{Logical; is the match from decoy. The default is FALSE.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{filename}{An output file name for saving the the plot of MS2 ions.}
}
\description{
Secondary ions of b0, y0 etc. were also used in ion matches though not
directly documented.
}
\examples{
\donttest{
ans <- mapMS2ions(out_path = "a/proteoM/output/folder",
                  scan = 9933,
                  raw_file = "a-raw-file-name.raw",
                  rank = 1L,
                  is_decoy = FALSE)

# Custom plots
library(ggplot2)

mgf <- ans$mgf
duo <- ans$duo

ggplot() +
geom_segment(mgf, mapping = aes(x = ms2_moverz, y = ms2_int,
                                         xend = ms2_moverz, yend = 0),
                      color = "gray", size = .1) +
  geom_segment(duo, mapping = aes(x = ms2_moverz, y = ms2_int,
                                           xend = ms2_moverz, yend = 0,
                                           color = type),
                        size = .4, show.legend = FALSE) +
  geom_text(duo, mapping = aes(x = ms2_moverz, y = ms2_int,
                                        label = label, color = type),
                     size = 3, alpha = .5, hjust = 0, angle = 90, vjust = 0,
                     nudge_x = 0.05, nudge_y = 0.05, na.rm = TRUE,
                     show.legend = FALSE) +
  labs(x = "m/z", y = "Intensity")
}
}
