% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{match_aas_indexes}
\alias{match_aas_indexes}
\title{Matches the indexes of amino-acid residues to cached results.}
\usage{
match_aas_indexes(X, Vec)
}
\arguments{
\item{X}{Lists of cached results.}

\item{Vec}{A vector of names (lower-case vec for sites).}
}
\description{
Matches the indexes of amino-acid residues to cached results.
}
