% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{match_ex2th2}
\alias{match_ex2th2}
\title{Matches between secondary experimentals and theoreticals.}
\usage{
match_ex2th2(expt, theo, min_ms2mass = 115L, d = 1e-05)
}
\arguments{
\item{expt}{A vector of experimental m-over-z values.}

\item{theo}{A vector of theoretical m-over-z values.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{d}{Bin size, e.g., \eqn{20 ppm / 2 * 1E-6}.}
}
\description{
Matches between secondary experimentals and theoreticals.
}
