% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{match_mgf_path}
\alias{match_mgf_path}
\title{Matches the value of mgf_path.}
\usage{
match_mgf_path(out_path)
}
\arguments{
\item{out_path}{An output path.}
}
\description{
Matches the value of mgf_path.
}
