% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms1_labels.R
\name{match_mvmods}
\alias{match_mvmods}
\title{Matches to the pre-calculated labels of combinatorial MS1 variable
modifications.}
\usage{
match_mvmods(aas = NULL, ms1vmods = NULL, amods = NULL)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{ms1vmods}{The i-th result from lapply(aa_masses_all, make_ms1vmod_i).}

\item{amods}{\code{Anywhere} variable modifications.}
}
\description{
No checking of conflicting [NC]-term and Anywhere sites. Even so, it only
sets a overall limit on the counts. Still need to check the MS2 permutation
with indexes in aas for exact exclusion of certain permutations.
}
\examples{
\donttest{
library(proteoM)

fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)",
               "Carbamidomethyl (C)")

varmods <- c("Acetyl (Protein N-term)", "Oxidation (M)",
             "Deamidated (N)",
             "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods,
                               varmods = varmods,
                               maxn_vmods_setscombi = 64,
                               out_path = NULL)

maxn_vmods_per_pep <- 5L
maxn_sites_per_vmod <- 3L

ms1vmods_all <- lapply(aa_masses_all, proteoM:::make_ms1vmod_i,
                       maxn_vmods_per_pep = maxn_vmods_per_pep,
                       maxn_sites_per_vmod = maxn_sites_per_vmod)

i <- 11L
aa_masses <- aa_masses_all[[i]]
amods <- attr(aa_masses, "amods")
ms1vmods <- ms1vmods_all[[i]]

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

vmods_combi <- proteoM:::match_mvmods(aas = aas, ms1vmods = ms1vmods, amods = amods)

## M and N
fixedmods = c("TMT6plex (K)", "dHex (S)")
varmods = c("Carbamidomethyl (M)", "Carbamyl (M)",
            "Deamidated (N)", "Acetyl (Protein N-term)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

maxn_vmods_per_pep <- 5L
maxn_sites_per_vmod <- 3L

ms1vmods_all <- lapply(aa_masses_all, proteoM:::make_ms1vmod_i,
                       maxn_vmods_per_pep = maxn_vmods_per_pep,
                       maxn_sites_per_vmod = maxn_sites_per_vmod)

i <- 16L
aa_masses <- aa_masses_all[[i]]
amods <- attr(aa_masses, "amods")
ntmod <- attr(aa_masses, "ntmod")
ctmod <- attr(aa_masses, "ctmod")

ms1vmods <- ms1vmods_all[[i]]

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

vmods_combi <- proteoM:::match_mvmods(aas = aas, ms1vmods = ms1vmods, amods = amods)

stopifnot(sapply(vmods_combi$ms1, function (x) all(names(amods) \%in\% x)),
          length(vmods_combi$ms1) == 6L)

}
}
