% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{match_raw_id}
\alias{match_raw_id}
\title{Matches the id of raw_file.}
\usage{
match_raw_id(mgf_path, raw_file)
}
\arguments{
\item{mgf_path}{An MGF path.}

\item{raw_file}{Character string; a RAW file name.}
}
\description{
Matches the id of raw_file.
}
