% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{matchMS_ms1calib}
\alias{matchMS_ms1calib}
\title{Wrapper of \link{matchMS}.}
\usage{
matchMS_ms1calib(...)
}
\arguments{
\item{...}{Arguments for \link{matchMS}.}
}
\description{
With the calibration of precursor masses.
}
