% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{matchMS_NES}
\alias{matchMS_NES}
\title{Match MS at no enzymatic specificity.}
\usage{
matchMS_NES(...)
}
\arguments{
\item{...}{Variable arguments.}
}
\description{
Match MS at no enzymatic specificity.
}
