% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{matchMS_noenzyme}
\alias{matchMS_noenzyme}
\title{Noenzyme search.}
\usage{
matchMS_noenzyme(
  this_call = NULL,
  min_len = 7L,
  max_len = 40L,
  fasta = NULL,
  out_path = NULL,
  mgf_path = NULL,
  noenzyme_maxn = 0L,
  quant = "none",
  sys_ram = 32L
)
}
\arguments{
\item{this_call}{An expression from match.call.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded. The default is 40.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The experimenter needs to supply the files.}

\item{out_path}{A file path of outputs.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).

  Individuality in MGF files are slightly preferred to take advantage of
  parallel reading of the files.}

\item{noenzyme_maxn}{Non-negative integer; the maximum number of peptide
lengths for sectional searches at \code{noenzyme} specificity. The argument
may be used to guard against RAM exhaustion. At the zero default, The
peptide lengths from \code{min_len} to \code{max_len} will be broken
automatically into continuous sections. At value 1, searches will be
performed against individual peptide lengths; at value 2, two adjacent
lengths will be taken at a time, etc.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{sys_ram}{For Mac OS or Linux. The amount of system RAM in GB.}
}
\description{
Noenzyme search.
}
