% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{matchMS_par_groups}
\alias{matchMS_par_groups}
\title{Searches by sets of parameters.}
\usage{
matchMS_par_groups(
  par_groups = NULL,
  grp_args = NULL,
  mgf_paths = NULL,
  this_call = NULL,
  out_path = NULL
)
}
\arguments{
\item{par_groups}{Parameter(s) of \code{matchMS} multiplied by sets of values
in groups. Multiple searches will be performed separately against the
parameter groups. For instance with one set of samples in SILAC light and
the other in SILAC heavy, the experimenters may specify two arguments for
parameter \code{mgf_path} and two arguments for parameter \code{fixedmods}
that link to the respective samples. In this way, there is no need to
search against, e.g. heavy-isotope-labeled K8R10 with the light samples and
vice versa. Note that results will be combined at the end, with the group
names indicated under column \code{pep_group}. The default is NULL without
grouped searches. See the examples under SILAC and Group searches.}

\item{grp_args}{The names of arguments in \code{par_groups}.}

\item{mgf_paths}{The paths to MGF (with group searches).}

\item{this_call}{An expression from match.call.}

\item{out_path}{A file path of outputs.}
}
\description{
Searches by sets of parameters.
}
