% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{matchMS_silac_mix}
\alias{matchMS_silac_mix}
\title{SILAC}
\usage{
matchMS_silac_mix(
  silac_mix = list(base = NULL, heavy = c("K8 (K)", "R10 (R)")),
  this_call,
  out_path,
  mgf_path,
  aa_masses
)
}
\arguments{
\item{silac_mix}{A list of labels indicating SILAC groups in samples. The
parameter is most relevant for SILAC experiments where peptides of heavy,
light etc. were \emph{mixed} into one sample. The default is NULL
indicating a none mixed-SILAC experiment. See also the examples under
SILAC.}

\item{this_call}{An expression from match.call.}

\item{out_path}{A file path of outputs.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}

\item{aa_masses}{An amino acid look-ups.}
}
\description{
Searches against mixed SILAC groups (heave and light mixed into one sample).
}
