% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{mframes_adv}
\alias{mframes_adv}
\title{Frame advancing}
\usage{
mframes_adv(
  mgf_frames = NULL,
  theopeps = NULL,
  aa_masses_all,
  funs_ms2,
  ms1vmods_all,
  ms2vmods_all,
  mod_indexes = NULL,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_fnl_per_seq = 3L,
  maxn_vnl_per_seq = 3L,
  maxn_vmods_sitescombi_per_pep = 64L,
  minn_ms2 = 6L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  index_mgf_ms2 = FALSE
)
}
\arguments{
\item{mgf_frames}{A group of mgf frames (from chunk splitting). A
\code{mgf_frames[[i]]} contains one to multiple MGFs whose MS1 masses are
in the same interval. The \code{mgf_frames} are ordered by increasing
values in \code{frame} for progressive searches.}

\item{theopeps}{Binned theoretical peptides corresponding to an i-th
\code{aa_masses}.}

\item{aa_masses_all}{A list of amino acid lookups for all the combination of
fixed and variable modifications.}

\item{funs_ms2}{The functions in correspondence to \code{aa_masses_all} for
calculating MS2 ion series.}

\item{ms1vmods_all}{All possible labels of MS1 variable modifications in
correspondence to \code{aa_masses_all}}

\item{ms2vmods_all}{All possible labels of MS2 variable modifications in
correspondence to \code{aa_masses_all}}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{maxn_fnl_per_seq}{A non-negative integer; the maximum number of
permutative neutral losses per peptide sequence for fixed modifications. To
bypass the combinatorial of neutral losses, set \code{maxn_fnl_per_seq = 1}
or \code{0}.}

\item{maxn_vnl_per_seq}{A non-negative integer; the maximum number of
permutative neutral losses per peptide sequence for variable modifications.
To bypass the combinatorial of neutral losses, set \code{maxn_vnl_per_seq =
1} or \code{0}.}

\item{maxn_vmods_sitescombi_per_pep}{A non-negative integer; the maximum
number of combinatorial variable modifications per peptide sequence (per
module). The ways include the permutations in neutral losses and
modifications (e.g., \code{Acetyl (K) and TMT (K)}).}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{index_mgf_ms2}{A low-priority feature. Logical; if TRUE, converts
  upfrontly MS2 m-over-z values from numeric to integers as opposed to
  \emph{in-situ} conversion during ion matches. The default is FALSE. The
  \code{index_mgf_ms2 = TRUE} might be useful for very large MS files by
  reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}
}
\value{
Matches to each MGF as a list elements. The length of the output is
  equal to the number of MGFs in the given frame.
}
\description{
For all modules of fixed and variable modifications
}
