% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{mmake_noenzpeps}
\alias{mmake_noenzpeps}
\title{Helper: multiple applications of \link{make_noenzpeps}.}
\usage{
mmake_noenzpeps(
  fasta_db = NULL,
  min_len = 7L,
  max_len = 40L,
  aa_masses = NULL,
  ftmass = 18.010565
)
}
\arguments{
\item{fasta_db}{Fasta database(s).}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{aa_masses}{A lookup table of the masses of amino-acid residues.}

\item{ftmass}{The sum of masses of \code{fixed} N-term and C-term
modifications.}
}
\description{
For \code{no enzyme} searches.
}
