% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{move_scfiles}
\alias{move_scfiles}
\title{Moves score-related files from \code{tempdir} to \code{sc_path}}
\usage{
move_scfiles(type = "list_table", tempdir, sc_path)
}
\arguments{
\item{type}{The type of files}

\item{tempdir}{A temporary directory containing the files}

\item{sc_path}{A target directory}
}
\description{
Moves score-related files from \code{tempdir} to \code{sc_path}
}
