% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{mprepBrukerMGF}
\alias{mprepBrukerMGF}
\title{Parallel \link{prepBrukerMGF}}
\usage{
mprepBrukerMGF(filepath, n_cores = 48L)
}
\arguments{
\item{filepath}{A file path to MGF.}

\item{n_cores}{The number of CPU cores.}
}
\description{
Parallel \link{prepBrukerMGF}
}
