% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{ms1_a1_vnl0_fnl0}
\alias{ms1_a1_vnl0_fnl0}
\title{Helper by individual peptides.}
\usage{
ms1_a1_vnl0_fnl0(
  mass,
  aa_seq,
  amods,
  aa_masses,
  vmods_nl = NULL,
  fmods_nl = NULL,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  ms1vmods = NULL,
  min_mass = 700L,
  max_mass = 4500L,
  digits = 4L
)
}
\arguments{
\item{mass}{The mass of a peptide.}

\item{aa_seq}{Character string; a peptide sequence with one-letter
representation of amino acids.}

\item{amods}{\code{Anywhere} variable modifications.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{vmods_nl}{The attribute of \code{vmods_nl} from an \code{aa_masses}.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{ms1vmods}{The set of all possible MS1 vmod labels at a given aa_masses.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value. The primary guard
against low molecular-weight precursors is \code{min_len}.}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation. The default is 4500.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
(7) "amods+ tmod- vnl- fnl-"; (8) "amods+ tmod+ vnl- fnl-".
}
\examples{
\donttest{
m0 <- calc_monopeptide("HQGVMCNVGMGQKMNSC", NULL, NULL)
stopifnot(unlist(m0$mass, use.names = FALSE) - 1822.7405 < 1e-4)

m1 <- calc_monopeptide("HQGVMCNVGMGQKMNSC", "TMT6plex (N-term)", NULL)
stopifnot(unlist(m1$mass, use.names = FALSE) - 2051.9035 < 1e-4)

# (7) "amods+ tmod- vnl- fnl-"
aa_masses_all <- calc_aamasses(fixedmods = c("TMT6plex (N-term)"),
                               varmods = c("Deamidated (N)",
                                           "Carbamidomethyl (C)"))

pep <- c("HQGVMCNVGMGQKMNSC" = 2051.90346)
amods <- list(`Deamidated (N)` = c("Anywhere" = "N"),
              `Carbamidomethyl (C)` = c("Anywhere" = "C"))

x <- ms1_a1_vnl0_fnl0(pep, names(pep), amods, aa_masses_all[[4]])

stopifnot(x[[1]] - 2109.9089 < 1e-4,
          x[[2]] - 2166.9304 < 1e-4,
          x[[3]] - 2110.8930 < 1e-4,
          x[[4]] - 2167.9144 < 1e-4)


# (8) "amods+ tmod+ vnl- fnl-"
aa_masses_all <- calc_aamasses(fixedmods = "TMT6plex (K)",
                               varmods = c("Deamidated (N)",
                                           "Carbamidomethyl (S)",
                                           "Acetyl (Protein N-term)"))

pep <- c("HQGVMNVGMGQKSMNS" = 1932.9171) # + TMT6plex (K)
amods <- list(`Deamidated (N)` = c("Anywhere" = "N"),
              `Carbamidomethyl (S)` = c("Anywhere" = "S"))

x <- ms1_a1_vnl0_fnl0(pep, names(pep), amods, aa_masses_all[[8]])

stopifnot(x[[1]] - 1990.9226 < 1e-4,
          x[[2]] - 1991.9066 < 1e-4,
          x[[3]] - 2047.9440 < 1e-4,
          x[[4]] - 2048.9281 < 1e-4)


# (8-b)
.ms1_vmodsets <- make_ms1_vmodsets(aa_masses_all = aa_masses_all, 
                                   maxn_vmods_per_pep = 5L, 
                                   maxn_sites_per_vmod = 3L)
.base_ent <- lapply(.ms1_vmodsets, `[[`, 1)

 x <- ms1_a1_vnl0_fnl0(pep, names(pep), amods, aa_masses_all[[8]], 
                      .ms1_vmodsets = .ms1_vmodsets, 
                      .base_ent = .base_ent)

# (8-c)
fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)")

varmods = c("Acetyl (Protein N-term)", "Oxidation (M)", "Deamidated (N)",
            "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods,
                               varmods = varmods,
                               maxn_vmods_setscombi = 64,
                               out_path = NULL)

ms1vmods_all <- lapply(aa_masses_all, make_ms1vmod_i)

i <- 10
aa_masses <- aa_masses_all[[i]]
ms1vmods <- ms1vmods_all[[i]]

pep <- c("HQGVMCNVGMGQKMNSC" = 2051.90346)
amods <- attr(aa_masses, "amods")

x <- ms1_a1_vnl0_fnl0(pep, names(pep), amods, aa_masses_all[[i]], 
                      ms1vmods = ms1vmods)

# y <- ms1_a1_vnl0_fnl0(pep, names(pep), amods, aa_masses_all[[i]], 
#                       ms1vmods = NULL)

# identical(x, y)
}
}
