% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{ms1masses_bare_noenz}
\alias{ms1masses_bare_noenz}
\title{Calculates masses of peptides.}
\usage{
ms1masses_bare_noenz(x, aa_masses, ftmass = 18.010565)
}
\arguments{
\item{x}{A vector of peptide sequences.}

\item{aa_masses}{A lookup table of the masses of amino-acid residues.}

\item{ftmass}{The sum of masses of \code{fixed} N-term and C-term
modifications.}
}
\description{
For no enzyme workflow.
}
\examples{
\donttest{
library(mzion)

fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", 
               "Carbamidomethyl (C)")

varmods <- c("Acetyl (Protein N-term)", "Oxidation (M)", 
             "Deamidated (N)","Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods, varmods = varmods)
aa_masses <- aa_masses_all[[1]]

x <- c("MSSKQHC", "MSSKQHCV", "MSSKQHCVK", "MSSKQHCVKL")
masses1 <- mzion:::ms1masses_bare_noenz(x, aa_masses)

## input peptides should contain neither N- or C-term "-" tag
# y <- c("MSSKQHC", "MSSKQHCV", "MSSKQHCVK", "MSSKQHCVKL-")
# masses2 <- mzion:::ms1masses_bare_noenz(y, aa_masses)
# stopifnot(identical(unname(masses1), unname(masses2)))
}
}
