% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{ms1masses_noterm}
\alias{ms1masses_noterm}
\title{Helper of \link{ms1masses_bare}.}
\usage{
ms1masses_noterm(aa_seqs, aa_masses)
}
\arguments{
\item{aa_seqs}{Character string; a vector of peptide sequences with
one-letter representation of amino acids.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
For either forward or reversed sequences.
}
