% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_ladder.R
\name{ms2ions_by_type}
\alias{ms2ions_by_type}
\title{Helper: switches among ion types for calculating MS2 masses.}
\usage{
ms2ions_by_type(aam, ntmass, ctmass, type_ms2ions = "by", digits = 4L)
}
\arguments{
\item{aam}{A sequence of amino-acid residues with \emph{masses}. Residues
are in names and masses in values (note that argument \code{aas}
corresponds to residues without masses).}

\item{ntmass}{The mass of a fixed or variable N-term modification.}

\item{ctmass}{The mass of a fixed or variable C-term modification.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper: switches among ion types for calculating MS2 masses.
}
