% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{msub_protpep}
\alias{msub_protpep}
\title{Helper of \link{sub_protpep}}
\usage{
msub_protpep(prps, uniq_peps)
}
\arguments{
\item{prps}{A protein-peptide lookup.}

\item{uniq_peps}{A non-redundant list of peptides from search results.}
}
\description{
Subsets the protein-peptide lookup (prps) against the uniq_peps list from
search results.
}
