% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzion.R
\docType{package}
\name{mzion}
\alias{mzion}
\title{mzion: A package for proteomic database searches of mass spectrometry data.}
\description{
The mzion package provides utilities for (1) database searches. The
companion package proteoQ provides utilities for (2) data preprocessing, (3)
annotation, (4) quality assurance assessments and (5) informatic analysis.
}
\section{Main utility in database searches}{
 matchMS (additional UI tools:
  mapMS2ions, load_fasta2, parse_unimod, table_unimods, add_unimod, 
  calc_monopeptide, calc_ms2ionseries)
}

