% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{order_fracs3}
\alias{order_fracs3}
\title{Order fractions}
\usage{
order_fracs3(type = "list_table", tempdir)
}
\arguments{
\item{type}{The type of files.}

\item{tempdir}{A temporary directory containing the files.}
}
\description{
Three layers: module_notch_frac
}
