% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{pad_zeroes}
\alias{pad_zeroes}
\title{Adds padding zero.}
\usage{
pad_zeroes(vec, npl = 2)
}
\arguments{
\item{vec}{A numeric vector.}

\item{npl}{The number of places for padding.}
}
\description{
Adds padding zero.
}
