% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{padd_ion_mathces}
\alias{padd_ion_mathces}
\title{Adds ion matches (primary).}
\usage{
padd_ion_mathces(df, out_path, mgf_path)
}
\arguments{
\item{df}{A data frame of psmC.txt.}

\item{out_path}{A file path of outputs.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).

  Individuality in MGF files are slightly preferred to take advantage of
  parallel reading of the files.}
}
\description{
Currently only used with psmC.txt during matchMS.
}
