% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{pair_mgftheo}
\alias{pair_mgftheo}
\title{Pairs MGF queries to theoretical MS1 masses and peptide sequences.}
\usage{
pair_mgftheo(mgf_path, n_modules, .path_bin, by_modules = TRUE)
}
\arguments{
\item{mgf_path}{The path to MGF files}

\item{n_modules}{The number of modules (\code{length(aa_masses_all)}) or one}

\item{.path_bin}{The path to binned theoretical masses}

\item{by_modules}{Logical; if TRUE, results are saved with one mgf to one
theo module. At FALSE, results are saved with one mgf paired to all theo
modules}
}
\description{
Pairs MGF queries to theoretical MS1 masses and peptide sequences.
}
