% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{pair_mgftheos}
\alias{pair_mgftheos}
\title{Pairs MGF queries to theoretical MS1 masses and peptide sequences.}
\usage{
pair_mgftheos(
  mgf_path,
  n_modules,
  ms1_offsets = 0,
  by_modules = TRUE,
  min_mass = 200L,
  max_mass = 4500L,
  ppm_ms1_bin = 10L,
  .path_bin,
  reframe_mgfs = FALSE,
  first_search = FALSE
)
}
\arguments{
\item{mgf_path}{The path to MGF files}

\item{n_modules}{The number of modules (\code{length(aa_masses_all)}) or one}

\item{ms1_offsets}{Off-sets in precursor masses (in relative to the values in
the original MGFs).}

\item{by_modules}{Logical; if TRUE, results are saved with one mgf to one
theo module. At FALSE, results are saved with one mgf paired to all theo
modules}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{ppm_ms1_bin}{The tolerance in precursor mass error after mass binning.}

\item{.path_bin}{The path to binned theoretical masses}

\item{reframe_mgfs}{Logical; if TRUE, recalculates the frame indexes of MGFs.}

\item{first_search}{Logical; is the first search (for MGF mass calibration)
or not.}
}
\description{
Pairs MGF queries to theoretical MS1 masses and peptide sequences.
}
