% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{parse_unimod_composition}
\alias{parse_unimod_composition}
\title{Parses A unimod position.}
\usage{
parse_unimod_composition(composition = "H(4) C O S")
}
\arguments{
\item{composition}{A chemical composition.}
}
\description{
Parses A unimod position.
}
