% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{pasefMS1xyz0}
\alias{pasefMS1xyz0}
\title{De-isotoping PASEF DDA-MS1}
\usage{
pasefMS1xyz0(
  msx_moverzs,
  msx_ints,
  ms1_fr,
  ms_level,
  orig_scan,
  iso_ctr = NULL,
  iso_lwr = NULL,
  iso_upr = NULL,
  slice_start = NULL,
  slice_end = NULL,
  mobility = NULL,
  ms1_moverzs = NULL,
  ms1_charges = NULL,
  ms1_ints = NULL,
  maxn_mdda_precurs = 1L,
  n_mdda_flanks = 0L,
  topn_ms2ions = 150L,
  max_ms1_charge = 4L,
  ppm_ms1_deisotope = 8L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  use_defpeaks = FALSE,
  min_mass = 115L,
  filename = NULL
)
}
\arguments{
\item{msx_moverzs}{Full-spectrum MS1 or MS2 m-over-z values.}

\item{msx_ints}{Full-spectrum MS1 or MS2 intensities.}

\item{ms1_fr}{MS1 frame numbers.}

\item{ms_level}{Vectors of MS levels.}

\item{orig_scan}{Original scan numbers.}

\item{iso_ctr}{A vector of isolation centers.}

\item{iso_lwr}{A vector of isolation lowers.}

\item{iso_upr}{A vector of isolation uppers.}

\item{slice_start}{A vector of indexes of slice starts.}

\item{slice_end}{A vector of indexes of slice ends.}

\item{mobility}{A vector of mobilities.}

\item{ms1_moverzs}{MS1 moverz values.}

\item{ms1_charges}{MS1 charge states.}

\item{ms1_ints}{MS1 intensity values.}

\item{maxn_mdda_precurs}{Maximum number of precursors for consideration in a
multi-precursor DDA scan. Note that at \code{maxn_mdda_precurs = 1}, it is
equivalent to DDA with precursor re-deisotoping. At \code{maxn_mdda_precurs
= 0}, it is equivalent to DDA using the original monoisotopic masses and
charge states provided by a peak-picking algorithm (e.g., MSConvert).}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{max_ms1_charge}{A positive integer; the maximum MS1 charge state for
considerations. The default is 6.}

\item{ppm_ms1_deisotope}{Mass error tolerance in MS1 deisotoping.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{A multiplication factor for the fuzzy discrimination of a
secondary precursor in parallel to a primary precursor in an isotopic
envelop. The smaller the value, the more stringent it is in calling an
adjacent peak being a mono-isotopic precursor}

\item{use_defpeaks}{Depreciated. Logical; if TRUE, uses MS1 m-over-z's,
intensities and charge states pre-calculated by other peak-picking
algorithms.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{filename}{A filename for logging.}
}
\description{
Depreciated; for PASEF with MS1 combining of all slices.
}
