% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{pcollapse_sortpeps}
\alias{pcollapse_sortpeps}
\title{Helper of \link{map_pepprot}.}
\usage{
pcollapse_sortpeps(Mat, ncol = NULL, peps = NULL, fct = 4L)
}
\arguments{
\item{Mat}{A sparse matrix.}

\item{ncol}{The number of columns in \code{mat}.}

\item{peps}{Peptide sequences as the row names of \code{mat}.}

\item{fct}{A factor for data splitting into chunks.}
}
\description{
Parallel version of \link{collapse_sortpeps}.
}
