% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolator.R
\name{perco_svm}
\alias{perco_svm}
\title{Percolator}
\usage{
perco_svm(
  prob_cos = NULL,
  out_path = NULL,
  df = NULL,
  target_fdr = 0.01,
  fdr_type = "protein",
  min_len = 7L,
  max_len = 40L,
  max_pepscores_co = 50,
  min_pepscores_co = 0,
  enzyme = "trypsin_p",
  fdr_group = "base",
  nes_fdr_group = "base",
  fct_score = 10,
  k = 10,
  cross_valid = FALSE,
  costs = c(0.1, 0.3, 1, 3, 10),
  def_cost = 1L,
  svm_kernel = "radial",
  svm_feats = c("pep_score", "pep_ret_range", "pep_delta", "pep_n_ms2", "pep_expect",
    "pep_exp_mz", "pep_exp_mr", "pep_tot_int", "pep_n_matches2", "pep_ms2_deltas_mean"),
  svm_iters = 10L,
  svm_tol = 1e-04,
  ...
)
}
\arguments{
\item{prob_cos}{Probability cot-offs (as a function of pep_len).}

\item{out_path}{A file path of outputs.}

\item{df}{A data frame of \code{psmC.txt}.}

\item{target_fdr}{A numeric; the targeted false-discovery rate (FDR) at the
levels of PSM, peptide or protein. The default is 0.01. See also argument
\code{fdr_type}.}

\item{fdr_type}{A character string; the type of FDR control. The value is in
  one of c("protein", "peptide", "psm"). The default is \code{protein}.

  Note that \code{fdr_type = protein} is comparable to \code{fdr_type =
  peptide} with the additional filtration of data at \code{prot_tier == 1}.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded. The default is 40.}

\item{max_pepscores_co}{A positive numeric; the upper limit in the cut-offs
of peptide scores for discriminating significant and insignificant
identities.}

\item{min_pepscores_co}{A non-negative numeric; the lower limit in the
cut-offs of peptide scores for discriminating significant and insignificant
identities.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from protein entries. The
default is \code{Trypsin_P}. See also parameter \code{custom_enzyme}.}

\item{fdr_group}{A character string; the modification group(s) for uses in
peptide FDR controls. The value is in one of \code{c("all", "base")}. The
\code{base} corresponds to the modification group with the largest number
of matches.}

\item{nes_fdr_group}{A character string in one of \code{c("all",
"all_cterm_tryptic", "all_cterm_nontryptic", "base", "base_cterm_tryptic",
"base_cterm_nontryptic")}. All peptides will be used in the classifications
of targets and decoys at \code{"all"}. Peptides with the chemistry of
C-terminal K or R will be used at \code{"all_cterm_tryptic"} (peptides from
protein C-terminals being excluded). Peptides without C-terminal K or R
will be used at \code{"all_cterm_nontryptic"}. The same applied to
\code{"base_cterm_tryptic"} and \code{"base_cterm_nontryptic"} with the
difference of only peptides from the \code{base} group being used. See also
parameter \code{fdr_group}.}

\item{fct_score}{The factor in converting probability p-values to scores. The
value is always 10.}

\item{k}{The k-folds for cross validation.}

\item{cross_valid}{Logical; to perform cross validations or not.}

\item{costs}{The costs for cross validations.}

\item{def_cost}{The default cost.}

\item{svm_kernel}{The SVM kernel. See also \link[e1071]{svm}.}

\item{svm_feats}{Features used for SVM classifications.}

\item{svm_iters}{The number of iterations.}

\item{svm_tol}{Tolerance in FDR.}

\item{...}{Not currently used.}
}
\description{
Percolator
}
