% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{plotMS2ions}
\alias{plotMS2ions}
\title{Plots matched MS2 ions}
\usage{
plotMS2ions(
  duos,
  out_path = "~",
  out_name = "bar.png",
  width = 12.5,
  height = 9
)
}
\arguments{
\item{duos}{The duo of MGF data and matched data.}

\item{out_path}{An output path.}

\item{out_name}{An output file name for saving the the plot of MS2 ions.}

\item{width}{Plot width.}

\item{height}{Plot height.}
}
\description{
Plots matched MS2 ions
}
