% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{post_calib}
\alias{post_calib}
\title{Post MS1 calibrations}
\usage{
post_calib(mgfs, min_mass, max_mass, mgf_path, filename, islist = FALSE)
}
\arguments{
\item{mgfs}{MGF data.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}

\item{filename}{An MGF file name.}

\item{islist}{Logical; are the ms1_mass values in list format or not.}
}
\description{
Post MS1 calibrations
}
