% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{post_ms2match}
\alias{post_ms2match}
\title{Post processing after ms2match.}
\usage{
post_ms2match(df, i, aa_masses, out_path)
}
\arguments{
\item{df}{An output from various ms2match(es).}

\item{i}{Integer; the index for a set of corresponding aa_masses and
theoretical peptides.}

\item{aa_masses}{An amino-acid mass lookup.}

\item{out_path}{A file path of outputs.}
}
\description{
Post processing after ms2match.
}
