% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{post_pepfdr}
\alias{post_pepfdr}
\title{Post processing after peptide FDR.}
\usage{
post_pepfdr(prob_cos = NULL, out_path = NULL)
}
\arguments{
\item{prob_cos}{Probability cut-offs (in data frame).}

\item{out_path}{An output path.}
}
\description{
Post processing after peptide FDR.
}
