% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{post_psmC2Q}
\alias{post_psmC2Q}
\title{Post \link{psmC2Q}.}
\usage{
post_psmC2Q(df, dfC, tier = NULL)
}
\arguments{
\item{df}{A data frame of protein-peptide map.}

\item{dfC}{A \code{psmQ} data with the removal of non-significant peptides
etc.}

\item{tier}{The tier of proteins in \code{df}.}
}
\description{
Post \link{psmC2Q}.
}
