% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{post_readmgf}
\alias{post_readmgf}
\title{Post-processing of MGF or mzML}
\usage{
post_readmgf(df, raw_id, mgf_path)
}
\arguments{
\item{df}{A data frame of processed peak lists.}

\item{raw_id}{An ID to replace the original RAW file name.}

\item{mgf_path}{A path to mzML or MGF files.}
}
\description{
Calculates mass \code{frame}s etc.
}
