% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{predCalibMS}
\alias{predCalibMS}
\title{Predictions of MS calibration}
\usage{
predCalibMS(col, lens, mgfs, fit, tr_min, tr_max)
}
\arguments{
\item{col}{The column key in mgfs.}

\item{lens}{The lengths of each entry in \code{col}.}

\item{mgfs}{A data frame of MGF data.}

\item{fit}{A fit model.}

\item{tr_min}{The min value in train data.}

\item{tr_max}{The max value in train data}
}
\description{
Predictions of MS calibration
}
