% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{predeisoDDA}
\alias{predeisoDDA}
\title{Helper of \link{deisoDDA}.}
\usage{
predeisoDDA(
  filename = NULL,
  temp_dir = NULL,
  mzml_type = "raw",
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  maxn_mdda_precurs = 5L,
  topn_ms2ions = 150L,
  n_mdda_flanks = 6L,
  n_dia_scans = 4L,
  min_mass = 200L,
  max_mass = 4500L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  min_ms1_charge = 2L,
  max_ms1_charge = 4L,
  min_ret_time = 0,
  max_ret_time = Inf,
  min_scan_num = 1L,
  max_scan_num = .Machine$integer.max,
  deisotope_ms2 = TRUE,
  max_ms2_charge = 3L,
  ppm_ms1_deisotope = 8L,
  ppm_ms2_deisotope = 8L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  quant = "none",
  use_lfq_intensity = TRUE,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  use_defpeaks = FALSE,
  is_pasef = FALSE,
  n_para = 1L,
  debug = FALSE
)
}
\arguments{
\item{dedug}{Logical; debug mode or not.}
}
\description{
Helper of \link{deisoDDA}.
}
