% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{prepBrukerMGF}
\alias{prepBrukerMGF}
\title{Preprocessing Bruker's MGF}
\usage{
prepBrukerMGF(file = NULL, begin_offset = 5L, charge_offset = 5L)
}
\arguments{
\item{file}{A file name}

\item{begin_offset}{The number of lines before a BEGIN line.}

\item{charge_offset}{The number lines after a BEGIN line to a following
CHARGE line.}
}
\description{
Some entries may have no CHARGE line
}
