% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{probco_bypeplen}
\alias{probco_bypeplen}
\title{Helper of \link{calc_pepfdr}.}
\usage{
probco_bypeplen(
  len,
  td,
  fdr_type = "protein",
  target_fdr = 0.01,
  max_pepscores_co = 50,
  min_pepscores_co = 0,
  out_path
)
}
\arguments{
\item{len}{Numeric; the length of peptides.}

\item{td}{A target-decoy pair.}

\item{fdr_type}{A character string; the type of FDR control. The value is in
  one of c("protein", "peptide", "psm"). The default is \code{protein}.

  Note that \code{fdr_type = protein} is comparable to \code{fdr_type =
  peptide} with the additional filtration of data at \code{prot_tier == 1}.}

\item{target_fdr}{A numeric; the targeted false-discovery rate (FDR) at the
levels of PSM, peptide or protein. The default is 0.01. See also argument
\code{fdr_type}.}

\item{max_pepscores_co}{A positive numeric; the upper limit in the cut-offs
of peptide scores for discriminating significant and insignificant
identities. Note that a probability p-value of, e.g., \eqn{1e-20} may not
be interpreted as more probable than \eqn{1e-10} (beyond the precision of a
probability test). Also note that the numeric limit when converting
probability to scores (\eqn{-10 \times log10(p)}); e.g. \eqn{-10 \times
log10(e-324)} yielded \code{Inf} instead of 3240. For reasons like these,
a high value of \code{max_pepscores_co} is not recommended.}

\item{min_pepscores_co}{A non-negative numeric; the lower limit in the
cut-offs of peptide scores for discriminating significant and insignificant
identities.}

\item{out_path}{A file path of outputs.}
}
\description{
Calculates the probability cut-off for target-decoy pairs at a given peptide
length.
}
