% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{proc_dda}
\alias{proc_dda}
\title{Proc mzML data for DDA workflows.}
\usage{
proc_dda(
  spec,
  raw_file,
  idx_sc = 3L,
  idx_osc = 3L,
  idx_mslev = 2L,
  idx_title = 10L,
  idx_scanList_2 = 11L,
  idx_rt_2 = 2L,
  idx_scan_start_2 = 1L,
  idx_precursor_2 = 12L,
  idx_isolationWindow = 1L,
  idx_selectedIonList = 2L,
  idx_moverz = 1L,
  idx_charge = 2L,
  idx_ms1int = 3L,
  idx_bin_2 = 13L,
  deisotope_ms2 = TRUE,
  max_ms2_charge = 3L,
  ppm_ms2_deisotope = 10L,
  topn_ms2ions = 100L,
  quant = "none",
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  index_mgf_ms2 = FALSE
)
}
\arguments{
\item{spec}{Spectrum list.}

\item{raw_file}{The RAW file name of an mzML.}

\item{idx_sc}{Index of scan numbers.}

\item{idx_osc}{Index of original scan numbers.}

\item{idx_mslev}{Index of MS levels.}

\item{idx_title}{Index of scan titles.}

\item{idx_scanList_2}{Index of MS2 scanList.}

\item{idx_rt_2}{Index of MS2 retention times.}

\item{idx_scan_start_2}{Index of MS2 scan start.}

\item{idx_precursor_2}{Index of MS2 precursor.}

\item{idx_isolationWindow}{Index of isolationWindow.}

\item{idx_selectedIonList}{Index of selectedIonList.}

\item{idx_moverz}{Index of m-over-z sequences.}

\item{idx_charge}{Index of charge states.}

\item{idx_ms1int}{Index of MS1 intensities.}

\item{idx_bin_2}{Index of MS2 binary data.}

\item{deisotope_ms2}{Logical; to deisotope MS2 features or not.}

\item{max_ms2_charge}{Maximum charge states for consideration with MS2
deisotoping.}

\item{ppm_ms2_deisotope}{Mass error tolerance in MS2 deisotoping.}

\item{topn_ms2ions}{Top-n MS2 ions for consideration.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{index_mgf_ms2}{Depreciated. A low-priority feature. Logical; if TRUE,
  converts up-frontly MS2 m-over-z values from numeric to integers as opposed
  to \emph{on-the-fly} conversion during ion matches. The default is FALSE.
  The \code{index_mgf_ms2 = TRUE} might be useful for very large MS files by
  reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}
}
\description{
Proc mzML data for DDA workflows.
}
