% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{proc_mdda}
\alias{proc_mdda}
\title{Proc mzML data for mDDA workflows.}
\usage{
proc_mdda(
  spec,
  raw_file,
  idx_sc = 3L,
  idx_osc = 3L,
  idx_mslev = 2L,
  idx_title = 10L,
  idx_scanList_2 = 11L,
  idx_rt_2 = 2L,
  idx_scan_start_2 = 1L,
  idx_precursor_2 = 12L,
  idx_isolationWindow = 1L,
  idx_ctrmz = 1L,
  idx_lwrmz = 2L,
  idx_uprmz = 3L,
  idx_selectedIonList = 2L,
  idx_moverz = 1L,
  idx_charge = 2L,
  idx_ms1int = 3L,
  idx_bin_2 = 13L,
  maxn_precurs = 5L,
  max_ms1_charge = 4L,
  topn_ms2ions = 300L,
  idx_scanList_1 = 11L,
  idx_rt_1 = 2L,
  idx_scan_start_1 = 1L,
  idx_bin_1 = 12L,
  deisotope_ms2 = TRUE,
  ppm_ms1_deisotope = 10L,
  ppm_ms2_deisotope = 10L,
  n_mdda_flanks = 6L,
  use_defpeaks = FALSE
)
}
\arguments{
\item{spec}{Spectrum list.}

\item{raw_file}{The RAW file name of an mzML.}

\item{idx_sc}{Index of scan numbers.}

\item{idx_osc}{Index of original scan numbers.}

\item{idx_mslev}{Index of MS levels.}

\item{idx_title}{Index of scan titles.}

\item{idx_scanList_2}{Index of MS2 scanList.}

\item{idx_rt_2}{Index of MS2 retention times.}

\item{idx_scan_start_2}{Index of MS2 scan start.}

\item{idx_precursor_2}{Index of MS2 precursor.}

\item{idx_isolationWindow}{Index of isolationWindow.}

\item{idx_ctrmz}{Index of center m-over-z.}

\item{idx_lwrmz}{Index of lower m-over-z.}

\item{idx_uprmz}{Index of upper m-over-z.}

\item{idx_selectedIonList}{Index of selectedIonList.}

\item{idx_moverz}{Index of m-over-z sequences.}

\item{idx_charge}{Index of charge states.}

\item{idx_ms1int}{Index of MS1 intensities.}

\item{idx_bin_2}{Index of MS2 binary data.}

\item{maxn_precurs}{Maximum number of chimeric precursors.}

\item{max_ms1_charge}{Maximum precursor charge states.}

\item{topn_ms2ions}{Top-n MS2 ions for consideration.}

\item{idx_scanList_1}{Index of MS1 scanList.}

\item{idx_rt_1}{Index of precursor retention times.}

\item{idx_scan_start_1}{Index of MS1 scan starts.}

\item{idx_bin_1}{Index of MS1 binary data.}

\item{deisotope_ms2}{Logical; to deisotope MS2 features or not.}

\item{ppm_ms1_deisotope}{Mass error tolerance in MS1 deisotoping.}

\item{ppm_ms2_deisotope}{Mass error tolerance in MS2 deisotoping.}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{use_defpeaks}{Logical; if TRUE, uses MS1 m-over-z's, intensities and
charge states pre-calculated by other peak-picking algorithms.}
}
\description{
Proc mzML data for mDDA workflows.
}
