% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{proc_mgfs}
\alias{proc_mgfs}
\title{Helper in processing MGF entries in chunks.}
\usage{
proc_mgfs(
  lines,
  topn_ms2ions = 100L,
  ms1_charge_range = c(2L, 6L),
  ms1_scan_range = c(1L, .Machine$integer.max),
  ret_range = c(0, Inf),
  min_mass = 700L,
  max_mass = 4500L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  type_mgf = "msconv_thermo",
  n_bf_begin = 0L,
  n_spacer = 0L,
  n_hdr = 5L,
  n_to_pepmass = 3L,
  n_to_title = 1L,
  n_to_scan = 0L,
  n_to_rt = 2L,
  n_to_charge = 4L,
  sep_ms2s = " ",
  nfields_ms2s = 2L,
  sep_pepmass = " ",
  nfields_pepmass = 2L,
  raw_file = NULL,
  digits = 5L
)
}
\arguments{
\item{lines}{Lines of MGF.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ms1_charge_range}{The range of MS1 charge states.}

\item{ms1_scan_range}{The range of MS1 scan numbers.}

\item{ret_range}{The range of retention time in seconds.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoSeqs}.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation. The default is 4500.}

\item{mgf_cutmzs}{Cut points of MS1 moverz values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{type_mgf}{The type of MGF format.}

\item{n_bf_begin}{The number of lines before \code{BEGIN IONS}. Zero for PD
and MSConvert.}

\item{n_spacer}{The number of spacer lines between the preceding line END
IONS and the following line BEGIN IONS. The value is 1 for Proteome
Discoverer and 0 for MSConvert.}

\item{n_hdr}{The number of lines before MS2 data in an MGF. The value is +6
for PD and +5 for MSConvert.}

\item{n_to_pepmass}{The number of lines from BEGIN to PEPMASS.}

\item{n_to_title}{The number of lines from BEGIN to TITLE. The value is the
same between PD and MSConvert.}

\item{n_to_scan}{The number of lines from BEGIN to SCANS. The value is +5 for
PD.}

\item{n_to_rt}{The number of lines from BEGIN to RTINSECONDS.}

\item{n_to_charge}{The number of lines from BEGIN to CHARGE.}

\item{sep_ms2s}{The separation character between MS2 m/z and intensity
values.}

\item{nfields_ms2s}{The number of fields in MS2 entries. Mostely two and can
be three for some Bruker MGFs.}

\item{sep_pepmass}{The separation character between MS1 m/z and intensity
values.}

\item{nfields_pepmass}{The number of fields in \code{PEPMASS}.}

\item{raw_file}{The raw file name. Is NULL for PD and MSConvert.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper in processing MGF entries in chunks.
}
