% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{proc_mzml}
\alias{proc_mzml}
\title{Helper of \link{readmzML}}
\usage{
proc_mzml(
  file,
  ms1_charge_range = c(2L, 6L),
  ret_range = c(0, Inf),
  min_mass = 700L,
  max_mass = 4500L,
  topn_ms2ions = 100L,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  min_ms2mass = 115L,
  max_ms2mass = 4500L
)
}
\arguments{
\item{file}{A file name to mzML with a prepending path.}

\item{ms1_charge_range}{The range of MS1 charge states.}

\item{ret_range}{The range of retention time in seconds.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoSeqs}.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{mgf_cutmzs}{Cut points of MS1 moverz values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation. The default is 4500.}
}
\description{
Helper of \link{readmzML}
}
