% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{proc_pasefs}
\alias{proc_pasefs}
\title{Helper in processing MGF entries in chunks.}
\usage{
proc_pasefs(raw_file, mgf_path, temp_dir, topn_ms2ions = 150L)
}
\arguments{
\item{raw_file}{The file name of RAW MS data.}

\item{mgf_path}{The file path of MGF files.}

\item{temp_dir}{A file path for temporary files.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}
}
\description{
Helper in processing MGF entries in chunks.
}
