% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfilereader.R
\name{proc_raws}
\alias{proc_raws}
\title{Helper in processing MGF entries in chunks.}
\usage{
proc_raws(raw_file, mgf_path, temp_dir)
}
\arguments{
\item{raw_file}{The file name of RAW MS data.}

\item{mgf_path}{The file path of MGF files.}

\item{temp_dir}{A file path for temporary files.}
}
\description{
Helper in processing MGF entries in chunks.
}
