% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{prot_pep_tbl}
\alias{prot_pep_tbl}
\title{A lookup table between prot_acc and pep_seqs.}
\usage{
prot_pep_tbl(seqs)
}
\arguments{
\item{seqs}{Results from \link{ms1masses_bare}.}
}
\description{
Not currently used.
}
\examples{
\donttest{
prot_pep_tbl(fwd_peps)
}
}
