% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoM.R
\docType{package}
\name{proteoM}
\alias{proteoM}
\title{proteoM: A package for database searches of mass spectrometry data.}
\description{
The proteoM package provides utilities for (1) database searches. The
companion package proteoQ provides utilities for (2) data preprocessing, (3)
annotation, (4) quality assurance assessments and (5) informatic analysis.
}
\section{Main utility in database searches}{
 matchMS (additional tool kits:
  load_fasta2, parse_unimod, calc_monopeptide, calc_ms2ionseries)
}

