% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{psmC2Q}
\alias{psmC2Q}
\title{From \code{psmC.txt} to \code{psmQ.txt}.}
\usage{
psmC2Q(
  df = NULL,
  out_path = NULL,
  fdr_type = "protein",
  combine_tier_three = FALSE,
  max_n_prots = 60000L,
  fct = 4L
)
}
\arguments{
\item{df}{A result of \code{psmC.txt} with the removals of non-significant
or decoy peptides, as well as decoy proteins.}

\item{out_path}{A file path of outputs.}

\item{fdr_type}{A character string; the type of FDR control. The value is in
  one of c("protein", "peptide", "psm"). The default is \code{protein}.

  Note that \code{fdr_type = protein} is comparable to \code{fdr_type =
  peptide} with the additional filtration of data at \code{prot_tier == 1}.}

\item{combine_tier_three}{Logical; if TRUE, combines search results at tier-3
  to tier-1 to form the single output of \code{psmQ.txt}. The default is
  FALSE in that data will be segregated into the three quality tiers (shown
  below) by the choice of \code{fdr_type}. Note that the argument affects
  only at the \code{fdr_type} of \code{psm} or \code{peptide} where there are
  no tier-2 outputs. In general, the tier-3 results correspond to
  one-hit-wonders and setting \code{combine_tier_three = TRUE} is
  discouraged.

  In subproteome analysis, such as phosphoproteome analysis, some proteins
  may be well established globally, but fail the significance assessment by
  protein FDR on the local scale. In situations like this, it may be suitable
  to apply \code{fdr_type = "peptide"} or \code{fdr_type = "psm"} other than
  incurring \code{combine_tier_three = TRUE}.

  Tier-1: both proteins and peptides with scores above significance
  thresholds.

  Tier-2: \eqn{\ge} 2 significant peptides but protein scores below
  significance thresholds.

  Tier-3: one significant peptide per protein and protein scores below
  significance thresholds.}

\item{max_n_prots}{Softly depreciated. A positive integer to threshold the
  maximum number of protein entries before coercing \code{fdr_type} from
  \code{psm} or \code{peptide} to \code{protein}. The argument has no effect
  if \code{fdr_type} is already \code{protein}. In general, there is no need
  to change the default.

  Note that for memory efficiency proteins at tiers 1, 2 and 3 are grouped
  separately. Further note that there is no tier-2 proteins at
  \code{fdr_type} of \code{psm} or \code{peptide}. For very large data sets,
  a lower value of \code{max_n_prots} can be used to reduce the chance of
  memory exhaustion by setting aside some protein entries from tier 1 to 2.}

\item{fct}{A factor for data splitting into chunks. May consider a greater
value for a larger data set.}
}
\description{
Non-significant and decoy peptides should have been removed from the input
\code{df}, as well as decoy proteins.
}
