% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{purge_decoys}
\alias{purge_decoys}
\title{Purges decoy pep_seq(s) that are also found in target fasta.}
\usage{
purge_decoys(target, decoy)
}
\arguments{
\item{target}{A target data frame with column \code{pep_seq}.}

\item{decoy}{A decoy data frame with column \code{pep_seq}.}
}
\description{
Decoy sequences may be present in targets and thus removed.
}
\examples{
## found in both forward and reverse fastas
# pep_seq  prot_acc
# RQEEELR NP_064522
# RQEEELR NP_997555
# RQEEELR NP_997553
# 
# pep_seq      prot_acc
# RQEEELR -NP_001073379
# RQEEELR -NP_001157031
# RQEEELR    -NP_796085
# RQEEELR    -NP_083410

}
