% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{purge_iso_info}
\alias{purge_iso_info}
\title{Removes \code{iso_info} rows not found in MS data.}
\usage{
purge_iso_info(msdata, iso_info)
}
\arguments{
\item{msdata}{MS data.}

\item{iso_info}{A subset of metadata paired with \code{msdata}.}
}
\description{
Removes \code{iso_info} rows not found in MS data.
}
