% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{purge_search_space}
\alias{purge_search_space}
\title{Subsets the search space.}
\usage{
purge_search_space(
  i,
  aa_masses,
  mgf_path,
  n_cores,
  ppm_ms1 = 20L,
  fmods_nl = NULL
)
}
\arguments{
\item{i}{Integer; the index for a set of corresponding aa_masses and
theoretical peptides.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required (don't think it
  can be altered by users, but just in case).}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}
}
\description{
Subsets the search space.
}
