% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{quick_join}
\alias{quick_join}
\title{Joining of two data frames.}
\usage{
quick_join(x, y, by = NULL, type = "left")
}
\arguments{
\item{x}{The left data frame.}

\item{y}{The right data frame.}

\item{by}{The key.}

\item{type}{Left or right (joining).}
}
\description{
Not a general purpose utility. For simple circumstances that the number of
rows in the output is equal to that in \code{x}, and can be use as an
alternative to \link[dplyr]{left_join}.
}
\examples{
\donttest{
df1 <- data.frame(A = c("a", "b", "c"), B = c(1, 1, 1))
df2 <- data.frame(A = c("a", "c", "d"), C = c(2, 2, "3"))

x1 <- mzion:::quick_join(df1, df2, by = "A")
x1 <- x1[, c("A", "B", "C")]
rownames(x1) <- seq_len(nrow(x1))

x2 <- dplyr::left_join(df1, df2, by = "A")

stopifnot(identical(x1, x2))

df1 <- data.frame(A = c("a", "b", "c"), B = c(1, 1, 1))
df2 <- data.frame(A = c("a", "c", "d"), C = c(2, 2, "3"))


x1 <- mzion:::quick_join(df1, df2, by = "A", type = "right")
x1 <- x1[, c("A", "B", "C")]
rownames(x1) <- seq_len(nrow(x1))

x2 <- dplyr::right_join(df1, df2, by = "A")
x2 <- x2[, c("A", "B", "C")]

# dplyr::right_join with attr(,".match.hash")
x2$A <- as.character(x2$A)
mapply(identical, x1, x2)

# row order may be different
x1 <- mzion:::quick_join(df2, df1, by = "A", type = "right")
x1 <- x1[, c("A", "B", "C")]
rownames(x1) <- seq_len(nrow(x1))

x2 <- dplyr::right_join(df2, df1, by = "A")
x2 <- x2[, c("A", "B", "C")]

# FALSE
!identical(x1, x2)
}
}
