% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{read_mgf_chunks}
\alias{read_mgf_chunks}
\title{Reads MGFs in chunks.}
\usage{
read_mgf_chunks(
  filepath,
  temp_dir,
  raw_id = 1L,
  topn_ms2ions = 150L,
  min_ms1_charge = 2L,
  max_ms1_charge = 4L,
  min_scan_num = 1L,
  max_scan_num = .Machine$integer.max,
  min_ret_time = 0,
  max_ret_time = Inf,
  min_mass = 200L,
  max_mass = 4500L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  type_mgf = "msconv_thermo",
  n_bf_begin = 0L,
  n_spacer = 0L,
  n_hdr = 5L,
  n_to_pepmass = 3L,
  n_to_title = 1L,
  n_to_scan = 0L,
  n_to_rt = 2L,
  n_to_charge = 4L,
  sep_ms2s = " ",
  nfields_ms2s = 2L,
  sep_pepmass = " ",
  nfields_pepmass = 2L,
  raw_file = NULL,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  deisotope_ms2 = TRUE,
  max_ms2_charge = 3L,
  maxn_dia_precurs = 1000L,
  use_defpeaks = FALSE,
  maxn_mdda_precurs = 5L,
  n_mdda_flanks = 6L,
  ppm_ms1_deisotope = 10L,
  ppm_ms2_deisotope = 10L,
  quant = "none",
  digits = 4L
)
}
\arguments{
\item{filepath}{The file path to a list of MGF or mzML files.}

\item{temp_dir}{A temporary path of MGFs.}

\item{raw_id}{An ID to RAW file name.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{min_ms1_charge}{A positive integer; the minimum MS1 charge state for
considerations. The default is 2.}

\item{max_ms1_charge}{A positive integer; the maximum MS1 charge state for
considerations. The default is 6.}

\item{min_scan_num}{Depreciated. A positive integer; the minimum scan number
for considerations. The default is 1. The setting only applies to MGFs with
numeric scan numbers. For example, it has no effects on Bruker's timsTOF
data.}

\item{max_scan_num}{Depreciated. A positive integer; the maximum scan number
for considerations. The default is the maximum machine integer. The setting
only applies to MGFs with numeric scan numbers.}

\item{min_ret_time}{A non-negative numeric; the minimum retention time in
seconds for considerations. The default is 0.}

\item{max_ret_time}{Numeric; the maximum retention time in seconds for
considerations. The default is \code{Inf}. At a negative value, e.g. -800,
data in the last 800s of LC retention time will be excluded.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoSeqs}.}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{mgf_cutmzs}{Cut points of MS1 m-over-z values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{type_mgf}{The type of MGF format.}

\item{n_bf_begin}{The number of lines before \code{BEGIN IONS}. Zero for PD
and MSConvert.}

\item{n_spacer}{The number of spacer lines between the preceding line END
IONS and the following line BEGIN IONS. The value is 1 for Proteome
Discoverer and 0 for MSConvert.}

\item{n_hdr}{The number of lines before MS2 data in an MGF. The value is +6
for PD and +5 for MSConvert.}

\item{n_to_pepmass}{The number of lines from BEGIN to PEPMASS.}

\item{n_to_title}{The number of lines from BEGIN to TITLE. The value is the
same between PD and MSConvert.}

\item{n_to_scan}{The number of lines from BEGIN to SCANS. The value is +5 for
PD.}

\item{n_to_rt}{The number of lines from BEGIN to RTINSECONDS.}

\item{n_to_charge}{The number of lines from BEGIN to CHARGE.}

\item{sep_ms2s}{The separation character between MS2 m/z and intensity
values.}

\item{nfields_ms2s}{The number of fields in MS2 entries. Mostely two and can
be three for some Bruker MGFs.}

\item{sep_pepmass}{The separation character between MS1 m/z and intensity
values.}

\item{nfields_pepmass}{The number of fields in \code{PEPMASS}.}

\item{raw_file}{The raw file name. Is NULL for PD and MSConvert.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{deisotope_ms2}{Logical; if TRUE, de-isotope MS2 features.}

\item{max_ms2_charge}{Maximum charge states for consideration with MS2
deisotoping.}

\item{maxn_dia_precurs}{Maximum number of DIA precursors.}

\item{use_defpeaks}{Depreciated. Logical; if TRUE, uses MS1 m-over-z's,
intensities and charge states pre-calculated by other peak-picking
algorithms.}

\item{maxn_mdda_precurs}{Maximum number of precursors for consideration in a
multi-precursor DDA scan. Note that at \code{maxn_mdda_precurs = 1}, it is
equivalent to DDA with precursor re-deisotoping. At \code{maxn_mdda_precurs
= 0}, it is equivalent to DDA using the original monoisotopic masses and
charge states provided by a peak-picking algorithm (e.g., MSConvert).}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{ppm_ms1_deisotope}{Mass error tolerance in MS1 deisotoping.}

\item{ppm_ms2_deisotope}{Mass error tolerance in MS2 deisotoping.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Reads MGFs in chunks.
}
