% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{read_mzml}
\alias{read_mzml}
\title{Reads mzML from MSConvert}
\usage{
read_mzml(
  xml_file,
  topn_ms2ions = 100L,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  max_ms1_charge = 4L,
  deisotope_ms2 = TRUE,
  max_ms2_charge = 3L,
  use_defpeaks = FALSE,
  maxn_dia_precurs = 300L,
  maxn_mdda_precurs = 5L,
  n_mdda_flanks = 6L,
  ppm_ms1_deisotope = 10L,
  ppm_ms2_deisotope = 10L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  quant = "none",
  digits = 4L
)
}
\arguments{
\item{xml_file}{A file name of mzML.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation.}

\item{max_ms1_charge}{A positive integer; the maximum MS1 charge state for
considerations. The default is 6.}

\item{deisotope_ms2}{Logical; if TRUE, de-isotope MS2 features.}

\item{max_ms2_charge}{Maximum charge states for consideration with MS2
deisotoping.}

\item{use_defpeaks}{Logical; if TRUE, uses MS1 m-over-z's, intensities and
charge states pre-calculated by other peak-picking algorithms.}

\item{maxn_dia_precurs}{Maximum number of precursors for consideration in a
DIA scan.}

\item{maxn_mdda_precurs}{Maximum number of precursors for consideration in a
multi-precursor DDA scan. Note that at \code{maxn_mdda_precurs = 1}, it is
equivalent to DDA with precursor re-deisotoping. At \code{maxn_mdda_precurs
= 0}, it is equivalent to DDA using the original monoisotopic masses and
charge states provided by a peak-picking algorithm (e.g., MSConvert).}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{ppm_ms1_deisotope}{Mass error tolerance in MS1 deisotoping.}

\item{ppm_ms2_deisotope}{Mass error tolerance in MS2 deisotoping.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{A multiplication factor for the fuzzy discrimination of a
secondary precursor in parallel to a primary precursor in an isotopic
envelop. The smaller the value, the more stringent it is in calling an
adjacent peak being a mono-isotopic precursor}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Only for Thermo's RAW format.
}
