% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{readlineMGFs}
\alias{readlineMGFs}
\title{Helper of \link{readMGF}.}
\usage{
readlineMGFs(i, file, filepath, raw_file)
}
\arguments{
\item{i}{An index of the i-th file.}

\item{file}{An MGF file name.}

\item{filepath}{The file path to a list of MGF or mzML files.}

\item{raw_file}{The raw file name. Is NULL for PD and MSConvert.}
}
\value{
Updated raw_file (for Bruker's timsTOF). Otherwise, raw_file remains
  NULL.
}
\description{
Helper of \link{readMGF}.
}
