% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{readMGF}
\alias{readMGF}
\title{Reads MGF files in chunks.}
\usage{
readMGF(
  filepath = "~/proteoM/mgf",
  min_mass = 700L,
  max_mass = 4500L,
  min_ms2mass = 115L,
  topn_ms2ions = 100L,
  ms1_charge_range = c(2L, 6L),
  ms1_scan_range = c(1L, .Machine$integer.max),
  ret_range = c(0, Inf),
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  index_ms2 = FALSE,
  out_path = file.path(filepath, "mgf_queries.rds")
)
}
\arguments{
\item{filepath}{The file path to a list of mgf files.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoSeqs}.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ms1_charge_range}{The range of MS1 charge states.}

\item{ms1_scan_range}{The range of MS1 scan numbers.}

\item{ret_range}{The range of retention time in seconds.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 25.}

\item{index_ms2}{Logical; if TRUE, converts MS2 m/z values to indices.}

\item{out_path}{An output path.}
}
\description{
Reads MGF files in chunks.
}
\examples{
\donttest{
mgf_queries <- proteoM:::readMGF()
}
}
