% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{readPASEF}
\alias{readPASEF}
\title{Reads Bruker's PASEF files.}
\usage{
readPASEF(mgf_path = NULL, filelist = NULL, topn_ms2ions = 150L)
}
\arguments{
\item{mgf_path}{A file path RAW MS files.}

\item{filelist}{A list of RAW MS files.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}
}
\description{
Reads Bruker's PASEF files.
}
