% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfilereader.R
\name{readRAW}
\alias{readRAW}
\title{Reads Thermo RAW files.}
\usage{
readRAW(mgf_path = NULL, filelist = NULL)
}
\arguments{
\item{mgf_path}{A file path RAW MS files.}

\item{filelist}{A list of RAW MS files.}
}
\description{
Reads Thermo RAW files.
}
