% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{remove_unimod_title}
\alias{remove_unimod_title}
\title{Removes an existing modification.}
\usage{
remove_unimod_title(title = NULL)
}
\arguments{
\item{title}{The title of a modification.}
}
\description{
Removes an existing modification.
}
\examples{
\dontrun{
library(proteoM)
x <- remove_unimod_title("Oxi+Carbamidomethyl")
}
}
