% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{reproc_psmC}
\alias{reproc_psmC}
\title{Reprocessing of \code{psmC.txt}.}
\usage{
reproc_psmC(out_path = NULL, fdr_type = "protein", combine_tier_three = FALSE)
}
\arguments{
\item{out_path}{A file path of outputs.}

\item{fdr_type}{Character string; the type of FDR control. The value is in
one of c("psm", "peptide", "protein"). Note that \code{fdr_type = protein}
is equivalent to \code{fdr_type = peptide} with the additional filtration
of data at \code{prot_tier == 1}. A variant is to set \code{fdr_type =
psm}, followed by a data filtration at \code{prot_tier == 1}.}

\item{combine_tier_three}{Logical; if TRUE, combines search results at tiers
  1, 2 and 3 to the single output of \code{psmQ.txt}. The default is FALSE in
  that data will be segregated into the three quality tiers according to the
  choice of \code{fdr_type}. The (convenience) parameter matters since
  \href{http://github.com/qzhang503/proteoQ}{proteoQ} will only look for the
  inputs of \code{psmQ[...].txt}.

  For instance, if the aim is to bypass the constraint by protein FDR and
  focus on PSMs that have met the cut-offs specified by \code{target_fdr},
  experimenters may set \code{combine_tier_three = TRUE} and hence pool all
  significant peptides in \code{psmQ.txt} for downstream proteoQ.

  Tier-1: both proteins and peptides with scores above significance
  thresholds.

  Tier-2: \eqn{\ge} 2 significant peptides but protein scores below
  significance thresholds.

  Tier-3: one significant peptide and protein scores below significance
  thresholds.}
}
\description{
Protein grouping from \code{psmC.txt} to \code{psmQ.txt}.
}
\details{
May solve some memory shortage issues for large data sets (e.g., over a
million peptide sequences * 35000 proteins from \code{psmC.txt}).
}
