% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{reset_rettimes}
\alias{reset_rettimes}
\title{Resets the range of retention times by experimental observations.}
\usage{
reset_rettimes(ret_times, min_ret_time = 0, max_ret_time = Inf)
}
\arguments{
\item{ret_times}{A vector of experimental retention times.}

\item{min_ret_time}{A non-negative numeric; the minimum retention time in
seconds for considerations. The default is 0.}

\item{max_ret_time}{Numeric; the maximum retention time in seconds for
considerations. The default is \code{Inf}. At a negative value, e.g. -800,
data in the last 800s of LC retention time will be excluded.}
}
\description{
Needed since \code{max_ret_time} can be negative.
}
