% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{reverse_peps_in_frame}
\alias{reverse_peps_in_frame}
\title{Helper of \link{reverse_seqs}.}
\usage{
reverse_peps_in_frame(pep_frame)
}
\arguments{
\item{pep_frame}{A frame of data.}
}
\description{
Reverses `pep_seq` in a frame.
}
\examples{
## pep_frame
#                                  pep_seq      mass  frame    prot_acc
# 2148391     CSCNNGEMCDRFQGCLCSPGWQGLQCER 3694.4923 100001   TIE2_HUMAN
# 2148392 EGSARASEQPENAESPDNEDGDCEETTEEAGR 3694.5248 100001  TXLNB_HUMAN
}
