% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{rm_char_in_nfirst}
\alias{rm_char_in_nfirst}
\title{Removes a starting character from the first \code{n} entries.}
\usage{
rm_char_in_nfirst(x, char = "-", n = (max_miss + 1L) * 2L, max_len = 40L)
}
\arguments{
\item{x}{A list of character strings. Peptide sequences in names and masses
in values.}

\item{char}{A starting character to be removed.}

\item{n}{The number of beginning entries to be considered.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded.}
}
\description{
Removes a starting character from the first \code{n} entries.
}
