% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{s_readRDS}
\alias{s_readRDS}
\title{Helper of \link{binTheoPeps_bysets}.}
\usage{
s_readRDS(file, out_path)
}
\arguments{
\item{file}{A file name.}

\item{out_path}{An output path.}
}
\description{
Reads single rds.
}
