% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{sadd_ion_matches}
\alias{sadd_ion_matches}
\title{Helper of a single row of \link{add_ion_matches}.}
\usage{
sadd_ion_matches(df)
}
\arguments{
\item{df}{A data frame of psmC.txt subset at a given pep_grp.}
}
\description{
Helper of a single row of \link{add_ion_matches}.
}
