% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{sadd_ion_mathces}
\alias{sadd_ion_mathces}
\title{Helper of a single row of \link{add_ion_mathces}.}
\usage{
sadd_ion_mathces(df)
}
\arguments{
\item{df}{A data frame of psmC.txt subset at a given pep_grp.}
}
\description{
Helper of a single row of \link{add_ion_mathces}.
}
