% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{save_mod_indexes}
\alias{save_mod_indexes}
\title{Saves mod_indexes.txt}
\usage{
save_mod_indexes(out_path = NULL, fixedmods, varmods, f_to_v)
}
\arguments{
\item{out_path}{An output path}

\item{fixedmods}{Fixed modifications}

\item{varmods}{Variable modifications}

\item{f_to_v}{Coerced fixed to variable modifications}
}
\description{
Saves mod_indexes.txt
}
