% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{save_ms1calib}
\alias{save_ms1calib}
\title{Saves the \code{ppm_ms1} before and after calibration.}
\usage{
save_ms1calib(ppm_ms1, ppm_ms1calib, mgf_path)
}
\arguments{
\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms1calib}{The mass error after calibration in ppm.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}
}
\description{
Saves the \code{ppm_ms1} before and after calibration.
}
