% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scalc_pepprobs}
\alias{scalc_pepprobs}
\title{Calculates peptide scores by single MGF entries (scan numbers).}
\usage{
scalc_pepprobs(
  entry,
  topn_ms2ions = 150L,
  type_ms2ions = "by",
  ppm_ms2 = 20L,
  minn_ms2 = 6L,
  soft_secions = FALSE,
  max_len = 40L,
  n_ms2_bg = 20000L,
  min_ms2mass = 115L,
  d2 = 1e-05,
  tally_ms2ints = TRUE,
  digits = 4L
)
}
\arguments{
\item{entry}{A row of data.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{minn_ms2}{A positive integer; the minimum number of matched MS2 ions
for consideration as a hit. Counts of secondary ions, e.g. b0, b* etc., are
not part of the threshold.}

\item{soft_secions}{Logical; if TRUE, collapses the intensities of secondary
ions to primary ions even when the primaries are absent. The default is
FALSE. For instance, the signal of \code{b5^*} will be ignored if its
primary ion \code{b5} is not matched. The impacts of \code{soft_secions =
TRUE} on search performance has not yet been assessed.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded. The default is 40.}

\item{n_ms2_bg}{The count of MS2 background noises.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{d2}{Bin width in ppm divided by 1E6.}

\item{tally_ms2ints}{Logical; tally MS2 intensities or not.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Each entry corresponds to a row in \code{ion_matches.rds}.
}
