% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{sep_df1_byRTs}
\alias{sep_df1_byRTs}
\title{Separate MS1 data}
\usage{
sep_df1_byRTs(df1, col_rt = "ret_time")
}
\arguments{
\item{df1}{A data frame of MS1.}

\item{col_rt}{The key of retention time column.}

\item{gap}{A gap size for forward and backward looking of precursors.}
}
\description{
Separate MS1 data
}
