% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{sep_df1_byRTs_v0}
\alias{sep_df1_byRTs_v0}
\title{Separate MS1 data}
\usage{
sep_df1_byRTs_v0(df1, col_rt = "ret_time", gap = 120, n_chunks = 4L)
}
\arguments{
\item{df1}{A data frame of MS1.}

\item{col_rt}{The key of retention time column.}

\item{gap}{A gap size for forward and backward looking of precursors.}

\item{n_chunks}{The number of chunks.}
}
\description{
Depreciated.
}
