% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{sep_pasef_ms2info}
\alias{sep_pasef_ms2info}
\title{Separates PASEF MS2Info}
\usage{
sep_pasef_ms2info(data, iso_info)
}
\arguments{
\item{data}{Line data of MS frames (containing both MS1 and MS2).}

\item{iso_info}{Data frame of ms2info.}
}
\description{
Separates PASEF MS2Info
}
