% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{set_bin_ncores}
\alias{set_bin_ncores}
\title{Sets the number of CPU cores for precursor mass binning.}
\usage{
set_bin_ncores(len_m, enzyme)
}
\arguments{
\item{len_m}{The number of \code{aa_masses} modules.}

\item{enzyme}{The assume enzymatic activity.}
}
\description{
Sets the number of CPU cores for precursor mass binning.
}
