% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{sim_combn}
\alias{sim_combn}
\title{From \link[utils]{combn}.}
\usage{
sim_combn(x, m)
}
\arguments{
\item{x}{Integer; the vector source for combinations.}

\item{m}{Integer; the number of elements to choose.}
}
\description{
To avoid the conversion, e.g., \code{3} to \code{1:3}.
}
