% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{sim_outer}
\alias{sim_outer}
\title{From \link[base]{outer}.}
\usage{
sim_outer(X, Y, FUN = "*", ...)
}
\arguments{
\item{X}{A numeric vector.}

\item{Y}{A numeric vector.}

\item{FUN}{A function to be used for the outer products.}

\item{...}{Optional arguments for \code{FUN}.}
}
\description{
From \link[base]{outer}.
}
