% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{simple_prots_peps}
\alias{simple_prots_peps}
\title{A lookup between prot_acc and pep_seqs.}
\usage{
simple_prots_peps(seqs)
}
\arguments{
\item{seqs}{Results from \link{distri_peps}.}
}
\description{
For targets peptides.
}
\details{
Columns: \code{is_pnt}, is Protein N-term; \code{is_pct}, is Protein C-term.
For example, MEYEWKPDEQGLQQILQLLK: NP_002261, pep_start 9; NP_694858:
pep_start 1. Without the information of protein terminals, the peptide
sequence with the modification of Acetyl Protein N-term will match to both
protein accessions, with the match to NP_002261 being inconsistent.
}
\examples{
\donttest{
# library(mzion)
# mzion:::simple_prots_peps(fwd_peps[[1]])
}
}
