% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{sparseD_fourquad}
\alias{sparseD_fourquad}
\title{Builds manually distance sparse matrix.}
\usage{
sparseD_fourquad(M_ul, ncols_ur = 0)
}
\arguments{
\item{M_ul}{The upper-left matrix.}

\item{ncols_ur}{The number of columns for the matrix block on the upper
right.}
}
\description{
Not yet used.
}
\examples{
m <- sparseD_fourquad(ul, 6)
}
