% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{split_fastaseqs_noenz}
\alias{split_fastaseqs_noenz}
\title{Splits fasta sequences by proteins.}
\usage{
split_fastaseqs_noenz(
  fasta = NULL,
  acc_type = "uniprot_acc",
  acc_pattern = NULL,
  maxn_fasta_seqs = 200000L,
  min_len = 7L,
  max_len = 40L,
  aa_masses = NULL,
  ftmass = 18.010565
)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The experimenter needs to supply the files.}

\item{acc_type}{Character string(s); the types of protein accessions in one
of c("uniprot_acc", "uniprot_id", "refseq_acc", "other"). For custom names,
the corresponding regular expressions need to be supplied via argument
\code{acc_pattern}.}

\item{acc_pattern}{Regular expression(s) describing the patterns in
separating the header lines of fasta entries. At the \code{NULL} default,
the pattern will be automated when \code{acc_type} are among
c("uniprot_acc", "uniprot_id", "refseq_acc", "other"). See also
\link{load_fasta2} for custom examples.}

\item{maxn_fasta_seqs}{Integer; the maximum number of protein sequences in
fasta files.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{aa_masses}{A lookup table of the masses of amino-acid residues.}

\item{ftmass}{The sum of masses of \code{fixed} N-term and C-term
modifications.}
}
\description{
For \code{no enzyme} searches.
}
