% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{split_im}
\alias{split_im}
\title{Split \code{ion_matches_.rds}}
\usage{
split_im(files, sc_path, tempdir, max_size = 1e+07)
}
\arguments{
\item{files}{File names such as \code{ion_matches_1.rds, ion_matches_2.rds}}

\item{sc_path}{A file path to \code{ion_matches_1.rds}}

\item{tempdir}{A temporary directory}

\item{max_size}{The maximum file size}
}
\description{
For RAM efficiency
}
