% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{split_matrix}
\alias{split_matrix}
\title{Splits a matrix}
\usage{
split_matrix(M, by = "row")
}
\arguments{
\item{M}{A matrix.}

\item{by}{Split by matrix rows or columns.}
}
\description{
Splits a matrix
}
