% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{split_paseflines_by_frames}
\alias{split_paseflines_by_frames}
\title{Splits single \code{spec_[...].txt} by frames}
\usage{
split_paseflines_by_frames(file, path)
}
\arguments{
\item{file}{A file name of \code{spec_[...].txt}.}

\item{path}{A path to a \code{.d} folder.}
}
\description{
Splits single \code{spec_[...].txt} by frames
}
