% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{split_vec}
\alias{split_vec}
\title{Split a vector by values}
\usage{
split_vec(vec)
}
\arguments{
\item{vec}{A vector.}
}
\description{
Split a vector by values
}
\examples{
\donttest{
## M
vec <- c(`Carbamidomethyl (M)` = "M", 
         `Carbamyl (M)` = "M")

x <- split_vec(vec)
y <- split(vec, vec)

stopifnot(identical(x, y))

microbenchmark::microbenchmark(split_vec(vec), split(vec, vec))

## N
vec <- c(`Deamidated (N)` = "N")

x <- split_vec(vec)
y <- split(vec, vec)

stopifnot(identical(x, y))

microbenchmark::microbenchmark(split_vec(vec), 
                               split(vec, vec))

## M, N
vec <- c(`Carbamidomethyl (M)` = "M", 
         `Carbamyl (M)` = "M", 
         `Deamidated (N)` = "N")

x <- split_vec(vec)
y <- split(vec, vec)

stopifnot(identical(x, y))

microbenchmark::microbenchmark(split_vec(vec), split(vec, vec))
}

}
