% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{spreadMS}
\alias{spreadMS}
\title{Spreads MS data across adjacent scans.}
\usage{
spreadMS(matx, maty, matz, ns, ps)
}
\arguments{
\item{matx}{The data matrix of moverzs.}

\item{maty}{The data matrix of intensities.}

\item{matz}{The data matrix of charge states.}
}
\description{
Spreads MS data across adjacent scans.
}
