% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{spreadMS_v1}
\alias{spreadMS_v1}
\title{Spreads MS data across adjacent scans.}
\usage{
spreadMS_v1(matx, maty, matz, ns, ps)
}
\arguments{
\item{matx}{The data matrix of moverzs.}

\item{maty}{The data matrix of intensities.}

\item{matz}{The data matrix of charge states.}

\item{ns}{The number of observing spectra that have contributed to an MS
feature.}

\item{ps}{The row positions (along LC scans) of features.}
}
\description{
Not used.
}
