% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{spreadMSohw}
\alias{spreadMSohw}
\title{Spreads one-hit wonders to adjacent scans.}
\usage{
spreadMSohw(matx, maty, matz, ns, ps, gap = 2L, join_ms = FALSE)
}
\arguments{
\item{matx}{The data matrix of moverzs.}

\item{maty}{The data matrix of intensities.}

\item{matz}{The data matrix of charge states.}

\item{ns}{The number of observing spectra that have contributed to an MS
feature. One-hit wonders correspond to \code{ns == 1L}.}

\item{ps}{The row positions (along LC scans) of features.}

\item{gap}{The gap width for one-hit wonders.}

\item{join_ms}{Logical; if TRUE, combine adjacent entries.}

\item{neigh_start}{The starting point of neighbors for the data spreading.}
}
\description{
Spreads one-hit wonders to adjacent scans.
}
