% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{standardize_unimod_ps}
\alias{standardize_unimod_ps}
\title{Standardizes the site and position of a modification.}
\usage{
standardize_unimod_ps(x)
}
\arguments{
\item{x}{A name string. Note that \code{x} is either a site or a position.}
}
\description{
Standardizes the site and position of a modification.
}
