% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{sub_pasef_ms2info}
\alias{sub_pasef_ms2info}
\title{Separates PASEF MS2Info by frame ranges}
\usage{
sub_pasef_ms2info(data, iso_info, offset = 3L)
}
\arguments{
\item{data}{Line data of MS frames (containing both MS1 and MS2).}

\item{iso_info}{A data frame containing ms2info.}

\item{offset}{The offset from an empty line to the FRAME-number line.}
}
\description{
Separates PASEF MS2Info by frame ranges
}
