% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{sub_protpep}
\alias{sub_protpep}
\title{Subsets the protein-peptide lookups.}
\usage{
sub_protpep(prp, mt, nm)
}
\arguments{
\item{prp}{A vector of peptides under a protein.}

\item{mt}{A vector of matches. The values are NA if no matches. The matched
integers are the indexes in the list of uniq_peps for protein annotations.}

\item{nm}{The names of \code{prp}.}
}
\description{
Duplicated peptide sequences with a protein also removed.
}
