% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{subDIA_MS1}
\alias{subDIA_MS1}
\title{Get precursor mass candidates for each DIA window.}
\usage{
subDIA_MS1(filename = NULL)
}
\arguments{
\item{filename}{A filename with prepending path.}
}
\description{
Subset by MS2 isolation windows.
}
