% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{subMSfull}
\alias{subMSfull}
\title{Subsets full MS by the universe of monoisotopic moverzs.}
\usage{
subMSfull(xs, ys, unv, from = 200L, step = 1e-05)
}
\arguments{
\item{xs}{Vectors of moverzs values.}

\item{ys}{Vectors of intensity values.}

\item{unv}{The universe.}

\item{from}{The starting mozerz for calculating bin indexes.}

\item{step}{The size of bins.}
}
\description{
Some non-monoisotopic moverzs can be removed.
}
