% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{subpeps_by_vmods}
\alias{subpeps_by_vmods}
\title{Subsets peptides by variable modifications.}
\usage{
subpeps_by_vmods(aa_masses, prps)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{prps}{Lists of peptide sequences with a one-letter representation
of amino acid residues. Each list is named by protein accession.}
}
\description{
From N-term to C-term.
}
