% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{subset_by_prps}
\alias{subset_by_prps}
\title{Subsets peptides}
\usage{
subset_by_prps(ps, s, motifs)
}
\arguments{
\item{ps}{A list of peptides under a protein}

\item{s}{A pattern}

\item{motifs}{A list of motifs of amino-acide residues. for example,
"MN|NG". See also \link{matchMS}.}
}
\description{
By individual proteins.
}
