% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{subset_pasefms}
\alias{subset_pasefms}
\title{Subset PASEF MS data by frame numbers}
\usage{
subset_pasefms(mdata, ms_frs, iso_info, ms_col = "MS1Frame")
}
\arguments{
\item{mdata}{A list of PASEF frames. Each list entry corresponding to one MS
frame.}

\item{ms_frs}{MS frame numbers corresponding to \code{mdata}.}

\item{iso_info}{The metadata os slice starts etc.}

\item{ms_col}{The criteria column for subsetting.}
}
\description{
Subset PASEF MS data by frame numbers
}
\seealso{
\link{bracket_pasef_ms2info} that brackets \code{iso_info} by the
  first and the last frames of \code{mdata}.
}
