% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{substract_ms1mass}
\alias{substract_ms1mass}
\title{Subtract ms1_mass values in mass calibrations}
\usage{
substract_ms1mass(vals, mdiff, islist = FALSE)
}
\arguments{
\item{vals}{MS1 mass values.}

\item{islist}{Logical; is the vals in list for scalar.}
}
\description{
Subtract ms1_mass values in mass calibrations
}
