% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{sum_mono_yints}
\alias{sum_mono_yints}
\title{Specialty summing of monotonically decreased or increased intensities.}
\usage{
sum_mono_yints(xs, ys, reso2 = 24000, down = TRUE)
}
\arguments{
\item{xs}{A vector of moverz values}

\item{ys}{A vector of intensity values.}

\item{reso2}{Adjusted resolution.}

\item{down}{Logical; are the \code{ys} monotonically decreased or not.}
}
\description{
Specialty summing of monotonically decreased or increased intensities.
}
