% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{table_unimods}
\alias{table_unimods}
\title{Parses \href{https://www.unimod.org/}{Unimod} entries.}
\usage{
table_unimods(
  file = system.file("extdata", "master.xml", package = "proteoM"),
  out_nm = "~/proteoM/unimods.txt"
)
}
\arguments{
\item{file}{A file path to a Unimod ".xml".}

\item{out_nm}{A name to outputs.}
}
\description{
For convenience findings of the \code{title}, \code{site} and
\code{position}.
}
\examples{
\donttest{
ans <- table_unimods()

ans[with(ans, title == "TMTpro"), ]
this_mod <- parse_unimod("TMTpro (Anywhere = K)")

ans[with(ans, grepl("^Gln->pyro", title)), ]
this_mod <- parse_unimod("Gln->pryo-Glu (N-term = Q)")
}
}
\seealso{
\link{find_unimod}, \link{parse_unimod}.
}
