% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{trace_ms}
\alias{trace_ms}
\title{Traces MS peaks.}
\usage{
trace_ms(xs, ys, zs, from = 115L, step = 1e-05, gap = 4L, reord = TRUE)
}
\arguments{
\item{xs}{Vectors of moverzs.}

\item{ys}{Vectors of intensities.}

\item{zs}{Vectors of charge states.}

\item{from}{The starting value for mass binning.}

\item{step}{A step size for mass binning.}

\item{gap}{The gap size for collapsing moverzs.}

\item{reord}{Logical; re-order data or not.}
}
\description{
For both moverzs and intensities.
}
