% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{traceLCMS}
\alias{traceLCMS}
\title{Traces LCMS peaks.}
\usage{
traceLCMS(
  xs,
  ys,
  zs,
  n_dia_scans = 4L,
  from = 115L,
  step = 8e-06,
  icenter = 1L,
  ms_lev = 2L,
  temp_dir = NULL,
  filename = NULL,
  reord = TRUE,
  cleanup = TRUE,
  replace_ms1_by_apex = FALSE,
  direct_out = FALSE
)
}
\arguments{
\item{xs}{Vectors of moverzs at a given isolation center (\code{icenter}).
Each entry corresponds to a full vector of MS1 or MS2.}

\item{ys}{Vectors of intensities.}

\item{zs}{Vectors of charge states.}

\item{n_dia_scans}{The number of adjacent MS scans for constructing a peak
profile and thus for determining the apex scan number of an moverz value
along LC.}

\item{from}{The starting value for mass binning.}

\item{step}{A step size for mass binning.}

\item{icenter}{The index of an isolation center.}

\item{ms_lev}{The level of MS.}

\item{temp_dir}{A temp_dir to the filename.}

\item{filename}{A peaklist filename.}

\item{reord}{Logical; re-order data or not.}

\item{cleanup}{Logical; cleans up xs, ys and zs or not. Set the value to
FALSE to maintain one-to-one correspondence between input (data frame) and
the outputs. This will help, e.g., keep track of scan numbers in the input.}

\item{replace_ms1_by_apex}{Logical; if TRUE, fill all entries within a gate
by its apex values.}

\item{direct_out}{Logical; if TRUE, return outputs directly.}
}
\description{
For both moverzs and intensities.
}
