% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{traceMS}
\alias{traceMS}
\title{Traces MS peaks.}
\usage{
traceMS(
  xs,
  ys,
  zs,
  n_dia_scans = 4L,
  from = 115L,
  step = 1e-05,
  icenter = 1L,
  ms_lev = 2L,
  temp_dir = NULL,
  reord = TRUE
)
}
\arguments{
\item{xs}{Vectors of moverzs at a given isolation center (\code{icenter}).
Each entry corresponds to a full vector of MS1 or MS2.}

\item{ys}{Vectors of intensities.}

\item{zs}{Vectors of charge states.}

\item{n_dia_scans}{The number of adjacent MS scans for constructing a peak
profile and thus for determining the apex scan number of an moverz value
along LC.}

\item{from}{The starting value for mass binning.}

\item{step}{A step size for mass binning.}

\item{icenter}{The index of an isolation center.}

\item{ms_lev}{The level of MS.}

\item{temp_dir}{A temporary directory.}

\item{reord}{Logical; re-order data or not.}
}
\description{
For both moverzs and intensities.
}
