% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{traceMS1}
\alias{traceMS1}
\title{Helper of MS1 tracing.}
\usage{
traceMS1(
  df,
  min_mass = 200L,
  step = 8e-06,
  n_ms1peakpicking_flanks = 4L,
  replace_ms1_by_apex = TRUE,
  filename = NULL,
  temp_dir = NULL
)
}
\arguments{
\item{df}{MS1 data.}

\item{step}{Step size.}

\item{n_ms1peakpicking_flanks}{The number of flanking MS1 scans for precursor
peak picking.}

\item{replace_ms1_by_apex}{Logical; if TRUE, fill all entries within a gate
by its apex values.}

\item{filename}{A peaklist filename.}

\item{temp_dir}{A temp_dir to the filename.}
}
\description{
Not yet used.
}
