% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{traceXY}
\alias{traceXY}
\title{Helper of MS1 tracing.}
\usage{
traceXY(
  xs,
  ys,
  n_dia_scans = 4L,
  from = 115L,
  step = 1e-05,
  reord = TRUE,
  cleanup = TRUE,
  replace_ms1_by_apex = FALSE
)
}
\arguments{
\item{xs}{moverzs.}

\item{ys}{intensities.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}

\item{from}{The starting point for mass binning.}

\item{step}{A step size for mass binning.}

\item{reord}{Logical; re-order data or not.}

\item{cleanup}{Logical; cleans up xs, ys and zs or not. Set the value to
FALSE to maintain one-to-one correspondence between input (data frame) and
the outputs. This will help, e.g., keep track of scan numbers in the input.}

\item{replace_ms1_by_apex}{Logical; if TRUE, fill all entries within a gate
by its apex values.}
}
\description{
Helper of MS1 tracing.
}
