% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{tsoutliers}
\alias{tsoutliers}
\title{Identifies and replace outliers in a time series.}
\usage{
tsoutliers(x, iterate = 2, lambda = NULL)
}
\arguments{
\item{x}{Time series}

\item{iterate}{The number of iterations.}

\item{lambda}{Box-Cox transformation parameter. If lambda="auto", then a
transformation is automatically selected using BoxCox.lambda. The
transformation is ignored if NULL. Otherwise, data transformed before model
is estimated.}
}
\description{
From \code{forecast}.
}
