% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{unique_mvmods}
\alias{unique_mvmods}
\title{The unique entries of variable modifications (multiple sites)}
\usage{
unique_mvmods(
  amods,
  ntmod,
  ctmod,
  aa_masses,
  aas,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  digits = 5L
)
}
\arguments{
\item{amods}{Anywhere modifications.}

\item{ntmod}{The attribute \code{ntmod} from a \code{aa_masses} (for MS1
calculations).}

\item{ctmod}{The attribute \code{ctmod} from a \code{aa_masses} (for MS1
calculations).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\value{
Lists by residues in \code{amods}.
}
\description{
For all the \code{Anywhere} modifications specified in \code{amods}.
}
