% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{updateMS1Int}
\alias{updateMS1Int}
\title{Updates MS1 intensity with apex values.}
\usage{
updateMS1Int(
  df,
  matx,
  maty,
  unv,
  ss,
  ts,
  row_sta,
  row_end,
  from = 200L,
  step = 6e-06,
  min_y = 0,
  yco = 500,
  n_dia_scans = 6L
)
}
\arguments{
\item{df}{A data frame.}

\item{matx}{The matrix of m-over-z's Y: by masses; X: by LC scans.}

\item{maty}{The matrix of intensities. Y: by masses; X: by LC scans.}

\item{unv}{The universe of m-over-z bins.}

\item{ss}{A vector of scan numbers.}

\item{ts}{A vector of retention times.}

\item{row_sta}{The starting row of \code{matx}.}

\item{row_end}{The ending row of \code{matx}.}

\item{from}{The starting point for mass binning.}

\item{step}{A step size for mass binning.}

\item{min_y}{The cut-off of intensity values.}

\item{yco}{An intensity cut-off.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}
}
\description{
Including forward and backward looking.
}
