% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{updateMS1Int2}
\alias{updateMS1Int2}
\title{Updates MS1 intensity with apex values.}
\usage{
updateMS1Int2(
  df,
  matx,
  maty,
  row_sta,
  row_end,
  scan_apexs,
  rngs,
  from = 200L,
  step = 1e-05
)
}
\arguments{
\item{df}{A data frame.}

\item{matx}{The matrix of moverzs Y: by masses; X: by LC scans.}

\item{maty}{The matrix of intensities. Y: by masses; X: by LC scans.}

\item{row_sta}{The starting row of \code{matx}.}

\item{row_end}{The ending row of \code{matx}.}

\item{scan_apexs}{The vectors of apex scan number. Each vector corresponds to
a mass in \code{matx}.}

\item{rngs}{The ranges of row indexes for each column in \code{matx}. The
\code{rngs[[i]]} corresponds to \code{matx[, i]}. The argument is currently
used for debugging.}

\item{from}{The starting point for mass binning.}

\item{step}{A step size for mass binning.}
}
\description{
Including forward and backward looking.
}
