% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{vec_to_list}
\alias{vec_to_list}
\title{Split a named character vector to lists.}
\usage{
vec_to_list(x)
}
\arguments{
\item{x}{A named character vector.}
}
\description{
Split a named character vector to lists.
}
\examples{
\donttest{
x <- c(`Deamidated (N)` = "N", `Carbamidomethyl (S)` = "S")

identical(vec_to_list(x), split(x, x))
microbenchmark::microbenchmark(vec_to_list(x), split(x, x))
}
}
