% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{vmods_elements}
\alias{vmods_elements}
\title{Find the sets of variable modifications.}
\usage{
vmods_elements(
  aas,
  residue_mods,
  ntmod,
  ctmod,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  .ms1_vmodsets = NULL,
  .base_ent = NULL,
  digits = 4L
)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{residue_mods}{Amino-acid residues with Unimod names. For example
rownames of \code{Carbamidomethyl (M)} and \code{Oxidation (M)} and a
column residues of \code{M, M}.}

\item{ntmod}{The attribute \code{ntmod} from a \code{aa_masses} (for MS1
calculations).}

\item{ctmod}{The attribute \code{ctmod} from a \code{aa_masses} (for MS1
calculations).}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{.ms1_vmodsets}{Not used.}

\item{.base_ent}{Not used.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
The same residue, e.g. M, at different modifications, c("Carbamyl (M",
"Oxidation (M)")).
}
\details{
Excluding position differences, i.e., \code{A, B} and \code{B, A} is the
same set.
}
\examples{
\donttest{
library(mzion)

ntmod <- list(`Acetyl (Protein N-term)` = c(`Protein N-term` = "N-term"))

ctmod <- list()
names(ctmod) <- character()

aas <- unlist(strsplit("HQGVMNVGMGQKSMNS", ""))
residue_mods <- c(`Carbamidomethyl (M)` = "M", `Carbamyl (M)` = "M")

x <- mzion:::vmods_elements(aas, residue_mods, ntmod, ctmod)
}
}
