% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{which_topx}
\alias{which_topx}
\title{Finds the indexes of top-n entries without re-ordering.}
\usage{
which_topx(x, n = 50L, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of top entries to keep.}

\item{...}{Additional arguments for base function sort.}
}
\value{
The indexes of the top-n entries.
}
\description{
At length(x) >= n, the length of output may be shorter than n with ties.
}
\examples{
\donttest{
library(proteoM)

proteoM:::which_topx(c(1:5), 50)

length(proteoM:::which_topx(sample(5000, 500), 100))

length(proteoM:::which_topx(sample(100, 100, replace = TRUE), 100))
}
}
