% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{which_topx3}
\alias{which_topx3}
\title{Finds the indexes of top-n entries without re-ordering.}
\usage{
which_topx3(x, n = 50L, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of top entries to keep.}

\item{...}{Additional arguments for base function sort.}
}
\value{
The indexes of the top-n entries.
}
\description{
Only used with \link{calc_peploc}: \link{which_topx2} with additional
handling of NA.
}
\details{
For speed, codes duplicated from \link{which_topx2} except for the anyNA
checks.
}
