% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{accumulate_char}
\alias{accumulate_char}
\title{Accumulates character strings.}
\usage{
accumulate_char(x, f)
}
\arguments{
\item{x}{A vector of character strings.}

\item{f}{A function.}
}
\description{
Accumulates character strings.
}
\examples{
\donttest{
library(mzion)

x <- c("aa", "bb", "cc")
mzion:::accumulate_char(x, paste0)

x <- c(a = 1, b = 2, c = 3)
mzion:::accumulate_char(x, paste0)

x <- "-EDEIQDXI-"
mzion:::accumulate_char(x, paste0)
}

}
