% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{add_primatches}
\alias{add_primatches}
\title{Adds sequences of primary and secondary matches.}
\usage{
add_primatches(
  file = NULL,
  tempdir = NULL,
  add_ms2theos = FALSE,
  add_ms2theos2 = FALSE,
  add_ms2moverzs = FALSE,
  add_ms2ints = FALSE,
  index_mgf_ms2 = FALSE
)
}
\arguments{
\item{file}{The file name of a temporary score file.}

\item{tempdir}{The temporary directory for score results.}

\item{add_ms2theos}{Logical. If true, adds the sequence of primary
  theoretical MS2 m/z values (\code{pep_ms2_theos}). The sequence order at a
  given \code{type_ms2ions} is:

  \tabular{ll}{ \strong{Type}   \tab \strong{Sequence}\cr by \tab \eqn{b1,
  b2..., y1, y2...} \cr ax \tab \eqn{a1, a2..., x1, x2...} \cr cz \tab
  \eqn{c1, c2..., z1, z2...} \cr }}

\item{add_ms2theos2}{Logical. If true, adds the sequence of secondary
  theoretical MS2 m/z values (\code{pep_ms2_theos2}). The sequence order at a
  given \code{type_ms2ions} is:

  \tabular{ll}{ \strong{Type}   \tab \strong{Order of sequences}\cr by \tab
  \eqn{b2, b*, b*2, b0, b02, y2, y*, y*2, y0, y02} \cr ax \tab \eqn{a2, a*,
  a*2, a0, a02, x2} \cr cz \tab \eqn{c2, z2} \cr }}

\item{add_ms2moverzs}{Logical; if TRUE, adds the sequence of experimental
\eqn{m/z} values (\code{pep_ms2_moverzs}).}

\item{add_ms2ints}{Logical; if TRUE, adds the sequence of experimental MS2
intensity values (\code{pep_ms2_ints}).}

\item{index_mgf_ms2}{Depreciated. A low-priority feature. Logical; if TRUE,
  converts up-frontly MS2 m-over-z values from numeric to integers as opposed
  to \emph{on-the-fly} conversion during ion matches. The default is FALSE.
  The \code{index_mgf_ms2 = TRUE} might be useful for very large MS files by
  reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}
}
\description{
Applied to both targets and decoys as feature "pep_ms2_deltas_mean" may be
used in SVM-Percolator.
}
