% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{add_unimod}
\alias{add_unimod}
\title{Adds or modifies a Unimod entry.}
\usage{
add_unimod(
  header = c(title = "Foo", full_name = "Foo bar"),
  specificity = c(site = "C", position = "Anywhere"),
  delta = c(mono_mass = "42.010565", avge_mass = "42.0367", composition =
    "H(2) C(2) O"),
  neuloss = c(mono_mass = "0", avge_mass = "0", composition = "0")
)
}
\arguments{
\item{header}{The header of a \href{https://www.unimod.org/}{Unimod}. The
fields of \code{title} and \code{full_name} are required. No spaces are
allowed in \code{title}.}

\item{specificity}{The specificity of \code{site} and \code{position} of a
  modification. The value of \code{site} is an upper-case, one-letter
  representation of an amino-acid residue, or "N-term" or "C-term". The value
  of \code{position} is one of "Anywhere", "Protein N-term", "Protein
  C-term", "Any N-term" or "Any C-term".

  See also \link{remove_unimod} for a wildcard approach of \code{site = "."}
  and \code{position = "."} to remove all sites and positions under a title.}

\item{delta}{The mass and composition difference of a modification:
  \code{mono_mass}, the difference in mono-isotopic mass; \code{avge_mass},
  the difference in average mass; \code{composition}, the difference in
  chemical composition.

  The precision of \code{mono_mass} is typically set with a decimal place of
  \eqn{\ge 5}.}

\item{neuloss}{The mass and composition \emph{loss} of a neutral species atop
  of the \code{delta}. Like many other search engines, \code{proteoM} adapts
  the common convention where the losses of masses and compositions are
  expressed in \emph{positive} forms for both the masses and the
  compositions.

  See also \link{remove_unimod} for a wildcard approach of \code{neuloss =
  c(mono_mass = ".", ...)} to remove all neutral losses under a site and a
  position for a given title.}
}
\value{
An xml object.
}
\description{
Changes are made to system files.
}
\examples{
\dontrun{
library(proteoM)

# To avoid unsound chemistries, proteoM prohibits
#   additive modifications to the same site.
# To enable cumulative effects, the solution is to
#   devise a "merged" modification.

(system.file("extdata", "master.xml", package = "proteoM"))
(system.file("extdata", "custom.xml", package = "proteoM"))

# Additive N-terminal modifications (though not chemically sound)
masses <- calc_unimod_compmass("H(17) C(8) 13C(4) 15N O(2)")
mono_mass <- masses$mono_mass
avge_mass <- masses$avge_mass

x <- add_unimod(header      = c(title       = "TMT10plexNterm+Gln->pyro-Glu",
                                full_name   = "Additive N-term TMT10plex and Gln->pyro-Glu"),

                # think about this: site = "Q", not "N-term"
                specificity = c(site        = "Q",
                                position    = "Any N-term"),
                delta       = c(mono_mass   = "212.136383",
                                avge_mass   = "212.2329",
                                composition = "H(17) C(8) 13C(4) 15N O(2)"),
                neuloss     = c(mono_mass   = "0",
                                avge_mass   = "0",
                                composition = "0"))

(ans <- table_unimods())
(ans[with(ans, title == "TMT10plexNterm+Gln->pyro-Glu"), ])

# site `C`: Oxiation + Carbamidomethyl
# (without neutral losses)
x <- add_unimod(header      = c(title       = "Oxi+Carbamidomethyl",
                                full_name   = "Oxidation and iodoacetamide derivative"),
                specificity = c(site        = "C",
                                position    = "Anywhere"),
                delta       = c(mono_mass   = "73.016379",
                                avge_mass   = "73.0507",
                                composition = "H(3) C(2) N O(2)"),
                neuloss     = c(mono_mass   = "0",
                                avge_mass   = "0",
                                composition = "0"))

# site `M`: Oxiation + Carbamidomethyl
# (with neutral losses)
x <- add_unimod(header      = c(title       = "Oxi+Carbamidomethyl",
                                full_name   = "Oxidation and iodoacetamide derivative"),
                specificity = c(site        = "M",
                                position    = "Anywhere"),
                delta       = c(mono_mass   = "73.016379",
                                avge_mass   = "73.0507",
                                composition = "H(3) C(2) N O(2)"),
                neuloss     = c(mono_mass   = "63.998285",
                                avge_mass   = "64.1069",
                                composition = "H(4) C O S"))
}

## Heavy isotopes
# Lysine
K8 <- calc_unimod_compmass("13C(6) C(-6) 15N(2) N(-2)")

mono_mass <- K8$mono_mass
avge_mass <- K8$avge_mass

x <- add_unimod(header      = c(title       = "K8",
                                full_name   = "Heavy lysine 13C(6) 15N(2)"),
                specificity = c(site        = "K",
                                position    = "Anywhere"),
                delta       = c(mono_mass   = "8.0142",
                                avge_mass   = "7.94272",
                                composition = "13C(6) C(-6) 15N(2) N(-2)"),
                neuloss     = c(mono_mass   = "0",
                                avge_mass   = "0",
                                composition = "0"))

# Arginine
R10 <- calc_unimod_compmass("13C(6) C(-6) 15N(4) N(-4)")

mono_mass <- R10$mono_mass
avge_mass <- R10$avge_mass

x <- add_unimod(header      = c(title       = "R10",
                                full_name   = "Heavy arginine 13C(6) 15N(4)"),
                specificity = c(site        = "R",
                                position    = "Anywhere"),
                delta       = c(mono_mass   = "10.00827",
                                avge_mass   = "9.92954",
                                composition = "13C(6) C(-6) 15N(4) N(-4)"),
                neuloss     = c(mono_mass   = "0",
                                avge_mass   = "0",
                                composition = "0"))

}
\seealso{
\link{table_unimods}, \link{parse_unimod}, \link{find_unimod}.
}
