% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{aggr_prot_es}
\alias{aggr_prot_es}
\title{Helper for descriptive statstics}
\usage{
aggr_prot_es(f)
}
\arguments{
\item{f}{One of \code{max, mean, median, min}.}
}
\description{
Helper for descriptive statstics
}
