% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_labels.R
\name{bacth_vmods_combi}
\alias{bacth_vmods_combi}
\title{Helper in cycling through \code{2:maxn_vmods_per_pep} numbers of positions.}
\usage{
bacth_vmods_combi(
  resmods = NULL,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L
)
}
\arguments{
\item{resmods}{A vector of amino-acid residues with Unimod in names. For
example, c(`Oxidation (M)` = "M", `Carbamyl` = "M").The residues are the
same and the names are different.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}
}
\description{
For the same residue (M) at different names of modifications.
}
\examples{
\donttest{
resmods <- c(`Oxidation (M)` = "M", `Carbamidomethyl (M)` = "M")
ans <- bacth_vmods_combi(resmods)
}
}
