% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{bin_ms1masses}
\alias{bin_ms1masses}
\title{Helper in binning precursor masses.}
\usage{
bin_ms1masses(
  res = NULL,
  min_mass = 500L,
  max_mass = 6000L,
  ppm_ms1 = 20L,
  use_ms1_cache = TRUE,
  .path_cache = NULL,
  .path_ms1masses = NULL,
  is_ms1_three_frame = TRUE
)
}
\arguments{
\item{res}{The results from \link{calc_pepmasses2}.}

\item{min_mass}{A minimum mass of precursors.}

\item{max_mass}{A maximum mass of precursors.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{use_ms1_cache}{Logical; if TRUE, use cached precursor masses.}

\item{.path_cache}{The file path of cached search parameters. At the NULL
default, the path is \code{"~/proteoM/.MSearches/Cache/Calls/"}. The
parameter is for users' awareness of the structure of file folders and the
default is suggested. Occasionally experimenters may remove the file folder
for disk space or (hopefully not) in events of (improper) structural change
incurred by the developer.}

\item{.path_ms1masses}{The file path to the theoretical masses of MS1
precursors.}

\item{is_ms1_three_frame}{Logical; is the searches by the three frames of
preceeding, current and following.}
}
\description{
For target peptides.
}
