% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_masses.R
\name{bin_ms1masses}
\alias{bin_ms1masses}
\title{Helper in binning precursor masses.}
\usage{
bin_ms1masses(
  res = NULL,
  min_mass = 200L,
  max_mass = 4500L,
  min_len = 7L,
  max_len = 40L,
  ppm_ms1 = 20L,
  use_ms1_cache = TRUE,
  .path_cache = NULL,
  .path_ms1masses = NULL,
  is_ms1_three_frame = TRUE,
  out_path = NULL,
  enzyme = "trypsin_p",
  sys_ram = 24L
)
}
\arguments{
\item{res}{The results from \link{calc_pepmasses2}.}

\item{min_mass}{A minimum mass of precursors.}

\item{max_mass}{A maximum mass of precursors.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded. The default is 40.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{use_ms1_cache}{Logical; at the TRUE default, use cached precursor
  masses.

  Set \code{use_ms1_cache = TRUE} for reprocessing of data, e.g., from
  \code{fdr_type = psm} to \code{fdr_type = protein}.}

\item{.path_cache}{The file path of cached search parameters. The parameter
is for the users' awareness of the underlying structure of file folders and
the use of default is suggested. Occasionally experimenters may remove the
file folder for disk space or under infrequent events of modified framework
incurred by the developer.}

\item{.path_ms1masses}{The file path to the theoretical masses of MS1
precursors.}

\item{is_ms1_three_frame}{Logical; is the searches by the three frames of
preceding, current and following.}

\item{out_path}{A file path of outputs.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from protein entries. The
default is \code{Trypsin_P}. See also parameter \code{custom_enzyme}.}

\item{sys_ram}{A putative value of system RAM.}
}
\description{
For target peptides.
}
