% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_ladder.R
\name{byions}
\alias{byions}
\alias{czions}
\alias{axions}
\alias{bions_base}
\alias{yions_base}
\alias{b2ions_base}
\alias{bstarions}
\alias{bstar2ions}
\alias{b0ions}
\alias{b02ions}
\alias{y2ions}
\alias{ystarions}
\alias{ystar2ions}
\alias{y0ions}
\alias{y02ions}
\alias{cions_base}
\alias{c2ions}
\alias{zions_base}
\alias{z2ions}
\alias{aions_base}
\alias{a2ions}
\alias{astarions}
\alias{astar2ions}
\alias{a0ions}
\alias{a02ions}
\alias{xions_base}
\alias{x2ions}
\title{Masses of singly-charged b- and y-ions.}
\usage{
byions(ntmass, ctmass, aam, digits = 4L)

czions(ntmass, ctmass, aam, digits = 4L)

axions(ntmass, ctmass, aam, digits = 4L)

bions_base(aam, tmass, digits = 4L)

yions_base(aam, tmass, digits = 4L)

b2ions_base(aam, tmass, digits = 4L, n = 2L)

bstarions(aam, tmass, digits = 4L)

bstar2ions(aam, tmass, digits = 4L, n = 2L)

b0ions(aam, tmass, digits = 4L)

b02ions(aam, tmass, digits = 4L, n = 2L)

y2ions(aam, tmass, digits = 4L, n = 2L)

ystarions(aam, tmass, digits = 4L)

ystar2ions(aam, tmass, digits = 4L, n = 2L)

y0ions(aam, tmass, digits = 4L)

y02ions(aam, tmass, digits = 4L, n = 2L)

cions_base(aam, tmass, digits = 4L)

c2ions(aam, tmass, digits = 4L, n = 2L)

zions_base(aam, tmass, digits = 4L)

z2ions(aam, tmass, digits = 4L, n = 2L)

aions_base(aam, tmass, digits = 4L)

a2ions(aam, tmass, digits = 4L, n = 2L)

astarions(aam, tmass, digits = 4L)

astar2ions(aam, tmass, digits = 4L, n = 2L)

a0ions(aam, tmass, digits = 4L)

a02ions(aam, tmass, digits = 4L, n = 2L)

xions_base(aam, tmass, digits = 4L)

x2ions(aam, tmass, digits = 4L, n = 2L)
}
\arguments{
\item{ntmass}{The mass of a fixed or variable N-term modification.}

\item{ctmass}{The mass of a fixed or variable C-term modification.}

\item{aam}{A sequence of amino-acid residues with \emph{masses}. Residues
  are in names and masses in values.

  The masses reflects fixed/variable modifications, and/or fixed/variable
  neutral losses.}

\item{digits}{Integer; the number of decimal places to be used.}

\item{tmass}{The mass of a fixed or variable N-term or C-term modification.}

\item{n}{The charge state.}
}
\description{
For (1) "amods- tmod- vnl- fnl-", (2) "amods- tmod+ vnl- fnl-".

\code{H2O = 18.010565}.
}
\examples{
\donttest{
library(mzion)
library(stringr)

## (1) "amods- tmod- vnl- fnl-"
# (Fixed N-term mods; also for no N-term mod)

fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("Oxidation (M)", "Deamidated (N)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

aa_masses <- aa_masses_all[[1]]

ntmod <- attr(aa_masses, "ntmod", exact = TRUE)
ctmod <- attr(aa_masses, "ctmod", exact = TRUE)

if (!length(ntmod)) {
  ntmass <- aa_masses["N-term"] - 0.000549 # - electron
} else {
  ntmass <- aa_masses[names(ntmod)] - 0.000549
}

if (!length(ctmod)) {
  ctmass <- aa_masses["C-term"] + 2.01510147 # + (H) + (H+)
} else {
  ctmass <- aa_masses[names(ctmod)] + 2.01510147
}

aa_seq <- "MAKEMASSPECFUN"
aas <- stringr::str_split(aa_seq, "", simplify = TRUE)
aam <- aa_masses[aas]

b <- mzion:::bions_base(aam, ntmass)
y <- mzion:::yions_base(aam, ctmass)


## (2) "amods- tmod+ vnl- fnl-"
# (2a, N-term)
fixedmods <- c("TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("TMT6plex (N-term)", "Acetyl (Protein N-term)", "Oxidation (M)",
             "Deamidated (N)", "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

aa_masses <- aa_masses_all[[3]]

# (Fixed or variable C-term mods +/- makes no difference on b-ions;
# and vice versa for y-ions)
ntmod <- attr(aa_masses, "ntmod", exact = TRUE)
ctmod <- attr(aa_masses, "ctmod", exact = TRUE)

if (!length(ntmod)) {
  ntmass <- aa_masses["N-term"] - 0.000549
} else {
  ntmass <- aa_masses[names(ntmod)] - 0.000549
}

if (!length(ctmod)) {
  ctmass <- aa_masses["C-term"] + 2.01510147
} else {
  ctmass <- aa_masses[names(ctmod)] + 2.01510147
}

aa_seq <- "MAKEMASSPECFUN"
aas <- stringr::str_split(aa_seq, "", simplify = TRUE)
aam <- aa_masses[aas]

b <- mzion:::bions_base(aam, ntmass)
y <- mzion:::yions_base(aam, ctmass)


# (2b, C-term)
fixedmods <- c("TMT6plex (K)", "Carbamidomethyl (C)")
varmods <- c("TMT6plex (N-term)", "Amidated (Protein C-term)", "Oxidation (M)",
             "Deamidated (N)", "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods, varmods)

# `TMT6plex (N-term)`; `Amidated (Protein C-term)`
aa_masses <- aa_masses_all[[7]]

ntmod <- attr(aa_masses, "ntmod", exact = TRUE)
ctmod <- attr(aa_masses, "ctmod", exact = TRUE)

if (!length(ntmod)) {
  ntmass <- aa_masses["N-term"] - 0.000549
} else {
  ntmass <- aa_masses[names(ntmod)] - 0.000549
}

if (!length(ctmod)) {
  ctmass <- aa_masses["C-term"] + 2.01510147
} else {
  ctmass <- aa_masses[names(ctmod)] + 2.01510147
}

b <- mzion:::bions_base(aam, ntmass)
y <- mzion:::yions_base(aam, ctmass)

}
}
