% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calc_fwxm}
\alias{calc_fwxm}
\title{Calculate FWXM}
\usage{
calc_fwxm(ys, ts, ymax, ymin, len, imax, h = 0.5)
}
\arguments{
\item{ys}{A vector of intensity values.}

\item{ts}{A vector of retention times corresponding to \code{ys}.}

\item{ymax}{The maximum of \code{ys}.}

\item{ymin}{The minimum of \code{ys}.}

\item{len}{The length of \code{ys}.}

\item{imax}{The index of \code{ymax} along \code{ys}.}

\item{h}{The height (from the apex); the height from baseline: \code{1 - h}.
The value must be \eqn{\le{.5}}.}
}
\description{
For a height \eqn{\ge{.5}}.
}
