% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calc_ms1xys}
\alias{calc_ms1xys}
\title{Makes MS1 matrices of moverzs and intensities}
\usage{
calc_ms1xys(matx, maty)
}
\arguments{
\item{matx}{A matrix of moverzs onto the data universe.}

\item{maty}{Vectors of intensities onto the data universe.}
}
\value{
A list of x: the weighted means of moverzs; y: the means of
  intensities; n: the numbers of observations.
}
\description{
Makes MS1 matrices of moverzs and intensities
}
