% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calc_ms1xys}
\alias{calc_ms1xys}
\title{Makes MS1 matrixes of moverzs and intensities}
\usage{
calc_ms1xys(xs, ys)
}
\arguments{
\item{xs}{Vectors of moverzs}

\item{ys}{Vectors of intensities}
}
\description{
Makes MS1 matrixes of moverzs and intensities
}
