% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calc_ms1xys}
\alias{calc_ms1xys}
\title{Collapse MS1 matrices of X and Y}
\usage{
calc_ms1xys(matx, maty)
}
\arguments{
\item{matx}{A matrix of m-over-z values.}

\item{maty}{A matrix of intensity values.}
}
\value{
A list of x: the weighted means of m-over-z values; y: the sum of
  intensities; n: the numbers of observations.
}
\description{
Collapse MS1 matrices of X and Y
}
