% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl1_fnl0.R
\name{calc_ms2ions_a1_vnl1_fnl0}
\alias{calc_ms2ions_a1_vnl1_fnl0}
\title{Calculates MS2 ions.}
\usage{
calc_ms2ions_a1_vnl1_fnl0(
  vmods_combi,
  vnl_combi,
  aas2,
  aa_masses,
  ntmass,
  ctmass,
  type_ms2ions,
  digits
)
}
\arguments{
\item{vmods_combi}{Lists of variable modifications.}

\item{vnl_combi}{Lists of combinations of neutral losses for corresponding
\code{vmods_combi}. Each list contains a table where each column
corresponds to a set of neutral loss. The first column corresponds to the
combination without NLs.}

\item{aas2}{A sequence of amino-acid residues with \emph{masses}. Residues
are in names and masses in values (note that argument \code{aas}
corresponds to residues without masses).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{ntmass}{The mass of a fixed or variable N-term modification.}

\item{ctmass}{The mass of a fixed or variable C-term modification.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
Calculates MS2 ions.
}
