% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepfdr}
\alias{calc_pepfdr}
\title{Calculates the cut-off score at a peptide FDR.}
\usage{
calc_pepfdr(
  out,
  nms = "rev_2",
  target_fdr = 0.01,
  fdr_type = "psm",
  min_len = 7L,
  max_len = 50L,
  out_path
)
}
\arguments{
\item{out}{The output from \link{calc_pepscores}.}

\item{nms}{The name(s) of \code{out} that correspond(s) to decoy results.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{fdr_type}{Character string; the type of FDR for controlling.}

\item{min_len}{Integer; the minimum length of peptides. Shorter peptides will
be excluded.}

\item{max_len}{Integer; the maximum length of peptides. Longer peptides will
be excluded.}

\item{out_path}{A file path of outputs.}
}
\description{
Needs \code{min_len} and \code{max_len} since the target-decoy pair may not
cover all \code{pep_len} values.
}
\examples{
\donttest{
out_path <- "~/proteoM/bi_1"

# Reload decoy
rev <- local({
  files <- list.files(path = file.path(out_path, "temp"), 
                      pattern = "^pepscores_rev_\\\\d+\\\\.rds$")
  
  if (!length(files)) stop("Decoy scores not found.", call. = FALSE)
  
  nms <- gsub("^pepscores_rev_(\\\\d+)\\\\.rds$", "\\\\1", files)
  
  ord <- order(as.integer(nms))
  nms <- nms[ord]
  files <- files[ord]
  
  out <- lapply(files, function (x) readRDS(file.path(out_path, "temp", x)))
  names(out) <- paste0("rev_", nms)
  
  out
})


# Reload targets and concatenate decoy
out <- local({
  files <- list.files(path = file.path(out_path, "temp"), 
                      pattern = "^pepscores_\\\\d+\\\\.rds$")
  
  if (!length(files)) stop("Target scores not found.", call. = FALSE)
  
  nms <- gsub("^pepscores_(\\\\d+)\\\\.rds$", "\\\\1", files)
  
  ord <- order(as.integer(nms))
  nms <- nms[ord]
  files <- files[ord]
  
  out <- lapply(files, function (x) readRDS(file.path(out_path, "temp", x)))
  names(out) <- nms
  
  out <- c(out, rev)
})

prob_cos <- calc_pepfdr(out, 
                        nms = "rev_1", 
                        target_fdr = .01, 
                        fdr_type = "protein", 
                        min_len = 7L, 
                        max_len = 50L, 
                        out_path = "~/proteoM/bi_1")

}
}
