% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepfdr}
\alias{calc_pepfdr}
\title{Calculates the cut-off score at a peptide FDR.}
\usage{
calc_pepfdr(
  target_fdr = 0.01,
  fdr_type = "psm",
  min_len = 7L,
  max_len = 40L,
  max_pepscores_co = 50,
  min_pepscores_co = 0,
  enzyme = "trypsin_p",
  fdr_group = "all",
  nes_fdr_group = "all",
  out_path
)
}
\arguments{
\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{fdr_type}{Character string; the type of FDR for controlling.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded.}

\item{max_pepscores_co}{A positive numeric; the upper limit in the cut-offs
of peptide scores for discriminating significant and insignificant
identities. The default is changed from \code{Inf} to 50 from version
1.1.9.2 on.}

\item{min_pepscores_co}{A non-negative numeric; the lower limit in the
cut-offs of peptide scores for discriminating significant and insignificant
identities.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from protein entries. The
default is \code{Trypsin_P}. See also parameter \code{custom_enzyme}.}

\item{fdr_group}{A character string; the modification group(s) for uses in
peptide FDR controls. The value is in one of c("all", "base"). The
\code{base} corresponds to the modification group with the largest number
of matches.}

\item{nes_fdr_group}{A character string in one of \code{c("all",
"all_cterm_tryptic", "all_cterm_nontryptic", "base", "base_cterm_tryptic",
"base_cterm_nontryptic")}. All peptides will be used in the classifications
of targets and decoys at \code{"all"}. Peptides with the chemistry of
C-terminal K or R will be used at \code{"all_cterm_tryptic"} (peptides from
protein C-terminals being excluded). Peptides without C-terminal K or R
will be used at \code{"all_cterm_nontryptic"}. The same applied to
\code{"base_cterm_tryptic"} and \code{"base_cterm_nontryptic"} with the
difference of only peptides from the \code{base} group being used. See also
parameter \code{fdr_group}.}

\item{out_path}{A file path of outputs.}
}
\description{
Needs \code{min_len} and \code{max_len} since the target-decoy pair may not
cover all \code{pep_len} values.
}
\examples{
\donttest{
library(proteoM)

if (FALSE) {
  prob_cos <- calc_pepfdr(target_fdr = .01, 
                          fdr_type = "protein", 
                          min_len = 7L, 
                          max_len = 50L, 
                          out_path = "~/proteoM/bi_1")
}

}
}
