% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepfdr}
\alias{calc_pepfdr}
\title{Calculates the cut-off score at a peptide FDR.}
\usage{
calc_pepfdr(
  target_fdr = 0.01,
  fdr_type = "psm",
  min_len = 7L,
  max_len = 40L,
  max_pepscores_co = 50,
  min_pepscores_co = 0,
  enzyme = "trypsin_p",
  use_nontryptic_fdr = FALSE,
  out_path
)
}
\arguments{
\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{fdr_type}{Character string; the type of FDR for controlling.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded.}

\item{max_pepscores_co}{A positive numeric; the upper limit in the cut-offs
of peptide scores for discriminating significant and insignificant
identities. The default is changed from \code{Inf} to 50 from version
1.1.9.2 on.}

\item{min_pepscores_co}{A non-negative numeric; the lower limit in the
cut-offs of peptide scores for discriminating significant and insignificant
identities.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from protein entries. The
default is \code{Trypsin_P}. See also parameter \code{custom_enzyme}.}

\item{use_nontryptic_fdr}{Logical; if TRUE, use only peptide sequences not
ending with C-terminal K or R in peptide FDR estimates. The argument only
affects searches at \code{enzyme = noenzyme}. The default is FALSE. May
consider TRUE, which may lead to more hits of non-tryptic sequences.}

\item{out_path}{A file path of outputs.}
}
\description{
Needs \code{min_len} and \code{max_len} since the target-decoy pair may not
cover all \code{pep_len} values.
}
\examples{
\donttest{
library(proteoM)

if (FALSE) {
  prob_cos <- calc_pepfdr(target_fdr = .01, 
                          fdr_type = "protein", 
                          min_len = 7L, 
                          max_len = 50L, 
                          out_path = "~/proteoM/bi_1")
}

}
}
