% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pepfdr}
\alias{calc_pepfdr}
\title{Calculates the cut-off score at a peptide FDR.}
\usage{
calc_pepfdr(
  target_fdr = 0.01,
  fdr_type = "psm",
  min_len = 7L,
  max_len = 50L,
  match_pepfdr = TRUE,
  max_pepscores_co = Inf,
  out_path
)
}
\arguments{
\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{fdr_type}{Character string; the type of FDR for controlling.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded. The default is 40.}

\item{match_pepfdr}{Logical; if TRUE, matches empirically the highest peptide
probability (corresponding to the lowest score) cut-offs to the pre-defined
level of \code{target_fdr}. The default if TRUE. Choose \code{match_pepfdr
= FALSE} for higher data quality (at a price of fewer hits).}

\item{max_pepscores_co}{A positive numeric; the upper limit in the cut-offs
of peptide scores for discriminating significant and insignificant
identities. For higher quality and data-driven thresholds, choose the
default \code{max_pepscores_co = Inf}.}

\item{out_path}{A file path of outputs.}
}
\description{
Needs \code{min_len} and \code{max_len} since the target-decoy pair may not
cover all \code{pep_len} values.
}
\examples{
\donttest{
prob_cos <- calc_pepfdr(target_fdr = .01, 
                        fdr_type = "protein", 
                        min_len = 7L, 
                        max_len = 50L, 
                        out_path = "~/proteoM/bi_1")
}
}
