% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_peploc}
\alias{calc_peploc}
\title{Calculates the delta scores of `pep_seq`.}
\usage{
calc_peploc(
  x = NULL,
  out_path = NULL,
  topn_mods_per_seq = 3L,
  topn_seqs_per_query = 3L
)
}
\arguments{
\item{x}{The result from \link{calc_pepscores}.}

\item{out_path}{An output path.}
}
\description{
A score delta between the best and the second best.
}
\details{
There should not be any duplicated rows by the combination of
c("pep_isdecoy", "scan_num", "raw_file", "pep_seq", "pep_ivmod")
}
