% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_probi}
\alias{calc_probi}
\title{Helper for score calculations}
\usage{
calc_probi(
  mts,
  expt_moverzs,
  expt_ints,
  N,
  type_ms2ions = "by",
  topn_ms2ions = 100L,
  ppm_ms2 = 25L,
  soft_secions = FALSE,
  digits = 5L
)
}
\arguments{
\item{mts}{Nested data frame of \code{theo} and matched \code{expt} m-over-z.}

\item{expt_moverzs}{Nested list of match and unmatched experimental m-over-z.}

\item{expt_ints}{Nested list of match and unmatched experimental intensity.}

\item{N}{Numeric; the number of MS2 features in an MGF query.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches. The default is to use the top-100 ions in an MS2 event.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 25.}

\item{soft_secions}{Logical; if TRUE, collapses the intensities of secondary
ions to primary ions at the absence of the primaries. The default is FALSE.
For instance, the signal of \code{b5^*} will be ignored if its primary ion
\code{b5} is not matched.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper for score calculations
}
