% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_probi_bypep}
\alias{calc_probi_bypep}
\title{Helper for score calculations}
\usage{
calc_probi_bypep(
  mts,
  nms,
  expt_moverzs,
  expt_ints,
  N = 500L,
  type_ms2ions = "by",
  topn_ms2ions = 100L,
  ppm_ms2 = 20L,
  soft_secions = FALSE,
  min_ms2mass = 115L,
  d2 = 1e-05,
  tally_ms2ints = TRUE,
  digits = 4L
)
}
\arguments{
\item{mts}{Nested data frame of \code{theo} and matched \code{expt} m-over-z.}

\item{nms}{The names (of peptides).}

\item{expt_moverzs}{Nested list of match and unmatched experimental m-over-z.}

\item{expt_ints}{Nested list of match and unmatched experimental intensity.}

\item{N}{Numeric; the number of MS2 features in an MGF query.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{soft_secions}{Logical; if TRUE, collapses the intensities of secondary
ions to primary ions even when the primaries are absent. The default is
FALSE. For instance, the signal of \code{b5^*} will be ignored if its
primary ion \code{b5} is not matched. The impacts of \code{soft_secions =
TRUE} on search performance has not yet been assessed.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{d2}{Bin width in ppm divided by 1E6.}

\item{tally_ms2ints}{Logical; tally MS2 intensities or not.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
By peptides.
}
