% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_protfdr}
\alias{calc_protfdr}
\title{Calculates the cut-offs of protein scores.}
\usage{
calc_protfdr(
  df = NULL,
  target_fdr = 0.01,
  max_protscores_co = 50,
  out_path = NULL
)
}
\arguments{
\item{df}{An output from upstream steps.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{max_protscores_co}{Numeric; the upper limit in the cut-offs of protein
scores for discriminating significant and insignificant identities.  For
higher quality and data-driven thresholds, choose \code{max_protscores_co =
Inf}.}

\item{out_path}{A file path of outputs.}

\item{outpath}{An output path.}
}
\description{
Calculates the cut-offs of protein scores.
}
