% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_protfdr}
\alias{calc_protfdr}
\title{Calculates the cut-offs of protein scores.}
\usage{
calc_protfdr(df, target_fdr = 0.01, out_path)
}
\arguments{
\item{df}{An output from upstream steps.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}

\item{out_path}{A file path of outputs.}

\item{outpath}{An output path.}
}
\description{
Calculates the cut-offs of protein scores.
}
