% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_protfdr_i}
\alias{calc_protfdr_i}
\title{Helper of \link{calc_protfdr}.}
\usage{
calc_protfdr_i(td, target_fdr = 0.01)
}
\arguments{
\item{td}{A data frame with paired target-decoys at prot_n_pep = i.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}
}
\value{
A score cut-off at a given prot_n_pep.
}
\description{
For prot_n_pep at value i.
}
