% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calcAUC}
\alias{calcAUC}
\title{Calculate area under a curve}
\usage{
calcAUC(
  ys,
  ts,
  rng,
  yco = 100,
  ytot_co = 2e+05,
  min_n = 15L,
  max_fwhm = 25,
  y_rpl = 2,
  err = 2
)
}
\arguments{
\item{ys}{Y values around an LC peak. Should contain no NA values in the
sequence.}

\item{ts}{Time values corresponding to \code{ys}.}

\item{rng}{The range indexes of \code{ys}.}

\item{yco}{An intensity cut-off.}

\item{ytot_co}{A cut-off in peak area.}

\item{min_n}{The minimum points across the peak profile of \code{ys} for
consideration.}

\item{max_fwhm}{The maximum FWHM for considering as a peak.}

\item{y_rpl}{A replacement value of intensities.}

\item{err}{An error tolerance of retention times between max and median (in
seconds).}
}
\description{
Calculate area under a curve
}
