% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calcAUC}
\alias{calcAUC}
\title{Calculate area under a curve}
\usage{
calcAUC(ys, ts)
}
\arguments{
\item{ys}{Y values around an LC peak. Should contain no NA values in the
sequence.}

\item{ts}{Time values corresponding to \code{ys}.}
}
\description{
Calculate area under a curve
}
