% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{calcFWHM}
\alias{calcFWHM}
\title{Calculate FWHM}
\usage{
calcFWHM(ys, ts, yco = 100, min_n = 15L, max_fwhm = 25, min_fwhm = 0)
}
\arguments{
\item{ys}{A vector of Y values.}

\item{ts}{A vector of T values.}

\item{yco}{The cut-off in Y values.}

\item{min_n}{The minimum number of points in \code{ys} for consideration.}

\item{max_fwhm}{The maximum FWHM for considering as a peak.}

\item{min_fwhm}{The minimum width of FWHM (in seconds) for consideration. A
value of \code{0} for uses within Mzion; otherwise for proteoQ.}
}
\description{
Do not alter \code{length(ys)} since need \code{ista} and \code{iend}.
}
