% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calcpepsc}
\alias{calcpepsc}
\title{Helper of \link{calc_pepscores}.}
\usage{
calcpepsc(
  file,
  topn_ms2ions = 100L,
  type_ms2ions = "by",
  ppm_ms2 = 20L,
  soft_secions = FALSE,
  out_path = NULL,
  min_ms2mass = 115L,
  d2 = 1e-05,
  index_mgf_ms2 = FALSE,
  add_ms2theos = FALSE,
  add_ms2theos2 = FALSE,
  add_ms2moverzs = FALSE,
  add_ms2ints = FALSE,
  n_cores = 16L,
  digits = 4L
)
}
\arguments{
\item{file}{A file name of \code{ion_matches_}.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{soft_secions}{Depreciated. Logical; if TRUE, collapses the intensities
of secondary ions to primary ions at the absence of the primaries. The
default is FALSE. For instance, the signal of \code{b5^*} will be ignored
if its primary ion \code{b5} is not matched.}

\item{out_path}{A file path of outputs.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{d2}{Bin width in ppm divided by 1E6.}

\item{index_mgf_ms2}{Logical; if TRUE, converts upfrontly MS2 m-over-z values
  from numeric to integers as opposed to in-situ conversion during ion
  matches. The default is FALSE. The \code{index_mgf_ms2 = TRUE} might be
  useful for very large MS files by reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}

\item{add_ms2theos}{Logical. If true, adds the sequence of primary
  theoretical MS2 m/z values (\code{pep_ms2_theos}). The sequence order at a
  given \code{type_ms2ions} is:

  \tabular{ll}{ \strong{Type}   \tab \strong{Sequence}\cr by \tab \eqn{b1,
  b2..., y1, y2...} \cr ax \tab \eqn{a1, a2..., x1, x2...} \cr cz \tab
  \eqn{c1, c2..., z1, z2...} \cr }}

\item{add_ms2theos2}{Logical. If true, adds the sequence of secondary
  theoretical MS2 m/z values (\code{pep_ms2_theos2}). The sequence order at a
  given \code{type_ms2ions} is:

  \tabular{ll}{ \strong{Type}   \tab \strong{Order of sequences}\cr by \tab
  \eqn{b2, b*, b*2, b0, b02, y2, y*, y*2, y0, y02} \cr ax \tab \eqn{a2, a*,
  a*2, a0, a02, x2} \cr cz \tab \eqn{c2, z2} \cr }}

\item{add_ms2moverzs}{Logical; if TRUE, adds the sequence of experimental
\eqn{m/z} values (\code{pep_ms2_moverzs}).}

\item{add_ms2ints}{Logical; if TRUE, adds the sequence of experimental MS2
intensity values (\code{pep_ms2_ints}).}

\item{n_cores}{The number of CPU cores.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper of \link{calc_pepscores}.
}
