% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calcpepsc}
\alias{calcpepsc}
\title{Helper of \link{calc_pepscores}.}
\usage{
calcpepsc(
  file,
  im_path,
  pep_fmod_all,
  pep_vmod_all,
  maxn_mdda_precurs = 1L,
  max_len = 40L,
  n_ms2_bg = 20000L,
  topn_ms2ions = 150L,
  type_ms2ions = "by",
  ppm_ms2 = 20L,
  soft_secions = FALSE,
  out_path = NULL,
  min_ms2mass = 115L,
  d2 = 1e-05,
  tally_ms2ints = TRUE,
  add_ms2theos = FALSE,
  add_ms2theos2 = FALSE,
  add_ms2moverzs = FALSE,
  add_ms2ints = FALSE,
  quant = "none",
  ppm_reporters = 10,
  digits = 4L
)
}
\arguments{
\item{file}{A file name of \code{ion_matches_}}

\item{im_path}{A temporary file path to \code{ion_matches_}}

\item{pep_fmod_all}{Attributes of \code{pep_fmod} from \code{aa_masses_all}}

\item{pep_vmod_all}{Attributes of \code{pep_vmod} from \code{aa_masses_all}}

\item{maxn_mdda_precurs}{Maximum number of precursors for consideration in a
multi-precursor DDA scan. Note that at \code{maxn_mdda_precurs = 1}, it is
equivalent to DDA with precursor re-deisotoping. At \code{maxn_mdda_precurs
= 0}, it is equivalent to DDA using the original monoisotopic masses and
charge states provided by a peak-picking algorithm (e.g., MSConvert).}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded. The default is 40.}

\item{n_ms2_bg}{The count of MS2 background noises.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{soft_secions}{Logical; if TRUE, collapses the intensities of secondary
ions to primary ions even when the primaries are absent. The default is
FALSE. For instance, the signal of \code{b5^*} will be ignored if its
primary ion \code{b5} is not matched. The impacts of \code{soft_secions =
TRUE} on search performance has not yet been assessed.}

\item{out_path}{A file path of outputs.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{d2}{Bin width in ppm divided by 1E6}

\item{tally_ms2ints}{Logical; tally MS2 intensities or not.}

\item{add_ms2theos}{Logical. If true, adds the sequence of primary
  theoretical MS2 m/z values (\code{pep_ms2_theos}). The sequence order at a
  given \code{type_ms2ions} is:

  \tabular{ll}{ \strong{Type}   \tab \strong{Sequence}\cr by \tab \eqn{b1,
  b2..., y1, y2...} \cr ax \tab \eqn{a1, a2..., x1, x2...} \cr cz \tab
  \eqn{c1, c2..., z1, z2...} \cr }}

\item{add_ms2theos2}{Logical. If true, adds the sequence of secondary
  theoretical MS2 m/z values (\code{pep_ms2_theos2}). The sequence order at a
  given \code{type_ms2ions} is:

  \tabular{ll}{ \strong{Type}   \tab \strong{Order of sequences}\cr by \tab
  \eqn{b2, b*, b*2, b0, b02, y2, y*, y*2, y0, y02} \cr ax \tab \eqn{a2, a*,
  a*2, a0, a02, x2} \cr cz \tab \eqn{c2, z2} \cr }}

\item{add_ms2moverzs}{Logical; if TRUE, adds the sequence of experimental
\eqn{m/z} values (\code{pep_ms2_moverzs}).}

\item{add_ms2ints}{Logical; if TRUE, adds the sequence of experimental MS2
intensity values (\code{pep_ms2_ints}).}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{ppm_reporters}{A positive integer; the mass tolerance of MS2 reporter
ions. The default is 10.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper of \link{calc_pepscores}.
}
