% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{calib_ms1}
\alias{calib_ms1}
\title{Calibrates precursor masses (by individual RAW_Files)}
\usage{
calib_ms1(
  filename,
  df = NULL,
  mgf_path = NULL,
  out_path = NULL,
  ppm_ms1 = 20L,
  min_mass = 200L,
  max_mass = 4500L,
  range = 3:5,
  is_tmt = FALSE
)
}
\arguments{
\item{filename}{A peaklist file name.}

\item{df}{A data frame of \code{ion_matches_1.rds}.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).}

\item{out_path}{A file path of outputs.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{range}{The range of spine knots.}

\item{is_tmt}{is a TMT experiment or not.}
}
\description{
Calibrates precursor masses (by individual RAW_Files)
}
