% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{check_aaseq}
\alias{check_aaseq}
\title{Checks the validity of a peptide sequence and dispatched it by fixedmods and
varmods.}
\usage{
check_aaseq(aa_seq, aa_masses_all, fixedmods, varmods)
}
\arguments{
\item{aa_seq}{Character string; a peptide sequences with one-letter
representation of amino acids.}

\item{aa_masses_all}{All the amino acid lookup tables.}

\item{fixedmods}{Character string(s) of fixed modifications.}

\item{varmods}{Character string(s) of variable modifications. Multiple
  modifications to the same residue are allowed, for example, both
  \code{Carbamyl (M)} and \code{Oxidation (M)}.

  For both \code{fixedmods} and \code{varmods}, the modification title,
  \code{TMT6plex}, applies to all of TMT-6, TMT-10, TMT-11. It is also
  possible to use aliased: (1) \code{TMT10plex} for TMT-10 and
  \code{TMT11plex} for TMT-11 and (2) \code{TMT16plex} for TMTpro. See also
  \link{parse_unimod} for grammars of modification \code{title},
  \code{position} and \code{site}.}
}
\description{
A sequence may be invalid at a given set of fixedmods and varmods. For
example, "MAKEMASSPECFUN" cannot have a mod of "Gln->pyro-Glu (N-term = Q)".
}
