% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_ms2ions.R
\name{check_ms2frames}
\alias{check_ms2frames}
\title{Checks for pre-existed MS2 bins.}
\usage{
check_ms2frames(.path_fasta, ms1_time, .path_cache, path_bin2)
}
\arguments{
\item{.path_fasta}{The parent file path to the theoretical masses of MS1
precursors. At the NULL default, the path is \code{gsub("(.*)\\.[^\\.]*$",
"\\1", get("fasta", envir = environment())[1])}. The parameter is for
users' awareness of the structure of file folders and the default is
suggested.}

\item{ms1_time}{A cached MS1 time (directory).}

\item{.path_cache}{The file path of cached search parameters. At the NULL
default, the path is \code{"~/proteoM/.MSearches/Cache/Calls/"}. The
parameter is for users' awareness of the structure of file folders and the
default is suggested.}

\item{path_bin2}{A file path to binned MS2 results.}
}
\description{
Checks for pre-existed MS2 bins.
}
