% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{check_tmt_pars}
\alias{check_tmt_pars}
\title{Checks the compatibility of TMT names and plexes.}
\usage{
check_tmt_pars(fixedmods, varmods, quant)
}
\arguments{
\item{fixedmods}{A character vector of fixed modifications. See also
\link{parse_unimod} for grammars.}

\item{varmods}{A character vector of variable modifications.}

\item{quant}{A quantitation method. The default is "none". Additional choices
include \code{tmt6} etc. For other multiplicities of \code{tmt}, use the
compatible higher plexes, for example, \code{tmt16} for \code{tmt12} etc.
and \code{tmt10} for \code{tmt8} etc.}
}
\description{
Checks the compatibility of TMT names and plexes.
}
