% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{check_tmt_pars}
\alias{check_tmt_pars}
\title{Checks the compatibility of TMT names and plexes.}
\usage{
check_tmt_pars(fixedmods, varmods, quant)
}
\arguments{
\item{fixedmods}{Character string(s) of fixed modifications.}

\item{varmods}{Character string(s) of variable modifications. Multiple
  modifications to the same residue are allowed, for example, both
  \code{Carbamyl (M)} and \code{Oxidation (M)}.

  For both \code{fixedmods} and \code{varmods}, the modification title,
  \code{TMT6plex}, applies to all of TMT-6, TMT-10, TMT-11. It is also
  possible to use aliased: (1) \code{TMT10plex} for TMT-10 and
  \code{TMT11plex} for TMT-11 and (2) \code{TMT16plex} for TMTpro. See also
  \link{parse_unimod} for grammars of modification \code{title},
  \code{position} and \code{site}.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}
}
\description{
Checks the compatibility of TMT names and plexes.
}
