% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{collapse_pasef_xys}
\alias{collapse_pasef_xys}
\title{Collapse PASEF X and Y values}
\usage{
collapse_pasef_xys(xs, ys)
}
\arguments{
\item{xs}{Lists of m-over-z vectors.}

\item{ys}{Lists of intensity vectors.}
}
\description{
Collapse PASEF X and Y values
}
