% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{collapse_vecs}
\alias{collapse_vecs}
\title{Pastes vectors to character strings.}
\usage{
collapse_vecs(vecs, nm = "theo", sep = ";", digits = 4L)
}
\arguments{
\item{vecs}{A list of vectors.}

\item{nm}{The name of sub list in \code{vecs}.}

\item{sep}{A separator.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Pastes vectors to character strings.
}
