% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{collapse_xyz}
\alias{collapse_xyz}
\title{Gather MS data}
\usage{
collapse_xyz(
  xs = NULL,
  ys = NULL,
  zs = NULL,
  temp_dir = NULL,
  icenter = 1L,
  ms_lev = 2L,
  lwr = 115L,
  step = 1e-05,
  cleanup = TRUE,
  direct_out = FALSE
)
}
\arguments{
\item{xs}{Vectors of m-over-z values within a given isolation center
(\code{icenter}). Each entry corresponds to a full vector of MS1 or MS2.}

\item{ys}{Vectors of intensities.}

\item{zs}{Vectors of charge states.}

\item{temp_dir}{A temporary directory.}

\item{icenter}{The index of an isolation center.}

\item{ms_lev}{The level of MS.}

\item{lwr}{A lower bound as the starting point in mass binning.}

\item{step}{A step size for mass binning.}

\item{cleanup}{Logical; cleans up xs, ys and zs or not. Set the value to
FALSE to maintain one-to-one correspondence between input (data frame) and
the outputs. This will help, e.g., keep track of scan numbers in the input.}

\item{direct_out}{Logical; if TRUE, return outputs directly.}
}
\value{
Matrices of moverzs, intensities and charge states. Columns: masses;
  rows: scans.
}
\description{
Similar to \link{collapse_mms1ints} with the addition of zs.
}
\details{
Values of \code{xs} can be smaller than \code{lwr} since no ms2_mass cut-offs
(some z undetermined).
}
