% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{collapse_xyz}
\alias{collapse_xyz}
\title{Collapse MS data.}
\usage{
collapse_xyz(xs = NULL, ys = NULL, zs = NULL, lwr = 115L, step = 1e-05)
}
\arguments{
\item{xs}{Vectors of SORTED m-over-z values.}

\item{ys}{Vectors of intensity values corresponding to xs.}

\item{zs}{Vectors of charge states.}

\item{lwr}{The lower mass limit.}

\item{step}{The bin size in converting numeric m-over-z values to integers.}
}
\description{
Almost identical to \link{collapse_mms1ints} with the addition of zs.
}
