% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{deisoDDAMS2}
\alias{deisoDDAMS2}
\title{Deisotopes DDA-MS2}
\usage{
deisoDDAMS2(
  ms2_moverzs,
  ms2_ints,
  topn_ms2ions = 150L,
  ppm_ms2_deisotope = 8L,
  max_ms2_charge = 3L,
  n_fwd = 10L,
  n_bwd = 10L,
  offset_upr = 30L,
  offset_lwr = 30L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  quant = "none",
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  n_para = 1L
)
}
\arguments{
\item{ms2_moverzs}{MS2 m-over-z values.}

\item{ms2_ints}{MS2 intensities.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{ppm_ms2_deisotope}{Mass error tolerance in MS2 deisotoping.}

\item{max_ms2_charge}{Maximum charge states for consideration with MS2
deisotoping.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries. The default is
20 for MS1 and 10 for MS2.}

\item{n_bwd}{Backward looking up to \code{n_bwd} mass entries.}

\item{offset_upr}{A cardinal number of upper mass off-sets.}

\item{offset_lwr}{A cardinal number of lower mass off-sets.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{A multiplication factor for the fuzzy discrimination of a
secondary precursor in parallel to a primary precursor in an isotopic
envelop. The smaller the value, the more stringent it is in calling an
adjacent peak being a mono-isotopic precursor}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{n_para}{The allowance of parallel processing.}
}
\description{
Deisotopes DDA-MS2
}
