% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotope.R
\name{deisotope}
\alias{deisotope}
\title{De-isotopes precursor masses.}
\usage{
deisotope(
  moverzs,
  msxints,
  exclude_reporter_region = FALSE,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  center = 650,
  ppm = 6L,
  ms_lev = 1L,
  maxn_feats = 300L,
  max_charge = 4L,
  n_fwd = 20L,
  offset_upr = 30L,
  offset_lwr = 30L,
  order_mz = TRUE,
  step = ppm/1e+06,
  backward_mass_co = 800/ms_lev,
  iso_ratio = 5,
  bound = FALSE
)
}
\arguments{
\item{moverzs}{Mass-to-charge ratios.}

\item{msxints}{MS1 or MS2 peak intensities.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{center}{The mass center of an isolation window (only for MS1).}

\item{ppm}{Allowance in mass error when deisotoping.}

\item{ms_lev}{MS level.}

\item{maxn_feats}{The maximum number of MS features.}

\item{max_charge}{The maximum charge state.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries.}

\item{offset_upr}{A cardinal number of upper mass off-sets.}

\item{offset_lwr}{A cardinal number of lower mass off-sets.}

\item{order_mz}{Logical; if TRUE, orders peaks from low to high m-over-z's.}

\item{step}{Step size for mass binning.}

\item{backward_mass_co}{A mass cut-off to initiate backward looking of an
isotope envelop.}

\item{iso_ratio}{Ratio threshold between two adjacent isotope peaks
(current/before).}

\item{bound}{Not yet used. Logical; if TRUE, removes precursors outside of
the boundary of isolation window.}
}
\description{
De-isotopes precursor masses.
}
\examples{
\donttest{
library(mzion)
moverzs <- c(881 + 1:10*.1, 882.0674, 882.0981, 882.4034, 882.60, 882.7372)
msxints <- c(1000 * 1:10, 1652869, 882.0981, 2043015, 2314111, 4314111)

# out <- mzion:::deisotope(moverzs, msxints, ppm = 10L, ms_lev = 1L,
#                          maxn_feats = 5L, max_charge = 4L, offset_upr = 8L,
#                          offset_lwr = 8L, order_mz = TRUE, bound = FALSE)
}
}
