% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{distri_peps}
\alias{distri_peps}
\title{Distributes peptides by variable modifications.}
\usage{
distri_peps(peps, aa_masses_all, max_miss)
}
\arguments{
\item{peps}{Lists of peptides, either "fwds" or "revs", from
\link{split_fastaseqs}.}

\item{aa_masses_all}{All the amino acid lookup tables.}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}
}
\description{
Distributes peptides by variable modifications.
}
