% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{distri_peps}
\alias{distri_peps}
\title{Distributes peptides by variable modifications.}
\usage{
distri_peps(prps, aa_masses_all, max_miss = 2L)
}
\arguments{
\item{prps}{Lists of peptide sequences with a one-letter representation
of amino acid residues. Each list is named by protein accession.}

\item{aa_masses_all}{All the amino acid look-up tables.}

\item{max_miss}{The maximum number of mis-cleavages per peptide sequence.}
}
\description{
Distributes peptides by variable modifications.
}
