% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfilereader.R
\name{exeReadRAW}
\alias{exeReadRAW}
\title{Helper in the execution of CSharp-compiled ReadRAW.exe}
\usage{
exeReadRAW(raw_file, mgf_path, type = "scans")
}
\arguments{
\item{raw_file}{A RAW file name.}

\item{mgf_path}{A file path to peak lists or RAW MS data.}

\item{type}{The type of MSFileReader request.}
}
\description{
Helper in the execution of CSharp-compiled ReadRAW.exe
}
