% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{expand_grid_rows}
\alias{expand_grid_rows}
\title{Expands grids.}
\usage{
expand_grid_rows(..., nmax = 3L, use.names = TRUE)
}
\arguments{
\item{...}{Lists of data.}

\item{nmax}{The maximum number of combinations allowed.}

\item{use.names}{Logical; uses names or not.}
}
\description{
Outputs are vectors corresponding to rows in the the data.frame from the
original expand.grid.
}
\details{
Some overhead in making row vectors but can avoid the expensive row
subsetting from a data.frame.
}
\examples{
library(mzion)

x <- list(`Oxidation (M)` = c(0.000000, 63.998285),
          `Carbamidomethyl (M)` = c(0.000000, 105.024835),
          `Oxidation (M)` = c(0.000000, 63.998285))

mzion:::expand_grid_rows(x)

x <- list(`Bar (M)` = c(0, 3),
          `Foo (M)` = c(0, 5, 7),
          `Bar (M)` = c(0, 3))

mzion:::expand_grid_rows(x)

x <- list(`Bar (M)` = c(0, 3))
mzion:::expand_grid_rows(x)
}
