% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{extract_pasef_frame}
\alias{extract_pasef_frame}
\title{Extracts data from a PASEF frame.}
\usage{
extract_pasef_frame(data)
}
\arguments{
\item{data}{A PASEF frame (with multiple scans).}
}
\description{
Extracts data from a PASEF frame.
}
