% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{fill_probco_nas}
\alias{fill_probco_nas}
\title{Fills NA in probability cutoffs}
\usage{
fill_probco_nas(
  prob_nas = NULL,
  prob_no_uses = NULL,
  prob_cos,
  target_fdr = 0.01
)
}
\arguments{
\item{prob_nas}{Named vector; peptide length in names and NA in values.}

\item{prob_no_uses}{Named vector; peptide length in names and NA in values.}

\item{prob_cos}{Named vector; non-NA probability cutoffs.}

\item{target_fdr}{Numeric; the levels of false-discovery rate (FDR).}
}
\description{
Fills NA in probability cutoffs
}
