% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{finalize_aamasses}
\alias{finalize_aamasses}
\title{Finalizes \code{aa_masses_all}}
\usage{
finalize_aamasses(aa_masses_i, aa_masses, varlabs = NULL, mod_motifs = NULL)
}
\arguments{
\item{aa_masses_i}{The i-th aa_masses_all.}

\item{aa_masses}{The original look-ups of AA masses.}

\item{varlabs}{Character string(s) of variable isotopic labels. See examples
of SILAC for details. Can be but not typically used in standard alone
searches of labeled residues.}

\item{mod_motifs}{The motifs to restrict \code{Anywhere} variable
  modification. For example, provided the \code{Anywhere} variable
  modifications containing \code{c("Oxidation (M)", "Deamidated (N)")} and

  \code{mod_motifs = list(`Deamidated (N)` = c("NG", "NM"), `Oxidation (M)` =
  c("NM", "MP"))}

  variable modifications will only be considered at sites that satisfy the
  motifs.}
}
\description{
Replaces interim fixed and variable modifications with the finals. Results in
correct attributes in aa_masses_i.
}
