% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{find_aamasses_vmodscombi}
\alias{find_aamasses_vmodscombi}
\title{Finds the combination of varmods}
\usage{
find_aamasses_vmodscombi(
  varmods = NULL,
  fixednt_coercion = TRUE,
  fixedct_coercion = FALSE,
  anywhere_coerce_sites = NULL
)
}
\arguments{
\item{varmods}{A character vector of variable modifications.}

\item{fixednt_coercion}{Logical; with coercion of N-term fixed to variable
modifications.}

\item{fixedct_coercion}{Logical; with coercion of C-term fixed to variable
modifications.}

\item{anywhere_coerce_sites}{\code{Anywhere} sites coerced from fixed to
variable modifications.}
}
\description{
Finds the combination of varmods
}
