% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{find_best_apex}
\alias{find_best_apex}
\title{Find tentatively the apex under a mass column}
\usage{
find_best_apex(
  xvs,
  yvs,
  ss,
  yints,
  aps,
  rts,
  rngs,
  fwhms,
  xref,
  ret_ms1,
  step_tr = 6e-06,
  min_y = 0,
  min_n1 = 10L,
  min_n2 = 20L,
  min_n3 = 15L,
  max_scan_delta = 3500,
  max_rt_delta = 180
)
}
\arguments{
\item{xvs}{X values.}

\item{yvs}{Y values.}

\item{ss}{All of the scan numbers.}

\item{yints}{Peak areas under under each gate.}

\item{aps}{All apexes scan numbers under a mass column.}

\item{rts}{All apexes retention times under a mass column.}

\item{rngs}{The scan ranges of apexes under column k.}

\item{fwhms}{All FWHMs under a column.}

\item{xref}{The experimental X value for tracing.}

\item{ret_ms1}{The preceding MS1 retention time of an MS2 event at
\code{xref}. Or may be the MS2 retention time of \code{xref}.}

\item{step_tr}{A step size.}

\item{min_y}{The cut-off of intensity values.}

\item{max_scan_delta}{The maximum difference in scan numbers between an apex
and \code{scan_ms1}.}

\item{max_rt_delta}{The maximum difference in retention times between an apex
and \code{ret_ms1}.}
}
\description{
Also remove low-quality apexes.
}
