% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{find_callarg_vals}
\alias{find_callarg_vals}
\title{Finds the values of a list of arguments.}
\usage{
find_callarg_vals(
  time = NULL,
  path = NULL,
  fun = NULL,
  args = NULL,
  new_args = NULL
)
}
\arguments{
\item{time}{The time stamp.}

\item{path}{A (system) file path.}

\item{fun}{The name of function being saved.}

\item{args}{Arguments to be matched.}

\item{new_args}{Named vector; new arguments that are not in earlier versions.}
}
\value{
The time stamp of a matched cache results.
}
\description{
Finds the values of a list of arguments.
}
