% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_cterm_mass}
\alias{find_cterm_mass}
\title{Finds C-terminal mass.}
\usage{
find_cterm_mass(aa_masses)
}
\arguments{
\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
Not yet used.
}
