% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{find_free_mem}
\alias{find_free_mem}
\title{Finds the amount of free system memory.}
\usage{
find_free_mem(sys_ram = 32L)
}
\arguments{
\item{sys_ram}{The amount of system RAM; only for uses with Linux or Mas OS.}
}
\description{
Outputs of free RAM in the unit of MB.
}
