% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_gate_edges}
\alias{find_gate_edges}
\title{Helper to find the positions of logical gates.}
\usage{
find_gate_edges(vals)
}
\arguments{
\item{vals}{A logical vector.}
}
\description{
Helper to find the positions of logical gates.
}
