% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{find_group_breaks}
\alias{find_group_breaks}
\title{Chunksplits by groups.}
\usage{
find_group_breaks(vec, n_chunks = 5L)
}
\arguments{
\item{vec}{A sorted vector.}

\item{n_chunks}{The number of chunks}
}
\description{
Chunksplits by groups.
}
\examples{
\donttest{
vec <- rep(LETTERS[1:5], 1:5)
vec <- sort(vec)

find_group_breaks(vec, 3)
}
}
