% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a1_vnl0_fnl0.R
\name{find_intercombi_p2}
\alias{find_intercombi_p2}
\title{Finds the combinations of positions and sites across residues.}
\usage{
find_intercombi_p2(intra_combis, maxn_vmods_sitescombi_per_pep = 32L)
}
\arguments{
\item{intra_combis}{Inter-residue combinations.}

\item{maxn_vmods_sitescombi_per_pep}{Integer; the maximum number of
combinatorial variable modifications per peptide sequence. The default is
64. May consider a smaller value, i.e. 32, when searching against
phosphopeptide data.}
}
\description{
For multiple residues (each residue one to multiple modifications).
}
\examples{
N <- list(c(`7` = "Deamidated (N)"), 
          c(`15` = "Deamidated (N)"), 
          c(`7` = "Deamidated (N)", `15` = "Deamidated (N)"))

S <- c(`16` = "Carbamidomethyl (S)")

intra_combis <- list(N = N, S = S)

ans <- find_intercombi_p2(intra_combis)
}
