% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_lc_gates}
\alias{find_lc_gates}
\title{Finds the gates of retention times.}
\usage{
find_lc_gates(ys, n_dia_scans = 4L)
}
\arguments{
\item{ys}{A vector of intensity value for the same (approximate) mass along
LC.}

\item{n_dia_scans}{The number of adjacent MS scans for constructing a peak
profile and thus for determining the apex scan number of an moverz value
along LC.}
}
\value{
Scan indexes of LC peaks.
}
\description{
Finds the gates of retention times.
}
\examples{
library(mzion)

find_lc_gates(c(10,0,0,0,11,15,15,0,0,12,0,10,0,0,10), 2)

# find_lc_gates(c(rep(0, 7), 100, 101, rep(0, 2), seq(200, 500, 100), rep(0, 1), 20, 50))
# find_lc_gates(c(rep(0, 7), 100, 101, rep(0, 2), seq(200, 500, 100), rep(0, 4), 20, 50))

# all discrete
# find_lc_gates(c(rep(0, 5), 100, rep(0, 6), 200, rep(0, 4), 50))
}
