% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_lc_gates}
\alias{find_lc_gates}
\title{Finds the gates of retention times.}
\usage{
find_lc_gates(ys, gap = 4L)
}
\arguments{
\item{ys}{A vector of corresponding signal intensities.}

\item{gap}{The size of gap for LC peaks collapsion.}
}
\description{
Finds the gates of retention times.
}
\examples{
library(mzion)

# find_lc_gates(c(rep(0, 7), 100, 101, rep(0, 2), seq(200, 500, 100), rep(0, 1), 20, 50))
# find_lc_gates(c(rep(0, 7), 100, 101, rep(0, 2), seq(200, 500, 100), rep(0, 4), 20, 50))

# all discrete
# find_lc_gates(c(rep(0, 5), 100, rep(0, 6), 200, rep(0, 4), 50))
}
