% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_lc_gates}
\alias{find_lc_gates}
\title{Finds the gates of retention times.}
\usage{
find_lc_gates(
  xs = NULL,
  ys = NULL,
  ts = NULL,
  n_dia_scans = 4L,
  y_rpl = 2,
  step = 5e-06
)
}
\arguments{
\item{xs}{A vector of m-over-z value for the same (approximate) mass along
LC.}

\item{ys}{A vector of intensity value for the same (approximate) mass along
LC.}

\item{ts}{The retention times corrresponding to \code{ys}.}

\item{n_dia_scans}{The number of adjacent MS scans for constructing a peak
profile and thus for determining the apex scan number of an moverz value
along LC.}

\item{y_rpl}{A replacement value of intensities.}
}
\value{
Scan indexes of LC peaks.
}
\description{
Finds the gates of retention times.
}
\examples{
mzion:::find_lc_gates(xs = rep_len(100, 15), ys = c(10,0,0,0,11,15,15,0,0,12,0,10,0,0,10), ts = seq_len(15), n_dia_scans = 2)
mzion:::find_lc_gates(xs = rep_len(100, 18), ys = c(rep(0, 7), 100, 101, rep(0, 2), seq(200, 500, 100), rep(0, 1), 20, 50), ts = seq_len(18))
mzion:::find_lc_gates(xs = rep_len(100, 21), ys = c(rep(0, 7), 100, 101, rep(0, 2), seq(200, 500, 100), rep(0, 4), 20, 50), ts = seq_len(21))

# all discrete
mzion:::find_lc_gates(xs = rep_len(100, 18), ys = c(rep(0, 5), 100, rep(0, 6), 200, rep(0, 4), 50), ts = seq_len(18))

# two passes
ys <- c(rep_len(0, 10), rep_len(100, 50), rep_len(0, 10), rep_len(100, 50), rep_len(3, 5), rep_len(50, 80), rep_len(0, 10))
xs <- rep(5, length(ys))
ts <- seq_along(ys)
mzion:::find_lc_gates(xs, ys, ts)
}
