% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{find_lc_gates2}
\alias{find_lc_gates2}
\title{Find peak edges by intensity threshold relative to the base peak}
\usage{
find_lc_gates2(perc = 0.02, ys, sta, n_dia_scans = 4L, y_rpl = 2, min_n = 5L)
}
\arguments{
\item{perc}{The percentage to the base-peak intensity for cut-offs.}

\item{ys}{A vector of intensity values.}

\item{sta}{The starting point (off-set) of the current ys (ys[ri]) in the
whole ys.}

\item{n_dia_scans}{The number of adjacent MS scans for constructing a peak
profile and thus for determining the apex scan number of an moverz value
along LC.}

\item{y_rpl}{A replacement value of intensities.}

\item{min_n}{The minimum number of \code{ys} to define an LC gate.}
}
\description{
Find peak edges by intensity threshold relative to the base peak
}
