% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_mdda_mms1s}
\alias{find_mdda_mms1s}
\title{Finds mDDA precursors.}
\usage{
find_mdda_mms1s(
  df1,
  df2,
  stas1,
  stas2,
  ends2,
  ppm = 10L,
  maxn_precurs = 5L,
  max_ms1_charge = 4L,
  n_fwd = 20L,
  grad_isotope = 2.5,
  width = 2.01,
  step = ppm/1e+06
)
}
\arguments{
\item{df1}{A sub data frame of MS1.}

\item{df2}{A sub data frame of MS2.}

\item{stas1}{Indexes of MS1 starting lines (in the full data frame).}

\item{stas2}{Indexes of MS2 starting lines.}

\item{ends2}{Indexes of MS2 ending lines.}

\item{ppm}{Mass error tolerance.}

\item{maxn_precurs}{Maximum number of precursors for consideration.}

\item{max_ms1_charge}{Maximum charge state of precursors for consideration.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries.}

\item{width}{The width of an MS1 window. A wide window is used for containing
isotope envelops.}

\item{step}{The bin size in converting numeric m-over-z values to integers.}
}
\description{
Averages of multiple MS1 scans.
}
