% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_mdda_mms1s}
\alias{find_mdda_mms1s}
\title{Finds mDDA precursors.}
\usage{
find_mdda_mms1s(
  df1,
  df2,
  ppm = 10L,
  maxn_precurs = 5L,
  max_ms1_charge = 4L,
  n_fwd = 20L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  use_defpeaks = FALSE,
  width = 2.01,
  step = ppm/1e+06
)
}
\arguments{
\item{df1}{A sub data frame of MS1.}

\item{df2}{A sub data frame of MS2.}

\item{ppm}{Mass error tolerance.}

\item{maxn_precurs}{Maximum number of precursors for consideration.}

\item{max_ms1_charge}{Maximum charge state of precursors for consideration.}

\item{n_fwd}{Forward looking up to \code{n_fwd} mass entries.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{A multiplication factor for the fuzzy discrimination of a
secondary precursor in parallel to a primary precursor in an isotopic
envelop. The smaller the value, the more stringent it is in calling an
adjacent peak being a mono-isotopic precursor}

\item{use_defpeaks}{Use default peak info or not.}

\item{width}{The width of an MS1 window. A wide window is used for containing
isotope envelops.}
}
\description{
Averages of multiple MS1 scans.
}
