% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{find_mdda_ms1s}
\alias{find_mdda_ms1s}
\title{Finds chimeric precursors.}
\usage{
find_mdda_ms1s(
  df1,
  df2,
  stas1,
  stas2,
  ends2,
  ppm = 6L,
  maxn_precurs = 5L,
  max_ms1_charge = 4L
)
}
\arguments{
\item{df1}{A one-row data frame of precursor data.}

\item{df2}{The data frame of MS2 corresponding to df1.}

\item{stas1}{The indexes of df1 starting lines.}

\item{stas2}{The indexes of df2 starting lines.}

\item{ends2}{The indexes of df2 ending lines.}

\item{ppm}{Mass error tolerance.}

\item{maxn_precurs}{The maximum number of chimeric precursors.}

\item{max_ms1_charge}{The maximum charge state of precursors.}
}
\description{
Not used; by single MS1 scan.
}
