% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_base.R
\name{find_ms2_bypep}
\alias{find_ms2_bypep}
\title{Helper: matches between theoretical and experimental MS2 ions.}
\usage{
find_ms2_bypep(
  theos = NULL,
  expts = NULL,
  ppm_ms2 = 25L,
  min_ms2mass = 110L,
  minn_ms2 = 6L
)
}
\arguments{
\item{theos}{Numeric vector; one to multiple series of theoretical MS2s.}

\item{expts}{Numeric vector; one series of experimental MS2s.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}
}
\description{
Run time about 45 us.
}
\examples{
\donttest{
# experimental `322.18704` fit to both b- and y- ions
expts <- c(101.07140,102.05540,107.04956,110.07165,111.07500,
           112.08729,113.07130,115.08693,116.07092,120.08105,
           121.08428,126.12794,127.12494,127.13121,128.12825,
           128.13452,129.13164,129.13789,130.06493,130.09772,
           130.13495,130.14124,131.13831,132.14160,134.07635,
           136.07573,157.10826,158.09238,159.09117,170.12067,
           173.12846,173.14980,175.11896,176.12245,176.15956,
           184.11806,186.15303,188.15977,190.09743,193.63914,
           207.11292,210.12289,229.16670,230.17030,231.17410,
           235.10779,240.14383,248.18073,262.15305,265.67874,
           273.21210,285.13416,301.20700,305.16055,312.17740,
           314.69138,322.18704,365.23856,369.24496,371.70316,
           374.18283,376.27573,392.19308,393.23337,394.23743,
           399.20920,400.27567,400.72501,401.22650,401.27139,
           401.58405,401.72778,409.21918,410.22241,423.24564,
           433.29709,452.27066,462.25473,480.26578,481.26828,
           498.27670,530.35126,572.28278,573.28625,599.33899,
           600.34039,609.32202,626.29218,627.29303,627.33948,
           628.33417,629.30463,630.30219,643.31946,644.31763,
           644.35913,645.32520,646.32880,647.32825,648.32892)

theos <- c(114.0913,251.1503,322.1874,423.2350,480.2565,
           627.3249,783.4260,175.1190,322.1874,379.2088,
           480.2565,551.2936,688.3525,801.4366)

pep <- "ABCDEFG"
nms <- unlist(stringr::str_split(pep, ""))

names(theos) <- c(nms, rev(nms))
theos <- list(`0000000` = theos)

x <- find_ms2_bypep(theos, expts)

## Both expts 74953 and 74955 fit to theos 74954
pep <- "DIAVEEDLSSTPLFKDLLALMR"
nms <- unlist(stringr::str_split(pep, ""))

theos <- c(158.0448,271.1288,342.1660,441.2344,570.2770,699.3196,
           814.3465,927.4306,1094.4289,1181.4610,1282.5086,1379.5614,
           1492.6455,1639.7139,1996.9718,2111.9987,2225.0828,2338.1668,
           2409.2040,2522.2880,2653.3285,2809.4296,175.1190,306.1594,
           419.2435,490.2806,603.3647,716.4487,831.4757,1188.7336,
           1335.8020,1448.8861,1545.9388,1646.9865,1734.0185,1901.0169,
           2014.1010,2129.1279,2258.1705,2387.2131,2486.2815,2557.3186,
           2670.4027,2785.4296)
names(theos) <- c(nms, rev(nms))
theos <- list(`0000000070000000000000 (1)` = theos)

expts <- c(126.12768, 127.13107, 128.12868, 128.13484, 130.13541, 
           130.14117, 136.07610, 167.08173, 178.27960, 228.13425, 
           230.17053, 238.11922, 248.18092, 256.12946, 257.12473, 
           276.10172, 278.15823, 283.14044, 321.21228, 327.13010, 
           358.71371, 368.22955, 376.27615, 394.73138, 396.15076, 
           396.22488, 400.27673, 407.24142, 414.16251, 426.76105, 
           445.29724, 447.31360, 458.76022, 486.28998, 490.79019, 
           491.29218, 500.26715, 514.27716, 516.33514, 522.78906, 
           523.28961, 528.33496, 535.27863, 543.27417, 549.31561, 
           572.32416, 572.39752, 572.82666, 576.35638, 603.31360, 
           603.36700, 604.36945, 619.36670, 631.30743, 632.30969, 
           641.41919, 642.42230, 646.36780, 647.36938, 658.36029, 
           675.03418, 675.36835, 681.03772, 681.37146, 690.27856, 
           701.34332, 707.45813, 707.69342, 708.02728, 716.41888, 
           716.45227, 726.35089, 726.85413, 760.40997, 787.48865, 
           788.45386, 789.45624, 796.44757, 813.47363, 814.47614, 
           815.48041, 874.51715, 906.91064, 907.41144, 916.51337, 
           918.38928, 919.39429, 926.55829, 953.56543, 971.57239, 
           972.57562, 1031.47534, 1044.57398, 1069.55029, 1085.45569, 
           1143.64014, 1144.64404, 1214.67590, 1321.77869, 1322.78186)

x <- find_ms2_bypep(theos, expts)
}
}
