% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_base.R
\name{find_ms2_bypep}
\alias{find_ms2_bypep}
\title{Helper: matches between theoretical and experimental MS2 ions.}
\usage{
find_ms2_bypep(
  theos = NULL,
  expts = NULL,
  ex = NULL,
  d = NULL,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  minn_ms2 = 6L
)
}
\arguments{
\item{theos}{Numeric vector; one to multiple series of theoretical MS2s.}

\item{expts}{Numeric vector; one series of experimental MS2s.}

\item{ex}{Converted expts as integers.}

\item{d}{ppm_ms2 divided by 1E6.}

\item{ppm_ms2}{The mass tolerance of MS2 species.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{minn_ms2}{Integer; the minimum number of MS2 ions for consideration as
a hit.}
}
\value{
Lists of (1) theo, (2) expt, (3) ith, (4) iex and (5) m.
}
\description{
Helper: matches between theoretical and experimental MS2 ions.
}
\examples{
\donttest{
## Experimental 322.18704 fit to both b- and y- ions
#  (one expt to multiple theos)
expts <- c(101.07140,102.05540,107.04956,110.07165,111.07500,
           112.08729,113.07130,115.08693,116.07092,120.08105,
           121.08428,126.12794,127.12494,127.13121,128.12825,
           128.13452,129.13164,129.13789,130.06493,130.09772,
           130.13495,130.14124,131.13831,132.14160,134.07635,
           136.07573,157.10826,158.09238,159.09117,170.12067,
           173.12846,173.14980,175.11896,176.12245,176.15956,
           184.11806,186.15303,188.15977,190.09743,193.63914,
           207.11292,210.12289,229.16670,230.17030,231.17410,
           235.10779,240.14383,248.18073,262.15305,265.67874,
           273.21210,285.13416,301.20700,305.16055,312.17740,
           314.69138,322.18704,365.23856,369.24496,371.70316,
           374.18283,376.27573,392.19308,393.23337,394.23743,
           399.20920,400.27567,400.72501,401.22650,401.27139,
           401.58405,401.72778,409.21918,410.22241,423.24564,
           433.29709,452.27066,462.25473,480.26578,481.26828,
           498.27670,530.35126,572.28278,573.28625,599.33899,
           600.34039,609.32202,626.29218,627.29303,627.33948,
           628.33417,629.30463,630.30219,643.31946,644.31763,
           644.35913,645.32520,646.32880,647.32825,648.32892)

theos <- c(114.0913,251.1503,322.1874,423.2350,480.2565,
           627.3249,783.4260,175.1190,322.1874,379.2088,
           480.2565,551.2936,688.3525,801.4366)

ppm_ms2 <- 25L
min_ms2mass <- 115L

d <- ppm_ms2/1E6
ex <- ceiling(log(expts/min_ms2mass)/log(1+d))

pep <- "PEPTIDE"
nms <- unlist(stringr::str_split(pep, ""))

names(theos) <- c(nms, rev(nms))
theos <- list(`0000000` = theos)

x1 <- find_ms2_bypep(theos, expts, ex, d)

## Both expts 74953 and 74955 fit to theos 74954
#  (multiple expts to one theo)
pep <- "DIAVEEDLSSTPLFKDLLALMR"
nms <- unlist(stringr::str_split(pep, ""))

theos <- c(158.0448,271.1288,342.1660,441.2344,570.2770,699.3196,
           814.3465,927.4306,1094.4289,1181.4610,1282.5086,1379.5614,
           1492.6455,1639.7139,1996.9718,2111.9987,2225.0828,2338.1668,
           2409.2040,2522.2880,2653.3285,2809.4296,175.1190,306.1594,
           419.2435,490.2806,603.3647,716.4487,831.4757,1188.7336,
           1335.8020,1448.8861,1545.9388,1646.9865,1734.0185,1901.0169,
           2014.1010,2129.1279,2258.1705,2387.2131,2486.2815,2557.3186,
           2670.4027,2785.4296)
names(theos) <- c(nms, rev(nms))
theos <- list(`0000000070000000000000 (1)` = theos)

expts <- c(126.12768, 127.13107, 128.12868, 128.13484, 130.13541,
           130.14117, 136.07610, 167.08173, 178.27960, 228.13425,
           230.17053, 238.11922, 248.18092, 256.12946, 257.12473,
           276.10172, 278.15823, 283.14044, 321.21228, 327.13010,
           358.71371, 368.22955, 376.27615, 394.73138, 396.15076,
           396.22488, 400.27673, 407.24142, 414.16251, 426.76105,
           445.29724, 447.31360, 458.76022, 486.28998, 490.79019,
           491.29218, 500.26715, 514.27716, 516.33514, 522.78906,
           523.28961, 528.33496, 535.27863, 543.27417, 549.31561,
           572.32416, 572.39752, 572.82666, 576.35638, 603.31360,
           603.36700, 604.36945, 619.36670, 631.30743, 632.30969,
           641.41919, 642.42230, 646.36780, 647.36938, 658.36029,
           675.03418, 675.36835, 681.03772, 681.37146, 690.27856,
           701.34332, 707.45813, 707.69342, 708.02728, 716.41888,
           716.45227, 726.35089, 726.85413, 760.40997, 787.48865,
           788.45386, 789.45624, 796.44757, 813.47363, 814.47614,
           815.48041, 874.51715, 906.91064, 907.41144, 916.51337,
           918.38928, 919.39429, 926.55829, 953.56543, 971.57239,
           972.57562, 1031.47534, 1044.57398, 1069.55029, 1085.45569,
           1143.64014, 1144.64404, 1214.67590, 1321.77869, 1322.78186)

ppm_ms2 <- 25L
min_ms2mass <- 115L

d <- ppm_ms2/1E6
ex <- ceiling(log(expts/min_ms2mass)/log(1+d))

x2 <- find_ms2_bypep(theos, expts, ex, d)

## Experimental 317.20001 & 317.19315 match to theoreitcal 317.2022 & 317.1932;
#  experimental 959.48468 is also multiple dipping
#  (multiple to multiple)
aas <- unlist(strsplit("SLAAEEEAAR", ""))
theos <- c(317.2022, 430.2863, 501.3234, 572.3605, 701.4031, 
           830.4457, 959.4883, 1030.5254, 1101.5625, 1257.6636, 
           175.1190, 246.1561, 317.1932, 446.2358, 575.2784, 
           704.3210, 775.3581, 846.3952, 959.4793, 1046.5113)
names(theos) <- c(aas, rev(aas))
theos <- list(`0000000000` = theos)

expts <- c(102.05550, 110.07173, 112.08743, 114.06662, 115.08708, 116.07106, 
           120.08112, 126.12806, 127.12510, 127.13136, 128.12843, 128.13467, 
           129.13181, 129.13805, 130.13509, 130.14134, 131.13843, 132.14159, 
           133.04318, 136.07600, 143.08151, 157.10843, 158.09259, 173.14983, 
           175.11917, 176.15997, 186.15306, 188.15982, 201.08725, 229.12970, 
           230.17041, 231.17366, 241.08228, 246.15643, 248.18086, 255.17380, 
           259.09235, 289.20801, 300.16696, 315.25952, 317.19315, 317.20001, 
           343.25476, 351.20572, 367.22961, 376.27621, 402.29141, 430.28644,
           438.26654, 446.23587, 501.32376, 502.32788, 523.34491, 537.33649, 
           556.83990, 557.34125, 557.84283, 572.36127, 575.27863, 590.31671, 
           605.84045, 629.33453, 637.86853, 638.33740, 638.83984, 661.35687, 
           667.39923, 673.40417, 701.40350, 702.40936, 770.42517, 775.35840, 
           776.37720, 802.44659, 830.44556, 831.44958, 846.39349, 847.39465, 
           931.49158, 932.48187, 933.48218, 954.54474, 955.54791, 957.55493, 
           958.55646, 959.48468, 960.48773, 1030.52563, 1046.50830, 1047.50684, 
           1100.53101, 1101.54858, 1103.53040, 1116.59692, 1117.54749, 
           1118.54334, 1119.54565, 1120.55347, 1121.55408, 1122.55737)

ppm_ms2 <- 13L
min_ms2mass <- 115L
d <- ppm_ms2/1E6
ex <- ceiling(log(expts/min_ms2mass)/log(1+d))

x3 <- find_ms2_bypep(theos, expts, ex, d)

## 7 b-ions and 9 y-bios 
#  (an even total, n_ps matched but identities off)
theos <- c(344.2131,504.2438,617.3278,732.3548,845.4389,946.4865,
           1003.5080,1102.5764,1258.6775,175.1190,274.1874,331.2088,
           432.2565,545.3406,660.3675,773.4516,933.4822,1047.5252)
names(theos) <- c("N","C","I","D","I","T","G","V","R",
                  "R","V","G","T","I","D","I","C","N")
theos <- list(`000000000` = theos)

expts <- c(115.08701,116.07098,120.08125,126.12807,127.12508,127.13135,
           128.12840,128.13467,129.10252,129.13177,129.13803,130.09766,
           130.13507,130.14136,131.13843,136.07590,142.09769,157.09749,
           157.10825,158.09258,173.14984,175.11916,175.15634,176.15977,
           186.15321,188.15988,215.13939,230.17043,231.17412,247.19698,
           248.18083,254.16139,255.14516,257.16092,272.17194,273.21265,
           295.17020,314.18243,316.21832,331.21051,331.21661,344.21338,
           345.19727,376.27597,377.27975,389.14813,400.27567,415.22980,
           432.25732,445.25986,458.28168,462.25909,473.74744,476.24893,
           491.30301,497.73514,501.31610,504.24396,506.24860,520.24512,
           545.34082,546.34387,584.76740,588.35773,589.32990,606.37183,
           616.83417,617.32935,629.88574,630.33142,638.32935,638.38837,
           638.84747,638.89111,639.34772,639.39099,660.36957,672.38916,
           715.32959,717.40179,732.35510,733.35760,773.45404,845.43958,
           900.49872,901.49738,902.42206,933.48560,946.48560,1003.50885,
           1047.52295,1102.53992,1104.54358,1118.53979,1118.65076,
           1119.55566,1120.54919,1121.56397,1122.56702,1123.57056)

ppm_ms2 <- 13L
min_ms2mass <- 115L
d <- ppm_ms2/1E6
ex <- ceiling(log(expts/min_ms2mass)/log(1+d))

x4 <- find_ms2_bypep(theos, expts, ex, d)


## fewer matches with "find_ppm_outer_bycombi" and check minn_ms2 again
#  (doesn't really return NULL since now with "ppm_ms2 * 2")
pep <- "EFINSLRLYR"
nms <- unlist(stringr::str_split(pep, ""))

theos <- c(359.2128,506.2812,619.3653,734.3922,821.4243,934.5083,1090.6094,
           1203.6935,1366.7568,1522.8579,175.1190,338.1823,451.2663,607.3675,
           720.4515,807.4835,922.5105,1035.5946,1182.6630,1311.7056)

names(theos) <- c(nms, rev(nms))
theos <- list(`0005000000` = theos)

expts <- c(126.12811,127.12556,127.13139,128.12862,128.13455,129.13194,
           129.13786,130.13542,130.14139,131.13852,136.07597,173.14980,
           175.11916,175.15663,176.15979,186.15309,188.15987,215.13940,
           227.10303,230.17044,231.17381,247.19708,248.18105,249.18443,
           273.21262,316.21869,344.21350,345.19760,345.21689,353.19775,
           358.22925,364.14948,376.27615,377.27942,397.20880,479.28204,
           480.28650,507.27704,507.31665,508.28040,508.31964,550.31934,
           578.31403,579.31818,601.33063,602.33466,620.40070,621.40363,
           679.36194,680.40875,680.90851,690.32965,707.35663,708.35950,
           721.44830,722.45142,736.94965,761.93817,762.43274,762.93304,
           770.79028,770.94440,771.44598,786.48285,786.98523,792.44543,
           820.44061,821.44318,834.53247,835.53638,877.45514,893.49561,
           903.47803,904.48071,921.48846,922.49133,945.56488,963.57489,
           964.57990,1006.57715,1016.55975,1017.56122,1034.60925,1035.57776,
           1035.61169,1165.61340,1166.61731,1197.67346,1198.67700,1311.71948,
           1312.72559,1366.73694,1368.74927,1369.74463,1382.75659,1383.75610,
           1384.75915,1385.76355,1386.76611,1387.76599)

ppm_ms2 <- 13L
min_ms2mass <- 115L
d <- ppm_ms2/1E6
ex <- ceiling(log(expts/min_ms2mass)/log(1+d))

x5 <- find_ms2_bypep(theos, expts, ex, d, ppm_ms2)

}

}
