% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{find_nmodtree}
\alias{find_nmodtree}
\title{Find and subset peptides.}
\usage{
find_nmodtree(prps, min_n_res, vmods, posns, len)
}
\arguments{
\item{prps}{Lists of peptide sequences with a one-letter representation
of amino acid residues. Each list is named by protein accession.}

\item{min_n_res}{The minimum numbers of residues for a given set of sites.}

\item{vmods}{A named list of variable modifications. See also
\link{find_protntsite} for examples of \code{vmods}.}

\item{posns}{The position (e.g., \code{Protein N-term}, \code{Anywhere},
etc.) of \code{vmods}. The argument can be obtained from \code{vmods} but
passed as a parameter for vectorization.}

\item{len}{The count of \code{vmods} (passed as a parameter for
vectorization).}
}
\description{
(1-nt) Protein N-term + site -> (2-nt) Any protein N-term -> (3-nt) Any
N-term + site -> (4-nt) Any N-term -> (5) Anywhere + site.
}
