% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{find_perm_sets}
\alias{find_perm_sets}
\title{A (faster alternative) to \link[gtools]{permutations} with duplicated labels.}
\usage{
find_perm_sets(labs = c("A", "A", "A", "B", "B", "C"))
}
\arguments{
\item{labs}{A vector of labels.}
}
\description{
A (faster alternative) to \link[gtools]{permutations} with duplicated labels.
}
\examples{
\dontrun{
library(gtools)
library(mzion)
library(dplyr)

# four positions for four balls in three colors
labs  <- c("A", "A", "B", "C")
len   <- length(labs)
g <- unique(permutations(len, len, labs, set = FALSE, repeats.allowed = FALSE))
m <- mzion:::find_perm_sets(labs)

cns <- paste0("p", 1:len)
g <- data.frame(g)
m <- data.frame(m)
colnames(m) <- colnames(g) <- cns

stopifnot(identical(dplyr::arrange_all(g), dplyr::arrange_all(m)))

# example-2
labs <- c("A", "A", "A", "B", "B", "C")
m <- mzion:::find_perm_sets(labs)

# Compares to gtools::permutations
len <- length(labs)
g <- unique(permutations(len, len, labs, set = FALSE))

cns <- paste0("p", 1:len)
g <- data.frame(g)
m <- data.frame(m)
colnames(m) <- colnames(g) <- cns

stopifnot(identical(dplyr::arrange_all(g), dplyr::arrange_all(m)))

# Compares to the theoretical number of entries
len_labs <- length(labs)
ns <- mzion:::count_elements(labs)
theo_cts <- factorial(len_labs)/prod(factorial(ns))
stopifnot(nrow(m) == theo_cts)

## Others
x <- mzion:::find_perm_sets(rep("A", 3))
}
}
