% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{find_vmodposM}
\alias{find_vmodposM}
\title{Helper of \link{combi_namesiteM} (by each rows of labels).}
\usage{
find_vmodposM(Vec, vec, ps, aas)
}
\arguments{
\item{Vec}{A vector of names (lower-case vec for sites).}

\item{vec}{A vector of labels.}

\item{ps}{Named vector; counts for each site. Sites in names and counts in
values.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}
}
\description{
Multiple Names to the same Site.
}
