% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms2_labels.R
\name{find_vmodposU}
\alias{find_vmodposU}
\title{Helper of \link{combi_namesiteU}.}
\usage{
find_vmodposU(vec, ps, aas)
}
\arguments{
\item{vec}{A vector of labels.}

\item{ps}{Named vector; counts for each site. Sites in names and counts in
values.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}
}
\description{
One-to-one correspondence between Names and Sites.
}
\details{
Custom functions: vec_to_list.
}
\examples{
\donttest{
library(mzion)

m <- c("M", "M", "N", "N")
ps <- c(2, 2)
names(ps) <- c("M", "N")

aa_seq <- "MHQGVMNVNMGQKMNS"
aas <- .Internal(strsplit(aa_seq, "", fixed = TRUE, perl = FALSE, useBytes = FALSE))
aas <- .Internal(unlist(aas, recursive = FALSE, use.names = FALSE))
ans <- mzion:::find_vmodposU(m, ps, aas)
}
}
