% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{findLocFracsDF}
\alias{findLocFracsDF}
\title{Finds the localization fractions of STY.}
\usage{
findLocFracsDF(df, phosmods = NULL)
}
\arguments{
\item{df}{A data frame containing columns \code{pep_ivmod2} and
\code{pep_ms2_ideltas.}.}

\item{phosmods}{A vector to the modification indexes of phosphorylations.}
}
\description{
Counting statistics.
}
\examples{
phosmods <- c("8", "9", "a")
df <- data.frame(pep_ivmod2 = c("0080000", "0009000"), pep_ms2_ideltas. = NA)
df$pep_ms2_ideltas.[1] <- list(c(1,2,3,4,5,6,8,9,10,11,12,13,14))
df$pep_ms2_ideltas.[2] <- list(c(1,2,3,4,5,6,8,9,10,12,13,14))
}
