% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{finds_uniq_vec}
\alias{finds_uniq_vec}
\title{Finds unique elements including the same name in a vector.}
\usage{
finds_uniq_vec(x)
}
\arguments{
\item{x}{A named vector.}
}
\description{
The same value at different name will be treated as different elements.
}
