% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{fuzzy_match_one2}
\alias{fuzzy_match_one2}
\title{Fuzzy matches with a +/-1 window.}
\usage{
fuzzy_match_one2(x, y)
}
\arguments{
\item{x}{A vector to be matched.}

\item{y}{A vector to be matched against.}
}
\description{
No multiple dipping of \code{y} matches. A \code{y} value will be removed (or
became 0) if matched,
}
\examples{
library(mzion)

ans1 <- mzion:::fuzzy_match_one2(c(74953, 74955), rep(74954, 2))
ans2 <- mzion:::fuzzy_match_one2(c(74953, 74955), 74954)

stopifnot(identical(ans1, ans2))
stopifnot(ans1 == c(FALSE, TRUE))

ans3 <- mzion:::fuzzy_match_one2(c(74953, 74955, 80000), c(74955, 80000))

## The x3 example from "find_ms2_bypep"
x <- c(-9185, -3369, -1973, -626, 59, 714, 3326, 7106, 7711, 7715, 8316, 8320, 
       8916, 8920, 9511, 9515, 10102, 10688, 11211, 12945, 16807, 24001, 24481, 
       31480, 32350, 32805, 37050, 37875, 42986, 53028, 53377, 53711, 56940, 58542, 
       59172, 61310, 62482, 70941, 73801, 77575, 78046, 78047, 84120, 85881, 89313, 
       91185, 96328, 101503, 102916, 104302, 113257, 113411, 116563, 118593, 
       121336, 121405, 121474, 123450, 123841, 125826, 127823, 130750, 131786, 
       131842, 131903, 134568, 135267, 135956, 139090, 139200, 146310, 146801, 
       146902, 149442, 152081, 152174, 153544, 153635, 160913, 160995, 161078, 
       162794, 162875, 163036, 163117, 163191, 163271, 168686, 169869, 169943, 
       173741, 173812, 173951, 174856, 174922, 174990, 175059, 175128, 175197, 
       175266)

aas <- unlist(strsplit("SLAAEEEAAR", ""))

y <- c(317.2022, 430.2863, 501.3234, 572.3605, 701.4031, 
       830.4457, 959.4883, 1030.5254, 1101.5625, 1257.6636, 
       175.1190, 246.1561, 317.1932, 446.2358, 575.2784, 
       704.3210, 775.3581, 846.3952, 959.4793, 1046.5113)

names(y) <- c(aas, rev(aas))

ppm_ms2 <- 13L
min_ms2mass <- 115L
d <- ppm_ms2/1E6
y <- ceiling(log(y/min_ms2mass)/log(1+d))

ans <- mzion:::fuzzy_match_one2(x, y)
}
