% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{getMSrowIndexes}
\alias{getMSrowIndexes}
\title{Obtains the indexes of MS1 and MS2 starts and ends.}
\usage{
getMSrowIndexes(ms_level, pad_nas = FALSE)
}
\arguments{
\item{ms_level}{A vector of MS levels.}

\item{pad_nas}{Logical; if TRUE, adds padding values to keep the same length.}
}
\description{
Obtains the indexes of MS1 and MS2 starts and ends.
}
