% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{group_ms2pasef_by_precursors}
\alias{group_ms2pasef_by_precursors}
\title{Collapses PASEF MS2 data (belonging to the same \code{MS1Frame}) by
\code{Precursor} IDs.}
\usage{
group_ms2pasef_by_precursors(dat, lwr = 115L, step = 1.6e-05, title = "")
}
\arguments{
\item{dat}{MS2 data under the same precursor ID.}

\item{lwr}{A lower bound as the starting point in mass binning.}

\item{step}{A step size for mass binning.}

\item{title}{The title (.d file  name).}
}
\description{
The same \code{Precursor} ID must have the same \code{MS1Frame}. The same
precursor at different \code{MS1Frame}s has different \code{Precursor} IDs.
}
