% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{groupProts}
\alias{groupProts}
\title{Groups proteins by shared peptides.}
\usage{
groupProts(df, out_path = NULL)
}
\arguments{
\item{df}{Interim results from \link{matchMS}.}

\item{out_path}{The output path.}
}
\description{
Adds columns \code{prot_hit_num} and \code{prot_family_member} to
\code{psm.txt}.
}
\examples{
\donttest{
df <- data.frame(prot_acc = character(2000), pep_seq = character(2000))
set.seed(100)
df$prot_acc <- sample(LETTERS[1:20], 2000, replace = TRUE)
df$pep_seq <- sample(letters[1:26], 20, replace = TRUE)
df <- df[!duplicated(df), ]

out <- proteoM:::groupProts(df)

# One peptide, multiple proteins
df <- data.frame(prot_acc = LETTERS[1:3], pep_seq = rep("X", 3))
out <- proteoM:::groupProts(df)
stopifnot(nrow(out) == 3L)

# One peptide, one proteins
df <- data.frame(prot_acc = "A", pep_seq = "X")
out <- proteoM:::groupProts(df)
stopifnot(nrow(out) == 1L)

# One proteins
df <- data.frame(prot_acc = rep("A", 3), pep_seq = LETTERS[24:26])
out <- proteoM:::groupProts(df)
stopifnot(nrow(out) == 3L)
}
}
