% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{hannot_decoys}
\alias{hannot_decoys}
\title{Helper of annotating decoy peptides.}
\usage{
hannot_decoys(df, prps)
}
\arguments{
\item{df}{A data frame.}

\item{prps_fwd}{The look-ups of forward protein and peptides.}

\item{prps_rev}{The look-ups of reversed protein and peptides.}
}
\description{
Helper of annotating decoy peptides.
}
