% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{hash_frame_nums}
\alias{hash_frame_nums}
\title{Makes hash table.}
\usage{
hash_frame_nums(data, r = 1.5)
}
\arguments{
\item{data}{A named vector or list.}

\item{r}{Numeric; bucket-to-key ratio.}
}
\description{
Not yet used.
}
