% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{hcalc_tmtint}
\alias{hcalc_tmtint}
\title{Helper of calc_tmtint}
\usage{
hcalc_tmtint(df, quant, ppm_reporters, i, out_path, index_mgf_ms2 = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{ppm_reporters}{A positive integer; the mass tolerance of MS2 reporter
ions. The default is 10.}

\item{i}{The i-th module.}

\item{out_path}{A file path of outputs.}

\item{index_mgf_ms2}{Logical; if TRUE, converts upfrontly MS2 m-over-z values
  from numeric to integers as opposed to in-situ conversion during ion
  matches. The default is FALSE. The \code{index_mgf_ms2 = TRUE} might be
  useful for very large MS files by reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}
}
\description{
Helper of calc_tmtint
}
