% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{hcalc_tmtint}
\alias{hcalc_tmtint}
\title{Helper of calc_tmtint}
\usage{
hcalc_tmtint(
  df,
  quant = "tmt10",
  ppm_reporters = 10L,
  idx = 1L,
  ms1_offsets = 0,
  out_path = NULL
)
}
\arguments{
\item{df}{A data frame.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{ppm_reporters}{A positive integer; the mass tolerance of MS2 reporter
ions. The default is 10.}

\item{idx}{The i-th chunk}

\item{ms1_offsets}{Off-sets in precursor masses in relative to the original
values in MGFs.}

\item{out_path}{A file path of outputs.}
}
\description{
Helper of calc_tmtint
}
