% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{hextract_pasef}
\alias{hextract_pasef}
\title{Helper of \link{extract_pasef_frame}}
\usage{
hextract_pasef(
  mdata,
  iso_info,
  keys = c("msx_moverzs", "msx_ints", "msx_ns", "ms1_fr", "ms1_moverzs", "ms1_ints",
    "ms1_charges", "scan_title", "ms_level", "ret_time", "scan_num", "orig_scan",
    "iso_ctr", "iso_lwr", "iso_upr", "mobility"),
  step = 1.6e-05
)
}
\arguments{
\item{mdata}{A list of PASEF frames. Each list entry corresponding to one
frame (over the ion-mobility dimension).}

\item{iso_info}{The information of MS2 isolation windows etc.}

\item{keys}{The key names of output lists.}

\item{step}{A step size for mass binning.}
}
\description{
(1) Pairs MS2 data and iso_info by frame IDs; (2) Collapses data by Precursor
groups.
}
