% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{hms1_a0_vnl0_fnl1}
\alias{hms1_a0_vnl0_fnl1}
\title{Helper of \link{ms1_a0_vnl0_fnl1}.}
\usage{
hms1_a0_vnl0_fnl1(
  masses,
  fnl_combi,
  aa_masses,
  min_mass = 700,
  max_mass = 4500L,
  digits = 4L
)
}
\arguments{
\item{masses}{A named list of peptide masses.}

\item{fnl_combi}{A data.frame of combinations of neutral losses for fixed
modifications. Each row corresponds to a set of neutral loss. The first row
corresponds to the combination without NLs (all zeros).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value. The primary guard
against low molecular-weight precursors is \code{min_len}.}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper of \link{ms1_a0_vnl0_fnl1}.
}
