% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{hpair_mgths}
\alias{hpair_mgths}
\title{Helper of \link{pair_mgftheos}.}
\usage{
hpair_mgths(
  ms1_offset = 0,
  notch = NULL,
  mgfs,
  n_modules,
  by_modules = TRUE,
  mgf_path,
  quant = "none",
  min_mass = 200L,
  max_mass = 4500L,
  ppm_ms1_bin = 10L,
  .path_bin,
  data_type = "mzML"
)
}
\arguments{
\item{ms1_offset}{The ms1 offset.}

\item{notch}{The index assigned to an ms1_offset.}

\item{mgfs}{MGF data.}

\item{n_modules}{The number of modules (\code{length(aa_masses_all)}) or one}

\item{by_modules}{Logical; if TRUE, results are saved with one mgf to one
theo module. At FALSE, results are saved with one mgf paired to all theo
modules}

\item{mgf_path}{The path to MGF files}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{ppm_ms1_bin}{The tolerance in precursor mass error after mass binning.}

\item{.path_bin}{The path to binned theoretical masses}

\item{data_type}{The type of peak lists in one of raw, mzML or mgf. The
argument is used to handle list(NULL) at an mzML format.}
}
\description{
Helper of \link{pair_mgftheos}.
}
