% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{hsemipeps_byprots}
\alias{hsemipeps_byprots}
\title{Helper of \link{semipeps_byprots}.}
\usage{
hsemipeps_byprots(prots, min_len = 7L, max_len = 40L, aa_masses)
}
\arguments{
\item{prots}{Lists of proteins with full-enzymatic sequences. For each entry
under a protein, the value is a mass and the name is a peptide sequence.}

\item{min_len}{The minimum length of peptide sequences for consideration.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}
}
\description{
Helper of \link{semipeps_byprots}.
}
