% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{htraceXY}
\alias{htraceXY}
\title{Helper of MS1 tracing.}
\usage{
htraceXY(
  xs,
  ys,
  ss,
  ts,
  df,
  gap_bf = 256L,
  gap_af = 256L,
  out_name = NULL,
  n_dia_scans = 6L,
  from = 200L,
  step_tr = 6e-06,
  tol_ms1 = 1e-05,
  y_perc = 0.01,
  yco = 100,
  min_y = 2e+06,
  ytot_co = 2e+05,
  sum_y = FALSE,
  min_n1 = 10L,
  min_n2 = 20L,
  min_n3 = 15L,
  path_ms1 = NULL,
  out_cols = c("ms_level", "ms1_moverz", "ms1_int", "orig_scan", "apex_scan_num",
    "apex_xs", "apex_ys", "apex_ps", "apex_ts")
)
}
\arguments{
\item{xs}{Vectors of full-spectrum MS1 m/z values.}

\item{ys}{Vectors of full-spectrum MS1 intensities.}

\item{ss}{Vectors of MS1 (original) scan numbers.}

\item{ts}{Vectors of MS1 retention times (for calculating area-under-a-peak).}

\item{df}{A data frame of MS1 and MS2 corresponding to \code{xs}.}

\item{gap_bf}{A preceding gap size.}

\item{gap_af}{A following gap size.}

\item{out_name}{A file name of output.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}

\item{from}{The starting point in mass binning.}

\item{step_tr}{The bin size of MS1 tracing}

\item{tol_ms1}{The tolerance in MS1 errors \code{ppm_ms1 * 1E-6}.}

\item{y_perc}{The cut-off in intensity values in relative to the base peak.}

\item{yco}{The cut-off in y values.}

\item{min_y}{A more strict cut-off in peak area for finding tentatively the
best peak.}

\item{ytot_co}{A more permissive cut-off in peak area for finding all peaks.}

\item{sum_y}{Logical; sum Y values at near the same X values or not. Mostly
FALSE with Thermo's data.}

\item{min_n1}{The first cut-off of a minimum number of points across an MS1
peak.}

\item{min_n2}{The second cut-off of a minimum number of points across an MS1
peak.}

\item{min_n3}{The third cut-off of a minimum number of points across an MS1
peak.}

\item{path_ms1}{The path for saving the output of MS1 data.}

\item{out_cols}{The output column keys.}
}
\description{
Helper of MS1 tracing.
}
