% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{htraceXY}
\alias{htraceXY}
\title{Helper of \link{traceXY}.}
\usage{
htraceXY(
  xs,
  ys,
  ss,
  ts,
  df,
  gap_bf = 256L,
  gap_af = 256L,
  n_mdda_flanks = 6L,
  from = 200L,
  step = 6e+06,
  y_perc = 0.01,
  yco = 500,
  look_back = TRUE
)
}
\arguments{
\item{xs}{Vectors of full-spectrum MS1 m/z values.}

\item{ys}{Vectors of full-spectrum MS1 intensities.}

\item{ss}{Vectors of MS1 (original) scan numbers.}

\item{ts}{Vectors of MS1 retention times (for calculating area-under-a-peak).}

\item{df}{A data frame of MS1 and MS2 corresponding to \code{xs}.}

\item{gap_bf}{A preceding gap size.}

\item{gap_af}{A following gap size.}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{from}{The starting point for mass binning.}

\item{step}{The step size for mass binning.}

\item{y_perc}{The cut-off in intensity values in relative to the base peak.}

\item{yco}{The cut-off in y values.}

\item{look_back}{Logical; look up the preceding MS bin or not.}
}
\description{
Helper of \link{traceXY}.
}
