% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{htraceXY}
\alias{htraceXY}
\title{Helper of \link{traceXY}.}
\usage{
htraceXY(
  xs,
  ys,
  ss,
  ts,
  df,
  gap_bf = 256L,
  gap_af = 256L,
  out_name = NULL,
  n_dia_scans = 6L,
  from = 200L,
  step = 5e-06,
  y_perc = 0.01,
  yco = 500,
  look_back = TRUE,
  min_y = 0,
  path_ms1 = NULL,
  out_cols = c("ms_level", "ms1_moverz", "ms1_int", "orig_scan", "apex_scan_num",
    "apex_xs", "apex_ys", "apex_ps", "apex_ts")
)
}
\arguments{
\item{xs}{Vectors of full-spectrum MS1 m/z values.}

\item{ys}{Vectors of full-spectrum MS1 intensities.}

\item{ss}{Vectors of MS1 (original) scan numbers.}

\item{ts}{Vectors of MS1 retention times (for calculating area-under-a-peak).}

\item{df}{A data frame of MS1 and MS2 corresponding to \code{xs}.}

\item{gap_bf}{A preceding gap size.}

\item{gap_af}{A following gap size.}

\item{out_name}{A file name of output.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}

\item{from}{The starting point for mass binning.}

\item{step}{The step size for mass binning.}

\item{y_perc}{The cut-off in intensity values in relative to the base peak.}

\item{yco}{The cut-off in y values.}

\item{look_back}{Logical; look up the preceding MS bin or not.}

\item{min_y}{The cut-off of intensity values.}

\item{path_ms1}{The path for saving the output of MS1 data.}

\item{out_cols}{The output column keys.}
}
\description{
Helper of \link{traceXY}.
}
