% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{index_mz}
\alias{index_mz}
\title{Converts ms2_moverzs to integers.}
\usage{
index_mz(x, from = 115L, d = 1e-05)
}
\arguments{
\item{x}{Numeric; MS2 mass.}

\item{from}{Numeric; the starting MS1 mass.}

\item{d}{Numeric; \eqn{ppm * 10E-6}.}
}
\description{
Note that as.integer is needed. When an indexed x equals 0 but is double, it
can cause %fin% to fail.
}
\details{
Changed from \code{ceiling} to \code{floor} on 2024-08-22. Reason: bins
collapse from left to right since v1.4.2.8.
}
\examples{
xs <- ceiling(log(c(114.0916, 114.9999)/115)/log(1+8E-6))
match(xs, xs); fastmatch::fmatch(xs, xs)

ys <- as.integer(xs)
match(ys, ys); fastmatch::fmatch(ys, ys)
}
