% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{load_cache_info}
\alias{load_cache_info}
\title{Loads variable to the global environment.}
\usage{
load_cache_info(cache = NULL, overwrite = TRUE)
}
\arguments{
\item{overwrite}{If TRUE, overwrite the pre-existed global values with the
cache values.}

\item{file}{An file name as an alternative for looking up}
}
\description{
Not yet used.
}
\examples{
\donttest{
out_path <- "~/proteoM/example/"
cache <- file.path(out_path, "Calls/.cache_info.rds")
load_cache_info(cache)
}
}
\seealso{
\link{find_ms1_times}
}
