% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastas.R
\name{load_fasta}
\alias{load_fasta}
\title{Loads fasta}
\usage{
load_fasta(fasta = NULL)
}
\arguments{
\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. There is no default and the experimenters need to
supply the files.}
}
\description{
Not used in proteoM.
}
\examples{
\donttest{
fasta_db <- load_fasta("~/proteoM/dbs/fasta/uniprot/uniprot_hs_2020_05.fasta")
}
}
