% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{load_mgfs}
\alias{load_mgfs}
\title{Helper in loading MGFs.}
\usage{
load_mgfs(
  out_path,
  mgf_path,
  min_mass = 700L,
  max_mass = 4500L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  topn_ms2ions = 100L,
  min_ms1_charge = 2L,
  max_ms1_charge = 6L,
  min_scan_num = 1L,
  max_scan_num = .Machine$integer.max,
  min_ret_time = 0,
  max_ret_time = Inf,
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  is_ms1_three_frame = TRUE,
  is_ms2_three_frame = TRUE,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  enzyme = "trypsin_p",
  digits = 5L
)
}
\arguments{
\item{out_path}{A file path of outputs.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).

  Individuality in MGF files are slightly preferred to take advantage of
  parallel reading of the files.}

\item{min_mass}{A minimum mass of precursors for considerations.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation. The default is 4500.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches. The default is to use the top-100 ions in an MS2 event.}

\item{min_ms1_charge}{A positive integer; the minimum MS1 charge state for
considerations. The default is 2.}

\item{max_ms1_charge}{A positive integer; the maximum MS1 charge state for
considerations. The default is 6.}

\item{min_scan_num}{Depreciated. A positive integer; the minimum scan number
for considerations. The default is 1. The setting only applies to MGFs with
numeric scan numbers. For example, it has no effects on Bruker's timsTOF
data.}

\item{max_scan_num}{Depreciated. A positive integer; the maximum scan number
for considerations. The default is the maximum machine integer. The setting
only applies to MGFs with numeric scan numbers.}

\item{min_ret_time}{A non-negative numeric; the minimum retention time in
seconds for considerations. The default is 0.}

\item{max_ret_time}{A non-negative numeric; the maximum retention time in
seconds for considerations. The default is \code{Inf}.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{is_ms1_three_frame}{Logical; is the searches by the three frames of
preceeding, current and following.}

\item{is_ms2_three_frame}{Logical; is the searches by the three frames of
preceeding, current and following.}

\item{mgf_cutmzs}{Cut points of MS1 moverz values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from protein entries. The
default is \code{Trypsin_P}. See also parameter \code{custom_enzyme}.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Helper in loading MGFs.
}
