% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmod_ms1_labels.R
\name{make_ms1_vmodsets}
\alias{make_ms1_vmodsets}
\title{Makes the sets of labels of variable modifications.}
\usage{
make_ms1_vmodsets(
  aa_masses_all = NULL,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L
)
}
\arguments{
\item{aa_masses_all}{All the amino acid lookup tables.}

\item{maxn_vmods_per_pep}{The maximum number of \code{Anywhere}
(non-terminal) variable modifications per peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
\code{Anywhere} (non-terminal) variable modifications per site in a per
peptide sequence.}
}
\description{
No position permutation (for MS1 masses).
}
