% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mztab.R
\name{make_mztab}
\alias{make_mztab}
\title{Makes an mzTab file.}
\usage{
make_mztab(out_path = stop("Provide the path.", call. = FALSE))
}
\arguments{
\item{out_path}{A parent path where the outputs of \code{PSM}, \code{Peptide}
and \code{Protein} files and folders are.}
}
\description{
With \code{mzion} searches and proteoQ preprocessing.
}
