% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{make_unique_sets}
\alias{make_unique_sets}
\title{Makes the unique sets of modifications by individual AA residues.}
\usage{
make_unique_sets(
  p = 5L,
  n = 2L,
  labs = c("X", "Y"),
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L
)
}
\arguments{
\item{p}{The number of open positions for filling.}

\item{n}{The number of labels.}

\item{labs}{The labels of balls.}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}
}
\description{
Makes the unique sets of modifications by individual AA residues.
}
\examples{
\donttest{
# multiple outs
p <- 5
labs <- c("Oxidation (M)", "Carbamidomethyl (M)", "Carbamyl (M)")
n <- length(labs)
stopifnot(n == length(labs))

ans5 <- make_unique_sets(5, n, labs)
ans7 <- make_unique_sets(7, n, labs)

stopifnot(identical(ans5, ans7))

# single out
p <- 3
labs <- c("Oxidation (M)", "Carbamidomethyl (M)", "Carbamyl (M)")
n <- length(labs)
stopifnot(n == length(labs))

ans <- make_unique_sets(p, n, labs)
}
}
