% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMS2ions.R
\name{mapMS2ions}
\alias{mapMS2ions}
\title{Visualization of matched MS2 ions.}
\usage{
mapMS2ions(
  out_path = NULL,
  in_name = "psmQ.txt",
  out_name = "bar.png",
  raw_file = "foo.raw",
  scan = 1234,
  rank = 1L,
  is_decoy = FALSE,
  type_ms2ions = "by",
  width = 12.5,
  height = 6
)
}
\arguments{
\item{out_path}{A file path where outputs of \link{matchMS} can be
identified.}

\item{in_name}{An input file name of PSMs, such as psmQ.txt, psmC.txt,
psmT2.txt, psmT3.txt etc.}

\item{out_name}{An output file name for saving the the plot of MS2 ions.}

\item{raw_file}{Character string; a RAW file name.}

\item{scan}{A scan number that can be found from the outputs. Positive
  integer, Thermo's Orbitrap data; character string, Bruker's timsTOF data.

  The searches of scan number will proceed automatically through
  \code{psmQ.txt},  \code{psmT2.txt}, \code{psmT3.txt} and/or
  \code{psmC.txt}.}

\item{rank}{Positive integer; the rank of a match. The default is one for the
best match.}

\item{is_decoy}{Logical; is the match from decoy. The default is FALSE.}

\item{width}{Plot width.}

\item{height}{Plot height.}
}
\description{
Visualization of matched MS2 ions.
}
\examples{
\dontrun{
library(mzion)

ans <- mapMS2ions(
    out_path = "a/mzion/output/folder", in_name = "psmQ.txt", 
    out_name = "bar.png", raw_file = "a-raw-file-name.raw", scan = 9933, 
    rank = 1L, is_decoy = FALSE, type_ms2ions = "by"
  )

# Custom plots
library(ggplot2)

mgf <- ans$mgf
duo <- ans$duo
duo2 <- ans$duo2

p <- ggplot() +
  geom_segment(mgf, mapping = aes(x = ms2_moverz, y = ms2_int,
                                         xend = ms2_moverz, yend = 0),
                      color = "gray", size = .1) +
  geom_segment(duo, mapping = aes(x = ms2_moverz, y = ms2_int,
                                           xend = ms2_moverz, yend = 0,
                                           color = type),
                        size = 1, show.legend = FALSE) +
  geom_text(duo, mapping = aes(x = ms2_moverz, y = ms2_int,
                                        label = label, color = type),
                     size = 6, alpha = .5, hjust = 0, angle = 90, vjust = 0,
                     nudge_x = 0.05, nudge_y = 0.05, na.rm = TRUE,
                     show.legend = FALSE) +
  labs(x = "m/z", y = "Intensity")

if (nrow(duo2)) {
  p <- p + 
    ggplot2::geom_segment(duo2, mapping = aes(x = ms2_moverz, y = ms2_int, 
                                              xend = ms2_moverz, yend = 0, 
                                              color = type), 
                          size = .5, show.legend = FALSE) + 
    ggplot2::geom_text(duo2, mapping = aes(x = ms2_moverz, y = ms2_int, 
                                           label = label, color = type),
                       size = 4, alpha = .5, hjust = 0, angle = 90, vjust = 0, 
                       nudge_x = 0.05, nudge_y = 0.05, na.rm = TRUE, 
                       show.legend = FALSE) 
}

pep <- paste0(duo$site[1:(nrow(duo)/2L)], collapse = " ")
p + annotate("text", -Inf, Inf, label = pep, hjust = -.2, vjust = 2)

}

}
