% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_os.R
\name{match_calltime}
\alias{match_calltime}
\title{Finds the time stamp of a matched call from cached results.}
\usage{
match_calltime(
  path = "~/mzion/.MSearches/Cache/Calls",
  fun = "calc_pepmasses2",
  nms = c("out_path"),
  type = c(TRUE, FALSE),
  new_args = NULL
)
}
\arguments{
\item{path}{A (system) file path.}

\item{fun}{The name of function being saved.}

\item{nms}{Names of arguments to be included in or excluded from matching.}

\item{type}{Logical; if TRUE, includes all arguments in \code{nms}. At FALSE,
excludes all \code{nms}.}

\item{new_args}{Named vector; new arguments that are not in earlier versions.}
}
\value{
An empty object if no matches.
}
\description{
Finds the time stamp of a matched call from cached results.
}
\examples{
\dontrun{
library(mzion)

.time_bin <- mzion:::match_calltime(
  path = file.path(.path_cache, "calc_pepmasses2", .time_stamp), 
  fun = fun,
  nms = c("min_mass", "max_mass", "ppm_ms1", "calib_masses"), 
  new_args = c(calib_masses = calib_masses)) 
}
}
