% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{match_ex2th2}
\alias{match_ex2th2}
\title{Matches between secondary experimentals and theoreticals.}
\usage{
match_ex2th2(expt, theo, min_ms2mass = 115L, d = 1e-05, index_mgf_ms2 = FALSE)
}
\arguments{
\item{expt}{A vector of experimental m-over-z values.}

\item{theo}{A vector of theoretical m-over-z values.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{d}{Bin size, e.g., \eqn{20 ppm / 2 * 1E-6}.}

\item{index_mgf_ms2}{A low-priority feature. Logical; if TRUE, converts
  upfrontly MS2 m-over-z values from numeric to integers as opposed to
  \emph{in-situ} conversion during ion matches. The default is FALSE. The
  \code{index_mgf_ms2 = TRUE} might be useful for very large MS files by
  reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}
}
\description{
At \code{index_mgf_ms2 = FALSE}.
}
