% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{match_mvmods}
\alias{match_mvmods}
\title{Matches to the pre-calculated labels of combinatorial MS1 variable
modifications.}
\usage{
match_mvmods(
  aas,
  aa_masses = NULL,
  amods = NULL,
  ms1vmods = NULL,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L
)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{amods}{\code{Anywhere} variable modifications.}

\item{ms1vmods}{The i-th result from lapply(aa_masses_all, make_ms1vmod_i).}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}
}
\description{
Matches to the pre-calculated labels of combinatorial MS1 variable
modifications.
}
\examples{
\donttest{
fixedmods <- c("TMT6plex (N-term)", "TMT6plex (K)", 
               "Carbamidomethyl (C)")

varmods <- c("Acetyl (Protein N-term)", "Oxidation (M)", 
             "Deamidated (N)",
             "Gln->pyro-Glu (N-term = Q)")

aa_masses_all <- calc_aamasses(fixedmods = fixedmods,
                               varmods = varmods,
                               maxn_vmods_setscombi = 64,
                               add_varmasses = FALSE,
                               add_nlmasses = FALSE,
                               exclude_phospho_nl = TRUE,
                               out_path = NULL)

maxn_vmods_per_pep <- 5L
maxn_sites_per_vmod <- 3L

ms1vmods_all <- lapply(aa_masses_all, make_ms1vmod_i,
                       maxn_vmods_per_pep = maxn_vmods_per_pep,
                       maxn_sites_per_vmod = maxn_sites_per_vmod)

i <- 11L
aa_masses <- aa_masses_all[[i]]
amods <- attr(aa_masses, "amods")
ms1vmods <- ms1vmods_all[[i]]

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

vmods_combi <- match_mvmods(aas, aa_masses, amods, ms1vmods)

## M and N
fixedmods = c("TMT6plex (K)", "dHex (S)")
varmods = c("Carbamidomethyl (M)", "Carbamyl (M)", 
            "Deamidated (N)", "Acetyl (Protein N-term)")

aa_masses_all <- calc_aamasses(fixedmods, varmods,
                               add_varmasses = FALSE,
                               add_nlmasses = FALSE)

maxn_vmods_per_pep <- 5L
maxn_sites_per_vmod <- 3L

ms1vmods_all <- lapply(aa_masses_all, make_ms1vmod_i,
                       maxn_vmods_per_pep = maxn_vmods_per_pep,
                       maxn_sites_per_vmod = maxn_sites_per_vmod)

i <- 16L
aa_masses <- aa_masses_all[[i]]
amods <- attr(aa_masses, "amods")
ms1vmods <- ms1vmods_all[[i]]

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

vmods_combi <- match_mvmods(aas, aa_masses, amods, ms1vmods)
}
}
