% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches2.R
\name{ms2match}
\alias{ms2match}
\title{Matches theoretical peptides (parallel by mgf chunks).}
\usage{
ms2match(
  mgf_path,
  aa_masses_all,
  out_path,
  mod_indexes,
  type_ms2ions,
  maxn_vmods_per_pep,
  maxn_sites_per_vmod,
  maxn_vmods_sitescombi_per_pep,
  minn_ms2,
  ppm_ms1,
  ppm_ms2,
  min_ms2mass,
  quant,
  ppm_reporters,
  use_first_rev = FALSE,
  fasta,
  acc_type,
  acc_pattern,
  topn_ms2ions,
  fixedmods,
  varmods,
  include_insource_nl,
  enzyme,
  maxn_fasta_seqs,
  maxn_vmods_setscombi,
  min_len,
  max_len,
  max_miss,
  digits
)
}
\arguments{
\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).

  Individuality in MGF files are slightly preferred to take advantage of
  parallel reading of the files.}

\item{aa_masses_all}{A list of amino acid lookups for all the combination of
fixed and variable modifications.}

\item{out_path}{A file path of outputs.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{maxn_vmods_sitescombi_per_pep}{A non-negative integer; the maximum
number of combinatorial variable modifications per peptide sequence. The
default is 64.}

\item{minn_ms2}{A positive integer; the minimum number of matched MS2 ions
for consideration as a hit. The default is 6.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 25.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro (7 * 13C + 2 * 15N). It
is also possible that an experimenter may construct a \code{tmt12} from a
18-plex TMTpro (8 *13C + 1 * 15N) and thus \code{quant = tmt18} is
suitable.}

\item{ppm_reporters}{A positive integer; the mass tolerance of MS2 reporter
ions. The default is 10.}

\item{use_first_rev}{Logical; if TRUE, uses the first set of \code{aa_masses}
for reversed searches; otherwise, uses the set with the maximum output.}

\item{fasta}{Character string(s) to the name(s) of fasta file(s) with
prepended directory path. The experimenter needs to supply the files.}

\item{acc_type}{Character string(s); the types of protein accessions in one
of c("uniprot_acc", "uniprot_id", "refseq_acc", "other"). For custom names,
the corresponding regular expressions need to be supplied via argument
\code{acc_pattern}.}

\item{acc_pattern}{Regular expression(s) describing the patterns to separate
the header lines of fasta entries. At the \code{NULL} default, the pattern
will be automated when \code{acc_type} are among c("uniprot_acc",
"uniprot_id", "refseq_acc", "other"). See also \link{load_fasta2} for
custom examples.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches. The default is to use the top-100 ions in an MS2 event.}

\item{fixedmods}{Character string(s) of fixed modifications.}

\item{varmods}{Character string(s) of variable modifications. Multiple
  modifications to the same residue are allowed, for example, both
  \code{Carbamyl (M)} and \code{Oxidation (M)}.

  For both \code{fixedmods} and \code{varmods}, the modification title,
  \code{TMT6plex}, applies to all of TMT-6, TMT-10, TMT-11. It is also
  possible to use aliased: (1) \code{TMT10plex} for TMT-10 and
  \code{TMT11plex} for TMT-11, (2) \code{TMT16plex} for TMTpro and (3)
  \code{TMT18plex} for TMTpro18. See also \link{parse_unimod} for grammars of
  modification \code{title}, \code{position} and \code{site}.}

\item{include_insource_nl}{Not yet used. Logical; if TRUE, includes MS1
precursor masses with the losses of neutral species prior to MS2
fragmentation. The default is FALSE. The setting at TRUE remains
experimenting by allowing additional masses in the universe of MS1
precursors.}

\item{enzyme}{A character string; the proteolytic specificity of the assumed
enzyme will be used to generate peptide sequences from proteins. The
default is \code{Trypsin_P}.}

\item{maxn_fasta_seqs}{A positive integer; the maximum number of protein
sequences in fasta files. The default is 200000.}

\item{maxn_vmods_setscombi}{A non-negative integer; the maximum number of
sets of combinatorial variable modifications. The default is 64.}

\item{min_len}{A positive integer; the minimum length of peptide sequences
for considerations. Shorter peptides will be excluded. The default is 7.}

\item{max_len}{A positive integer; the maximum length of peptide sequences
for considerations. Longer peptides will be excluded. The default is 40.}

\item{max_miss}{A non-negative integer; the maximum number of mis-cleavages
per peptide sequence for considerations. The default is 2.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
All files under `out_path` are removed if incur \code{calc_pepmasses} in the
upstream.
}
