% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_a0_vnl0_fnl1.R, R/ms2_a1_vnl0_fnl0.R,
%   R/ms2_a1_vnl0_fnl1.R, R/ms2_a1_vnl1_fnl0.R, R/ms2_base.R
\name{ms2match_a0_vnl0_fnl1}
\alias{ms2match_a0_vnl0_fnl1}
\alias{ms2match_a1_vnl0_fnl0}
\alias{ms2match_a1_vnl0_fnl1}
\alias{ms2match_a1_vnl1_fnl0}
\alias{ms2match_base}
\title{Matching MS2 ions.}
\usage{
ms2match_a0_vnl0_fnl1(
  i,
  aa_masses,
  ms1vmods,
  ms2vmods,
  ntmass,
  ctmass,
  fmods_nl,
  mod_indexes,
  mgf_path,
  out_path,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 64L,
  minn_ms2 = 6L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  index_mgf_ms2 = FALSE,
  df0 = NULL,
  digits = 4L
)

ms2match_a1_vnl0_fnl0(
  i,
  aa_masses,
  ms1vmods,
  ms2vmods,
  ntmod = NULL,
  ctmod = NULL,
  ntmass = NULL,
  ctmass = NULL,
  amods,
  mod_indexes,
  mgf_path,
  out_path,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 64L,
  minn_ms2 = 6L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  index_mgf_ms2 = FALSE,
  df0 = NULL,
  digits = 4L
)

ms2match_a1_vnl0_fnl1(
  i,
  aa_masses,
  ms1vmods,
  ms2vmods,
  ntmod = NULL,
  ctmod = NULL,
  ntmass = NULL,
  ctmass = NULL,
  amods,
  fmods_nl,
  mod_indexes,
  mgf_path,
  out_path,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 64L,
  minn_ms2 = 6L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  index_mgf_ms2 = FALSE,
  df0 = NULL,
  digits = 4L
)

ms2match_a1_vnl1_fnl0(
  i,
  aa_masses,
  ms1vmods,
  ms2vmods,
  ntmod = NULL,
  ctmod = NULL,
  ntmass = NULL,
  ctmass = NULL,
  amods,
  vmods_nl,
  mod_indexes,
  mgf_path,
  out_path,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 64L,
  minn_ms2 = 6L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  index_mgf_ms2 = FALSE,
  df0 = NULL,
  digits = 4L
)

ms2match_base(
  i,
  aa_masses,
  ms1vmods,
  ms2vmods,
  ntmass,
  ctmass,
  mod_indexes,
  mgf_path,
  out_path,
  type_ms2ions = "by",
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  maxn_vmods_sitescombi_per_pep = 64L,
  minn_ms2 = 6L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  index_mgf_ms2 = FALSE,
  df0 = NULL,
  digits = 4L
)
}
\arguments{
\item{i}{Integer; the index for a set of corresponding aa_masses and
theoretical peptides.}

\item{aa_masses}{An amino-acid mass lookup.}

\item{ms1vmods}{The set of all possible MS1 vmod labels at a given aa_masses.}

\item{ms2vmods}{Matrices of labels of variable modifications. Each
permutation in a row for each matrix.}

\item{ntmass}{The mass of N-terminal.}

\item{ctmass}{The mass of C-terminal.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}

\item{mod_indexes}{Integer; the indexes of fixed and/or variable
modifications.}

\item{mgf_path}{A file path to a list of MGF files. The experimenter needs to
  supply the files.

  The supported MGFs are in the formats of (1) MSConvert against \code{.raw}
  from Thermo's Orbitrap or \code{.d} from Bruker's timsTOF Pro, (2) Thermo's
  Proteome Discoverer or (3) Bruker's DataAnalysis.

  With MSConvert, the default \code{titleMaker} is required for correct
  parsing (don't think it can be altered by users, but just in case).

  Individuality in MGF files are slightly preferred to take advantage of
  parallel reading of the files.}

\item{out_path}{A file path of outputs.}

\item{type_ms2ions}{Character; the type of
\href{http://www.matrixscience.com/help/fragmentation_help.html}{ MS2
ions}. Values are in one of "by", "ax" and "cz". The default is "by" for b-
and y-ions.}

\item{maxn_vmods_per_pep}{A non-negative integer; the maximum number of
\code{Anywhere} (non-terminal) variable modifications per peptide. The
default is 5.}

\item{maxn_sites_per_vmod}{A non-negative integer; the maximum number of
  combinatorial \code{Anywhere} (non-terminal) variable modifications per
  site in a peptide sequence. The default is 3.

  For instance, variable modifications of \code{Carbamyl (M)} and
  \code{Oxidation (M)} both have site \code{M}. In order to have a
  combination of two \code{Carbamyl (M)} and two \code{Oxidation (M)} being
  considered, the value of \code{maxn_sites_per_vmod} needs to be four or
  greater.}

\item{maxn_vmods_sitescombi_per_pep}{A non-negative integer; the maximum
number of combinatorial variable modifications per peptide sequence. The
default is 64.}

\item{minn_ms2}{A positive integer; the minimum number of matched MS2 ions
for consideration as a hit. Counts of secondary ions, e.g. b0, b* etc., are
not part of the threshold.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{index_mgf_ms2}{Logical; if TRUE, converts upfrontly MS2 m-over-z values
  from numeric to integers as opposed to in-situ conversion during ion
  matches. The default is FALSE. The \code{index_mgf_ms2 = TRUE} might be
  useful for very large MS files by reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}

\item{df0}{A zero-row data frame that holds column names.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}

\item{ntmod}{The attribute \code{ntmod} from a \code{aa_masses} (for MS1
calculations).}

\item{ctmod}{The attribute \code{ctmod} from a \code{aa_masses} (for MS1
 calculations).
@rdname gen_ms2ions_base}

\item{amods}{The attribute of \code{amods} from an \code{aa_masses}.}

\item{vmods_nl}{The attribute of \code{vmods_nl} from an \code{aa_masses}.}
}
\description{
(5) "amods- tmod- vnl- fnl+", (6) "amods- tmod+ vnl- fnl+"

(7) "amods+ tmod- vnl- fnl-", (8) "amods+ tmod+ vnl- fnl-"

(11) "amods+ tmod- vnl- fnl+", (12) "amods+ tmod+ vnl- fnl+"

(9) "amods+ tmod- vnl+ fnl-", (10) "amods+ tmod+ vnl+ fnl-"

(1) "amods- tmod- vnl- fnl-", (2) "amods- tmod+ vnl- fnl-"
}
