% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{order_fracs}
\alias{order_fracs}
\title{Order fractions}
\usage{
order_fracs(type = "list_table", tempdir)
}
\arguments{
\item{type}{The type of files.}

\item{tempdir}{A temporary directory containing the files.}
}
\description{
Order fractions
}
