% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2frames.R
\name{pair_mgftheos}
\alias{pair_mgftheos}
\title{Pairs MGF queries to theoretical MS1 masses and peptide sequences.}
\usage{
pair_mgftheos(
  mgf_path,
  n_modules,
  ms1_offsets = 0,
  quant = "none",
  min_mass = 200L,
  max_mass = 4500L,
  ppm_ms1_bin = 10L,
  .path_bin,
  reframe_mgfs = FALSE,
  first_search = FALSE
)
}
\arguments{
\item{mgf_path}{The path to MGF files}

\item{n_modules}{The number of modules (\code{length(aa_masses_all)}) or one}

\item{ms1_offsets}{Off-sets in precursor masses (in relative to the values in
the original MGFs).}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{max_mass}{A positive integer; the maximum precursor mass for
interrogation.}

\item{ppm_ms1_bin}{The tolerance in precursor mass error after mass binning.}

\item{.path_bin}{The path to binned theoretical masses}

\item{reframe_mgfs}{Logical; if TRUE, recalculates the frame indexes of MGFs.}

\item{first_search}{Logical; is the first search (for MGF mass calibration)
or not.}
}
\description{
Pairs MGF queries to theoretical MS1 masses and peptide sequences.
}
