% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant2.R
\name{pcollapse_sortpeps}
\alias{pcollapse_sortpeps}
\title{Helper of \link{map_pepprot}.}
\usage{
pcollapse_sortpeps(mat, ncol = NULL, peps = NULL, n_cores = NULL)
}
\arguments{
\item{mat}{A dgCMatrix object. Column names are SORTED protein accessions.
Rownames are SORTED peptide sequences.}

\item{ncol}{The number of columns in \code{mat}.}

\item{peps}{Peptide sequences as the row names of \code{mat}.}

\item{n_cores}{The number of CPU cores.}
}
\description{
Parallel version of \link{collapse_sortpeps}.
}
