% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{post_readmgf}
\alias{post_readmgf}
\title{Post-processing of MGF or mzML}
\usage{
post_readmgf(
  df,
  raw_id,
  mgf_path,
  min_mass = 200L,
  max_mass = 4500L,
  ppm_ms1 = 10L
)
}
\arguments{
\item{df}{A data frame of processed peak lists.}

\item{raw_id}{An ID to replace the original RAW file name.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoSeqs}.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}
}
\description{
Calculates mass \code{frame}s etc.
}
