% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{pretraceXY}
\alias{pretraceXY}
\title{Splits \code{df} for parallel tracing.}
\usage{
pretraceXY(
  df,
  from = 200L,
  step = 1e-05,
  n_chunks = 4L,
  gap = 256L,
  n_dia_scans = 4L
)
}
\arguments{
\item{df}{A data frame of both MS1 and MS2.}

\item{from}{The starting point for mass binning.}

\item{step}{The step size for mass binning.}

\item{n_chunks}{The number of chunks.}

\item{gap}{The size of gapped entries.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}
}
\description{
Splits \code{df} for parallel tracing.
}
