% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{pretraceXY}
\alias{pretraceXY}
\title{Splits \code{df} for parallel tracing.}
\usage{
pretraceXY(df, from = 200L, step = 1e-05, n_chunks = 4L, gap = 256L)
}
\arguments{
\item{df}{A data frame of staggering MS1 and MS2 entries.}

\item{from}{The starting point for mass binning.}

\item{step}{The step size for mass binning.}

\item{n_chunks}{The number of chunks.}

\item{gap}{A gap size for forward and backward looking of precursors.}
}
\description{
Splits \code{df} for parallel tracing.
}
