% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{probco_bypeplen}
\alias{probco_bypeplen}
\title{Helper of \link{calc_pepfdr}.}
\usage{
probco_bypeplen(len, td, fdr_type = "psm", target_fdr = 0.01, out_path)
}
\arguments{
\item{len}{Numeric; the length of peptides.}

\item{td}{A target-decoy pair.}

\item{fdr_type}{A character string; the type of FDR control. The value is in
  one of c("psm", "peptide", "protein"). The current default is \code{psm}.

  Note that \code{fdr_type = protein} is equivalent to \code{fdr_type =
  peptide} with the additional filtration of data at \code{prot_tier == 1}. A
  variant is to set \code{fdr_type = psm}, followed by a data filtration at
  \code{prot_tier == 1}.}

\item{target_fdr}{A numeric; the targeted false-discovery rate (FDR) at the
levels of PSM, peptide or protein. The default is 0.01. See also argument
\code{fdr_type}.}

\item{out_path}{A file path of outputs.}
}
\description{
Calculates the probability cut-off for target-decoy pairs at a given peptide
length.
}
