% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{proc_mdda}
\alias{proc_mdda}
\title{Proc mzML data for mDDA workflows.}
\usage{
proc_mdda(
  spec,
  raw_file,
  idx_sc = 3L,
  idx_osc = 3L,
  idx_mslev = 2L,
  idx_title = 10L,
  idx_scanList_2 = 11L,
  idx_rt_2 = 2L,
  idx_scan_start_2 = 1L,
  idx_precursor_2 = 12L,
  idx_isolationWindow = 1L,
  idx_ctrmz = 1L,
  idx_lwrmz = 2L,
  idx_uprmz = 3L,
  idx_selectedIonList = 2L,
  idx_moverz = 1L,
  idx_charge = 2L,
  idx_ms1int = 3L,
  idx_bin_2 = 13L,
  maxn_precurs = 5L,
  max_ms1_charge = 4L,
  topn_ms2ions = 150L,
  quant = "none",
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  idx_scanList_1 = 11L,
  idx_rt_1 = 2L,
  idx_scan_start_1 = 1L,
  idx_bin_1 = 12L,
  deisotope_ms2 = TRUE,
  ppm_ms1_deisotope = 10L,
  ppm_ms2_deisotope = 10L,
  n_mdda_flanks = 6L,
  grad_isotope = 1.6,
  fct_iso2 = 3,
  use_defpeaks = FALSE
)
}
\arguments{
\item{spec}{Spectrum list.}

\item{raw_file}{The RAW file name of an mzML.}

\item{idx_sc}{Index of scan numbers.}

\item{idx_osc}{Index of original scan numbers.}

\item{idx_mslev}{Index of MS levels.}

\item{idx_title}{Index of scan titles.}

\item{idx_scanList_2}{Index of MS2 scanList.}

\item{idx_rt_2}{Index of MS2 retention times.}

\item{idx_scan_start_2}{Index of MS2 scan start.}

\item{idx_precursor_2}{Index of MS2 precursor.}

\item{idx_isolationWindow}{Index of isolationWindow.}

\item{idx_ctrmz}{Index of center m-over-z.}

\item{idx_lwrmz}{Index of lower m-over-z.}

\item{idx_uprmz}{Index of upper m-over-z.}

\item{idx_selectedIonList}{Index of selectedIonList.}

\item{idx_moverz}{Index of m-over-z sequences.}

\item{idx_charge}{Index of charge states.}

\item{idx_ms1int}{Index of MS1 intensities.}

\item{idx_bin_2}{Index of MS2 binary data.}

\item{maxn_precurs}{Maximum number of chimeric precursors.}

\item{max_ms1_charge}{Maximum precursor charge states.}

\item{topn_ms2ions}{Top-n MS2 ions for consideration.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{idx_scanList_1}{Index of MS1 scanList.}

\item{idx_rt_1}{Index of precursor retention times.}

\item{idx_scan_start_1}{Index of MS1 scan starts.}

\item{idx_bin_1}{Index of MS1 binary data.}

\item{deisotope_ms2}{Logical; to deisotope MS2 features or not.}

\item{ppm_ms1_deisotope}{Mass error tolerance in MS1 deisotoping.}

\item{ppm_ms2_deisotope}{Mass error tolerance in MS2 deisotoping.}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value, the more
stringent it is in calling an adjacent peak being a mono-isotopic
precursor.}

\item{fct_iso2}{A multiplication factor for the fuzzy discrimination of a
secondary precursor in parallel to a primary precursor in an isotopic
envelop. The smaller the value, the more stringent it is in calling an
adjacent peak being a mono-isotopic precursor}

\item{use_defpeaks}{Logical; if TRUE, uses MS1 m-over-z's, intensities and
charge states pre-calculated by other peak-picking algorithms.}
}
\description{
Proc mzML data for mDDA workflows.
}
