% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{purge_decoys}
\alias{purge_decoys}
\title{Purges decoy pep_seq(s) that are also found in target fasta.}
\usage{
purge_decoys(target, decoy)
}
\arguments{
\item{target}{A target data frame with column \code{pep_seq}.}

\item{decoy}{A decoy data frame with column \code{pep_seq}.}
}
\description{
Decoy sequences may be present in targets and thus removed.
}
