% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{purge_search_space}
\alias{purge_search_space}
\title{Subsets the search space.}
\usage{
purge_search_space(
  i,
  aa_masses,
  mgf_path,
  n_cores,
  ppm_ms1 = 20L,
  fmods_nl = NULL
)
}
\arguments{
\item{i}{Integer; the index for a set of corresponding aa_masses and
theoretical peptides.}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{mgf_path}{The file path to a list of MGF files. The experimenters need
to supply the files. Note that the supported MGFs are from MSConvert or
Proteome Discoverer.}

\item{ppm_ms1}{The mass tolerance of MS1 species.}

\item{fmods_nl}{The attribute of \code{fmods_nl} from an \code{aa_masses}.}
}
\description{
Subsets the search space.
}
