% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{read_mgf_chunks}
\alias{read_mgf_chunks}
\title{Reads mgfs in chunks.}
\usage{
read_mgf_chunks(
  filepath = "~/mzion/mgf/temp_1",
  topn_ms2ions = 100L,
  ms1_charge_range = c(2L, 6L),
  ms1_scan_range = c(1L, .Machine$integer.max),
  ret_range = c(0, Inf),
  min_mass = 200L,
  max_mass = 4500L,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  type_mgf = "msconv_thermo",
  n_bf_begin = 0L,
  n_spacer = 0L,
  n_hdr = 5L,
  n_to_pepmass = 3L,
  n_to_title = 1L,
  n_to_scan = 0L,
  n_to_rt = 2L,
  n_to_charge = 4L,
  sep_ms2s = " ",
  nfields_ms2s = 2L,
  sep_pepmass = " ",
  nfields_pepmass = 2L,
  raw_file = NULL,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  index_mgf_ms2 = FALSE,
  quant = "none",
  digits = 4L
)
}
\arguments{
\item{filepath}{The file path to a list of MGF or mzML files.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ms1_charge_range}{The range of MS1 charge states.}

\item{ms1_scan_range}{The range of MS1 scan numbers.}

\item{ret_range}{The range of retention time in seconds.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoSeqs}.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation.}

\item{mgf_cutmzs}{Cut points of MS1 m-over-z values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{type_mgf}{The type of MGF format.}

\item{n_bf_begin}{The number of lines before \code{BEGIN IONS}. Zero for PD
and MSConvert.}

\item{n_spacer}{The number of spacer lines between the preceding line END
IONS and the following line BEGIN IONS. The value is 1 for Proteome
Discoverer and 0 for MSConvert.}

\item{n_hdr}{The number of lines before MS2 data in an MGF. The value is +6
for PD and +5 for MSConvert.}

\item{n_to_pepmass}{The number of lines from BEGIN to PEPMASS.}

\item{n_to_title}{The number of lines from BEGIN to TITLE. The value is the
same between PD and MSConvert.}

\item{n_to_scan}{The number of lines from BEGIN to SCANS. The value is +5 for
PD.}

\item{n_to_rt}{The number of lines from BEGIN to RTINSECONDS.}

\item{n_to_charge}{The number of lines from BEGIN to CHARGE.}

\item{sep_ms2s}{The separation character between MS2 m/z and intensity
values.}

\item{nfields_ms2s}{The number of fields in MS2 entries. Mostely two and can
be three for some Bruker MGFs.}

\item{sep_pepmass}{The separation character between MS1 m/z and intensity
values.}

\item{nfields_pepmass}{The number of fields in \code{PEPMASS}.}

\item{raw_file}{The raw file name. Is NULL for PD and MSConvert.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{index_mgf_ms2}{A low-priority feature. Logical; if TRUE, converts
  upfrontly MS2 m-over-z values from numeric to integers as opposed to
  \emph{in-situ} conversion during ion matches. The default is FALSE. The
  \code{index_mgf_ms2 = TRUE} might be useful for very large MS files by
  reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Reads mgfs in chunks.
}
