% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{read_mzml}
\alias{read_mzml}
\title{Reads mzML from MSConvert}
\usage{
read_mzml(xml_file)
}
\arguments{
\item{xml_file}{A file name of mzML.}
}
\description{
zlib compression need to be disabled when creating mzML from MSConvert.
}
