% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{read_mzml}
\alias{read_mzml}
\title{Reads mzML from MSConvert}
\usage{
read_mzml(
  xml_file,
  topn_ms2ions = 100L,
  tmt_reporter_lower = 126.1,
  tmt_reporter_upper = 135.2,
  exclude_reporter_region = FALSE,
  index_mgf_ms2 = FALSE,
  ppm_ms1 = 10L,
  ppm_ms2 = 10L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  max_ms1_charge = 4L,
  is_mdda = FALSE,
  deisotope_ms2 = TRUE,
  max_ms2_charge = 3L,
  use_defpeaks = FALSE,
  maxn_dia_precurs = 300L,
  maxn_mdda_precurs = 5L,
  n_mdda_flanks = 6L,
  ppm_ms1_deisotope = 10L,
  ppm_ms2_deisotope = 10L,
  grad_isotope = 2.5,
  quant = "none",
  digits = 4L
)
}
\arguments{
\item{xml_file}{A file name of mzML.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{tmt_reporter_lower}{The lower bound of the region of TMT reporter ions.
The default is \eqn{126.1}.}

\item{tmt_reporter_upper}{The upper bound of the region of TMT reporter ions.
The default is \eqn{135.2}.}

\item{exclude_reporter_region}{Logical; if TRUE, excludes MS2 ions in the
region of TMT reporter ions. The default is FALSE. The corresponding range
of TMT reporter ions is informed by \code{tmt_reporter_lower} and
\code{tmt_reporter_upper}. The argument affects only TMT data.}

\item{index_mgf_ms2}{Depreciated. A low-priority feature. Logical; if TRUE,
  converts up-frontly MS2 m-over-z values from numeric to integers as opposed
  to \emph{on-the-fly} conversion during ion matches. The default is FALSE.
  The \code{index_mgf_ms2 = TRUE} might be useful for very large MS files by
  reducing RAM footprints.

  At \code{index_mgf_ms2 = TRUE}, the resolution of mass deltas between
  theoretical and experimental MS2 m-over-z values is limited by the
  \code{bin_width}, which is the ceiling half of the \code{ppm_ms2}. For
  instance, the \code{bin_width} is 10 ppm at the default \code{ppm_ms2 =
  20}. Due to the low resolution in mass deltas at \code{index_mgf_ms2 = TRUE},
  the fields of \code{pep_ms2_deltas, pep_ms2_deltas2, pep_ms2_deltas_mean,
  pep_ms2_deltas_sd} are nullified in the outputs.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation.}

\item{max_ms1_charge}{A positive integer; the maximum MS1 charge state for
considerations. The default is 6.}

\item{is_mdda}{Logical; if TRUE, consider the multiple (chimeric) precursors
in DDA.}

\item{deisotope_ms2}{Logical; if TRUE, de-isotope MS2 features.}

\item{max_ms2_charge}{Maximum charge states for consideration with MS2
deisotoping.}

\item{use_defpeaks}{Logical; if TRUE, uses MS1 m-over-z's, intensities and
charge states pre-calculated by other peak-picking algorithms.}

\item{maxn_dia_precurs}{Maximum number of precursors for consideration in a
DIA scan.}

\item{maxn_mdda_precurs}{Maximum number of precursors for consideration in a
multi-precursor DDA scan.}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{ppm_ms1_deisotope}{Mass error tolerance in MS1 deisotoping.}

\item{ppm_ms2_deisotope}{Mass error tolerance in MS2 deisotoping.}

\item{grad_isotope}{Positive numeric; the gradient threshold between two
adjacent peaks in an isotopic envelop. The smaller the value the more
stringent in calling a less abundant peak a preceding isotopic feature.}

\item{quant}{A character string; the quantitation method. The default is
"none". Additional choices include \code{tmt6, tmt10, tmt11, tmt16 and
tmt18}. For other multiplicities of \code{tmt}, use the compatible higher
plexes. For example, apply \code{tmt16} for \code{tmt12} provided a set of
12-plexes being constructed from a 16-plex TMTpro \eqn{(7 * 13C + 2 *
15N)}. It is also possible that an experimenter may construct a
\code{tmt12} from a 18-plex TMTpro \eqn{(8 *13C + 1 * 15N)} where
\code{quant = tmt18} is suitable.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Only for Thermo's RAW format.
}
