% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{readmzML}
\alias{readmzML}
\title{Reads mzML files.}
\usage{
readmzML(
  filepath = NULL,
  filelist = NULL,
  min_mass = 700L,
  max_mass = 4500L,
  min_ms2mass = 115L,
  max_ms2mass = 4500L,
  topn_ms2ions = 100L,
  ms1_charge_range = c(2L, 6L),
  ms1_scan_range = c(1L, .Machine$integer.max),
  ret_range = c(0, Inf),
  ppm_ms1 = 20L,
  ppm_ms2 = 25L,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  out_path,
  digits = 4L
)
}
\arguments{
\item{filepath}{The file path to a list of MGF or mzML files.}

\item{filelist}{A list of MGF or mzML files.}

\item{min_mass}{Numeric; the minimum mass of MS1 species. The value needs to
match the one in  \link{binTheoSeqs}.}

\item{max_mass}{A maximum mass of precursors for considerations.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation. The default is 4500.}

\item{topn_ms2ions}{A non-negative integer; the top-n species for uses in
MS2 ion searches. The default is to use the top-100 ions in an MS2 event.}

\item{ms1_charge_range}{The range of MS1 charge states.}

\item{ms1_scan_range}{The range of MS1 scan numbers.}

\item{ret_range}{The range of retention time in seconds.}

\item{ppm_ms1}{A positive integer; the mass tolerance of MS1 species. The
default is 20.}

\item{ppm_ms2}{A positive integer; the mass tolerance of MS2 species. The
default is 20.}

\item{mgf_cutmzs}{Cut points of MS1 moverz values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{out_path}{An output path.}

\item{digits}{A non-negative integer; the number of decimal places to be
used. The default is 4.}
}
\description{
Reads mzML files.
}
