% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasefreader.R
\name{readPASEF}
\alias{readPASEF}
\title{Reads Bruker's PASEF files.}
\usage{
readPASEF(
  mgf_path = NULL,
  filelist = NULL,
  pasef_slice_size = 1L,
  topn_ms2ions = 150L,
  bypass_rawexe = FALSE
)
}
\arguments{
\item{mgf_path}{A file path to .d folders.}

\item{filelist}{A list of .d folders.}

\item{pasef_slice_size}{The gap of scan indexes for data binning to a slice.}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{bypass_rawexe}{Logical; to bypass \link{exeReadPASEF} processing of
RAW .d files or not.}
}
\description{
Reads Bruker's PASEF files.
}
