% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimods.R
\name{remove_unimod}
\alias{remove_unimod}
\title{Removes an existing modification.}
\usage{
remove_unimod(title = NULL)
}
\arguments{
\item{title}{The title of a modification.}
}
\description{
Removes an existing modification.
}
