% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{rm_dup13c}
\alias{rm_dup13c}
\title{Removes duplicated entries in 13C results.}
\usage{
rm_dup13c(df, n_13c = 0L)
}
\arguments{
\item{df}{A data frame.}

\item{n_13c}{Number(s) of 13C off-sets in precursor masses, for example, over
the range of \code{-1:2}. The default is 0.}
}
\description{
Removes "n_13c != 0" entries with better matches at "n_13c == 0". By each
module as the same combination of raw_file + pep_scan_num can also be
duplicated across modification modules.
}
