% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{sep_vec}
\alias{sep_vec}
\title{Separates a vector into groups.}
\usage{
sep_vec(vec, n_chunks)
}
\arguments{
\item{vec}{A vector.}

\item{n_chunks}{The number of chunks.}
}
\description{
Separates a vector into groups.
}
