% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgfs.R
\name{sub_mgftopn}
\alias{sub_mgftopn}
\title{Subsets MGFs by top-n}
\usage{
sub_mgftopn(
  ms2_moverzs,
  ms2_ints,
  topn_ms2ions = 100L,
  mgf_cutmzs = numeric(),
  mgf_cutpercs = numeric(),
  min_ms2mass = 115L,
  max_ms2mass = 4500L
)
}
\arguments{
\item{ms2_moverzs}{Lists of MS2 moverz values.}

\item{ms2_ints}{Lists of MS2 intensities}

\item{topn_ms2ions}{A positive integer; the top-n species for uses in MS2 ion
searches.}

\item{mgf_cutmzs}{Cut points of MS1 m-over-z values in peak picking.}

\item{mgf_cutpercs}{The counts of MS2 features in each region of
\code{mgf_cutmzs}.}

\item{min_ms2mass}{A minimum m/z of MS2 ions for considerations.}

\item{max_ms2mass}{A positive integer; the maximum MS2 mass for
interrogation.}
}
\description{
\code{lens} after filtered by \code{min_ms2mass} but before subset by 
\code{topn_ms2ions} to reflect noise levels.
}
