% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{sub_td_byfdrtype}
\alias{sub_td_byfdrtype}
\title{Subsets targets and decoys by fdr_type.}
\usage{
sub_td_byfdrtype(td, fdr_type = "protein", fct_score = 5)
}
\arguments{
\item{td}{A data frame of targets and decoys.}

\item{fdr_type}{A character string; the type of FDR control. The value is in
  one of c("protein", "peptide", "psm"). The default is \code{protein}.

  Note that \code{fdr_type = protein} is comparable to \code{fdr_type =
  peptide} with the additional filtration of data at \code{prot_tier == 1}.}

\item{fct_score}{A factor to convert p-values to scores.}
}
\description{
Subsets targets and decoys by fdr_type.
}
