% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{subMSfull}
\alias{subMSfull}
\title{Subsets full MS1 by the universe of monoisotopic MS1 moverzs.}
\usage{
subMSfull(xs, ys, ms1s, from = 200L, step = 1e-05, gap = 256L)
}
\arguments{
\item{xs}{Vectors of MS1-X values.}

\item{ys}{Vectors of MS1-Y values.}

\item{ms1s}{Vectors of MS1 m/z values that are linked to the mono-isotopic
DDA-MS2 space.}

\item{from}{The starting m/z value for calculating bin indexes.}

\item{step}{The size of bins.}

\item{gap}{The gap of MS1 scans.}
}
\description{
Non-monoisotopic (13C_n) MS1-X are also removed.
}
