% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{subMSfull}
\alias{subMSfull}
\title{Subsets full MS by the universe of monoisotopic moverzs.}
\usage{
subMSfull(xs, ys, ms1s, from = 200L, step = 1e-05, gap = 256L)
}
\arguments{
\item{xs}{Vectors of full-spectrum moverz values.}

\item{ys}{Vectors of full-spectrum intensity values.}

\item{ms1s}{Vectors of MS2-specific MS1 moverz values.}

\item{from}{The starting movzerz for calculating bin indexes.}

\item{step}{The size of bins.}

\item{gap}{The gap of MS1 scans.}
}
\description{
Some non-monoisotopic moverzs can be removed.
}
