% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1_precursors.R
\name{tbl_prots_peps}
\alias{tbl_prots_peps}
\title{A lookup table between prot_acc and pep_seqs.}
\usage{
tbl_prots_peps(seqs, path)
}
\arguments{
\item{seqs}{Results from \link{distri_peps}.}

\item{path}{A file path.}
}
\description{
For both target and decoy peptides.
}
\examples{
\donttest{
tbl_prots_peps(fwd_peps[[1]])
}
}
