% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{traceDIA}
\alias{traceDIA}
\title{Traces DIA-MS2 against MS1.}
\usage{
traceDIA(
  df,
  min_mass = 200L,
  min_ms2mass = 115L,
  step1 = 1e-05,
  step2 = 1e-05,
  n_dia_ms2s = 0L
)
}
\arguments{
\item{df}{A data frame.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{min_ms2mass}{A positive integer; the minimum MS2 mass for
interrogation. The default is 110.}

\item{step1}{A step size for MS1.}

\item{step2}{A step size for MS2.}

\item{n_dia_ms2s}{Allowance for considering DIA-MS2 bins of retention times
being aligned with a DIA-MS1 bin of retention time. For instance, only the
center bin of MS2 will be considered at \code{n_dia_ms2s = 1}. The
preceding and following bins will also be considered at \code{n_dia_ms2s =
2}, etc. Setting \code{n_dia_ms2s = 0L} will bypass MS2 feature alignments.}
}
\description{
For a single MS1 range, e.g., 592.5 to 604.5.
}
