% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{traceMS1}
\alias{traceMS1}
\title{Helper of MS1 tracing.}
\usage{
traceMS1(df, min_mass = 200L, step = 8e-06, n_dia_scans = 4L)
}
\arguments{
\item{df}{MS1 data.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{step}{Step size.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}
}
\description{
Helper of MS1 tracing.
}
