% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{traceXY}
\alias{traceXY}
\title{Helper of MS1 tracing.}
\usage{
traceXY(
  xs,
  ys,
  ss,
  n_mdda_flanks = 6L,
  from = 115L,
  step = 1e-05,
  reord = TRUE,
  cleanup = FALSE,
  replace_ms1_by_apex = FALSE
)
}
\arguments{
\item{xs}{Vectors of full-spectrum MS1 moverzs.}

\item{ys}{Vectors of full-spectrum MS1 intensities.}

\item{ss}{Vectors of MS1 scan numbers.}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{from}{The starting point for mass binning.}

\item{step}{A step size for mass binning.}

\item{reord}{Logical; re-order data or not.}

\item{cleanup}{Logical; to clean up xs, ys and zs or not. Set the value to
FALSE to maintain one-to-one correspondence between input (data frame) and
the outputs. This will help, e.g., keep track of scan numbers in the input.}

\item{replace_ms1_by_apex}{Logical; if TRUE, fill all entries within a gate
by its apex values.}
}
\description{
Helper of MS1 tracing.
}
