% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{traceXY}
\alias{traceXY}
\title{Helper of MS1 tracing.}
\usage{
traceXY(
  xs,
  ys,
  ss,
  ts,
  n_mdda_flanks = 6L,
  from = 200L,
  step = 8e-06,
  reord = TRUE,
  cleanup = FALSE,
  replace_ms1_by_apex = FALSE,
  y_perc = 0.01,
  yco = 100,
  look_back = TRUE
)
}
\arguments{
\item{xs}{Lists of full-spectrum MS1 moverzs vectors.}

\item{ys}{Lists of full-spectrum MS1 intensities vectors.}

\item{ss}{Vectors of MS1 scan numbers.}

\item{ts}{Vectors of MS1 retention times (for calculating area-under-a-peak).}

\item{n_mdda_flanks}{The number of preceding and following MS1 scans for
consideration when averaging isotope envelops of precursors.}

\item{from}{The starting point for mass binning.}

\item{step}{A step size for mass binning.}

\item{reord}{Logical; re-order data or not.}

\item{cleanup}{Logical; to clean up xs, ys and zs or not. Set the value to
FALSE to maintain one-to-one correspondence between input (data frame) and
the outputs. This will help, e.g., keep track of scan numbers in the input.}

\item{replace_ms1_by_apex}{Logical; if TRUE, fill all entries within a gate
by its apex values.}

\item{y_perc}{The cut-off in intensity values in relative to the base peak.}

\item{yco}{The cut-off in y values. An arbitrary small number. The MS1 peaks
with high-resolution Thermo's instruments are not a continuum and the
maximum number of MS1 peaks seem \eqn{\le 1500}.}

\item{look_back}{Logical; look up the preceding MS bin or not.}
}
\description{
Helper of MS1 tracing.
}
