% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsmatches.R
\name{try_psmC2Q}
\alias{try_psmC2Q}
\title{Helper of \link{psmC2Q}.}
\usage{
try_psmC2Q(
  out = NULL,
  out_path = NULL,
  fdr_type = "protein",
  combine_tier_three = FALSE,
  max_n_prots = 40000L
)
}
\arguments{
\item{out}{A result of \code{psmC.txt}.}

\item{out_path}{A file path of outputs.}

\item{fdr_type}{A character string; the type of FDR control. The value is in
  one of c("psm", "peptide", "protein"). The current default is \code{psm}.

  Note that \code{fdr_type = protein} is equivalent to \code{fdr_type =
  peptide} with the additional filtration of data at \code{prot_tier == 1}. A
  variant is to set \code{fdr_type = psm}, followed by a data filtration at
  \code{prot_tier == 1}.}

\item{combine_tier_three}{Logical; if TRUE, combines search results at tiers
  1, 2 and 3 to the single output of \code{psmQ.txt}. The default is FALSE in
  that data will be segregated into the three quality tiers according to the
  choice of \code{fdr_type}. The (convenience) parameter matters since
  \href{http://github.com/qzhang503/proteoQ}{proteoQ} will only look for the
  inputs of \code{psmQ[...].txt}.

  For instance, if the aim is to bypass the constraint by protein FDR and
  focus on PSMs that have met the cut-offs specified by \code{target_fdr}, an
  experimenter may set \code{combine_tier_three = TRUE} and hence pool all
  significant peptides in \code{psmQ.txt} for downstream proteoQ.

  Tier-1: both proteins and peptides with scores above significance
  thresholds.

  Tier-2: \eqn{\ge} 2 significant peptides but protein scores below
  significance thresholds.

  Tier-3: one significant peptide per protein and protein scores below
  significance thresholds.}

\item{max_n_prots}{A positive integer to threshold the maximum number of
  protein entries before coercing \code{fdr_type} from \code{psm} or
  \code{peptide} to \code{protein}. The argument has no effect if
  \code{fdr_type} is already \code{protein}. In general, there is no need to
  change the default.

  Note that for memory efficiency proteins at tiers 1, 2 and 3 are grouped
  separately. Further note that there is no tier-2 proteins at
  \code{fdr_type} of \code{psm} or \code{peptide}. For very large data sets,
  a lower value of \code{max_n_prots} can be used to reduce the chance of
  memory exhaustion by setting aside some protein entries from tier 1 to 2.}
}
\description{
Helper of \link{psmC2Q}.
}
