% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{unique_mvmods}
\alias{unique_mvmods}
\title{The unique combinations of variable modifications.}
\usage{
unique_mvmods(
  amods,
  ntmod,
  ctmod,
  aa_masses,
  aas,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  .ms1_vmodsets = NULL,
  .base_ent = NULL,
  digits = 5L
)
}
\arguments{
\item{amods}{Anywhere modifications.}

\item{ntmod}{The attribute \code{ntmod} from a \code{aa_masses} (for MS1
calculations).}

\item{ctmod}{The attribute \code{ctmod} from a \code{aa_masses} (for MS1
calculations).}

\item{aa_masses}{A named list containing the (mono-isotopic) masses of amino
acid residues.}

\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{maxn_vmods_per_pep}{The maximum number of \code{Anywhere}
(non-terminal) variable modifications per peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
\code{Anywhere} (non-terminal) variable modifications per site in a per
peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\value{
Lists by residues in \code{amods}.
}
\description{
The same residue, e.g. M, at different modifications, c("Carbamyl (M",
"Oxidation (M)")).
}
\details{
Goes over all the \code{Anywhere} modifications specified in \code{amods} for
a given \code{aa_masses}.
}
\examples{
\donttest{
## M
fixedmods = c("TMT6plex (K)", "dHex (S)")
varmods = c("Carbamidomethyl (M)", "Carbamyl (M)", "Acetyl (Protein N-term)")

aa_masses_all <- calc_aamasses(fixedmods, varmods,
                               add_varmasses = FALSE,
                               add_nlmasses = FALSE)

aa_masses <- aa_masses_all[[8]]

amods <- list(`Carbamidomethyl (M)` = c(Anywhere = "M"),
              `Carbamyl (M)` = c(Anywhere = "M"))

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

ans <- unique_mvmods(amods = amods, ntmod = NULL, ctmod = NULL,
                     aa_masses = aa_masses, aas = aas)

stopifnot(length(ans) == 1L,
          length(ans[[1]]) == 3L)

## M and N
fixedmods = c("TMT6plex (K)", "dHex (S)")
varmods = c("Carbamidomethyl (M)", "Carbamyl (M)",
            "Deamidated (N)", "Acetyl (Protein N-term)")

aa_masses_all <- calc_aamasses(fixedmods, varmods,
                               add_varmasses = FALSE,
                               add_nlmasses = FALSE)

aa_masses <- aa_masses_all[[16]]

amods <- list(`Carbamidomethyl (M)` = c(Anywhere = "M"),
              `Carbamyl (M)` = c(Anywhere = "M"),
              `Deamidated (N)` = c(Anywhere = "N"))

aas <- unlist(strsplit("HQGVMNVGMGQKMNS", ""))

ans <- unique_mvmods(amods = amods, ntmod = NULL, ctmod = NULL,
                     aa_masses = aa_masses, aas = aas)

stopifnot(length(ans) == 2L,
          length(ans[[1]]) == 3L,
          length(ans[[2]]) == 2L)
}
}
\seealso{
\link{ms1_a1_vnl0_fnl0} for examples.
}
