% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml.R
\name{updateMS1Int}
\alias{updateMS1Int}
\title{Updates MS1 intensity with apex values.}
\usage{
updateMS1Int(df, matx, maty, min_mass = 200L, step = 8e-06)
}
\arguments{
\item{df}{A data frame.}

\item{matx}{The matrix of moverzs Y: by masses; X: by LC scans.}

\item{maty}{The matrix of intensities. Y: by masses; X: by LC scans.}

\item{min_mass}{A positive integer; the minimum precursor mass for
interrogation. The default is an arbitrarily low value (the primary guard
against low molecular-weight precursors is \code{min_len}).}

\item{step}{A step size for mass binning.}
}
\description{
Updates MS1 intensity with apex values.
}
