% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{updateMS1Int}
\alias{updateMS1Int}
\title{Updates MS1 intensity with apex values.}
\usage{
updateMS1Int(
  df,
  matx,
  maty,
  unv,
  ss,
  ts,
  row_sta,
  row_end,
  from = 200L,
  step_tr = 6e-06,
  tol_ms1 = 1e-05,
  min_y = 2e+06,
  yco = 100,
  ytot_co = 2e+05,
  min_n1 = 10L,
  min_n2 = 20L,
  min_n3 = 15L,
  n_dia_scans = 6L
)
}
\arguments{
\item{df}{A data frame.}

\item{matx}{The matrix of m-over-z's Y: by masses; X: by LC scans.}

\item{maty}{The matrix of intensities. Y: by masses; X: by LC scans.}

\item{unv}{The universe of m-over-z bins.}

\item{ss}{A vector of scan numbers.}

\item{ts}{A vector of retention times.}

\item{row_sta}{The starting row of \code{matx}.}

\item{row_end}{The ending row of \code{matx}.}

\item{from}{The starting point for mass binning.}

\item{step_tr}{A step size for mass binning.}

\item{tol_ms1}{The tolerance in MS1 errors \code{ppm_ms1 * 1E-6}.}

\item{min_y}{The cut-off in peak area.}

\item{yco}{An intensity (Y-value) cut-off.}

\item{ytot_co}{A more permissive cut-off in peak area for finding all peaks.}

\item{min_n1}{The first cut-off of a minimum number of points across an MS1
peak.}

\item{min_n2}{The second cut-off of a minimum number of points across an MS1
peak.}

\item{min_n3}{The third cut-off of a minimum number of points across an MS1
peak.}

\item{n_dia_scans}{Allowance in the number of adjacent MS scans for
constructing a peak profile and thus for determining the apex scan number
of an moverz value along LC.}
}
\description{
Including forward and backward looking.
}
