% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs.R
\name{vmods_elements}
\alias{vmods_elements}
\title{Find the sets of variable modifications.}
\usage{
vmods_elements(
  aas,
  residue_mods,
  ntmod,
  ctmod,
  maxn_vmods_per_pep = 5L,
  maxn_sites_per_vmod = 3L,
  .ms1_vmodsets = NULL,
  .base_ent = NULL,
  digits = 5L
)
}
\arguments{
\item{aas}{\code{aa_seq} split in a sequence of LETTERS.}

\item{residue_mods}{Amino-acid residues with Unimod names. For example
rownames of \code{Carbamidomethyl (M)} and \code{Oxidation (M)} and a
column residues of \code{M, M}.}

\item{ntmod}{The attribute \code{ntmod} from a \code{aa_masses} (for MS1
calculations).}

\item{ctmod}{The attribute \code{ctmod} from a \code{aa_masses} (for MS1
calculations).}

\item{maxn_vmods_per_pep}{The maximum number of variable modifications per
peptide.}

\item{maxn_sites_per_vmod}{Integer; the maximum number of combinatorial
variable modifications per site in a per peptide sequence.}

\item{digits}{Integer; the number of decimal places to be used.}
}
\description{
The same residue, e.g. M, at different modifications, c("Carbamyl (M",
"Oxidation (M)")).
}
\details{
Excluding position differences, i.e., \code{A, B} and \code{B, A} is the
same set.
}
\examples{
\donttest{
ntmod <- list(`Acetyl (Protein N-term)` = c(`Protein N-term` = "N-term"))

ctmod <- list()
names(ctmod) <- character()

aas <- unlist(strsplit("HQGVMNVGMGQKSMNS", ""))
residue_mods <- c(`Carbamidomethyl (M)` = "M", `Carbamyl (M)` = "M")

x <- vmods_elements(aas, residue_mods, ntmod, ctmod)
}
}
