% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engine.R
\name{which_topx2}
\alias{which_topx2}
\title{Finds the indexes of top-n entries without re-ordering.}
\usage{
which_topx2(x, n = 50L, vna = 0, ...)
}
\arguments{
\item{x}{A vector of non-negative numeric.}

\item{n}{A positive integer; the number of top entries to keep.}

\item{vna}{The value of NA replacement. The default is zero with the
assumption that \code{x >= 0}.}

\item{...}{Additional arguments for base function sort.}

\item{exclude_na}{Logical; removes NA or not.}
}
\value{
The indexes of the top-n entries.
}
\description{
Handles ties.
}
\examples{
library(mzion)

p <- 100
set.seed(1)
x <- sample(1:150, replace = TRUE)

# 103, not 100
xp <- sort(x, partial = p)[p]

# multiple ties
x <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
p <- 2L

x <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4, 5, 5)
set.seed <- (3)
x <- sample(x)
p <- 4L

## more NA values (3) than n (2)
x <- c(0.11, 0.11, NA, rep(0.11, 7), NA, NA)
ans <- mzion:::which_topx2(x, 2L, na.last = FALSE)

# OK
ans <- mzion:::which_topx2(x, 8L, na.last = FALSE)

# results can contain NA
x <- c(10, rep_len(NA_integer_, 3), 12, rep_len(NA_integer_, 3), 8, 10, 10)
ans <- mzion:::which_topx2(x, n = 7L)
x[ans]

# Bad
ans <- mzion:::which_topx2(x, 9L, na.last = FALSE)
# OK
ans <- mzion:::which_topx2(x, 9L, na.last = TRUE)

\donttest{
mzion:::which_topx2(5000, NA_integer_)
}
}
